@@ -1,14 +1,22 @@
-module github.com/360EntSecGroup-Skylar/excelize/v2
+module github.com/xuri/excelize/v2
 
-go 1.15
+go 1.20
 
 require (
-	github.com/mohae/deepcopy v0.0.0-20170929034955-c48cc78d4826
-	github.com/richardlehane/mscfb v1.0.3
-	github.com/stretchr/testify v1.6.1
-	github.com/xuri/efp v0.0.0-20210322160811-ab561f5b45e3
-	golang.org/x/crypto v0.0.0-20210513164829-c07d793c2f9a
-	golang.org/x/image v0.0.0-20210220032944-ac19c3e999fb
-	golang.org/x/net v0.0.0-20210610132358-84b48f89b13b
-	golang.org/x/text v0.3.6
+	github.com/richardlehane/mscfb v1.0.4
+	github.com/stretchr/testify v1.9.0
+	github.com/tiendc/go-deepcopy v1.2.0
+	github.com/xuri/efp v0.0.0-20241211021726-c4e992084aa6
+	github.com/xuri/nfp v0.0.0-20250111060730-82a408b9aa71
+	golang.org/x/crypto v0.32.0
+	golang.org/x/image v0.18.0
+	golang.org/x/net v0.34.0
+	golang.org/x/text v0.21.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/richardlehane/msoleps v1.0.4 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
