@@ -1,13 +1,13 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
@@ -16,7 +16,6 @@ import (
 	"encoding/xml"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"reflect"
 	"strconv"
@@ -26,70 +25,102 @@ import (
 
 // StreamWriter defined the type of stream writer.
 type StreamWriter struct {
-	File            *File
+	file            *File
 	Sheet           string
 	SheetID         int
 	sheetWritten    bool
-	cols            string
 	worksheet       *xlsxWorksheet
 	rawData         bufferedWriter
+	rows            int
 	mergeCellsCount int
-	mergeCells      string
+	mergeCells      strings.Builder
 	tableParts      string
 }
 
-// NewStreamWriter return stream writer struct by given worksheet name for
-// generate new worksheet with large amounts of data. Note that after set
-// rows, you must call the 'Flush' method to end the streaming writing
-// process and ensure that the order of line numbers is ascending, the common
-// API and stream API can't be work mixed to writing data on the worksheets,
-// you can't get cell value when in-memory chunks data over 16MB. For
-// example, set data for worksheet of size 102400 rows x 50 columns with
-// numbers and style:
+// NewStreamWriter returns stream writer struct by given worksheet name used for
+// writing data on a new existing empty worksheet with large amounts of data.
+// Note that after writing data with the stream writer for the worksheet, you
+// must call the 'Flush' method to end the streaming writing process, ensure
+// that the order of row numbers is ascending when set rows, and the normal
+// mode functions and stream mode functions can not be work mixed to writing
+// data on the worksheets. The stream writer will try to use temporary files on
+// disk to reduce the memory usage when in-memory chunks data over 16MB, and
+// you can't get cell value at this time. For example, set data for worksheet
+// of size 102400 rows x 50 columns with numbers and style:
 //
-//    file := excelize.NewFile()
-//    streamWriter, err := file.NewStreamWriter("Sheet1")
-//    if err != nil {
-//        fmt.Println(err)
-//    }
-//    styleID, err := file.NewStyle(`{"font":{"color":"#777777"}}`)
-//    if err != nil {
-//        fmt.Println(err)
-//    }
-//    if err := streamWriter.SetRow("A1", []interface{}{excelize.Cell{StyleID: styleID, Value: "Data"}}); err != nil {
-//        fmt.Println(err)
-//    }
-//    for rowID := 2; rowID <= 102400; rowID++ {
-//        row := make([]interface{}, 50)
-//        for colID := 0; colID < 50; colID++ {
-//            row[colID] = rand.Intn(640000)
-//        }
-//        cell, _ := excelize.CoordinatesToCellName(1, rowID)
-//        if err := streamWriter.SetRow(cell, row); err != nil {
-//            fmt.Println(err)
-//        }
-//    }
-//    if err := streamWriter.Flush(); err != nil {
-//        fmt.Println(err)
-//    }
-//    if err := file.SaveAs("Book1.xlsx"); err != nil {
-//        fmt.Println(err)
-//    }
+//	f := excelize.NewFile()
+//	defer func() {
+//	    if err := f.Close(); err != nil {
+//	        fmt.Println(err)
+//	    }
+//	}()
+//	sw, err := f.NewStreamWriter("Sheet1")
+//	if err != nil {
+//	    fmt.Println(err)
+//	    return
+//	}
+//	styleID, err := f.NewStyle(&excelize.Style{Font: &excelize.Font{Color: "777777"}})
+//	if err != nil {
+//	    fmt.Println(err)
+//	    return
+//	}
+//	if err := sw.SetRow("A1",
+//	    []interface{}{
+//	        excelize.Cell{StyleID: styleID, Value: "Data"},
+//	        []excelize.RichTextRun{
+//	            {Text: "Rich ", Font: &excelize.Font{Color: "2354e8"}},
+//	            {Text: "Text", Font: &excelize.Font{Color: "e83723"}},
+//	        },
+//	    },
+//	    excelize.RowOpts{Height: 45, Hidden: false}); err != nil {
+//	    fmt.Println(err)
+//	    return
+//	}
+//	for rowID := 2; rowID <= 102400; rowID++ {
+//	    row := make([]interface{}, 50)
+//	    for colID := 0; colID < 50; colID++ {
+//	        row[colID] = rand.Intn(640000)
+//	    }
+//	    cell, err := excelize.CoordinatesToCellName(1, rowID)
+//	    if err != nil {
+//	        fmt.Println(err)
+//	        break
+//	    }
+//	    if err := sw.SetRow(cell, row); err != nil {
+//	        fmt.Println(err)
+//	        break
+//	    }
+//	}
+//	if err := sw.Flush(); err != nil {
+//	    fmt.Println(err)
+//	    return
+//	}
+//	if err := f.SaveAs("Book1.xlsx"); err != nil {
+//	    fmt.Println(err)
+//	}
 //
 // Set cell value and cell formula for a worksheet with stream writer:
 //
-//    err := streamWriter.SetRow("A1", []interface{}{
-//        excelize.Cell{Value: 1},
-//        excelize.Cell{Value: 2},
-//        excelize.Cell{Formula: "SUM(A1,B1)"}});
+//	err := sw.SetRow("A1", []interface{}{
+//	    excelize.Cell{Value: 1},
+//	    excelize.Cell{Value: 2},
+//	    excelize.Cell{Formula: "SUM(A1,B1)"}});
 //
+// Set cell value and rows style for a worksheet with stream writer:
+//
+//	err := sw.SetRow("A1", []interface{}{
+//	    excelize.Cell{Value: 1}},
+//	    excelize.RowOpts{StyleID: styleID, Height: 20, Hidden: false});
 func (f *File) NewStreamWriter(sheet string) (*StreamWriter, error) {
+	if err := checkSheetName(sheet); err != nil {
+		return nil, err
+	}
 	sheetID := f.getSheetID(sheet)
 	if sheetID == -1 {
-		return nil, fmt.Errorf("sheet %s is not exist", sheet)
+		return nil, ErrSheetNotExist{sheet}
 	}
 	sw := &StreamWriter{
-		File:    f,
+		file:    f,
 		Sheet:   sheet,
 		SheetID: sheetID,
 	}
@@ -99,47 +130,48 @@ func (f *File) NewStreamWriter(sheet string) (*StreamWriter, error) {
 		return nil, err
 	}
 
-	sheetPath := f.sheetMap[trimSheetName(sheet)]
+	sheetXMLPath, _ := f.getSheetXMLPath(sheet)
 	if f.streams == nil {
 		f.streams = make(map[string]*StreamWriter)
 	}
-	f.streams[sheetPath] = sw
+	f.streams[sheetXMLPath] = sw
 
-	_, _ = sw.rawData.WriteString(XMLHeader + `<worksheet` + templateNamespaceIDMap)
-	bulkAppendFields(&sw.rawData, sw.worksheet, 2, 5)
+	_, _ = sw.rawData.WriteString(xml.Header + `<worksheet` + templateNamespaceIDMap)
+	bulkAppendFields(&sw.rawData, sw.worksheet, 2, 3)
 	return sw, err
 }
 
 // AddTable creates an Excel table for the StreamWriter using the given
-// coordinate area and format set. For example, create a table of A1:D5:
+// cell range and format set. For example, create a table of A1:D5:
 //
-//    err := sw.AddTable("A1", "D5", "")
+//	err := sw.AddTable(&excelize.Table{Range: "A1:D5"})
 //
 // Create a table of F2:H6 with format set:
 //
-//    err := sw.AddTable("F2", "H6", `{
-//        "table_name": "table",
-//        "table_style": "TableStyleMedium2",
-//        "show_first_column": true,
-//        "show_last_column": true,
-//        "show_row_stripes": false,
-//        "show_column_stripes": true
-//    }`)
+//	disable := false
+//	err := sw.AddTable(&excelize.Table{
+//	    Range:             "F2:H6",
+//	    Name:              "table",
+//	    StyleName:         "TableStyleMedium2",
+//	    ShowFirstColumn:   true,
+//	    ShowLastColumn:    true,
+//	    ShowRowStripes:    &disable,
+//	    ShowColumnStripes: true,
+//	})
 //
 // Note that the table must be at least two lines including the header. The
 // header cells must contain strings and must be unique.
 //
-// Currently only one table is allowed for a StreamWriter. AddTable must be
+// Currently, only one table is allowed for a StreamWriter. AddTable must be
 // called after the rows are written but before Flush.
 //
 // See File.AddTable for details on the table format.
-func (sw *StreamWriter) AddTable(hcell, vcell, format string) error {
-	formatSet, err := parseFormatTableSet(format)
+func (sw *StreamWriter) AddTable(table *Table) error {
+	options, err := parseTableOptions(table)
 	if err != nil {
 		return err
 	}
-
-	coordinates, err := areaRangeToCoordinates(hcell, vcell)
+	coordinates, err := rangeRefToCoordinates(options.Range)
 	if err != nil {
 		return err
 	}
@@ -150,8 +182,8 @@ func (sw *StreamWriter) AddTable(hcell, vcell, format string) error {
 		coordinates[3]++
 	}
 
-	// Correct table reference coordinate area, such correct C1:B3 to B1:C3.
-	ref, err := sw.File.coordinatesToAreaRef(coordinates)
+	// Correct table reference range, such correct C1:B3 to B1:C3.
+	ref, err := coordinatesToRangeRef(coordinates)
 	if err != nil {
 		return err
 	}
@@ -169,14 +201,14 @@ func (sw *StreamWriter) AddTable(hcell, vcell, format string) error {
 		}
 	}
 
-	tableID := sw.File.countTables() + 1
+	tableID := sw.file.countTables() + 1
 
-	name := formatSet.TableName
+	name := options.Name
 	if name == "" {
 		name = "Table" + strconv.Itoa(tableID)
 	}
 
-	table := xlsxTable{
+	tbl := xlsxTable{
 		XMLNS:       NameSpaceSpreadSheet.Value,
 		ID:          tableID,
 		Name:        name,
@@ -190,41 +222,42 @@ func (sw *StreamWriter) AddTable(hcell, vcell, format string) error {
 			TableColumn: tableColumn,
 		},
 		TableStyleInfo: &xlsxTableStyleInfo{
-			Name:              formatSet.TableStyle,
-			ShowFirstColumn:   formatSet.ShowFirstColumn,
-			ShowLastColumn:    formatSet.ShowLastColumn,
-			ShowRowStripes:    formatSet.ShowRowStripes,
-			ShowColumnStripes: formatSet.ShowColumnStripes,
+			Name:              options.StyleName,
+			ShowFirstColumn:   options.ShowFirstColumn,
+			ShowLastColumn:    options.ShowLastColumn,
+			ShowRowStripes:    *options.ShowRowStripes,
+			ShowColumnStripes: options.ShowColumnStripes,
 		},
 	}
 
 	sheetRelationshipsTableXML := "../tables/table" + strconv.Itoa(tableID) + ".xml"
-	tableXML := strings.Replace(sheetRelationshipsTableXML, "..", "xl", -1)
+	tableXML := strings.ReplaceAll(sheetRelationshipsTableXML, "..", "xl")
 
-	// Add first table for given sheet.
-	sheetPath := sw.File.sheetMap[trimSheetName(sw.Sheet)]
+	// Add first table for given sheet
+	sheetPath := sw.file.sheetMap[sw.Sheet]
 	sheetRels := "xl/worksheets/_rels/" + strings.TrimPrefix(sheetPath, "xl/worksheets/") + ".rels"
-	rID := sw.File.addRels(sheetRels, SourceRelationshipTable, sheetRelationshipsTableXML, "")
+	rID := sw.file.addRels(sheetRels, SourceRelationshipTable, sheetRelationshipsTableXML, "")
 
 	sw.tableParts = fmt.Sprintf(`<tableParts count="1"><tablePart r:id="rId%d"></tablePart></tableParts>`, rID)
 
-	sw.File.addContentTypePart(tableID, "table")
-
-	b, _ := xml.Marshal(table)
-	sw.File.saveFileList(tableXML, b)
-	return nil
+	if err = sw.file.addContentTypePart(tableID, "table"); err != nil {
+		return err
+	}
+	b, _ := xml.Marshal(tbl)
+	sw.file.saveFileList(tableXML, b)
+	return err
 }
 
 // Extract values from a row in the StreamWriter.
-func (sw *StreamWriter) getRowValues(hrow, hcol, vcol int) (res []string, err error) {
-	res = make([]string, vcol-hcol+1)
+func (sw *StreamWriter) getRowValues(hRow, hCol, vCol int) (res []string, err error) {
+	res = make([]string, vCol-hCol+1)
 
 	r, err := sw.rawData.Reader()
 	if err != nil {
 		return nil, err
 	}
 
-	dec := sw.File.xmlNewDecoder(r)
+	dec := sw.file.xmlNewDecoder(r)
 	for {
 		token, err := dec.Token()
 		if err == io.EOF {
@@ -233,7 +266,7 @@ func (sw *StreamWriter) getRowValues(hrow, hcol, vcol int) (res []string, err er
 		if err != nil {
 			return nil, err
 		}
-		startElement, ok := getRowElement(token, hrow)
+		startElement, ok := getRowElement(token, hRow)
 		if !ok {
 			continue
 		}
@@ -247,17 +280,17 @@ func (sw *StreamWriter) getRowValues(hrow, hcol, vcol int) (res []string, err er
 			if err != nil {
 				return nil, err
 			}
-			if col < hcol || col > vcol {
+			if col < hCol || col > vCol {
 				continue
 			}
-			res[col-hcol] = c.V
+			res[col-hCol], _ = c.getValueFrom(sw.file, nil, false)
 		}
 		return res, nil
 	}
 }
 
-// Check if the token is an XLSX row with the matching row number.
-func getRowElement(token xml.Token, hrow int) (startElement xml.StartElement, ok bool) {
+// Check if the token is an worksheet row with the matching row number.
+func getRowElement(token xml.Token, hRow int) (startElement xml.StartElement, ok bool) {
 	startElement, ok = token.(xml.StartElement)
 	if !ok {
 		return
@@ -272,7 +305,7 @@ func getRowElement(token xml.Token, hrow int) (startElement xml.StartElement, ok
 			continue
 		}
 		row, _ := strconv.Atoi(attr.Value)
-		if row == hrow {
+		if row == hRow {
 			ok = true
 			return
 		}
@@ -288,41 +321,110 @@ type Cell struct {
 	Value   interface{}
 }
 
-// SetRow writes an array to stream rows by giving a worksheet name, starting
-// coordinate and a pointer to an array of values. Note that you must call the
-// 'Flush' method to end the streaming writing process.
+// RowOpts define the options for the set row, it can be used directly in
+// StreamWriter.SetRow to specify the style and properties of the row.
+type RowOpts struct {
+	Height       float64
+	Hidden       bool
+	StyleID      int
+	OutlineLevel int
+}
+
+// marshalAttrs prepare attributes of the row.
+func (r *RowOpts) marshalAttrs() (strings.Builder, error) {
+	var (
+		err   error
+		attrs strings.Builder
+	)
+	if r == nil {
+		return attrs, err
+	}
+	if r.Height > MaxRowHeight {
+		err = ErrMaxRowHeight
+		return attrs, err
+	}
+	if r.OutlineLevel > 7 {
+		err = ErrOutlineLevel
+		return attrs, err
+	}
+	if r.StyleID > 0 {
+		attrs.WriteString(` s="`)
+		attrs.WriteString(strconv.Itoa(r.StyleID))
+		attrs.WriteString(`" customFormat="1"`)
+	}
+	if r.Height > 0 {
+		attrs.WriteString(` ht="`)
+		attrs.WriteString(strconv.FormatFloat(r.Height, 'f', -1, 64))
+		attrs.WriteString(`" customHeight="1"`)
+	}
+	if r.OutlineLevel > 0 {
+		attrs.WriteString(` outlineLevel="`)
+		attrs.WriteString(strconv.Itoa(r.OutlineLevel))
+		attrs.WriteString(`"`)
+	}
+	if r.Hidden {
+		attrs.WriteString(` hidden="1"`)
+	}
+	return attrs, err
+}
+
+// parseRowOpts provides a function to parse the optional settings for
+// *StreamWriter.SetRow.
+func parseRowOpts(opts ...RowOpts) *RowOpts {
+	options := &RowOpts{}
+	for _, opt := range opts {
+		options = &opt
+	}
+	return options
+}
+
+// SetRow writes an array to stream rows by giving starting cell reference and a
+// pointer to an array of values. Note that you must call the 'Flush' function
+// to end the streaming writing process.
 //
 // As a special case, if Cell is used as a value, then the Cell.StyleID will be
 // applied to that cell.
-func (sw *StreamWriter) SetRow(axis string, values []interface{}) error {
-	col, row, err := CellNameToCoordinates(axis)
+func (sw *StreamWriter) SetRow(cell string, values []interface{}, opts ...RowOpts) error {
+	col, row, err := CellNameToCoordinates(cell)
 	if err != nil {
 		return err
 	}
-	if !sw.sheetWritten {
-		if len(sw.cols) > 0 {
-			sw.rawData.WriteString("<cols>" + sw.cols + "</cols>")
-		}
-		_, _ = sw.rawData.WriteString(`<sheetData>`)
-		sw.sheetWritten = true
+	if row <= sw.rows {
+		return newStreamSetRowError(row)
 	}
-	fmt.Fprintf(&sw.rawData, `<row r="%d">`, row)
+	sw.rows = row
+	sw.writeSheetData()
+	options := parseRowOpts(opts...)
+	attrs, err := options.marshalAttrs()
+	if err != nil {
+		return err
+	}
+	_, _ = sw.rawData.WriteString(`<row r="`)
+	_, _ = sw.rawData.WriteString(strconv.Itoa(row))
+	_, _ = sw.rawData.WriteString(`"`)
+	_, _ = sw.rawData.WriteString(attrs.String())
+	_, _ = sw.rawData.WriteString(`>`)
 	for i, val := range values {
-		axis, err := CoordinatesToCellName(col+i, row)
+		if val == nil {
+			continue
+		}
+		ref, err := CoordinatesToCellName(col+i, row)
 		if err != nil {
 			return err
 		}
-		c := xlsxC{R: axis}
+		c := xlsxC{R: ref, S: sw.worksheet.prepareCellStyle(col, row, options.StyleID)}
+		var s int
 		if v, ok := val.(Cell); ok {
-			c.S = v.StyleID
-			val = v.Value
+			s, val = v.StyleID, v.Value
 			setCellFormula(&c, v.Formula)
 		} else if v, ok := val.(*Cell); ok && v != nil {
-			c.S = v.StyleID
-			val = v.Value
+			s, val = v.StyleID, v.Value
 			setCellFormula(&c, v.Formula)
 		}
-		if err = setCellValFunc(&c, val); err != nil {
+		if s > 0 {
+			c.S = s
+		}
+		if err = sw.setCellValFunc(&c, val); err != nil {
 			_, _ = sw.rawData.WriteString(`</row>`)
 			return err
 		}
@@ -332,119 +434,194 @@ func (sw *StreamWriter) SetRow(axis string, values []interface{}) error {
 	return sw.rawData.Sync()
 }
 
+// SetColStyle provides a function to set the style of a single column or
+// multiple columns for the StreamWriter. Note that you must call
+// the 'SetColStyle' function before the 'SetRow' function. For example set
+// style of column H on Sheet1:
+//
+//	err := sw.SetColStyle(8, 8, style)
+func (sw *StreamWriter) SetColStyle(minVal, maxVal, styleID int) error {
+	if sw.sheetWritten {
+		return ErrStreamSetColStyle
+	}
+	if minVal < MinColumns || minVal > MaxColumns || maxVal < MinColumns || maxVal > MaxColumns {
+		return ErrColumnNumber
+	}
+	if maxVal < minVal {
+		minVal, maxVal = maxVal, minVal
+	}
+	s, err := sw.file.stylesReader()
+	if err != nil {
+		return err
+	}
+	if styleID < 0 || s.CellXfs == nil || len(s.CellXfs.Xf) <= styleID {
+		return newInvalidStyleID(styleID)
+	}
+	sw.worksheet.setColStyle(minVal, maxVal, styleID)
+	return nil
+}
+
 // SetColWidth provides a function to set the width of a single column or
 // multiple columns for the StreamWriter. Note that you must call
 // the 'SetColWidth' function before the 'SetRow' function. For example set
 // the width column B:C as 20:
 //
-//    err := streamWriter.SetColWidth(2, 3, 20)
-//
-func (sw *StreamWriter) SetColWidth(min, max int, width float64) error {
+//	err := sw.SetColWidth(2, 3, 20)
+func (sw *StreamWriter) SetColWidth(minVal, maxVal int, width float64) error {
 	if sw.sheetWritten {
 		return ErrStreamSetColWidth
 	}
-	if min > TotalColumns || max > TotalColumns {
-		return ErrColumnNumber
-	}
-	if min < 1 || max < 1 {
+	if minVal < MinColumns || minVal > MaxColumns || maxVal < MinColumns || maxVal > MaxColumns {
 		return ErrColumnNumber
 	}
 	if width > MaxColumnWidth {
 		return ErrColumnWidth
 	}
-	if min > max {
-		min, max = max, min
+	if minVal > maxVal {
+		minVal, maxVal = maxVal, minVal
 	}
-	sw.cols += fmt.Sprintf(`<col min="%d" max="%d" width="%f" customWidth="1"/>`, min, max, width)
+	sw.worksheet.setColWidth(minVal, maxVal, width)
 	return nil
 }
 
-// MergeCell provides a function to merge cells by a given coordinate area for
+// InsertPageBreak creates a page break to determine where the printed page ends
+// and where begins the next one by a given cell reference, the content before
+// the page break will be printed on one page and after the page break on
+// another.
+func (sw *StreamWriter) InsertPageBreak(cell string) error {
+	return sw.worksheet.insertPageBreak(cell)
+}
+
+// SetPanes provides a function to create and remove freeze panes and split
+// panes by giving panes options for the StreamWriter. Note that you must call
+// the 'SetPanes' function before the 'SetRow' function.
+func (sw *StreamWriter) SetPanes(panes *Panes) error {
+	if sw.sheetWritten {
+		return ErrStreamSetPanes
+	}
+	return sw.worksheet.setPanes(panes)
+}
+
+// MergeCell provides a function to merge cells by a given range reference for
 // the StreamWriter. Don't create a merged cell that overlaps with another
 // existing merged cell.
-func (sw *StreamWriter) MergeCell(hcell, vcell string) error {
-	_, err := areaRangeToCoordinates(hcell, vcell)
+func (sw *StreamWriter) MergeCell(topLeftCell, bottomRightCell string) error {
+	_, err := cellRefsToCoordinates(topLeftCell, bottomRightCell)
 	if err != nil {
 		return err
 	}
 	sw.mergeCellsCount++
-	sw.mergeCells += fmt.Sprintf(`<mergeCell ref="%s:%s"/>`, hcell, vcell)
+	_, _ = sw.mergeCells.WriteString(`<mergeCell ref="`)
+	_, _ = sw.mergeCells.WriteString(topLeftCell)
+	_, _ = sw.mergeCells.WriteString(`:`)
+	_, _ = sw.mergeCells.WriteString(bottomRightCell)
+	_, _ = sw.mergeCells.WriteString(`"/>`)
 	return nil
 }
 
 // setCellFormula provides a function to set formula of a cell.
 func setCellFormula(c *xlsxC, formula string) {
 	if formula != "" {
-		c.F = &xlsxF{Content: formula}
+		c.T, c.F = "str", &xlsxF{Content: formula}
 	}
 }
 
+// setCellTime provides a function to set number of a cell with a time.
+func (sw *StreamWriter) setCellTime(c *xlsxC, val time.Time) error {
+	var date1904, isNum bool
+	wb, err := sw.file.workbookReader()
+	if err != nil {
+		return err
+	}
+	if wb != nil && wb.WorkbookPr != nil {
+		date1904 = wb.WorkbookPr.Date1904
+	}
+	if isNum, err = c.setCellTime(val, date1904); err == nil && isNum && c.S == 0 {
+		style, _ := sw.file.NewStyle(&Style{NumFmt: 22})
+		c.S = style
+	}
+	return nil
+}
+
 // setCellValFunc provides a function to set value of a cell.
-func setCellValFunc(c *xlsxC, val interface{}) (err error) {
+func (sw *StreamWriter) setCellValFunc(c *xlsxC, val interface{}) error {
+	var err error
 	switch val := val.(type) {
 	case int, int8, int16, int32, int64, uint, uint8, uint16, uint32, uint64:
-		err = setCellIntFunc(c, val)
+		setCellIntFunc(c, val)
 	case float32:
-		c.T, c.V = setCellFloat(float64(val), -1, 32)
+		c.setCellFloat(float64(val), -1, 32)
 	case float64:
-		c.T, c.V = setCellFloat(val, -1, 64)
+		c.setCellFloat(val, -1, 64)
 	case string:
-		c.T, c.V, c.XMLSpace = setCellStr(val)
+		c.setCellValue(val)
 	case []byte:
-		c.T, c.V, c.XMLSpace = setCellStr(string(val))
+		c.setCellValue(string(val))
 	case time.Duration:
 		c.T, c.V = setCellDuration(val)
 	case time.Time:
-		c.T, c.V, _, err = setCellTime(val)
+		err = sw.setCellTime(c, val)
 	case bool:
 		c.T, c.V = setCellBool(val)
 	case nil:
-		c.T, c.V, c.XMLSpace = setCellStr("")
+		return err
+	case []RichTextRun:
+		c.T, c.IS = "inlineStr", &xlsxSI{}
+		c.IS.R, err = setRichText(val)
 	default:
-		c.T, c.V, c.XMLSpace = setCellStr(fmt.Sprint(val))
+		c.setCellValue(fmt.Sprint(val))
 	}
 	return err
 }
 
 // setCellIntFunc is a wrapper of SetCellInt.
-func setCellIntFunc(c *xlsxC, val interface{}) (err error) {
+func setCellIntFunc(c *xlsxC, val interface{}) {
 	switch val := val.(type) {
 	case int:
-		c.T, c.V = setCellInt(val)
+		c.T, c.V = setCellInt(int64(val))
 	case int8:
-		c.T, c.V = setCellInt(int(val))
+		c.T, c.V = setCellInt(int64(val))
 	case int16:
-		c.T, c.V = setCellInt(int(val))
+		c.T, c.V = setCellInt(int64(val))
 	case int32:
-		c.T, c.V = setCellInt(int(val))
+		c.T, c.V = setCellInt(int64(val))
 	case int64:
-		c.T, c.V = setCellInt(int(val))
+		c.T, c.V = setCellInt(val)
 	case uint:
-		c.T, c.V = setCellInt(int(val))
+		c.T, c.V = setCellUint(uint64(val))
 	case uint8:
-		c.T, c.V = setCellInt(int(val))
+		c.T, c.V = setCellUint(uint64(val))
 	case uint16:
-		c.T, c.V = setCellInt(int(val))
+		c.T, c.V = setCellUint(uint64(val))
 	case uint32:
-		c.T, c.V = setCellInt(int(val))
+		c.T, c.V = setCellUint(uint64(val))
 	case uint64:
-		c.T, c.V = setCellInt(int(val))
-	default:
+		c.T, c.V = setCellUint(val)
 	}
-	return
 }
 
+// writeCell constructs a cell XML and writes it to the buffer.
 func writeCell(buf *bufferedWriter, c xlsxC) {
 	_, _ = buf.WriteString(`<c`)
 	if c.XMLSpace.Value != "" {
-		fmt.Fprintf(buf, ` xml:%s="%s"`, c.XMLSpace.Name.Local, c.XMLSpace.Value)
+		_, _ = buf.WriteString(` xml:`)
+		_, _ = buf.WriteString(c.XMLSpace.Name.Local)
+		_, _ = buf.WriteString(`="`)
+		_, _ = buf.WriteString(c.XMLSpace.Value)
+		_, _ = buf.WriteString(`"`)
 	}
-	fmt.Fprintf(buf, ` r="%s"`, c.R)
+	_, _ = buf.WriteString(` r="`)
+	_, _ = buf.WriteString(c.R)
+	_, _ = buf.WriteString(`"`)
 	if c.S != 0 {
-		fmt.Fprintf(buf, ` s="%d"`, c.S)
+		_, _ = buf.WriteString(` s="`)
+		_, _ = buf.WriteString(strconv.Itoa(c.S))
+		_, _ = buf.WriteString(`"`)
 	}
 	if c.T != "" {
-		fmt.Fprintf(buf, ` t="%s"`, c.T)
+		_, _ = buf.WriteString(` t="`)
+		_, _ = buf.WriteString(c.T)
+		_, _ = buf.WriteString(`"`)
 	}
 	_, _ = buf.WriteString(`>`)
 	if c.F != nil {
@@ -457,21 +634,76 @@ func writeCell(buf *bufferedWriter, c xlsxC) {
 		_ = xml.EscapeText(buf, []byte(c.V))
 		_, _ = buf.WriteString(`</v>`)
 	}
+	if c.IS != nil {
+		if len(c.IS.R) > 0 {
+			is, _ := xml.Marshal(c.IS.R)
+			_, _ = buf.WriteString(`<is>`)
+			_, _ = buf.Write(is)
+			_, _ = buf.WriteString(`</is>`)
+		}
+		if c.IS.T != nil {
+			_, _ = buf.WriteString(`<is><t`)
+			if c.IS.T.Space.Value != "" {
+				_, _ = buf.WriteString(` xml:`)
+				_, _ = buf.WriteString(c.IS.T.Space.Name.Local)
+				_, _ = buf.WriteString(`="`)
+				_, _ = buf.WriteString(c.IS.T.Space.Value)
+				_, _ = buf.WriteString(`"`)
+			}
+			_, _ = buf.WriteString(`>`)
+			_, _ = buf.Write([]byte(c.IS.T.Val))
+			_, _ = buf.WriteString(`</t></is>`)
+		}
+	}
 	_, _ = buf.WriteString(`</c>`)
+}
+
+// writeSheetData prepares the element preceding sheetData and writes the
+// sheetData XML start element to the buffer.
+func (sw *StreamWriter) writeSheetData() {
+	if !sw.sheetWritten {
+		bulkAppendFields(&sw.rawData, sw.worksheet, 4, 5)
+		if sw.worksheet.Cols != nil {
+			_, _ = sw.rawData.WriteString("<cols>")
+			for _, col := range sw.worksheet.Cols.Col {
+				sw.rawData.WriteString(`<col min="`)
+				sw.rawData.WriteString(strconv.Itoa(col.Min))
+				sw.rawData.WriteString(`" max="`)
+				sw.rawData.WriteString(strconv.Itoa(col.Max))
+				sw.rawData.WriteString(`"`)
+				if col.Width != nil {
+					sw.rawData.WriteString(` width="`)
+					sw.rawData.WriteString(strconv.FormatFloat(*col.Width, 'f', -1, 64))
+					sw.rawData.WriteString(`" customWidth="1"`)
+				}
+				if col.Style != 0 {
+					sw.rawData.WriteString(` style="`)
+					sw.rawData.WriteString(strconv.Itoa(col.Style))
+					sw.rawData.WriteString(`"`)
+				}
+				sw.rawData.WriteString(`/>`)
+			}
+			_, _ = sw.rawData.WriteString("</cols>")
+		}
+		_, _ = sw.rawData.WriteString(`<sheetData>`)
+		sw.sheetWritten = true
+	}
 }
 
 // Flush ending the streaming writing process.
 func (sw *StreamWriter) Flush() error {
-	if !sw.sheetWritten {
-		_, _ = sw.rawData.WriteString(`<sheetData>`)
-		sw.sheetWritten = true
-	}
+	sw.writeSheetData()
 	_, _ = sw.rawData.WriteString(`</sheetData>`)
 	bulkAppendFields(&sw.rawData, sw.worksheet, 8, 15)
+	mergeCells := strings.Builder{}
 	if sw.mergeCellsCount > 0 {
-		sw.mergeCells = fmt.Sprintf(`<mergeCells count="%d">%s</mergeCells>`, sw.mergeCellsCount, sw.mergeCells)
+		_, _ = mergeCells.WriteString(`<mergeCells count="`)
+		_, _ = mergeCells.WriteString(strconv.Itoa(sw.mergeCellsCount))
+		_, _ = mergeCells.WriteString(`">`)
+		_, _ = mergeCells.WriteString(sw.mergeCells.String())
+		_, _ = mergeCells.WriteString(`</mergeCells>`)
 	}
-	_, _ = sw.rawData.WriteString(sw.mergeCells)
+	_, _ = sw.rawData.WriteString(mergeCells.String())
 	bulkAppendFields(&sw.rawData, sw.worksheet, 17, 38)
 	_, _ = sw.rawData.WriteString(sw.tableParts)
 	bulkAppendFields(&sw.rawData, sw.worksheet, 40, 40)
@@ -480,10 +712,10 @@ func (sw *StreamWriter) Flush() error {
 		return err
 	}
 
-	sheetPath := sw.File.sheetMap[trimSheetName(sw.Sheet)]
-	delete(sw.File.Sheet, sheetPath)
-	delete(sw.File.checked, sheetPath)
-	delete(sw.File.XLSX, sheetPath)
+	sheetPath := sw.file.sheetMap[sw.Sheet]
+	sw.file.Sheet.Delete(sheetPath)
+	sw.file.checked.Delete(sheetPath)
+	sw.file.Pkg.Delete(sheetPath)
 
 	return nil
 }
@@ -509,12 +741,12 @@ type bufferedWriter struct {
 	buf bytes.Buffer
 }
 
-// Write to the in-memory buffer. The err is always nil.
+// Write to the in-memory buffer. The error is always nil.
 func (bw *bufferedWriter) Write(p []byte) (n int, err error) {
 	return bw.buf.Write(p)
 }
 
-// WriteString wites to the in-memory buffer. The err is always nil.
+// WriteString write to the in-memory buffer. The error is always nil.
 func (bw *bufferedWriter) WriteString(p string) (n int, err error) {
 	return bw.buf.WriteString(p)
 }
@@ -543,7 +775,7 @@ func (bw *bufferedWriter) Sync() (err error) {
 		return nil
 	}
 	if bw.tmp == nil {
-		bw.tmp, err = ioutil.TempFile(os.TempDir(), "excelize-")
+		bw.tmp, err = os.CreateTemp(os.TempDir(), "excelize-")
 		if err != nil {
 			// can not use local storage
 			return nil
