@@ -1,86 +1,112 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
 import (
-	"encoding/json"
 	"encoding/xml"
-	"errors"
 	"fmt"
 	"strconv"
 	"strings"
 )
 
-// This section defines the currently supported chart types.
+// ChartType is the type of supported chart types.
+type ChartType byte
+
+// This section defines the currently supported chart types enumeration.
 const (
-	Area                        = "area"
-	AreaStacked                 = "areaStacked"
-	AreaPercentStacked          = "areaPercentStacked"
-	Area3D                      = "area3D"
-	Area3DStacked               = "area3DStacked"
-	Area3DPercentStacked        = "area3DPercentStacked"
-	Bar                         = "bar"
-	BarStacked                  = "barStacked"
-	BarPercentStacked           = "barPercentStacked"
-	Bar3DClustered              = "bar3DClustered"
-	Bar3DStacked                = "bar3DStacked"
-	Bar3DPercentStacked         = "bar3DPercentStacked"
-	Bar3DConeClustered          = "bar3DConeClustered"
-	Bar3DConeStacked            = "bar3DConeStacked"
-	Bar3DConePercentStacked     = "bar3DConePercentStacked"
-	Bar3DPyramidClustered       = "bar3DPyramidClustered"
-	Bar3DPyramidStacked         = "bar3DPyramidStacked"
-	Bar3DPyramidPercentStacked  = "bar3DPyramidPercentStacked"
-	Bar3DCylinderClustered      = "bar3DCylinderClustered"
-	Bar3DCylinderStacked        = "bar3DCylinderStacked"
-	Bar3DCylinderPercentStacked = "bar3DCylinderPercentStacked"
-	Col                         = "col"
-	ColStacked                  = "colStacked"
-	ColPercentStacked           = "colPercentStacked"
-	Col3D                       = "col3D"
-	Col3DClustered              = "col3DClustered"
-	Col3DStacked                = "col3DStacked"
-	Col3DPercentStacked         = "col3DPercentStacked"
-	Col3DCone                   = "col3DCone"
-	Col3DConeClustered          = "col3DConeClustered"
-	Col3DConeStacked            = "col3DConeStacked"
-	Col3DConePercentStacked     = "col3DConePercentStacked"
-	Col3DPyramid                = "col3DPyramid"
-	Col3DPyramidClustered       = "col3DPyramidClustered"
-	Col3DPyramidStacked         = "col3DPyramidStacked"
-	Col3DPyramidPercentStacked  = "col3DPyramidPercentStacked"
-	Col3DCylinder               = "col3DCylinder"
-	Col3DCylinderClustered      = "col3DCylinderClustered"
-	Col3DCylinderStacked        = "col3DCylinderStacked"
-	Col3DCylinderPercentStacked = "col3DCylinderPercentStacked"
-	Doughnut                    = "doughnut"
-	Line                        = "line"
-	Pie                         = "pie"
-	Pie3D                       = "pie3D"
-	PieOfPieChart               = "pieOfPie"
-	BarOfPieChart               = "barOfPie"
-	Radar                       = "radar"
-	Scatter                     = "scatter"
-	Surface3D                   = "surface3D"
-	WireframeSurface3D          = "wireframeSurface3D"
-	Contour                     = "contour"
-	WireframeContour            = "wireframeContour"
-	Bubble                      = "bubble"
-	Bubble3D                    = "bubble3D"
+	Area ChartType = iota
+	AreaStacked
+	AreaPercentStacked
+	Area3D
+	Area3DStacked
+	Area3DPercentStacked
+	Bar
+	BarStacked
+	BarPercentStacked
+	Bar3DClustered
+	Bar3DStacked
+	Bar3DPercentStacked
+	Bar3DConeClustered
+	Bar3DConeStacked
+	Bar3DConePercentStacked
+	Bar3DPyramidClustered
+	Bar3DPyramidStacked
+	Bar3DPyramidPercentStacked
+	Bar3DCylinderClustered
+	Bar3DCylinderStacked
+	Bar3DCylinderPercentStacked
+	Col
+	ColStacked
+	ColPercentStacked
+	Col3D
+	Col3DClustered
+	Col3DStacked
+	Col3DPercentStacked
+	Col3DCone
+	Col3DConeClustered
+	Col3DConeStacked
+	Col3DConePercentStacked
+	Col3DPyramid
+	Col3DPyramidClustered
+	Col3DPyramidStacked
+	Col3DPyramidPercentStacked
+	Col3DCylinder
+	Col3DCylinderClustered
+	Col3DCylinderStacked
+	Col3DCylinderPercentStacked
+	Doughnut
+	Line
+	Line3D
+	Pie
+	Pie3D
+	PieOfPie
+	BarOfPie
+	Radar
+	Scatter
+	Surface3D
+	WireframeSurface3D
+	Contour
+	WireframeContour
+	Bubble
+	Bubble3D
+)
+
+// ChartLineType is the type of supported chart line types.
+type ChartLineType byte
+
+// This section defines the currently supported chart line types enumeration.
+const (
+	ChartLineUnset ChartLineType = iota
+	ChartLineSolid
+	ChartLineNone
+	ChartLineAutomatic
+)
+
+// ChartTickLabelPositionType is the type of supported chart tick label position
+// types.
+type ChartTickLabelPositionType byte
+
+// This section defines the supported chart tick label position types
+// enumeration.
+const (
+	ChartTickLabelNextToAxis ChartTickLabelPositionType = iota
+	ChartTickLabelHigh
+	ChartTickLabelLow
+	ChartTickLabelNone
 )
 
 // This section defines the default value of chart properties.
 var (
-	chartView3DRotX = map[string]int{
+	chartView3DRotX = map[ChartType]int{
 		Area:                        0,
 		AreaStacked:                 0,
 		AreaPercentStacked:          0,
@@ -123,10 +149,11 @@ var (
 		Col3DCylinderPercentStacked: 15,
 		Doughnut:                    0,
 		Line:                        0,
+		Line3D:                      20,
 		Pie:                         0,
 		Pie3D:                       30,
-		PieOfPieChart:               0,
-		BarOfPieChart:               0,
+		PieOfPie:                    0,
+		BarOfPie:                    0,
 		Radar:                       0,
 		Scatter:                     0,
 		Surface3D:                   15,
@@ -134,7 +161,7 @@ var (
 		Contour:                     90,
 		WireframeContour:            90,
 	}
-	chartView3DRotY = map[string]int{
+	chartView3DRotY = map[ChartType]int{
 		Area:                        0,
 		AreaStacked:                 0,
 		AreaPercentStacked:          0,
@@ -177,10 +204,11 @@ var (
 		Col3DCylinderPercentStacked: 20,
 		Doughnut:                    0,
 		Line:                        0,
+		Line3D:                      15,
 		Pie:                         0,
 		Pie3D:                       0,
-		PieOfPieChart:               0,
-		BarOfPieChart:               0,
+		PieOfPie:                    0,
+		BarOfPie:                    0,
 		Radar:                       0,
 		Scatter:                     0,
 		Surface3D:                   20,
@@ -188,17 +216,18 @@ var (
 		Contour:                     0,
 		WireframeContour:            0,
 	}
-	plotAreaChartOverlap = map[string]int{
+	plotAreaChartOverlap = map[ChartType]int{
 		BarStacked:        100,
 		BarPercentStacked: 100,
 		ColStacked:        100,
 		ColPercentStacked: 100,
 	}
-	chartView3DPerspective = map[string]int{
+	chartView3DPerspective = map[ChartType]int{
+		Line3D:           30,
 		Contour:          0,
 		WireframeContour: 0,
 	}
-	chartView3DRAngAx = map[string]int{
+	chartView3DRAngAx = map[ChartType]int{
 		Area:                        0,
 		AreaStacked:                 0,
 		AreaPercentStacked:          0,
@@ -241,10 +270,11 @@ var (
 		Col3DCylinderPercentStacked: 1,
 		Doughnut:                    0,
 		Line:                        0,
+		Line3D:                      0,
 		Pie:                         0,
 		Pie3D:                       0,
-		PieOfPieChart:               0,
-		BarOfPieChart:               0,
+		PieOfPie:                    0,
+		BarOfPie:                    0,
 		Radar:                       0,
 		Scatter:                     0,
 		Surface3D:                   0,
@@ -260,7 +290,7 @@ var (
 		"top":       "t",
 		"top_right": "tr",
 	}
-	chartValAxNumFmtFormatCode = map[string]string{
+	chartValAxNumFmtFormatCode = map[ChartType]string{
 		Area:                        "General",
 		AreaStacked:                 "General",
 		AreaPercentStacked:          "0%",
@@ -303,10 +333,11 @@ var (
 		Col3DCylinderPercentStacked: "0%",
 		Doughnut:                    "General",
 		Line:                        "General",
+		Line3D:                      "General",
 		Pie:                         "General",
 		Pie3D:                       "General",
-		PieOfPieChart:               "General",
-		BarOfPieChart:               "General",
+		PieOfPie:                    "General",
+		BarOfPie:                    "General",
 		Radar:                       "General",
 		Scatter:                     "General",
 		Surface3D:                   "General",
@@ -316,7 +347,7 @@ var (
 		Bubble:                      "General",
 		Bubble3D:                    "General",
 	}
-	chartValAxCrossBetween = map[string]string{
+	chartValAxCrossBetween = map[ChartType]string{
 		Area:                        "midCat",
 		AreaStacked:                 "midCat",
 		AreaPercentStacked:          "midCat",
@@ -359,10 +390,11 @@ var (
 		Col3DCylinderPercentStacked: "between",
 		Doughnut:                    "between",
 		Line:                        "between",
+		Line3D:                      "between",
 		Pie:                         "between",
 		Pie3D:                       "between",
-		PieOfPieChart:               "between",
-		BarOfPieChart:               "between",
+		PieOfPie:                    "between",
+		BarOfPie:                    "between",
 		Radar:                       "between",
 		Scatter:                     "between",
 		Surface3D:                   "midCat",
@@ -372,7 +404,7 @@ var (
 		Bubble:                      "midCat",
 		Bubble3D:                    "midCat",
 	}
-	plotAreaChartGrouping = map[string]string{
+	plotAreaChartGrouping = map[ChartType]string{
 		Area:                        "standard",
 		AreaStacked:                 "stacked",
 		AreaPercentStacked:          "percentStacked",
@@ -414,8 +446,9 @@ var (
 		Col3DCylinderStacked:        "stacked",
 		Col3DCylinderPercentStacked: "percentStacked",
 		Line:                        "standard",
+		Line3D:                      "standard",
 	}
-	plotAreaChartBarDir = map[string]string{
+	plotAreaChartBarDir = map[ChartType]string{
 		Bar:                         "bar",
 		BarStacked:                  "bar",
 		BarPercentStacked:           "bar",
@@ -451,6 +484,43 @@ var (
 		Col3DCylinderStacked:        "col",
 		Col3DCylinderPercentStacked: "col",
 		Line:                        "standard",
+		Line3D:                      "standard",
+	}
+	barColChartTypes = []ChartType{
+		Bar,
+		BarStacked,
+		BarPercentStacked,
+		Bar3DClustered,
+		Bar3DStacked,
+		Bar3DPercentStacked,
+		Bar3DConeClustered,
+		Bar3DConeStacked,
+		Bar3DConePercentStacked,
+		Bar3DPyramidClustered,
+		Bar3DPyramidStacked,
+		Bar3DPyramidPercentStacked,
+		Bar3DCylinderClustered,
+		Bar3DCylinderStacked,
+		Bar3DCylinderPercentStacked,
+		Col,
+		ColStacked,
+		ColPercentStacked,
+		Col3D,
+		Col3DClustered,
+		Col3DStacked,
+		Col3DPercentStacked,
+		Col3DCone,
+		Col3DConeStacked,
+		Col3DConeClustered,
+		Col3DConePercentStacked,
+		Col3DPyramid,
+		Col3DPyramidClustered,
+		Col3DPyramidStacked,
+		Col3DPyramidPercentStacked,
+		Col3DCylinder,
+		Col3DCylinderClustered,
+		Col3DCylinderStacked,
+		Col3DCylinderPercentStacked,
 	}
 	orientation = map[bool]string{
 		true:  "maxMin",
@@ -464,41 +534,71 @@ var (
 		true:  "r",
 		false: "l",
 	}
-	valTickLblPos = map[string]string{
+	tickLblPosVal = map[ChartTickLabelPositionType]string{
+		ChartTickLabelNextToAxis: "nextTo",
+		ChartTickLabelHigh:       "high",
+		ChartTickLabelLow:        "low",
+		ChartTickLabelNone:       "none",
+	}
+	tickLblPosNone = map[ChartType]string{
 		Contour:          "none",
 		WireframeContour: "none",
 	}
 )
 
-// parseFormatChartSet provides a function to parse the format settings of the
+// parseChartOptions provides a function to parse the format settings of the
 // chart with default value.
-func parseFormatChartSet(formatSet string) (*formatChart, error) {
-	format := formatChart{
-		Dimension: formatChartDimension{
-			Width:  480,
-			Height: 290,
-		},
-		Format: formatPicture{
-			FPrintsWithSheet: true,
-			FLocksWithSheet:  false,
-			NoChangeAspect:   false,
-			OffsetX:          0,
-			OffsetY:          0,
-			XScale:           1.0,
-			YScale:           1.0,
-		},
-		Legend: formatChartLegend{
-			Position:      "bottom",
-			ShowLegendKey: false,
-		},
-		Title: formatChartTitle{
-			Name: " ",
-		},
-		VaryColors:   true,
-		ShowBlanksAs: "gap",
+func parseChartOptions(opts *Chart) (*Chart, error) {
+	if opts == nil {
+		return nil, ErrParameterInvalid
 	}
-	err := json.Unmarshal([]byte(formatSet), &format)
-	return &format, err
+	if opts.Dimension.Width == 0 {
+		opts.Dimension.Width = defaultChartDimensionWidth
+	}
+	if opts.Dimension.Height == 0 {
+		opts.Dimension.Height = defaultChartDimensionHeight
+	}
+	if opts.Format.PrintObject == nil {
+		opts.Format.PrintObject = boolPtr(true)
+	}
+	if opts.Format.Locked == nil {
+		opts.Format.Locked = boolPtr(false)
+	}
+	if opts.Format.ScaleX == 0 {
+		opts.Format.ScaleX = defaultDrawingScale
+	}
+	if opts.Format.ScaleY == 0 {
+		opts.Format.ScaleY = defaultDrawingScale
+	}
+	if opts.Legend.Position == "" {
+		opts.Legend.Position = defaultChartLegendPosition
+	}
+	opts.parseTitle()
+	if opts.VaryColors == nil {
+		opts.VaryColors = boolPtr(true)
+	}
+	if opts.Border.Width == 0 {
+		opts.Border.Width = 0.75
+	}
+	if opts.ShowBlanksAs == "" {
+		opts.ShowBlanksAs = defaultChartShowBlanksAs
+	}
+	return opts, nil
+}
+
+// parseTitle parse the title settings of the chart with default value.
+func (opts *Chart) parseTitle() {
+	for i := range opts.Title {
+		if opts.Title[i].Font == nil {
+			opts.Title[i].Font = &Font{}
+		}
+		if opts.Title[i].Font.Color == "" {
+			opts.Title[i].Font.Color = "595959"
+		}
+		if opts.Title[i].Font.Size == 0 {
+			opts.Title[i].Font.Size = 14
+		}
+	}
 }
 
 // AddChart provides the method to add chart in a sheet by given chart format
@@ -506,205 +606,229 @@ func parseFormatChartSet(formatSet string) (*formatChart, error) {
 // properties set. For example, create 3D clustered column chart with data
 // Sheet1!$E$1:$L$15:
 //
-//    package main
+//	package main
 //
-//    import (
-//        "fmt"
+//	import (
+//	    "fmt"
 //
-//        "github.com/360EntSecGroup-Skylar/excelize/v2"
-//    )
+//	    "github.com/xuri/excelize/v2"
+//	)
 //
-//    func main() {
-//        categories := map[string]string{
-//            "A2": "Small", "A3": "Normal", "A4": "Large",
-//            "B1": "Apple", "C1": "Orange", "D1": "Pear"}
-//        values := map[string]int{
-//            "B2": 2, "C2": 3, "D2": 3, "B3": 5, "C3": 2, "D3": 4, "B4": 6, "C4": 7, "D4": 8}
-//        f := excelize.NewFile()
-//        for k, v := range categories {
-//            f.SetCellValue("Sheet1", k, v)
-//        }
-//        for k, v := range values {
-//            f.SetCellValue("Sheet1", k, v)
-//        }
-//        if err := f.AddChart("Sheet1", "E1", `{
-//            "type": "col3DClustered",
-//            "series": [
-//            {
-//                "name": "Sheet1!$A$2",
-//                "categories": "Sheet1!$B$1:$D$1",
-//                "values": "Sheet1!$B$2:$D$2"
-//            },
-//            {
-//                "name": "Sheet1!$A$3",
-//                "categories": "Sheet1!$B$1:$D$1",
-//                "values": "Sheet1!$B$3:$D$3"
-//            },
-//            {
-//                "name": "Sheet1!$A$4",
-//                "categories": "Sheet1!$B$1:$D$1",
-//                "values": "Sheet1!$B$4:$D$4"
-//            }],
-//            "title":
-//            {
-//                "name": "Fruit 3D Clustered Column Chart"
-//            },
-//            "legend":
-//            {
-//                "none": false,
-//                "position": "bottom",
-//                "show_legend_key": false
-//            },
-//            "plotarea":
-//            {
-//                "show_bubble_size": true,
-//                "show_cat_name": false,
-//                "show_leader_lines": false,
-//                "show_percent": true,
-//                "show_series_name": true,
-//                "show_val": true
-//            },
-//            "show_blanks_as": "zero",
-//            "x_axis":
-//            {
-//                "reverse_order": true
-//            },
-//            "y_axis":
-//            {
-//                "maximum": 7.5,
-//                "minimum": 0.5
-//            }
-//        }`); err != nil {
-//            fmt.Println(err)
-//            return
-//        }
-//        // Save spreadsheet by the given path.
-//        if err := f.SaveAs("Book1.xlsx"); err != nil {
-//            fmt.Println(err)
-//        }
-//    }
+//	func main() {
+//	    f := excelize.NewFile()
+//	    defer func() {
+//	        if err := f.Close(); err != nil {
+//	            fmt.Println(err)
+//	        }
+//	    }()
+//	    for idx, row := range [][]interface{}{
+//	        {nil, "Apple", "Orange", "Pear"}, {"Small", 2, 3, 3},
+//	        {"Normal", 5, 2, 4}, {"Large", 6, 7, 8},
+//	    } {
+//	        cell, err := excelize.CoordinatesToCellName(1, idx+1)
+//	        if err != nil {
+//	            fmt.Println(err)
+//	            return
+//	        }
+//	        f.SetSheetRow("Sheet1", cell, &row)
+//	    }
+//	    if err := f.AddChart("Sheet1", "E1", &excelize.Chart{
+//	        Type: excelize.Col3DClustered,
+//	        Series: []excelize.ChartSeries{
+//	            {
+//	                Name:       "Sheet1!$A$2",
+//	                Categories: "Sheet1!$B$1:$D$1",
+//	                Values:     "Sheet1!$B$2:$D$2",
+//	            },
+//	            {
+//	                Name:       "Sheet1!$A$3",
+//	                Categories: "Sheet1!$B$1:$D$1",
+//	                Values:     "Sheet1!$B$3:$D$3",
+//	            },
+//	            {
+//	                Name:       "Sheet1!$A$4",
+//	                Categories: "Sheet1!$B$1:$D$1",
+//	                Values:     "Sheet1!$B$4:$D$4",
+//	            },
+//	        },
+//	        Title: []excelize.RichTextRun{
+//	            {
+//	                Text: "Fruit 3D Clustered Column Chart",
+//	            },
+//	        },
+//	        Legend: excelize.ChartLegend{
+//	            ShowLegendKey: false,
+//	        },
+//	        PlotArea: excelize.ChartPlotArea{
+//	            ShowBubbleSize:  true,
+//	            ShowCatName:     false,
+//	            ShowLeaderLines: false,
+//	            ShowPercent:     true,
+//	            ShowSerName:     true,
+//	            ShowVal:         true,
+//	        },
+//	    }); err != nil {
+//	        fmt.Println(err)
+//	        return
+//	    }
+//	    // Save spreadsheet by the given path.
+//	    if err := f.SaveAs("Book1.xlsx"); err != nil {
+//	        fmt.Println(err)
+//	    }
+//	}
 //
 // The following shows the type of chart supported by excelize:
 //
-//     Type                        | Chart
-//    -----------------------------+------------------------------
-//     area                        | 2D area chart
-//     areaStacked                 | 2D stacked area chart
-//     areaPercentStacked          | 2D 100% stacked area chart
-//     area3D                      | 3D area chart
-//     area3DStacked               | 3D stacked area chart
-//     area3DPercentStacked        | 3D 100% stacked area chart
-//     bar                         | 2D clustered bar chart
-//     barStacked                  | 2D stacked bar chart
-//     barPercentStacked           | 2D 100% stacked bar chart
-//     bar3DClustered              | 3D clustered bar chart
-//     bar3DStacked                | 3D stacked bar chart
-//     bar3DPercentStacked         | 3D 100% stacked bar chart
-//     bar3DConeClustered          | 3D cone clustered bar chart
-//     bar3DConeStacked            | 3D cone stacked bar chart
-//     bar3DConePercentStacked     | 3D cone percent bar chart
-//     bar3DPyramidClustered       | 3D pyramid clustered bar chart
-//     bar3DPyramidStacked         | 3D pyramid stacked bar chart
-//     bar3DPyramidPercentStacked  | 3D pyramid percent stacked bar chart
-//     bar3DCylinderClustered      | 3D cylinder clustered bar chart
-//     bar3DCylinderStacked        | 3D cylinder stacked bar chart
-//     bar3DCylinderPercentStacked | 3D cylinder percent stacked bar chart
-//     col                         | 2D clustered column chart
-//     colStacked                  | 2D stacked column chart
-//     colPercentStacked           | 2D 100% stacked column chart
-//     col3DClustered              | 3D clustered column chart
-//     col3D                       | 3D column chart
-//     col3DStacked                | 3D stacked column chart
-//     col3DPercentStacked         | 3D 100% stacked column chart
-//     col3DCone                   | 3D cone column chart
-//     col3DConeClustered          | 3D cone clustered column chart
-//     col3DConeStacked            | 3D cone stacked column chart
-//     col3DConePercentStacked     | 3D cone percent stacked column chart
-//     col3DPyramid                | 3D pyramid column chart
-//     col3DPyramidClustered       | 3D pyramid clustered column chart
-//     col3DPyramidStacked         | 3D pyramid stacked column chart
-//     col3DPyramidPercentStacked  | 3D pyramid percent stacked column chart
-//     col3DCylinder               | 3D cylinder column chart
-//     col3DCylinderClustered      | 3D cylinder clustered column chart
-//     col3DCylinderStacked        | 3D cylinder stacked column chart
-//     col3DCylinderPercentStacked | 3D cylinder percent stacked column chart
-//     doughnut                    | doughnut chart
-//     line                        | line chart
-//     pie                         | pie chart
-//     pie3D                       | 3D pie chart
-//     pieOfPie                    | pie of pie chart
-//     barOfPie                    | bar of pie chart
-//     radar                       | radar chart
-//     scatter                     | scatter chart
-//     surface3D                   | 3D surface chart
-//     wireframeSurface3D          | 3D wireframe surface chart
-//     contour                     | contour chart
-//     wireframeContour            | wireframe contour chart
-//     bubble                      | bubble chart
-//     bubble3D                    | 3D bubble chart
+//	 ID | Enumeration                 | Chart
+//	----+-----------------------------+------------------------------
+//	 0  | Area                        | 2D area chart
+//	 1  | AreaStacked                 | 2D stacked area chart
+//	 2  | AreaPercentStacked          | 2D 100% stacked area chart
+//	 3  | Area3D                      | 3D area chart
+//	 4  | Area3DStacked               | 3D stacked area chart
+//	 5  | Area3DPercentStacked        | 3D 100% stacked area chart
+//	 6  | Bar                         | 2D clustered bar chart
+//	 7  | BarStacked                  | 2D stacked bar chart
+//	 8  | BarPercentStacked           | 2D 100% stacked bar chart
+//	 9  | Bar3DClustered              | 3D clustered bar chart
+//	 10 | Bar3DStacked                | 3D stacked bar chart
+//	 11 | Bar3DPercentStacked         | 3D 100% stacked bar chart
+//	 12 | Bar3DConeClustered          | 3D cone clustered bar chart
+//	 13 | Bar3DConeStacked            | 3D cone stacked bar chart
+//	 14 | Bar3DConePercentStacked     | 3D cone percent bar chart
+//	 15 | Bar3DPyramidClustered       | 3D pyramid clustered bar chart
+//	 16 | Bar3DPyramidStacked         | 3D pyramid stacked bar chart
+//	 17 | Bar3DPyramidPercentStacked  | 3D pyramid percent stacked bar chart
+//	 18 | Bar3DCylinderClustered      | 3D cylinder clustered bar chart
+//	 19 | Bar3DCylinderStacked        | 3D cylinder stacked bar chart
+//	 20 | Bar3DCylinderPercentStacked | 3D cylinder percent stacked bar chart
+//	 21 | Col                         | 2D clustered column chart
+//	 22 | ColStacked                  | 2D stacked column chart
+//	 23 | ColPercentStacked           | 2D 100% stacked column chart
+//	 24 | Col3DClustered              | 3D clustered column chart
+//	 25 | Col3D                       | 3D column chart
+//	 26 | Col3DStacked                | 3D stacked column chart
+//	 27 | Col3DPercentStacked         | 3D 100% stacked column chart
+//	 28 | Col3DCone                   | 3D cone column chart
+//	 29 | Col3DConeClustered          | 3D cone clustered column chart
+//	 30 | Col3DConeStacked            | 3D cone stacked column chart
+//	 31 | Col3DConePercentStacked     | 3D cone percent stacked column chart
+//	 32 | Col3DPyramid                | 3D pyramid column chart
+//	 33 | Col3DPyramidClustered       | 3D pyramid clustered column chart
+//	 34 | Col3DPyramidStacked         | 3D pyramid stacked column chart
+//	 35 | Col3DPyramidPercentStacked  | 3D pyramid percent stacked column chart
+//	 36 | Col3DCylinder               | 3D cylinder column chart
+//	 37 | Col3DCylinderClustered      | 3D cylinder clustered column chart
+//	 38 | Col3DCylinderStacked        | 3D cylinder stacked column chart
+//	 39 | Col3DCylinderPercentStacked | 3D cylinder percent stacked column chart
+//	 40 | Doughnut                    | doughnut chart
+//	 41 | Line                        | line chart
+//	 42 | Line3D                      | 3D line chart
+//	 43 | Pie                         | pie chart
+//	 44 | Pie3D                       | 3D pie chart
+//	 45 | PieOfPie                    | pie of pie chart
+//	 46 | BarOfPie                    | bar of pie chart
+//	 47 | Radar                       | radar chart
+//	 48 | Scatter                     | scatter chart
+//	 49 | Surface3D                   | 3D surface chart
+//	 50 | WireframeSurface3D          | 3D wireframe surface chart
+//	 51 | Contour                     | contour chart
+//	 52 | WireframeContour            | wireframe contour chart
+//	 53 | Bubble                      | bubble chart
+//	 54 | Bubble3D                    | 3D bubble chart
 //
-// In Excel a chart series is a collection of information that defines which data is plotted such as values, axis labels and formatting.
+// In Excel a chart series is a collection of information that defines which
+// data is plotted such as values, axis labels and formatting.
 //
 // The series options that can be set are:
 //
-//    name
-//    categories
-//    values
-//    line
-//    marker
+//	Name
+//	Categories
+//	Values
+//	Fill
+//	Line
+//	Marker
+//	DataLabel
+//	DataLabelPosition
 //
-// name: Set the name for the series. The name is displayed in the chart legend and in the formula bar. The name property is optional and if it isn't supplied it will default to Series 1..n. The name can also be a formula such as Sheet1!$A$1
+// Name: Set the name for the series. The name is displayed in the chart legend
+// and in the formula bar. The 'Name' property is optional and if it isn't
+// supplied it will default to Series 1..n. The name can also be a formula such
+// as Sheet1!$A$1
 //
-// categories: This sets the chart category labels. The category is more or less the same as the X axis. In most chart types the categories property is optional and the chart will just assume a sequential series from 1..n.
+// Categories: This sets the chart category labels. The category is more or less
+// the same as the X axis. In most chart types the 'Categories' property is
+// optional and the chart will just assume a sequential series from 1..n.
 //
-// values: This is the most important property of a series and is the only mandatory option for every chart object. This option links the chart with the worksheet data that it displays.
+// Values: This is the most important property of a series and is the only
+// mandatory option for every chart object. This option links the chart with
+// the worksheet data that it displays.
 //
-// line: This sets the line format of the line chart. The line property is optional and if it isn't supplied it will default style. The options that can be set is width. The range of width is 0.25pt - 999pt. If the value of width is outside the range, the default width of the line is 2pt.
+// Sizes: This sets the bubble size in a data series. The 'Sizes' property is
+// optional and the default value was same with 'Values'.
 //
-// marker: This sets the marker of the line chart and scatter chart. The range of optional field 'size' is 2-72 (default value is 5). The enumeration value of optional field 'symbol' are (default value is 'auto'):
+// Fill: This set the format for the data series fill. The 'Fill' property is
+// optional
 //
-//    circle
-//    dash
-//    diamond
-//    dot
-//    none
-//    picture
-//    plus
-//    square
-//    star
-//    triangle
-//    x
-//    auto
+// Line: This sets the line format of the line chart. The 'Line' property is
+// optional and if it isn't supplied it will default style. The options that
+// can be set are width and color. The range of width is 0.25pt - 999pt. If the
+// value of width is outside the range, the default width of the line is 2pt.
+//
+// Marker: This sets the marker of the line chart and scatter chart. The range
+// of optional field 'Size' is 2-72 (default value is 5). The enumeration value
+// of optional field 'Symbol' are (default value is 'auto'):
+//
+//	circle
+//	dash
+//	diamond
+//	dot
+//	none
+//	picture
+//	plus
+//	square
+//	star
+//	triangle
+//	x
+//	auto
+//
+// DataLabel: This sets the format of the chart series data label.
+//
+// DataLabelPosition: This sets the position of the chart series data label.
 //
 // Set properties of the chart legend. The options that can be set are:
 //
-//    none
-//    position
-//    show_legend_key
+//	Position
+//	ShowLegendKey
 //
-// none: Specified if show the legend without overlapping the chart. The default value is 'false'.
+// Position: Set the position of the chart legend. The default legend position
+// is bottom. The available positions are:
 //
-// position: Set the position of the chart legend. The default legend position is right. This parameter only takes effect when 'none' is false. The available positions are:
+//	none
+//	top
+//	bottom
+//	left
+//	right
+//	top_right
 //
-//    top
-//    bottom
-//    left
-//    right
-//    top_right
-//
-// show_legend_key: Set the legend keys shall be shown in data labels. The default value is false.
+// ShowLegendKey: Set the legend keys shall be shown in data labels. The default
+// value is false.
 //
 // Set properties of the chart title. The properties that can be set are:
 //
-//    title
+//	Title
 //
-// name: Set the name (title) for the chart. The name is displayed above the chart. The name can also be a formula such as Sheet1!$A$1 or a list with a sheetname. The name property is optional. The default is to have no chart title.
+// Title: Set the name (title) for the chart. The name is displayed above the
+// chart. The name can also be a formula such as Sheet1!$A$1 or a list with a
+// sheet name. The name property is optional. The default is to have no chart
+// title.
 //
-// Specifies how blank cells are plotted on the chart by show_blanks_as. The default value is gap. The options that can be set are:
+// Specifies how blank cells are plotted on the chart by 'ShowBlanksAs'. The
+// default value is gap. The options that can be set are:
 //
-//    gap
-//    span
-//    zero
+//	gap
+//	span
+//	zero
 //
 // gap: Specifies that blank values shall be left as a gap.
 //
@@ -712,189 +836,279 @@ func parseFormatChartSet(formatSet string) (*formatChart, error) {
 //
 // zero: Specifies that blank values shall be treated as zero.
 //
-// Specifies that each data marker in the series has a different color by vary_colors. The default value is true.
+// Specifies that each data marker in the series has a different color by
+// 'VaryColors'. The default value is true.
 //
-// Set chart offset, scale, aspect ratio setting and print settings by format, same as function AddPicture.
+// Set chart offset, scale, aspect ratio setting and print settings by 'Format',
+// same as function 'AddPicture'.
 //
-// Set the position of the chart plot area by plotarea. The properties that can be set are:
+// Set the position of the chart plot area by 'PlotArea'. The properties that
+// can be set are:
 //
-//    show_bubble_size
-//    show_cat_name
-//    show_leader_lines
-//    show_percent
-//    show_series_name
-//    show_val
+//	SecondPlotValues
+//	ShowBubbleSize
+//	ShowCatName
+//	ShowLeaderLines
+//	ShowPercent
+//	ShowSerName
+//	ShowVal
+//	NumFmt
 //
-// show_bubble_size: Specifies the bubble size shall be shown in a data label. The show_bubble_size property is optional. The default value is false.
+// SecondPlotValues: Specifies the values in second plot for the 'pieOfPie' and
+// 'barOfPie' chart.
 //
-// show_cat_name: Specifies that the category name shall be shown in the data label. The show_cat_name property is optional. The default value is true.
+// ShowBubbleSize: Specifies the bubble size shall be shown in a data label. The
+// 'ShowBubbleSize' property is optional. The default value is false.
 //
-// show_leader_lines: Specifies leader lines shall be shown for data labels. The show_leader_lines property is optional. The default value is false.
+// ShowCatName: Specifies that the category name shall be shown in the data
+// label. The 'ShowCatName' property is optional. The default value is true.
 //
-// show_percent: Specifies that the percentage shall be shown in a data label. The show_percent property is optional. The default value is false.
+// ShowLeaderLines: Specifies leader lines shall be shown for data labels. The
+// 'ShowLeaderLines' property is optional. The default value is false.
 //
-// show_series_name: Specifies that the series name shall be shown in a data label. The show_series_name property is optional. The default value is false.
+// ShowPercent: Specifies that the percentage shall be shown in a data label.
+// The 'ShowPercent' property is optional. The default value is false.
 //
-// show_val: Specifies that the value shall be shown in a data label. The show_val property is optional. The default value is false.
+// ShowSerName: Specifies that the series name shall be shown in a data label.
+// The 'ShowSerName' property is optional. The default value is false.
 //
-// Set the primary horizontal and vertical axis options by x_axis and y_axis. The properties of x_axis that can be set are:
+// ShowVal: Specifies that the value shall be shown in a data label.
+// The 'ShowVal' property is optional. The default value is false.
 //
-//    none
-//    major_grid_lines
-//    minor_grid_lines
-//    tick_label_skip
-//    reverse_order
-//    maximum
-//    minimum
+// NumFmt: Specifies that if linked to source and set custom number format code
+// for data labels. The 'NumFmt' property is optional. The default format code
+// is 'General'.
 //
-// The properties of y_axis that can be set are:
+// Set the primary horizontal and vertical axis options by 'XAxis' and 'YAxis'.
+// The properties of 'XAxis' that can be set are:
 //
-//    none
-//    major_grid_lines
-//    minor_grid_lines
-//    major_unit
-//    reverse_order
-//    maximum
-//    minimum
+//	None
+//	MajorGridLines
+//	MinorGridLines
+//	TickLabelSkip
+//	ReverseOrder
+//	Maximum
+//	Minimum
+//	Alignment
+//	Font
+//	NumFmt
+//	Title
 //
-// none: Disable axes.
+// The properties of 'YAxis' that can be set are:
 //
-// major_grid_lines: Specifies major gridlines.
+//	None
+//	MajorGridLines
+//	MinorGridLines
+//	MajorUnit
+//	Secondary
+//	ReverseOrder
+//	Maximum
+//	Minimum
+//	Alignment
+//	Font
+//	LogBase
+//	NumFmt
+//	Title
 //
-// minor_grid_lines: Specifies minor gridlines.
+// None: Disable axes.
 //
-// major_unit: Specifies the distance between major ticks. Shall contain a positive floating-point number. The major_unit property is optional. The default value is auto.
+// MajorGridLines: Specifies major grid lines.
 //
-// tick_label_skip: Specifies how many tick labels to skip between label that is drawn. The tick_label_skip property is optional. The default value is auto.
+// MinorGridLines: Specifies minor grid lines.
 //
-// reverse_order: Specifies that the categories or values on reverse order (orientation of the chart). The reverse_order property is optional. The default value is false.
+// MajorUnit: Specifies the distance between major ticks. Shall contain a
+// positive floating-point number. The 'MajorUnit' property is optional. The
+// default value is auto.
 //
-// maximum: Specifies that the fixed maximum, 0 is auto. The maximum property is optional. The default value is auto.
+// Secondary: Specifies the current series vertical axis as the secondary axis,
+// this only works for the second and later chart in the combo chart. The
+// default value is false.
 //
-// minimum: Specifies that the fixed minimum, 0 is auto. The minimum property is optional. The default value is auto.
+// TickLabelSkip: Specifies how many tick labels to skip between label that is
+// drawn. The 'TickLabelSkip' property is optional. The default value is auto.
 //
-// Set chart size by dimension property. The dimension property is optional. The default width is 480, and height is 290.
+// ReverseOrder: Specifies that the categories or values on reverse order
+// (orientation of the chart). The 'ReverseOrder' property is optional. The
+// default value is false.
 //
-// combo: Specifies the create a chart that combines two or more chart types
-// in a single chart. For example, create a clustered column - line chart with
+// Maximum: Specifies that the fixed maximum, 0 is auto. The 'Maximum' property
+// is optional. The default value is auto.
+//
+// Minimum: Specifies that the fixed minimum, 0 is auto. The 'Minimum' property
+// is optional. The default value is auto.
+//
+// Alignment: Specifies that the alignment of the horizontal and vertical axis.
+// The properties of alignment that can be set are:
+//
+//	TextRotation
+//	Vertical
+//
+// The value of 'TextRotation' that can be set from -90 to 90.
+//
+// The value of 'Vertical' that can be set are:
+//
+//	horz
+//	vert
+//	vert270
+//	wordArtVert
+//	eaVert
+//	mongolianVert
+//	wordArtVertRtl
+//
+// Font: Specifies that the font of the horizontal and vertical axis. The
+// properties of font that can be set are:
+//
+//	Bold
+//	Italic
+//	Underline
+//	Family
+//	Size
+//	Strike
+//	Color
+//	VertAlign
+//
+// LogBase: Specifies logarithmic scale base number of the vertical axis.
+//
+// NumFmt: Specifies that if linked to source and set custom number format code
+// for axis. The 'NumFmt' property is optional. The default format code is
+// 'General'.
+//
+// Title: Specifies that the primary horizontal or vertical axis title and
+// resize chart. The 'Title' property is optional.
+//
+// Set chart size by 'Dimension' property. The 'Dimension' property is optional.
+// The default width is 480, and height is 260.
+//
+// Set the bubble size in all data series for the bubble chart or 3D bubble
+// chart by 'BubbleSizes' property. The 'BubbleSizes' property is optional. The
+// default width is 100, and the value should be great than 0 and less or equal
+// than 300.
+//
+// Set the doughnut hole size in all data series for the doughnut chart by
+// 'HoleSize' property. The 'HoleSize' property is optional. The default width
+// is 75, and the value should be great than 0 and less or equal than 90.
+//
+// Set the gap with of the column and bar series chart by 'GapWidth' property.
+// The 'GapWidth' property is optional. The default width is 150, and the value
+// should be great or equal than 0 and less or equal than 500.
+//
+// Set series overlap of the column and bar series chart by 'Overlap' property.
+// The 'Overlap' property is optional. The default width is 0, and the value
+// should be great or equal than -100 and less or equal than 100.
+//
+// combo: Specifies the create a chart that combines two or more chart types in
+// a single chart. For example, create a clustered column - line chart with
 // data Sheet1!$E$1:$L$15:
 //
-//    package main
+//	package main
 //
-//    import (
-//        "fmt"
+//	import (
+//	    "fmt"
 //
-//        "github.com/360EntSecGroup-Skylar/excelize/v2"
-//    )
+//	    "github.com/xuri/excelize/v2"
+//	)
 //
-//    func main() {
-//        categories := map[string]string{
-//            "A2": "Small", "A3": "Normal", "A4": "Large",
-//            "B1": "Apple", "C1": "Orange", "D1": "Pear"}
-//        values := map[string]int{
-//            "B2": 2, "C2": 3, "D2": 3, "B3": 5, "C3": 2, "D3": 4, "B4": 6, "C4": 7, "D4": 8}
-//        f := excelize.NewFile()
-//        for k, v := range categories {
-//            f.SetCellValue("Sheet1", k, v)
-//        }
-//        for k, v := range values {
-//            f.SetCellValue("Sheet1", k, v)
-//        }
-//        if err := f.AddChart("Sheet1", "E1", `{
-//            "type": "col",
-//            "series": [
-//            {
-//                "name": "Sheet1!$A$2",
-//                "categories": "",
-//                "values": "Sheet1!$B$2:$D$2"
-//            },
-//            {
-//                "name": "Sheet1!$A$3",
-//                "categories": "Sheet1!$B$1:$D$1",
-//                "values": "Sheet1!$B$3:$D$3"
-//            }],
-//            "format":
-//            {
-//                "x_scale": 1.0,
-//                "y_scale": 1.0,
-//                "x_offset": 15,
-//                "y_offset": 10,
-//                "print_obj": true,
-//                "lock_aspect_ratio": false,
-//                "locked": false
-//            },
-//            "title":
-//            {
-//                "name": "Clustered Column - Line Chart"
-//            },
-//            "legend":
-//            {
-//                "position": "left",
-//                "show_legend_key": false
-//            },
-//            "plotarea":
-//            {
-//                "show_bubble_size": true,
-//                "show_cat_name": false,
-//                "show_leader_lines": false,
-//                "show_percent": true,
-//                "show_series_name": true,
-//                "show_val": true
-//            }
-//        }`, `{
-//            "type": "line",
-//            "series": [
-//            {
-//                "name": "Sheet1!$A$4",
-//                "categories": "Sheet1!$B$1:$D$1",
-//                "values": "Sheet1!$B$4:$D$4",
-//                "marker":
-//                {
-//                    "symbol": "none",
-//                    "size": 10
-//                }
-//            }],
-//            "format":
-//            {
-//                "x_scale": 1,
-//                "y_scale": 1,
-//                "x_offset": 15,
-//                "y_offset": 10,
-//                "print_obj": true,
-//                "lock_aspect_ratio": false,
-//                "locked": false
-//            },
-//            "legend":
-//            {
-//                "position": "right",
-//                "show_legend_key": false
-//            },
-//            "plotarea":
-//            {
-//                "show_bubble_size": true,
-//                "show_cat_name": false,
-//                "show_leader_lines": false,
-//                "show_percent": true,
-//                "show_series_name": true,
-//                "show_val": true
-//            }
-//        }`); err != nil {
-//            fmt.Println(err)
-//            return
-//        }
-//        // Save spreadsheet file by the given path.
-//        if err := f.SaveAs("Book1.xlsx"); err != nil {
-//            fmt.Println(err)
-//        }
-//    }
-//
-func (f *File) AddChart(sheet, cell, format string, combo ...string) error {
-	// Read sheet data.
+//	func main() {
+//	    f := excelize.NewFile()
+//	    defer func() {
+//	        if err := f.Close(); err != nil {
+//	            fmt.Println(err)
+//	        }
+//	    }()
+//	    for idx, row := range [][]interface{}{
+//	        {nil, "Apple", "Orange", "Pear"}, {"Small", 2, 3, 3},
+//	        {"Normal", 5, 2, 4}, {"Large", 6, 7, 8},
+//	    } {
+//	        cell, err := excelize.CoordinatesToCellName(1, idx+1)
+//	        if err != nil {
+//	            fmt.Println(err)
+//	            return
+//	        }
+//	        f.SetSheetRow("Sheet1", cell, &row)
+//	    }
+//	    enable, disable := true, false
+//	    if err := f.AddChart("Sheet1", "E1", &excelize.Chart{
+//	        Type: excelize.Col,
+//	        Series: []excelize.ChartSeries{
+//	            {
+//	                Name:       "Sheet1!$A$2",
+//	                Categories: "Sheet1!$B$1:$D$1",
+//	                Values:     "Sheet1!$B$2:$D$2",
+//	            },
+//	        },
+//	        Format: excelize.GraphicOptions{
+//	            ScaleX:          1,
+//	            ScaleY:          1,
+//	            OffsetX:         15,
+//	            OffsetY:         10,
+//	            PrintObject:     &enable,
+//	            LockAspectRatio: false,
+//	            Locked:          &disable,
+//	        },
+//	        Title: []excelize.RichTextRun{
+//	            {
+//	                Text: "Clustered Column - Line Chart",
+//	            },
+//	        },
+//	        Legend: excelize.ChartLegend{
+//	            Position:      "left",
+//	            ShowLegendKey: false,
+//	        },
+//	        PlotArea: excelize.ChartPlotArea{
+//	            ShowCatName:     false,
+//	            ShowLeaderLines: false,
+//	            ShowPercent:     true,
+//	            ShowSerName:     true,
+//	            ShowVal:         true,
+//	        },
+//	    }, &excelize.Chart{
+//	        Type: excelize.Line,
+//	        Series: []excelize.ChartSeries{
+//	            {
+//	                Name:       "Sheet1!$A$4",
+//	                Categories: "Sheet1!$B$1:$D$1",
+//	                Values:     "Sheet1!$B$4:$D$4",
+//	                Marker: excelize.ChartMarker{
+//	                    Symbol: "none", Size: 10,
+//	                },
+//	            },
+//	        },
+//	        Format: excelize.GraphicOptions{
+//	            ScaleX:          1,
+//	            ScaleY:          1,
+//	            OffsetX:         15,
+//	            OffsetY:         10,
+//	            PrintObject:     &enable,
+//	            LockAspectRatio: false,
+//	            Locked:          &disable,
+//	        },
+//	        Legend: excelize.ChartLegend{
+//	            Position:      "right",
+//	            ShowLegendKey: false,
+//	        },
+//	        PlotArea: excelize.ChartPlotArea{
+//	            ShowCatName:     false,
+//	            ShowLeaderLines: false,
+//	            ShowPercent:     true,
+//	            ShowSerName:     true,
+//	            ShowVal:         true,
+//	        },
+//	    }); err != nil {
+//	        fmt.Println(err)
+//	        return
+//	    }
+//	    // Save spreadsheet by the given path.
+//	    if err := f.SaveAs("Book1.xlsx"); err != nil {
+//	        fmt.Println(err)
+//	    }
+//	}
+func (f *File) AddChart(sheet, cell string, chart *Chart, combo ...*Chart) error {
+	// Read worksheet data
 	ws, err := f.workSheetReader(sheet)
 	if err != nil {
 		return err
 	}
-	formatSet, comboCharts, err := f.getFormatChart(format, combo)
+	opts, comboCharts, err := f.getChartOptions(chart, combo)
 	if err != nil {
 		return err
 	}
@@ -905,13 +1119,15 @@ func (f *File) AddChart(sheet, cell, format string, combo ...string) error {
 	drawingID, drawingXML = f.prepareDrawing(ws, drawingID, sheet, drawingXML)
 	drawingRels := "xl/drawings/_rels/drawing" + strconv.Itoa(drawingID) + ".xml.rels"
 	drawingRID := f.addRels(drawingRels, SourceRelationshipChart, "../charts/chart"+strconv.Itoa(chartID)+".xml", "")
-	err = f.addDrawingChart(sheet, drawingXML, cell, formatSet.Dimension.Width, formatSet.Dimension.Height, drawingRID, &formatSet.Format)
+	err = f.addDrawingChart(sheet, drawingXML, cell, int(opts.Dimension.Width), int(opts.Dimension.Height), drawingRID, &opts.Format)
 	if err != nil {
 		return err
 	}
-	f.addChart(formatSet, comboCharts)
-	f.addContentTypePart(chartID, "chart")
-	f.addContentTypePart(drawingID, "drawings")
+	f.addChart(opts, comboCharts)
+	if err = f.addContentTypePart(chartID, "chart"); err != nil {
+		return err
+	}
+	_ = f.addContentTypePart(drawingID, "drawings")
 	f.addSheetNameSpace(sheet, SourceRelationship)
 	return err
 }
@@ -920,12 +1136,16 @@ func (f *File) AddChart(sheet, cell, format string, combo ...string) error {
 // format set (such as offset, scale, aspect ratio setting and print settings)
 // and properties set. In Excel a chartsheet is a worksheet that only contains
 // a chart.
-func (f *File) AddChartSheet(sheet, format string, combo ...string) error {
+func (f *File) AddChartSheet(sheet string, chart *Chart, combo ...*Chart) error {
 	// Check if the worksheet already exists
-	if f.GetSheetIndex(sheet) != -1 {
-		return ErrExistsWorksheet
+	idx, err := f.GetSheetIndex(sheet)
+	if err != nil {
+		return err
 	}
-	formatSet, comboCharts, err := f.getFormatChart(format, combo)
+	if idx != -1 {
+		return ErrExistsSheet
+	}
+	opts, comboCharts, err := f.getChartOptions(chart, combo)
 	if err != nil {
 		return err
 	}
@@ -935,7 +1155,7 @@ func (f *File) AddChartSheet(sheet, format string, combo ...string) error {
 		},
 	}
 	f.SheetCount++
-	wb := f.workbookReader()
+	wb, _ := f.workbookReader()
 	sheetID := 0
 	for _, v := range wb.Sheets.Sheet {
 		if v.SheetID > sheetID {
@@ -944,19 +1164,23 @@ func (f *File) AddChartSheet(sheet, format string, combo ...string) error {
 	}
 	sheetID++
 	path := "xl/chartsheets/sheet" + strconv.Itoa(sheetID) + ".xml"
-	f.sheetMap[trimSheetName(sheet)] = path
-	f.Sheet[path] = nil
+	f.sheetMap[sheet] = path
+	f.Sheet.Store(path, nil)
 	drawingID := f.countDrawings() + 1
 	chartID := f.countCharts() + 1
 	drawingXML := "xl/drawings/drawing" + strconv.Itoa(drawingID) + ".xml"
 	f.prepareChartSheetDrawing(&cs, drawingID, sheet)
 	drawingRels := "xl/drawings/_rels/drawing" + strconv.Itoa(drawingID) + ".xml.rels"
 	drawingRID := f.addRels(drawingRels, SourceRelationshipChart, "../charts/chart"+strconv.Itoa(chartID)+".xml", "")
-	f.addSheetDrawingChart(drawingXML, drawingRID, &formatSet.Format)
-	f.addChart(formatSet, comboCharts)
-	f.addContentTypePart(chartID, "chart")
-	f.addContentTypePart(sheetID, "chartsheet")
-	f.addContentTypePart(drawingID, "drawings")
+	if err = f.addSheetDrawingChart(drawingXML, drawingRID, &opts.Format); err != nil {
+		return err
+	}
+	f.addChart(opts, comboCharts)
+	if err = f.addContentTypePart(chartID, "chart"); err != nil {
+		return err
+	}
+	_ = f.addContentTypePart(sheetID, "chartsheet")
+	_ = f.addContentTypePart(drawingID, "drawings")
 	// Update workbook.xml.rels
 	rID := f.addRels(f.getWorkbookRelsPath(), SourceRelationshipChartsheet, fmt.Sprintf("/xl/chartsheets/sheet%d.xml", sheetID), "")
 	// Update workbook.xml
@@ -967,59 +1191,61 @@ func (f *File) AddChartSheet(sheet, format string, combo ...string) error {
 	return err
 }
 
-// getFormatChart provides a function to check format set of the chart and
+// getChartOptions provides a function to check format set of the chart and
 // create chart format.
-func (f *File) getFormatChart(format string, combo []string) (*formatChart, []*formatChart, error) {
-	comboCharts := []*formatChart{}
-	formatSet, err := parseFormatChartSet(format)
+func (f *File) getChartOptions(opts *Chart, combo []*Chart) (*Chart, []*Chart, error) {
+	var comboCharts []*Chart
+	options, err := parseChartOptions(opts)
 	if err != nil {
-		return formatSet, comboCharts, err
+		return options, comboCharts, err
 	}
 	for _, comboFormat := range combo {
-		comboChart, err := parseFormatChartSet(comboFormat)
+		comboChart, err := parseChartOptions(comboFormat)
 		if err != nil {
-			return formatSet, comboCharts, err
+			return options, comboCharts, err
 		}
 		if _, ok := chartValAxNumFmtFormatCode[comboChart.Type]; !ok {
-			return formatSet, comboCharts, errors.New("unsupported chart type " + comboChart.Type)
+			return options, comboCharts, newUnsupportedChartType(comboChart.Type)
 		}
 		comboCharts = append(comboCharts, comboChart)
 	}
-	if _, ok := chartValAxNumFmtFormatCode[formatSet.Type]; !ok {
-		return formatSet, comboCharts, errors.New("unsupported chart type " + formatSet.Type)
+	if _, ok := chartValAxNumFmtFormatCode[options.Type]; !ok {
+		return options, comboCharts, newUnsupportedChartType(options.Type)
 	}
-	return formatSet, comboCharts, err
+	return options, comboCharts, err
 }
 
-// DeleteChart provides a function to delete chart in XLSX by given worksheet
-// and cell name.
-func (f *File) DeleteChart(sheet, cell string) (err error) {
+// DeleteChart provides a function to delete chart in spreadsheet by given
+// worksheet name and cell reference.
+func (f *File) DeleteChart(sheet, cell string) error {
 	col, row, err := CellNameToCoordinates(cell)
 	if err != nil {
-		return
+		return err
 	}
 	col--
 	row--
 	ws, err := f.workSheetReader(sheet)
 	if err != nil {
-		return
+		return err
 	}
 	if ws.Drawing == nil {
-		return
+		return err
 	}
-	drawingXML := strings.Replace(f.getSheetRelationshipsTargetByID(sheet, ws.Drawing.RID), "..", "xl", -1)
-	return f.deleteDrawing(col, row, drawingXML, "Chart")
+	drawingXML := strings.ReplaceAll(f.getSheetRelationshipsTargetByID(sheet, ws.Drawing.RID), "..", "xl")
+	_, err = f.deleteDrawing(col, row, drawingXML, "Chart")
+	return err
 }
 
 // countCharts provides a function to get chart files count storage in the
 // folder xl/charts.
 func (f *File) countCharts() int {
 	count := 0
-	for k := range f.XLSX {
-		if strings.Contains(k, "xl/charts/chart") {
+	f.Pkg.Range(func(k, v interface{}) bool {
+		if strings.Contains(k.(string), "xl/charts/chart") {
 			count++
 		}
-	}
+		return true
+	})
 	return count
 }
 
