@@ -1,13 +1,13 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
@@ -18,7 +18,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/mohae/deepcopy"
+	"github.com/tiendc/go-deepcopy"
 )
 
 // Define the default cell size and EMU unit of measurement.
@@ -32,36 +32,44 @@ const (
 
 // Cols defines an iterator to a sheet
 type Cols struct {
-	err                                  error
-	curCol, totalCol, stashCol, totalRow int
-	sheet                                string
-	f                                    *File
-	sheetXML                             []byte
+	err                                    error
+	curCol, totalCols, totalRows, stashCol int
+	rawCellValue                           bool
+	sheet                                  string
+	f                                      *File
+	sheetXML                               []byte
+	sst                                    *xlsxSST
 }
 
-// GetCols return all the columns in a sheet by given worksheet name (case
-// sensitive). For example:
+// GetCols gets the value of all cells by columns on the worksheet based on the
+// given worksheet name, returned as a two-dimensional array, where the value
+// of the cell is converted to the `string` type. If the cell format can be
+// applied to the value of the cell, the applied value will be used, otherwise
+// the original value will be used.
 //
-//    cols, err := f.GetCols("Sheet1")
-//    if err != nil {
-//        fmt.Println(err)
-//        return
-//    }
-//    for _, col := range cols {
-//        for _, rowCell := range col {
-//            fmt.Print(rowCell, "\t")
-//        }
-//        fmt.Println()
-//    }
+// For example, get and traverse the value of all cells by columns on a
+// worksheet named
+// 'Sheet1':
 //
-func (f *File) GetCols(sheet string) ([][]string, error) {
+//	cols, err := f.GetCols("Sheet1")
+//	if err != nil {
+//	    fmt.Println(err)
+//	    return
+//	}
+//	for _, col := range cols {
+//	    for _, rowCell := range col {
+//	        fmt.Print(rowCell, "\t")
+//	    }
+//	    fmt.Println()
+//	}
+func (f *File) GetCols(sheet string, opts ...Options) ([][]string, error) {
 	cols, err := f.Cols(sheet)
 	if err != nil {
 		return nil, err
 	}
 	results := make([][]string, 0, 64)
 	for cols.Next() {
-		col, _ := cols.Rows()
+		col, _ := cols.Rows(opts...)
 		results = append(results, col)
 	}
 	return results, nil
@@ -70,7 +78,7 @@ func (f *File) GetCols(sheet string) ([][]string, error) {
 // Next will return true if the next column is found.
 func (cols *Cols) Next() bool {
 	cols.curCol++
-	return cols.curCol <= cols.totalCol
+	return cols.curCol <= cols.totalCols
 }
 
 // Error will return an error when the error occurs.
@@ -79,17 +87,15 @@ func (cols *Cols) Error() error {
 }
 
 // Rows return the current column's row values.
-func (cols *Cols) Rows() ([]string, error) {
-	var (
-		err              error
-		inElement        string
-		cellCol, cellRow int
-		rows             []string
-	)
+func (cols *Cols) Rows(opts ...Options) ([]string, error) {
+	var rowIterator rowXMLIterator
 	if cols.stashCol >= cols.curCol {
-		return rows, err
+		return rowIterator.cells, rowIterator.err
 	}
-	d := cols.f.sharedStringsReader()
+	cols.rawCellValue = cols.f.getOptions(opts...).RawCellValue
+	if cols.sst, rowIterator.err = cols.f.sharedStringsReader(); rowIterator.err != nil {
+		return rowIterator.cells, rowIterator.err
+	}
 	decoder := cols.f.xmlNewDecoder(bytes.NewReader(cols.sheetXML))
 	for {
 		token, _ := decoder.Token()
@@ -98,42 +104,25 @@ func (cols *Cols) Rows() ([]string, error) {
 		}
 		switch xmlElement := token.(type) {
 		case xml.StartElement:
-			inElement = xmlElement.Name.Local
-			if inElement == "row" {
-				cellCol = 0
-				cellRow++
+			rowIterator.inElement = xmlElement.Name.Local
+			if rowIterator.inElement == "row" {
+				rowIterator.cellCol = 0
+				rowIterator.cellRow++
 				attrR, _ := attrValToInt("r", xmlElement.Attr)
 				if attrR != 0 {
-					cellRow = attrR
+					rowIterator.cellRow = attrR
 				}
 			}
-			if inElement == "c" {
-				cellCol++
-				for _, attr := range xmlElement.Attr {
-					if attr.Name.Local == "r" {
-						if cellCol, cellRow, err = CellNameToCoordinates(attr.Value); err != nil {
-							return rows, err
-						}
-					}
-				}
-				blank := cellRow - len(rows)
-				for i := 1; i < blank; i++ {
-					rows = append(rows, "")
-				}
-				if cellCol == cols.curCol {
-					colCell := xlsxC{}
-					_ = decoder.DecodeElement(&colCell, &xmlElement)
-					val, _ := colCell.getValueFrom(cols.f, d)
-					rows = append(rows, val)
-				}
+			if cols.rowXMLHandler(&rowIterator, &xmlElement, decoder); rowIterator.err != nil {
+				return rowIterator.cells, rowIterator.err
 			}
 		case xml.EndElement:
 			if xmlElement.Name.Local == "sheetData" {
-				return rows, err
+				return rowIterator.cells, rowIterator.err
 			}
 		}
 	}
-	return rows, err
+	return rowIterator.cells, rowIterator.err
 }
 
 // columnXMLIterator defined runtime use field for the worksheet column SAX parser.
@@ -157,7 +146,7 @@ func columnXMLHandler(colIterator *columnXMLIterator, xmlElement *xml.StartEleme
 				colIterator.row = colIterator.curRow
 			}
 		}
-		colIterator.cols.totalRow = colIterator.row
+		colIterator.cols.totalRows = colIterator.row
 		colIterator.cellCol = 0
 	}
 	if inElement == "c" {
@@ -169,42 +158,72 @@ func columnXMLHandler(colIterator *columnXMLIterator, xmlElement *xml.StartEleme
 				}
 			}
 		}
-		if colIterator.cellCol > colIterator.cols.totalCol {
-			colIterator.cols.totalCol = colIterator.cellCol
+		if colIterator.cellCol > colIterator.cols.totalCols {
+			colIterator.cols.totalCols = colIterator.cellCol
+		}
+	}
+}
+
+// rowXMLHandler parse the row XML element of the worksheet.
+func (cols *Cols) rowXMLHandler(rowIterator *rowXMLIterator, xmlElement *xml.StartElement, decoder *xml.Decoder) {
+	if rowIterator.inElement == "c" {
+		rowIterator.cellCol++
+		for _, attr := range xmlElement.Attr {
+			if attr.Name.Local == "r" {
+				if rowIterator.cellCol, rowIterator.cellRow, rowIterator.err = CellNameToCoordinates(attr.Value); rowIterator.err != nil {
+					return
+				}
+			}
+		}
+		blank := rowIterator.cellRow - len(rowIterator.cells)
+		for i := 1; i < blank; i++ {
+			rowIterator.cells = append(rowIterator.cells, "")
+		}
+		if rowIterator.cellCol == cols.curCol {
+			colCell := xlsxC{}
+			_ = decoder.DecodeElement(&colCell, xmlElement)
+			val, _ := colCell.getValueFrom(cols.f, cols.sst, cols.rawCellValue)
+			rowIterator.cells = append(rowIterator.cells, val)
 		}
 	}
 }
 
 // Cols returns a columns iterator, used for streaming reading data for a
-// worksheet with a large data. For example:
+// worksheet with a large data. This function is concurrency safe. For
+// example:
 //
-//    cols, err := f.Cols("Sheet1")
-//    if err != nil {
-//        fmt.Println(err)
-//        return
-//    }
-//    for cols.Next() {
-//        col, err := cols.Rows()
-//        if err != nil {
-//            fmt.Println(err)
-//        }
-//        for _, rowCell := range col {
-//            fmt.Print(rowCell, "\t")
-//        }
-//        fmt.Println()
-//    }
-//
+//	cols, err := f.Cols("Sheet1")
+//	if err != nil {
+//	    fmt.Println(err)
+//	    return
+//	}
+//	for cols.Next() {
+//	    col, err := cols.Rows()
+//	    if err != nil {
+//	        fmt.Println(err)
+//	    }
+//	    for _, rowCell := range col {
+//	        fmt.Print(rowCell, "\t")
+//	    }
+//	    fmt.Println()
+//	}
 func (f *File) Cols(sheet string) (*Cols, error) {
-	name, ok := f.sheetMap[trimSheetName(sheet)]
+	if err := checkSheetName(sheet); err != nil {
+		return nil, err
+	}
+	name, ok := f.getSheetXMLPath(sheet)
 	if !ok {
 		return nil, ErrSheetNotExist{sheet}
 	}
-	if f.Sheet[name] != nil {
-		output, _ := xml.Marshal(f.Sheet[name])
+	if worksheet, ok := f.Sheet.Load(name); ok && worksheet != nil {
+		ws := worksheet.(*xlsxWorksheet)
+		ws.mu.Lock()
+		defer ws.mu.Unlock()
+		output, _ := xml.Marshal(ws)
 		f.saveFileList(name, f.replaceNameSpaceBytes(name, output))
 	}
 	var colIterator columnXMLIterator
-	colIterator.cols.sheetXML = f.readXML(name)
+	colIterator.cols.sheetXML = f.readBytes(name)
 	decoder := f.xmlNewDecoder(bytes.NewReader(colIterator.cols.sheetXML))
 	for {
 		token, _ := decoder.Token()
@@ -220,7 +239,7 @@ func (f *File) Cols(sheet string) (*Cols, error) {
 		case xml.EndElement:
 			if xmlElement.Name.Local == "sheetData" {
 				colIterator.cols.f = f
-				colIterator.cols.sheet = trimSheetName(sheet)
+				colIterator.cols.sheet = sheet
 				return &colIterator.cols, nil
 			}
 		}
@@ -229,26 +248,28 @@ func (f *File) Cols(sheet string) (*Cols, error) {
 }
 
 // GetColVisible provides a function to get visible of a single column by given
-// worksheet name and column name. For example, get visible state of column D
-// in Sheet1:
+// worksheet name and column name. This function is concurrency safe. For
+// example, get visible state of column D in Sheet1:
 //
-//    visible, err := f.GetColVisible("Sheet1", "D")
-//
+//	visible, err := f.GetColVisible("Sheet1", "D")
 func (f *File) GetColVisible(sheet, col string) (bool, error) {
-	visible := true
 	colNum, err := ColumnNameToNumber(col)
 	if err != nil {
-		return visible, err
+		return true, err
 	}
-
+	f.mu.Lock()
 	ws, err := f.workSheetReader(sheet)
 	if err != nil {
+		f.mu.Unlock()
 		return false, err
 	}
+	f.mu.Unlock()
+	ws.mu.Lock()
+	defer ws.mu.Unlock()
 	if ws.Cols == nil {
-		return visible, err
+		return true, err
 	}
-
+	visible := true
 	for c := range ws.Cols.Col {
 		colData := &ws.Cols.Col[c]
 		if colData.Min <= colNum && colNum <= colData.Max {
@@ -259,18 +280,17 @@ func (f *File) GetColVisible(sheet, col string) (bool, error) {
 }
 
 // SetColVisible provides a function to set visible columns by given worksheet
-// name, columns range and visibility.
+// name, columns range and visibility. This function is concurrency safe.
 //
 // For example hide column D on Sheet1:
 //
-//    err := f.SetColVisible("Sheet1", "D", false)
+//	err := f.SetColVisible("Sheet1", "D", false)
 //
 // Hide the columns from D to F (included):
 //
-//    err := f.SetColVisible("Sheet1", "D:F", false)
-//
+//	err := f.SetColVisible("Sheet1", "D:F", false)
 func (f *File) SetColVisible(sheet, columns string, visible bool) error {
-	start, end, err := f.parseColRange(columns)
+	minVal, maxVal, err := f.parseColRange(columns)
 	if err != nil {
 		return err
 	}
@@ -278,10 +298,12 @@ func (f *File) SetColVisible(sheet, columns string, visible bool) error {
 	if err != nil {
 		return err
 	}
+	ws.mu.Lock()
+	defer ws.mu.Unlock()
 	colData := xlsxCol{
-		Min:         start,
-		Max:         end,
-		Width:       defaultColWidth, // default width
+		Min:         minVal,
+		Max:         maxVal,
+		Width:       float64Ptr(defaultColWidth),
 		Hidden:      !visible,
 		CustomWidth: true,
 	}
@@ -308,8 +330,7 @@ func (f *File) SetColVisible(sheet, columns string, visible bool) error {
 // column by given worksheet name and column name. For example, get outline
 // level of column D in Sheet1:
 //
-//    level, err := f.GetColOutlineLevel("Sheet1", "D")
-//
+//	level, err := f.GetColOutlineLevel("Sheet1", "D")
 func (f *File) GetColOutlineLevel(sheet, col string) (uint8, error) {
 	level := uint8(0)
 	colNum, err := ColumnNameToNumber(col)
@@ -333,20 +354,20 @@ func (f *File) GetColOutlineLevel(sheet, col string) (uint8, error) {
 }
 
 // parseColRange parse and convert column range with column name to the column number.
-func (f *File) parseColRange(columns string) (start, end int, err error) {
+func (f *File) parseColRange(columns string) (minVal, maxVal int, err error) {
 	colsTab := strings.Split(columns, ":")
-	start, err = ColumnNameToNumber(colsTab[0])
+	minVal, err = ColumnNameToNumber(colsTab[0])
 	if err != nil {
 		return
 	}
-	end = start
+	maxVal = minVal
 	if len(colsTab) == 2 {
-		if end, err = ColumnNameToNumber(colsTab[1]); err != nil {
+		if maxVal, err = ColumnNameToNumber(colsTab[1]); err != nil {
 			return
 		}
 	}
-	if end < start {
-		start, end = end, start
+	if maxVal < minVal {
+		minVal, maxVal = maxVal, minVal
 	}
 	return
 }
@@ -355,8 +376,7 @@ func (f *File) parseColRange(columns string) (start, end int, err error) {
 // column by given worksheet name and column name. The value of parameter
 // 'level' is 1-7. For example, set outline level of column D in Sheet1 to 2:
 //
-//    err := f.SetColOutlineLevel("Sheet1", "D", 2)
-//
+//	err := f.SetColOutlineLevel("Sheet1", "D", 2)
 func (f *File) SetColOutlineLevel(sheet, col string, level uint8) error {
 	if level > 7 || level < 1 {
 		return ErrOutlineLevel
@@ -395,32 +415,67 @@ func (f *File) SetColOutlineLevel(sheet, col string, level uint8) error {
 }
 
 // SetColStyle provides a function to set style of columns by given worksheet
-// name, columns range and style ID.
+// name, columns range and style ID. This function is concurrency safe. Note
+// that this will overwrite the existing styles for the columns, it won't
+// append or merge style with existing styles.
 //
 // For example set style of column H on Sheet1:
 //
-//    err = f.SetColStyle("Sheet1", "H", style)
+//	err = f.SetColStyle("Sheet1", "H", style)
 //
 // Set style of columns C:F on Sheet1:
 //
-//    err = f.SetColStyle("Sheet1", "C:F", style)
-//
+//	err = f.SetColStyle("Sheet1", "C:F", style)
 func (f *File) SetColStyle(sheet, columns string, styleID int) error {
-	start, end, err := f.parseColRange(columns)
+	minVal, maxVal, err := f.parseColRange(columns)
 	if err != nil {
+		return err
+	}
+	f.mu.Lock()
+	s, err := f.stylesReader()
+	if err != nil {
+		f.mu.Unlock()
 		return err
 	}
 	ws, err := f.workSheetReader(sheet)
 	if err != nil {
+		f.mu.Unlock()
 		return err
 	}
+	f.mu.Unlock()
+	s.mu.Lock()
+	if styleID < 0 || s.CellXfs == nil || len(s.CellXfs.Xf) <= styleID {
+		s.mu.Unlock()
+		return newInvalidStyleID(styleID)
+	}
+	s.mu.Unlock()
+	ws.mu.Lock()
+	ws.setColStyle(minVal, maxVal, styleID)
+	ws.mu.Unlock()
+	if rows := len(ws.SheetData.Row); rows > 0 {
+		for col := minVal; col <= maxVal; col++ {
+			from, _ := CoordinatesToCellName(col, 1)
+			to, _ := CoordinatesToCellName(col, rows)
+			err = f.SetCellStyle(sheet, from, to, styleID)
+		}
+	}
+	return err
+}
+
+// setColStyle provides a function to set the style of a single column or
+// multiple columns.
+func (ws *xlsxWorksheet) setColStyle(minVal, maxVal, styleID int) {
 	if ws.Cols == nil {
 		ws.Cols = &xlsxCols{}
 	}
+	width := defaultColWidth
+	if ws.SheetFormatPr != nil && ws.SheetFormatPr.DefaultColWidth > 0 {
+		width = ws.SheetFormatPr.DefaultColWidth
+	}
 	ws.Cols.Col = flatCols(xlsxCol{
-		Min:   start,
-		Max:   end,
-		Width: defaultColWidth,
+		Min:   minVal,
+		Max:   maxVal,
+		Width: float64Ptr(width),
 		Style: styleID,
 	}, ws.Cols.Col, func(fc, c xlsxCol) xlsxCol {
 		fc.BestFit = c.BestFit
@@ -432,46 +487,47 @@ func (f *File) SetColStyle(sheet, columns string, styleID int) error {
 		fc.Width = c.Width
 		return fc
 	})
-	return nil
 }
 
 // SetColWidth provides a function to set the width of a single column or
-// multiple columns. For example:
+// multiple columns. This function is concurrency safe. For example:
 //
-//    f := excelize.NewFile()
-//    err := f.SetColWidth("Sheet1", "A", "H", 20)
-//
-func (f *File) SetColWidth(sheet, startcol, endcol string, width float64) error {
-	min, err := ColumnNameToNumber(startcol)
-	if err != nil {
-		return err
-	}
-	max, err := ColumnNameToNumber(endcol)
+//	err := f.SetColWidth("Sheet1", "A", "H", 20)
+func (f *File) SetColWidth(sheet, startCol, endCol string, width float64) error {
+	minVal, maxVal, err := f.parseColRange(startCol + ":" + endCol)
 	if err != nil {
 		return err
 	}
 	if width > MaxColumnWidth {
 		return ErrColumnWidth
 	}
-	if min > max {
-		min, max = max, min
-	}
-
+	f.mu.Lock()
 	ws, err := f.workSheetReader(sheet)
 	if err != nil {
+		f.mu.Unlock()
 		return err
 	}
+	f.mu.Unlock()
+	ws.mu.Lock()
+	defer ws.mu.Unlock()
+	ws.setColWidth(minVal, maxVal, width)
+	return err
+}
+
+// setColWidth provides a function to set the width of a single column or
+// multiple columns.
+func (ws *xlsxWorksheet) setColWidth(minVal, maxVal int, width float64) {
 	col := xlsxCol{
-		Min:         min,
-		Max:         max,
-		Width:       width,
+		Min:         minVal,
+		Max:         maxVal,
+		Width:       float64Ptr(width),
 		CustomWidth: true,
 	}
 	if ws.Cols == nil {
 		cols := xlsxCols{}
 		cols.Col = append(cols.Col, col)
 		ws.Cols = &cols
-		return err
+		return
 	}
 	ws.Cols.Col = flatCols(col, ws.Cols.Col, func(fc, c xlsxCol) xlsxCol {
 		fc.BestFit = c.BestFit
@@ -482,15 +538,15 @@ func (f *File) SetColWidth(sheet, startcol, endcol string, width float64) error 
 		fc.Style = c.Style
 		return fc
 	})
-	return err
 }
 
 // flatCols provides a method for the column's operation functions to flatten
 // and check the worksheet columns.
 func flatCols(col xlsxCol, cols []xlsxCol, replacer func(fc, c xlsxCol) xlsxCol) []xlsxCol {
-	fc := []xlsxCol{}
+	var fc []xlsxCol
 	for i := col.Min; i <= col.Max; i++ {
-		c := deepcopy.Copy(col).(xlsxCol)
+		var c xlsxCol
+		deepcopy.Copy(&c, col)
 		c.Min, c.Max = i, i
 		fc = append(fc, c)
 	}
@@ -508,7 +564,8 @@ func flatCols(col xlsxCol, cols []xlsxCol, replacer func(fc, c xlsxCol) xlsxCol)
 				fc[idx] = replacer(fc[idx], column)
 				continue
 			}
-			c := deepcopy.Copy(column).(xlsxCol)
+			var c xlsxCol
+			deepcopy.Copy(&c, column)
 			c.Min, c.Max = i, i
 			fc = append(fc, c)
 		}
@@ -519,25 +576,25 @@ func flatCols(col xlsxCol, cols []xlsxCol, replacer func(fc, c xlsxCol) xlsxCol)
 // positionObjectPixels calculate the vertices that define the position of a
 // graphical object within the worksheet in pixels.
 //
-//          +------------+------------+
-//          |     A      |      B     |
-//    +-----+------------+------------+
-//    |     |(x1,y1)     |            |
-//    |  1  |(A1)._______|______      |
-//    |     |    |              |     |
-//    |     |    |              |     |
-//    +-----+----|    OBJECT    |-----+
-//    |     |    |              |     |
-//    |  2  |    |______________.     |
-//    |     |            |        (B2)|
-//    |     |            |     (x2,y2)|
-//    +-----+------------+------------+
+//	      +------------+------------+
+//	      |     A      |      B     |
+//	+-----+------------+------------+
+//	|     |(x1,y1)     |            |
+//	|  1  |(A1)._______|______      |
+//	|     |    |              |     |
+//	|     |    |              |     |
+//	+-----+----|    OBJECT    |-----+
+//	|     |    |              |     |
+//	|  2  |    |______________.     |
+//	|     |            |        (B2)|
+//	|     |            |     (x2,y2)|
+//	+-----+------------+------------+
 //
-// Example of an object that covers some of the area from cell A1 to B2.
+// Example of an object that covers some range reference from cell A1 to B2.
 //
 // Based on the width and height of the object we need to calculate 8 vars:
 //
-//    colStart, rowStart, colEnd, rowEnd, x1, y1, x2, y2.
+//	colStart, rowStart, colEnd, rowEnd, x1, y1, x2, y2.
 //
 // We also calculate the absolute x and y position of the top left vertex of
 // the object. This is required for images.
@@ -550,37 +607,36 @@ func flatCols(col xlsxCol, cols []xlsxCol, replacer func(fc, c xlsxCol) xlsxCol)
 // subtracting the width and height of the object from the width and
 // height of the underlying cells.
 //
-//    colStart        # Col containing upper left corner of object.
-//    x1              # Distance to left side of object.
+//	colStart        # Col containing upper left corner of object.
+//	x1              # Distance to left side of object.
 //
-//    rowStart        # Row containing top left corner of object.
-//    y1              # Distance to top of object.
+//	rowStart        # Row containing top left corner of object.
+//	y1              # Distance to top of object.
 //
-//    colEnd          # Col containing lower right corner of object.
-//    x2              # Distance to right side of object.
+//	colEnd          # Col containing lower right corner of object.
+//	x2              # Distance to right side of object.
 //
-//    rowEnd          # Row containing bottom right corner of object.
-//    y2              # Distance to bottom of object.
+//	rowEnd          # Row containing bottom right corner of object.
+//	y2              # Distance to bottom of object.
 //
-//    width           # Width of object frame.
-//    height          # Height of object frame.
-//
+//	width           # Width of object frame.
+//	height          # Height of object frame.
 func (f *File) positionObjectPixels(sheet string, col, row, x1, y1, width, height int) (int, int, int, int, int, int) {
+	colIdx, rowIdx := col-1, row-1
 	// Adjust start column for offsets that are greater than the col width.
-	for x1 >= f.getColWidth(sheet, col) {
-		x1 -= f.getColWidth(sheet, col)
-		col++
+	for x1 >= f.getColWidth(sheet, colIdx+1) {
+		colIdx++
+		x1 -= f.getColWidth(sheet, colIdx)
 	}
 
 	// Adjust start row for offsets that are greater than the row height.
-	for y1 >= f.getRowHeight(sheet, row) {
-		y1 -= f.getRowHeight(sheet, row)
-		row++
+	for y1 >= f.getRowHeight(sheet, rowIdx+1) {
+		rowIdx++
+		y1 -= f.getRowHeight(sheet, rowIdx)
 	}
 
-	// Initialise end cell to the same as the start cell.
-	colEnd := col
-	rowEnd := row
+	// Initialized end cell to the same as the start cell.
+	colEnd, rowEnd := colIdx, rowIdx
 
 	width += x1
 	height += y1
@@ -592,79 +648,125 @@ func (f *File) positionObjectPixels(sheet string, col, row, x1, y1, width, heigh
 	}
 
 	// Subtract the underlying cell heights to find end cell of the object.
-	for height >= f.getRowHeight(sheet, rowEnd) {
-		height -= f.getRowHeight(sheet, rowEnd)
+	for height >= f.getRowHeight(sheet, rowEnd+1) {
 		rowEnd++
+		height -= f.getRowHeight(sheet, rowEnd)
 	}
 
 	// The end vertices are whatever is left from the width and height.
-	x2 := width
-	y2 := height
-	return col, row, colEnd, rowEnd, x2, y2
+	return colIdx, rowIdx, colEnd, rowEnd, width, height
 }
 
 // getColWidth provides a function to get column width in pixels by given
-// sheet name and column index.
+// sheet name and column number.
 func (f *File) getColWidth(sheet string, col int) int {
-	xlsx, _ := f.workSheetReader(sheet)
-	if xlsx.Cols != nil {
+	ws, _ := f.workSheetReader(sheet)
+	ws.mu.Lock()
+	defer ws.mu.Unlock()
+	if ws.Cols != nil {
 		var width float64
-		for _, v := range xlsx.Cols.Col {
-			if v.Min <= col && col <= v.Max {
-				width = v.Width
+		for _, v := range ws.Cols.Col {
+			if v.Min <= col && col <= v.Max && v.Width != nil {
+				width = *v.Width
 			}
 		}
 		if width != 0 {
 			return int(convertColWidthToPixels(width))
 		}
 	}
-	// Optimisation for when the column widths haven't changed.
+	if ws.SheetFormatPr != nil && ws.SheetFormatPr.DefaultColWidth > 0 {
+		return int(convertColWidthToPixels(ws.SheetFormatPr.DefaultColWidth))
+	}
+	// Optimization for when the column widths haven't changed.
 	return int(defaultColWidthPixels)
 }
 
+// GetColStyle provides a function to get column style ID by given worksheet
+// name and column name. This function is concurrency safe.
+func (f *File) GetColStyle(sheet, col string) (int, error) {
+	var styleID int
+	colNum, err := ColumnNameToNumber(col)
+	if err != nil {
+		return styleID, err
+	}
+	f.mu.Lock()
+	ws, err := f.workSheetReader(sheet)
+	if err != nil {
+		f.mu.Unlock()
+		return styleID, err
+	}
+	f.mu.Unlock()
+	ws.mu.Lock()
+	defer ws.mu.Unlock()
+	if ws.Cols != nil {
+		for _, v := range ws.Cols.Col {
+			if v.Min <= colNum && colNum <= v.Max {
+				styleID = v.Style
+			}
+		}
+	}
+	return styleID, err
+}
+
 // GetColWidth provides a function to get column width by given worksheet name
-// and column index.
+// and column name. This function is concurrency safe.
 func (f *File) GetColWidth(sheet, col string) (float64, error) {
 	colNum, err := ColumnNameToNumber(col)
 	if err != nil {
 		return defaultColWidth, err
 	}
+	f.mu.Lock()
 	ws, err := f.workSheetReader(sheet)
 	if err != nil {
+		f.mu.Unlock()
 		return defaultColWidth, err
 	}
+	f.mu.Unlock()
+	ws.mu.Lock()
+	defer ws.mu.Unlock()
 	if ws.Cols != nil {
 		var width float64
 		for _, v := range ws.Cols.Col {
-			if v.Min <= colNum && colNum <= v.Max {
-				width = v.Width
+			if v.Min <= colNum && colNum <= v.Max && v.Width != nil {
+				width = *v.Width
 			}
 		}
 		if width != 0 {
 			return width, err
 		}
 	}
-	// Optimisation for when the column widths haven't changed.
+	if ws.SheetFormatPr != nil && ws.SheetFormatPr.DefaultColWidth > 0 {
+		return ws.SheetFormatPr.DefaultColWidth, err
+	}
+	// Optimization for when the column widths haven't changed.
 	return defaultColWidth, err
 }
 
-// InsertCol provides a function to insert a new column before given column
-// index. For example, create a new column before column C in Sheet1:
+// InsertCols provides a function to insert new columns before the given column
+// name and number of columns. For example, create two columns before column
+// C in Sheet1:
 //
-//    err := f.InsertCol("Sheet1", "C")
+//	err := f.InsertCols("Sheet1", "C", 2)
 //
-func (f *File) InsertCol(sheet, col string) error {
+// Use this method with caution, which will affect changes in references such
+// as formulas, charts, and so on. If there is any referenced value of the
+// worksheet, it will cause a file error when you open it. The excelize only
+// partially updates these references currently.
+func (f *File) InsertCols(sheet, col string, n int) error {
 	num, err := ColumnNameToNumber(col)
 	if err != nil {
 		return err
 	}
-	return f.adjustHelper(sheet, columns, num, 1)
+	if n < 1 || n > MaxColumns {
+		return ErrColumnNumber
+	}
+	return f.adjustHelper(sheet, columns, num, n)
 }
 
 // RemoveCol provides a function to remove single column by given worksheet
 // name and column index. For example, remove column C in Sheet1:
 //
-//    err := f.RemoveCol("Sheet1", "C")
+//	err := f.RemoveCol("Sheet1", "C")
 //
 // Use this method with caution, which will affect changes in references such
 // as formulas, charts, and so on. If there is any referenced value of the
@@ -693,7 +795,7 @@ func (f *File) RemoveCol(sheet, col string) error {
 	return f.adjustHelper(sheet, columns, num, -1)
 }
 
-// convertColWidthToPixels provieds function to convert the width of a cell
+// convertColWidthToPixels provides function to convert the width of a cell
 // from user's units to pixels. Excel rounds the column width to the nearest
 // pixel. If the width hasn't been set by the user we use the default value.
 // If the column is hidden it has a value of zero.
