@@ -1,13 +1,13 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
@@ -17,31 +17,34 @@ import (
 )
 
 const (
+	nanosInADay    = float64((24 * time.Hour) / time.Nanosecond)
 	dayNanoseconds = 24 * time.Hour
 	maxDuration    = 290 * 364 * dayNanoseconds
+	roundEpsilon   = 1e-9
 )
 
 var (
+	daysInMonth           = []int{31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31}
+	excel1900Epoc         = time.Date(1899, time.December, 30, 0, 0, 0, 0, time.UTC)
+	excel1904Epoc         = time.Date(1904, time.January, 1, 0, 0, 0, 0, time.UTC)
 	excelMinTime1900      = time.Date(1899, time.December, 31, 0, 0, 0, 0, time.UTC)
 	excelBuggyPeriodStart = time.Date(1900, time.March, 1, 0, 0, 0, 0, time.UTC).Add(-time.Nanosecond)
 )
 
 // timeToExcelTime provides a function to convert time to Excel time.
-func timeToExcelTime(t time.Time) (float64, error) {
-	// TODO in future this should probably also handle date1904 and like TimeFromExcelTime
-
-	if t.Before(excelMinTime1900) {
-		return 0.0, nil
+func timeToExcelTime(t time.Time, date1904 bool) (float64, error) {
+	date := excelMinTime1900
+	if date1904 {
+		date = excel1904Epoc
 	}
-
-	tt := t
-	diff := t.Sub(excelMinTime1900)
-	result := float64(0)
-
+	if t.Before(date) {
+		return 0, nil
+	}
+	tt, diff, result := t, t.Sub(date), 0.0
 	for diff >= maxDuration {
 		result += float64(maxDuration / dayNanoseconds)
 		tt = tt.Add(-maxDuration)
-		diff = tt.Sub(excelMinTime1900)
+		diff = tt.Sub(date)
 	}
 
 	rem := diff % dayNanoseconds
@@ -52,8 +55,8 @@ func timeToExcelTime(t time.Time) (float64, error) {
 	// Microsoft intentionally included this bug in Excel so that it would remain compatible with the spreadsheet
 	// program that had the majority market share at the time; Lotus 1-2-3.
 	// https://www.myonlinetraininghub.com/excel-date-and-time
-	if t.After(excelBuggyPeriodStart) {
-		result += 1.0
+	if !date1904 && t.After(excelBuggyPeriodStart) {
+		result++
 	}
 	return result, nil
 }
@@ -77,7 +80,6 @@ func shiftJulianToNoon(julianDays, julianFraction float64) (float64, float64) {
 // minutes, seconds and nanoseconds that comprised a given fraction of a day.
 // values would round to 1 us.
 func fractionOfADay(fraction float64) (hours, minutes, seconds, nanoseconds int) {
-
 	const (
 		c1us  = 1e3
 		c1s   = 1e9
@@ -112,7 +114,7 @@ func julianDateToGregorianTime(part1, part2 float64) time.Time {
 // "Communications of the ACM" in 1968 (published in CACM, volume 11, number
 // 10, October 1968, p.657). None of those programmers seems to have found it
 // necessary to explain the constants or variable names set out by Henry F.
-// Fliegel and Thomas C. Van Flandern.  Maybe one day I'll buy that jounal and
+// Fliegel and Thomas C. Van Flandern. Maybe one day I'll buy that journal and
 // expand an explanation here - that day is not today.
 func doTheFliegelAndVanFlandernAlgorithm(jd int) (day, month, year int) {
 	l := jd + 68569
@@ -131,12 +133,11 @@ func doTheFliegelAndVanFlandernAlgorithm(jd int) (day, month, year int) {
 // timeFromExcelTime provides a function to convert an excelTime
 // representation (stored as a floating point number) to a time.Time.
 func timeFromExcelTime(excelTime float64, date1904 bool) time.Time {
-	const MDD int64 = 106750 // Max time.Duration Days, aprox. 290 years
 	var date time.Time
-	var intPart = int64(excelTime)
+	wholeDaysPart := int(excelTime)
 	// Excel uses Julian dates prior to March 1st 1900, and Gregorian
 	// thereafter.
-	if intPart <= 61 {
+	if wholeDaysPart <= 61 {
 		const OFFSET1900 = 15018.0
 		const OFFSET1904 = 16480.0
 		const MJD0 float64 = 2400000.5
@@ -148,29 +149,96 @@ func timeFromExcelTime(excelTime float64, date1904 bool) time.Time {
 		}
 		return date
 	}
-	var floatPart = excelTime - float64(intPart)
-	var dayNanoSeconds float64 = 24 * 60 * 60 * 1000 * 1000 * 1000
+	floatPart := excelTime - float64(wholeDaysPart) + roundEpsilon
 	if date1904 {
-		date = time.Date(1904, 1, 1, 0, 0, 0, 0, time.UTC)
+		date = excel1904Epoc
 	} else {
-		date = time.Date(1899, 12, 30, 0, 0, 0, 0, time.UTC)
+		date = excel1900Epoc
 	}
-
-	// Duration is limited to aprox. 290 years
-	for intPart > MDD {
-		durationDays := time.Duration(MDD) * time.Hour * 24
-		date = date.Add(durationDays)
-		intPart = intPart - MDD
+	durationPart := time.Duration(nanosInADay * floatPart)
+	date = date.AddDate(0, 0, wholeDaysPart).Add(durationPart)
+	if date.Nanosecond()/1e6 > 500 {
+		return date.Round(time.Second)
 	}
-	durationDays := time.Duration(intPart) * time.Hour * 24
-	durationPart := time.Duration(dayNanoSeconds * floatPart)
-	return date.Add(durationDays).Add(durationPart)
+	return date.Truncate(time.Second)
 }
 
-// ExcelDateToTime converts a float-based excel date representation to a time.Time.
+// ExcelDateToTime converts a float-based Excel date representation to a time.Time.
 func ExcelDateToTime(excelDate float64, use1904Format bool) (time.Time, error) {
 	if excelDate < 0 {
 		return time.Time{}, newInvalidExcelDateError(excelDate)
 	}
 	return timeFromExcelTime(excelDate, use1904Format), nil
+}
+
+// isLeapYear determine if leap year for a given year.
+func isLeapYear(y int) bool {
+	if y == y/400*400 {
+		return true
+	}
+	if y == y/100*100 {
+		return false
+	}
+	return y == y/4*4
+}
+
+// getDaysInMonth provides a function to get the days by a given year and
+// month number.
+func getDaysInMonth(y, m int) int {
+	if m == 2 && isLeapYear(y) {
+		return 29
+	}
+	return daysInMonth[m-1]
+}
+
+// validateDate provides a function to validate if a valid date by a given
+// year, month, and day number.
+func validateDate(y, m, d int) bool {
+	if m < 1 || m > 12 {
+		return false
+	}
+	if d < 1 {
+		return false
+	}
+	return d <= getDaysInMonth(y, m)
+}
+
+// formatYear converts the given year number into a 4-digit format.
+func formatYear(y int) int {
+	if y < 1900 {
+		if y < 30 {
+			y += 2000
+		} else {
+			y += 1900
+		}
+	}
+	return y
+}
+
+// getDurationNumFmt returns most simplify numbers format code for time
+// duration type cell value by given worksheet name, cell reference and number.
+func getDurationNumFmt(d time.Duration) int {
+	if d >= time.Hour*24 {
+		return 46
+	}
+	// Whole minutes
+	if d.Minutes() == float64(int(d.Minutes())) {
+		return 20
+	}
+	return 21
+}
+
+// getTimeNumFmt returns most simplify numbers format code for time type cell
+// value by given worksheet name, cell reference and number.
+func getTimeNumFmt(t time.Time) int {
+	nextMonth := t.AddDate(0, 1, 0)
+	// Whole months
+	if t.Day() == 1 && nextMonth.Day() == 1 {
+		return 17
+	}
+	// Whole days
+	if t.Hour() == 0 && t.Minute() == 0 && t.Second() == 0 && t.Nanosecond() == 0 {
+		return 14
+	}
+	return 22
 }
