@@ -33,13 +33,14 @@ var excelTimeInputList = []dateTest{
 	{60.0, time.Date(1900, 2, 28, 0, 0, 0, 0, time.UTC)},
 	{61.0, time.Date(1900, 3, 1, 0, 0, 0, 0, time.UTC)},
 	{41275.0, time.Date(2013, 1, 1, 0, 0, 0, 0, time.UTC)},
+	{44450.3333333333, time.Date(2021, time.September, 11, 8, 0, 0, 0, time.UTC)},
 	{401769.0, time.Date(3000, 1, 1, 0, 0, 0, 0, time.UTC)},
 }
 
 func TestTimeToExcelTime(t *testing.T) {
 	for i, test := range trueExpectedDateList {
 		t.Run(fmt.Sprintf("TestData%d", i+1), func(t *testing.T) {
-			excelTime, err := timeToExcelTime(test.GoValue)
+			excelTime, err := timeToExcelTime(test.GoValue, false)
 			assert.NoError(t, err)
 			assert.Equalf(t, test.ExcelValue, excelTime,
 				"Time: %s", test.GoValue.String())
@@ -54,7 +55,7 @@ func TestTimeToExcelTime_Timezone(t *testing.T) {
 	}
 	for i, test := range trueExpectedDateList {
 		t.Run(fmt.Sprintf("TestData%d", i+1), func(t *testing.T) {
-			_, err := timeToExcelTime(test.GoValue.In(location))
+			_, err := timeToExcelTime(test.GoValue.In(location), false)
 			assert.NoError(t, err)
 		})
 	}
@@ -66,12 +67,38 @@ func TestTimeFromExcelTime(t *testing.T) {
 			assert.Equal(t, test.GoValue, timeFromExcelTime(test.ExcelValue, false))
 		})
 	}
+	for hour := 0; hour < 24; hour++ {
+		for minVal := 0; minVal < 60; minVal++ {
+			for sec := 0; sec < 60; sec++ {
+				date := time.Date(2021, time.December, 30, hour, minVal, sec, 0, time.UTC)
+				// Test use 1900 date system
+				excel1900Time, err := timeToExcelTime(date, false)
+				assert.NoError(t, err)
+				date1900Out := timeFromExcelTime(excel1900Time, false)
+				assert.EqualValues(t, hour, date1900Out.Hour())
+				assert.EqualValues(t, minVal, date1900Out.Minute())
+				assert.EqualValues(t, sec, date1900Out.Second())
+				// Test use 1904 date system
+				excel1904Time, err := timeToExcelTime(date, true)
+				assert.NoError(t, err)
+				date1904Out := timeFromExcelTime(excel1904Time, true)
+				assert.EqualValues(t, hour, date1904Out.Hour())
+				assert.EqualValues(t, minVal, date1904Out.Minute())
+				assert.EqualValues(t, sec, date1904Out.Second())
+			}
+		}
+	}
 }
 
 func TestTimeFromExcelTime_1904(t *testing.T) {
-	_, _ = shiftJulianToNoon(1, -0.6)
-	timeFromExcelTime(61, true)
-	timeFromExcelTime(62, true)
+	julianDays, julianFraction := shiftJulianToNoon(1, -0.6)
+	assert.Equal(t, julianDays, 0.0)
+	assert.Equal(t, julianFraction, 0.9)
+	julianDays, julianFraction = shiftJulianToNoon(1, 0.1)
+	assert.Equal(t, julianDays, 1.0)
+	assert.Equal(t, julianFraction, 0.6)
+	assert.Equal(t, timeFromExcelTime(61, true), time.Date(1904, time.March, 2, 0, 0, 0, 0, time.UTC))
+	assert.Equal(t, timeFromExcelTime(62, true), time.Date(1904, time.March, 3, 0, 0, 0, 0, time.UTC))
 }
 
 func TestExcelDateToTime(t *testing.T) {
@@ -85,5 +112,5 @@ func TestExcelDateToTime(t *testing.T) {
 	}
 	// Check error case
 	_, err := ExcelDateToTime(-1, false)
-	assert.EqualError(t, err, "invalid date value -1.000000, negative values are not supported supported")
+	assert.EqualError(t, err, newInvalidExcelDateError(-1).Error())
 }
