@@ -1,22 +1,20 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
 import (
 	"bytes"
 	"encoding/xml"
-	"fmt"
 	"io"
-	"log"
 	"reflect"
 	"strconv"
 	"strings"
@@ -27,13 +25,15 @@ import (
 func (f *File) prepareDrawing(ws *xlsxWorksheet, drawingID int, sheet, drawingXML string) (int, string) {
 	sheetRelationshipsDrawingXML := "../drawings/drawing" + strconv.Itoa(drawingID) + ".xml"
 	if ws.Drawing != nil {
-		// The worksheet already has a picture or chart relationships, use the relationships drawing ../drawings/drawing%d.xml.
-		sheetRelationshipsDrawingXML = f.getSheetRelationshipsTargetByID(sheet, ws.Drawing.RID)
+		// The worksheet already has a picture or chart relationships, use the
+		// relationships drawing ../drawings/drawing%d.xml or /xl/drawings/drawing%d.xml.
+		sheetRelationshipsDrawingXML = strings.ReplaceAll(f.getSheetRelationshipsTargetByID(sheet, ws.Drawing.RID), "/xl/drawings/", "../drawings/")
 		drawingID, _ = strconv.Atoi(strings.TrimSuffix(strings.TrimPrefix(sheetRelationshipsDrawingXML, "../drawings/drawing"), ".xml"))
-		drawingXML = strings.Replace(sheetRelationshipsDrawingXML, "..", "xl", -1)
+		drawingXML = strings.ReplaceAll(sheetRelationshipsDrawingXML, "..", "xl")
 	} else {
 		// Add first picture for given sheet.
-		sheetRels := "xl/worksheets/_rels/" + strings.TrimPrefix(f.sheetMap[trimSheetName(sheet)], "xl/worksheets/") + ".rels"
+		sheetXMLPath, _ := f.getSheetXMLPath(sheet)
+		sheetRels := "xl/worksheets/_rels/" + strings.TrimPrefix(sheetXMLPath, "xl/worksheets/") + ".rels"
 		rID := f.addRels(sheetRels, SourceRelationshipDrawingML, sheetRelationshipsDrawingXML, "")
 		f.addSheetDrawing(sheet, rID)
 	}
@@ -45,7 +45,8 @@ func (f *File) prepareDrawing(ws *xlsxWorksheet, drawingID int, sheet, drawingXM
 func (f *File) prepareChartSheetDrawing(cs *xlsxChartsheet, drawingID int, sheet string) {
 	sheetRelationshipsDrawingXML := "../drawings/drawing" + strconv.Itoa(drawingID) + ".xml"
 	// Only allow one chart in a chartsheet.
-	sheetRels := "xl/chartsheets/_rels/" + strings.TrimPrefix(f.sheetMap[trimSheetName(sheet)], "xl/chartsheets/") + ".rels"
+	sheetXMLPath, _ := f.getSheetXMLPath(sheet)
+	sheetRels := "xl/chartsheets/_rels/" + strings.TrimPrefix(sheetXMLPath, "xl/chartsheets/") + ".rels"
 	rID := f.addRels(sheetRels, SourceRelationshipDrawingML, sheetRelationshipsDrawingXML, "")
 	f.addSheetNameSpace(sheet, SourceRelationship)
 	cs.Drawing = &xlsxDrawing{
@@ -55,7 +56,7 @@ func (f *File) prepareChartSheetDrawing(cs *xlsxChartsheet, drawingID int, sheet
 
 // addChart provides a function to create chart as xl/charts/chart%d.xml by
 // given format sets.
-func (f *File) addChart(formatSet *formatChart, comboCharts []*formatChart) {
+func (f *File) addChart(opts *Chart, comboCharts []*Chart) {
 	count := f.countCharts()
 	xlsxChartSpace := xlsxChartSpace{
 		XMLNSa:         NameSpaceDrawingML.Value,
@@ -63,70 +64,12 @@ func (f *File) addChart(formatSet *formatChart, comboCharts []*formatChart) {
 		Lang:           &attrValString{Val: stringPtr("en-US")},
 		RoundedCorners: &attrValBool{Val: boolPtr(false)},
 		Chart: cChart{
-			Title: &cTitle{
-				Tx: cTx{
-					Rich: &cRich{
-						P: aP{
-							PPr: &aPPr{
-								DefRPr: aRPr{
-									Kern:   1200,
-									Strike: "noStrike",
-									U:      "none",
-									Sz:     1400,
-									SolidFill: &aSolidFill{
-										SchemeClr: &aSchemeClr{
-											Val: "tx1",
-											LumMod: &attrValInt{
-												Val: intPtr(65000),
-											},
-											LumOff: &attrValInt{
-												Val: intPtr(35000),
-											},
-										},
-									},
-									Ea: &aEa{
-										Typeface: "+mn-ea",
-									},
-									Cs: &aCs{
-										Typeface: "+mn-cs",
-									},
-									Latin: &aLatin{
-										Typeface: "+mn-lt",
-									},
-								},
-							},
-							R: &aR{
-								RPr: aRPr{
-									Lang:    "en-US",
-									AltLang: "en-US",
-								},
-								T: formatSet.Title.Name,
-							},
-						},
-					},
-				},
-				TxPr: cTxPr{
-					P: aP{
-						PPr: &aPPr{
-							DefRPr: aRPr{
-								Kern:   1200,
-								U:      "none",
-								Sz:     14000,
-								Strike: "noStrike",
-							},
-						},
-						EndParaRPr: &aEndParaRPr{
-							Lang: "en-US",
-						},
-					},
-				},
-				Overlay: &attrValBool{Val: boolPtr(false)},
-			},
+			Title: f.drawPlotAreaTitles(opts.Title, ""),
 			View3D: &cView3D{
-				RotX:        &attrValInt{Val: intPtr(chartView3DRotX[formatSet.Type])},
-				RotY:        &attrValInt{Val: intPtr(chartView3DRotY[formatSet.Type])},
-				Perspective: &attrValInt{Val: intPtr(chartView3DPerspective[formatSet.Type])},
-				RAngAx:      &attrValInt{Val: intPtr(chartView3DRAngAx[formatSet.Type])},
+				RotX:        &attrValInt{Val: intPtr(chartView3DRotX[opts.Type])},
+				RotY:        &attrValInt{Val: intPtr(chartView3DRotY[opts.Type])},
+				Perspective: &attrValInt{Val: intPtr(chartView3DPerspective[opts.Type])},
+				RAngAx:      &attrValInt{Val: intPtr(chartView3DRAngAx[opts.Type])},
 			},
 			Floor: &cThicknessSpPr{
 				Thickness: &attrValInt{Val: intPtr(0)},
@@ -139,34 +82,19 @@ func (f *File) addChart(formatSet *formatChart, comboCharts []*formatChart) {
 			},
 			PlotArea: &cPlotArea{},
 			Legend: &cLegend{
-				LegendPos: &attrValString{Val: stringPtr(chartLegendPosition[formatSet.Legend.Position])},
+				LegendPos: &attrValString{Val: stringPtr(chartLegendPosition[opts.Legend.Position])},
 				Overlay:   &attrValBool{Val: boolPtr(false)},
 			},
 
 			PlotVisOnly:      &attrValBool{Val: boolPtr(false)},
-			DispBlanksAs:     &attrValString{Val: stringPtr(formatSet.ShowBlanksAs)},
+			DispBlanksAs:     &attrValString{Val: stringPtr(opts.ShowBlanksAs)},
 			ShowDLblsOverMax: &attrValBool{Val: boolPtr(false)},
 		},
 		SpPr: &cSpPr{
 			SolidFill: &aSolidFill{
 				SchemeClr: &aSchemeClr{Val: "bg1"},
 			},
-			Ln: &aLn{
-				W:    9525,
-				Cap:  "flat",
-				Cmpd: "sng",
-				Algn: "ctr",
-				SolidFill: &aSolidFill{
-					SchemeClr: &aSchemeClr{Val: "tx1",
-						LumMod: &attrValInt{
-							Val: intPtr(15000),
-						},
-						LumOff: &attrValInt{
-							Val: intPtr(85000),
-						},
-					},
-				},
-			},
+			Ln: f.drawChartLn(&opts.Border),
 		},
 		PrintSettings: &cPrintSettings{
 			PageMargins: &cPageMargins{
@@ -179,7 +107,8 @@ func (f *File) addChart(formatSet *formatChart, comboCharts []*formatChart) {
 			},
 		},
 	}
-	plotAreaFunc := map[string]func(*formatChart) *cPlotArea{
+	xlsxChartSpace.SpPr = f.drawShapeFill(opts.Fill, xlsxChartSpace.SpPr)
+	plotAreaFunc := map[ChartType]func(pa *cPlotArea, opts *Chart) *cPlotArea{
 		Area:                        f.drawBaseChart,
 		AreaStacked:                 f.drawBaseChart,
 		AreaPercentStacked:          f.drawBaseChart,
@@ -222,22 +151,24 @@ func (f *File) addChart(formatSet *formatChart, comboCharts []*formatChart) {
 		Col3DCylinderPercentStacked: f.drawBaseChart,
 		Doughnut:                    f.drawDoughnutChart,
 		Line:                        f.drawLineChart,
-		Pie3D:                       f.drawPie3DChart,
+		Line3D:                      f.drawLine3DChart,
 		Pie:                         f.drawPieChart,
-		PieOfPieChart:               f.drawPieOfPieChart,
-		BarOfPieChart:               f.drawBarOfPieChart,
+		Pie3D:                       f.drawPie3DChart,
+		PieOfPie:                    f.drawPieOfPieChart,
+		BarOfPie:                    f.drawBarOfPieChart,
 		Radar:                       f.drawRadarChart,
 		Scatter:                     f.drawScatterChart,
 		Surface3D:                   f.drawSurface3DChart,
 		WireframeSurface3D:          f.drawSurface3DChart,
 		Contour:                     f.drawSurfaceChart,
 		WireframeContour:            f.drawSurfaceChart,
-		Bubble:                      f.drawBaseChart,
-		Bubble3D:                    f.drawBaseChart,
+		Bubble:                      f.drawBubbleChart,
+		Bubble3D:                    f.drawBubbleChart,
 	}
-	if formatSet.Legend.None {
+	if opts.Legend.Position == "none" {
 		xlsxChartSpace.Chart.Legend = nil
 	}
+	xlsxChartSpace.Chart.PlotArea.SpPr = f.drawShapeFill(opts.PlotArea.Fill, xlsxChartSpace.Chart.PlotArea.SpPr)
 	addChart := func(c, p *cPlotArea) {
 		immutable, mutable := reflect.ValueOf(c).Elem(), reflect.ValueOf(p).Elem()
 		for i := 0; i < mutable.NumField(); i++ {
@@ -245,14 +176,19 @@ func (f *File) addChart(formatSet *formatChart, comboCharts []*formatChart) {
 			if field.IsNil() {
 				continue
 			}
-			immutable.FieldByName(mutable.Type().Field(i).Name).Set(field)
+			fld := immutable.FieldByName(mutable.Type().Field(i).Name)
+			if field.Kind() == reflect.Slice && i < 16 { // All []*cCharts type fields
+				fld.Set(reflect.Append(fld, field.Index(0)))
+				continue
+			}
+			fld.Set(field)
 		}
 	}
-	addChart(xlsxChartSpace.Chart.PlotArea, plotAreaFunc[formatSet.Type](formatSet))
-	order := len(formatSet.Series)
+	addChart(xlsxChartSpace.Chart.PlotArea, plotAreaFunc[opts.Type](xlsxChartSpace.Chart.PlotArea, opts))
+	order := len(opts.Series)
 	for idx := range comboCharts {
 		comboCharts[idx].order = order
-		addChart(xlsxChartSpace.Chart.PlotArea, plotAreaFunc[comboCharts[idx].Type](comboCharts[idx]))
+		addChart(xlsxChartSpace.Chart.PlotArea, plotAreaFunc[comboCharts[idx].Type](xlsxChartSpace.Chart.PlotArea, comboCharts[idx]))
 		order += len(comboCharts[idx].Series)
 	}
 	chart, _ := xml.Marshal(xlsxChartSpace)
@@ -262,447 +198,536 @@ func (f *File) addChart(formatSet *formatChart, comboCharts []*formatChart) {
 
 // drawBaseChart provides a function to draw the c:plotArea element for bar,
 // and column series charts by given format sets.
-func (f *File) drawBaseChart(formatSet *formatChart) *cPlotArea {
-	c := cCharts{
-		BarDir: &attrValString{
-			Val: stringPtr("col"),
+func (f *File) drawBaseChart(pa *cPlotArea, opts *Chart) *cPlotArea {
+	c := []*cCharts{
+		{
+			BarDir: &attrValString{
+				Val: stringPtr("col"),
+			},
+			Grouping: &attrValString{
+				Val: stringPtr(plotAreaChartGrouping[opts.Type]),
+			},
+			VaryColors: &attrValBool{
+				Val: opts.VaryColors,
+			},
+			Ser:      f.drawChartSeries(opts),
+			Shape:    f.drawChartShape(opts),
+			DLbls:    f.drawChartDLbls(opts),
+			GapWidth: f.drawChartGapWidth(opts),
+			AxID:     f.genAxID(opts),
+			Overlap:  f.drawChartOverlap(opts),
 		},
-		Grouping: &attrValString{
-			Val: stringPtr("clustered"),
-		},
-		VaryColors: &attrValBool{
-			Val: boolPtr(formatSet.VaryColors),
-		},
-		Ser:   f.drawChartSeries(formatSet),
-		Shape: f.drawChartShape(formatSet),
-		DLbls: f.drawChartDLbls(formatSet),
-		AxID: []*attrValInt{
-			{Val: intPtr(754001152)},
-			{Val: intPtr(753999904)},
-		},
-		Overlap: &attrValInt{Val: intPtr(100)},
 	}
 	var ok bool
-	if *c.BarDir.Val, ok = plotAreaChartBarDir[formatSet.Type]; !ok {
-		c.BarDir = nil
+	if *c[0].BarDir.Val, ok = plotAreaChartBarDir[opts.Type]; !ok {
+		c[0].BarDir = nil
 	}
-	if *c.Grouping.Val, ok = plotAreaChartGrouping[formatSet.Type]; !ok {
-		c.Grouping = nil
-	}
-	if *c.Overlap.Val, ok = plotAreaChartOverlap[formatSet.Type]; !ok {
-		c.Overlap = nil
-	}
-	catAx := f.drawPlotAreaCatAx(formatSet)
-	valAx := f.drawPlotAreaValAx(formatSet)
-	charts := map[string]*cPlotArea{
-		"area": {
-			AreaChart: &c,
+	catAx := f.drawPlotAreaCatAx(pa, opts)
+	valAx := f.drawPlotAreaValAx(pa, opts)
+	charts := map[ChartType]*cPlotArea{
+		Area: {
+			AreaChart: c,
 			CatAx:     catAx,
 			ValAx:     valAx,
 		},
-		"areaStacked": {
-			AreaChart: &c,
+		AreaStacked: {
+			AreaChart: c,
 			CatAx:     catAx,
 			ValAx:     valAx,
 		},
-		"areaPercentStacked": {
-			AreaChart: &c,
+		AreaPercentStacked: {
+			AreaChart: c,
 			CatAx:     catAx,
 			ValAx:     valAx,
 		},
-		"area3D": {
-			Area3DChart: &c,
+		Area3D: {
+			Area3DChart: c,
 			CatAx:       catAx,
 			ValAx:       valAx,
 		},
-		"area3DStacked": {
-			Area3DChart: &c,
+		Area3DStacked: {
+			Area3DChart: c,
 			CatAx:       catAx,
 			ValAx:       valAx,
 		},
-		"area3DPercentStacked": {
-			Area3DChart: &c,
+		Area3DPercentStacked: {
+			Area3DChart: c,
 			CatAx:       catAx,
 			ValAx:       valAx,
 		},
-		"bar": {
-			BarChart: &c,
+		Bar: {
+			BarChart: c,
 			CatAx:    catAx,
 			ValAx:    valAx,
 		},
-		"barStacked": {
-			BarChart: &c,
+		BarStacked: {
+			BarChart: c,
 			CatAx:    catAx,
 			ValAx:    valAx,
 		},
-		"barPercentStacked": {
-			BarChart: &c,
+		BarPercentStacked: {
+			BarChart: c,
 			CatAx:    catAx,
 			ValAx:    valAx,
 		},
-		"bar3DClustered": {
-			Bar3DChart: &c,
+		Bar3DClustered: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"bar3DStacked": {
-			Bar3DChart: &c,
+		Bar3DStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"bar3DPercentStacked": {
-			Bar3DChart: &c,
+		Bar3DPercentStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"bar3DConeClustered": {
-			Bar3DChart: &c,
+		Bar3DConeClustered: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"bar3DConeStacked": {
-			Bar3DChart: &c,
+		Bar3DConeStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"bar3DConePercentStacked": {
-			Bar3DChart: &c,
+		Bar3DConePercentStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"bar3DPyramidClustered": {
-			Bar3DChart: &c,
+		Bar3DPyramidClustered: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"bar3DPyramidStacked": {
-			Bar3DChart: &c,
+		Bar3DPyramidStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"bar3DPyramidPercentStacked": {
-			Bar3DChart: &c,
+		Bar3DPyramidPercentStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"bar3DCylinderClustered": {
-			Bar3DChart: &c,
+		Bar3DCylinderClustered: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"bar3DCylinderStacked": {
-			Bar3DChart: &c,
+		Bar3DCylinderStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"bar3DCylinderPercentStacked": {
-			Bar3DChart: &c,
+		Bar3DCylinderPercentStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col": {
-			BarChart: &c,
+		Col: {
+			BarChart: c,
 			CatAx:    catAx,
 			ValAx:    valAx,
 		},
-		"colStacked": {
-			BarChart: &c,
+		ColStacked: {
+			BarChart: c,
 			CatAx:    catAx,
 			ValAx:    valAx,
 		},
-		"colPercentStacked": {
-			BarChart: &c,
+		ColPercentStacked: {
+			BarChart: c,
 			CatAx:    catAx,
 			ValAx:    valAx,
 		},
-		"col3D": {
-			Bar3DChart: &c,
+		Col3D: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DClustered": {
-			Bar3DChart: &c,
+		Col3DClustered: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DStacked": {
-			Bar3DChart: &c,
+		Col3DStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DPercentStacked": {
-			Bar3DChart: &c,
+		Col3DPercentStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DCone": {
-			Bar3DChart: &c,
+		Col3DCone: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DConeClustered": {
-			Bar3DChart: &c,
+		Col3DConeClustered: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DConeStacked": {
-			Bar3DChart: &c,
+		Col3DConeStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DConePercentStacked": {
-			Bar3DChart: &c,
+		Col3DConePercentStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DPyramid": {
-			Bar3DChart: &c,
+		Col3DPyramid: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DPyramidClustered": {
-			Bar3DChart: &c,
+		Col3DPyramidClustered: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DPyramidStacked": {
-			Bar3DChart: &c,
+		Col3DPyramidStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DPyramidPercentStacked": {
-			Bar3DChart: &c,
+		Col3DPyramidPercentStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DCylinder": {
-			Bar3DChart: &c,
+		Col3DCylinder: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DCylinderClustered": {
-			Bar3DChart: &c,
+		Col3DCylinderClustered: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DCylinderStacked": {
-			Bar3DChart: &c,
+		Col3DCylinderStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"col3DCylinderPercentStacked": {
-			Bar3DChart: &c,
+		Col3DCylinderPercentStacked: {
+			Bar3DChart: c,
 			CatAx:      catAx,
 			ValAx:      valAx,
 		},
-		"bubble": {
-			BubbleChart: &c,
+		Bubble: {
+			BubbleChart: c,
 			CatAx:       catAx,
 			ValAx:       valAx,
 		},
-		"bubble3D": {
-			BubbleChart: &c,
+		Bubble3D: {
+			BubbleChart: c,
 			CatAx:       catAx,
 			ValAx:       valAx,
 		},
 	}
-	return charts[formatSet.Type]
+	return charts[opts.Type]
 }
 
 // drawDoughnutChart provides a function to draw the c:plotArea element for
 // doughnut chart by given format sets.
-func (f *File) drawDoughnutChart(formatSet *formatChart) *cPlotArea {
+func (f *File) drawDoughnutChart(pa *cPlotArea, opts *Chart) *cPlotArea {
+	holeSize := 75
+	if opts.HoleSize > 0 && opts.HoleSize <= 90 {
+		holeSize = opts.HoleSize
+	}
+
 	return &cPlotArea{
-		DoughnutChart: &cCharts{
-			VaryColors: &attrValBool{
-				Val: boolPtr(formatSet.VaryColors),
+		DoughnutChart: []*cCharts{
+			{
+				VaryColors: &attrValBool{
+					Val: opts.VaryColors,
+				},
+				Ser:      f.drawChartSeries(opts),
+				HoleSize: &attrValInt{Val: intPtr(holeSize)},
 			},
-			Ser:      f.drawChartSeries(formatSet),
-			HoleSize: &attrValInt{Val: intPtr(75)},
 		},
 	}
 }
 
 // drawLineChart provides a function to draw the c:plotArea element for line
 // chart by given format sets.
-func (f *File) drawLineChart(formatSet *formatChart) *cPlotArea {
+func (f *File) drawLineChart(pa *cPlotArea, opts *Chart) *cPlotArea {
 	return &cPlotArea{
-		LineChart: &cCharts{
-			Grouping: &attrValString{
-				Val: stringPtr(plotAreaChartGrouping[formatSet.Type]),
-			},
-			VaryColors: &attrValBool{
-				Val: boolPtr(false),
-			},
-			Ser:   f.drawChartSeries(formatSet),
-			DLbls: f.drawChartDLbls(formatSet),
-			Smooth: &attrValBool{
-				Val: boolPtr(false),
-			},
-			AxID: []*attrValInt{
-				{Val: intPtr(754001152)},
-				{Val: intPtr(753999904)},
+		LineChart: []*cCharts{
+			{
+				Grouping: &attrValString{
+					Val: stringPtr(plotAreaChartGrouping[opts.Type]),
+				},
+				VaryColors: &attrValBool{
+					Val: boolPtr(false),
+				},
+				Ser:   f.drawChartSeries(opts),
+				DLbls: f.drawChartDLbls(opts),
+				AxID:  f.genAxID(opts),
 			},
 		},
-		CatAx: f.drawPlotAreaCatAx(formatSet),
-		ValAx: f.drawPlotAreaValAx(formatSet),
+		CatAx: f.drawPlotAreaCatAx(pa, opts),
+		ValAx: f.drawPlotAreaValAx(pa, opts),
+	}
+}
+
+// drawLine3DChart provides a function to draw the c:plotArea element for line
+// chart by given format sets.
+func (f *File) drawLine3DChart(pa *cPlotArea, opts *Chart) *cPlotArea {
+	return &cPlotArea{
+		Line3DChart: []*cCharts{
+			{
+				Grouping: &attrValString{
+					Val: stringPtr(plotAreaChartGrouping[opts.Type]),
+				},
+				VaryColors: &attrValBool{
+					Val: boolPtr(false),
+				},
+				Ser:   f.drawChartSeries(opts),
+				DLbls: f.drawChartDLbls(opts),
+				AxID:  f.genAxID(opts),
+			},
+		},
+		CatAx: f.drawPlotAreaCatAx(pa, opts),
+		ValAx: f.drawPlotAreaValAx(pa, opts),
 	}
 }
 
 // drawPieChart provides a function to draw the c:plotArea element for pie
 // chart by given format sets.
-func (f *File) drawPieChart(formatSet *formatChart) *cPlotArea {
+func (f *File) drawPieChart(pa *cPlotArea, opts *Chart) *cPlotArea {
 	return &cPlotArea{
-		PieChart: &cCharts{
-			VaryColors: &attrValBool{
-				Val: boolPtr(formatSet.VaryColors),
+		PieChart: []*cCharts{
+			{
+				VaryColors: &attrValBool{
+					Val: opts.VaryColors,
+				},
+				Ser: f.drawChartSeries(opts),
 			},
-			Ser: f.drawChartSeries(formatSet),
 		},
 	}
 }
 
 // drawPie3DChart provides a function to draw the c:plotArea element for 3D
 // pie chart by given format sets.
-func (f *File) drawPie3DChart(formatSet *formatChart) *cPlotArea {
+func (f *File) drawPie3DChart(pa *cPlotArea, opts *Chart) *cPlotArea {
 	return &cPlotArea{
-		Pie3DChart: &cCharts{
-			VaryColors: &attrValBool{
-				Val: boolPtr(formatSet.VaryColors),
+		Pie3DChart: []*cCharts{
+			{
+				VaryColors: &attrValBool{
+					Val: opts.VaryColors,
+				},
+				Ser: f.drawChartSeries(opts),
 			},
-			Ser: f.drawChartSeries(formatSet),
 		},
 	}
 }
 
 // drawPieOfPieChart provides a function to draw the c:plotArea element for
 // pie chart by given format sets.
-func (f *File) drawPieOfPieChart(formatSet *formatChart) *cPlotArea {
+func (f *File) drawPieOfPieChart(pa *cPlotArea, opts *Chart) *cPlotArea {
+	var splitPos *attrValInt
+	if opts.PlotArea.SecondPlotValues > 0 {
+		splitPos = &attrValInt{Val: intPtr(opts.PlotArea.SecondPlotValues)}
+	}
 	return &cPlotArea{
-		OfPieChart: &cCharts{
-			OfPieType: &attrValString{
-				Val: stringPtr("pie"),
+		OfPieChart: []*cCharts{
+			{
+				OfPieType: &attrValString{
+					Val: stringPtr("pie"),
+				},
+				VaryColors: &attrValBool{
+					Val: opts.VaryColors,
+				},
+				Ser:      f.drawChartSeries(opts),
+				SplitPos: splitPos,
+				SerLines: &attrValString{},
 			},
-			VaryColors: &attrValBool{
-				Val: boolPtr(formatSet.VaryColors),
-			},
-			Ser:      f.drawChartSeries(formatSet),
-			SerLines: &attrValString{},
 		},
 	}
 }
 
 // drawBarOfPieChart provides a function to draw the c:plotArea element for
 // pie chart by given format sets.
-func (f *File) drawBarOfPieChart(formatSet *formatChart) *cPlotArea {
+func (f *File) drawBarOfPieChart(pa *cPlotArea, opts *Chart) *cPlotArea {
+	var splitPos *attrValInt
+	if opts.PlotArea.SecondPlotValues > 0 {
+		splitPos = &attrValInt{Val: intPtr(opts.PlotArea.SecondPlotValues)}
+	}
 	return &cPlotArea{
-		OfPieChart: &cCharts{
-			OfPieType: &attrValString{
-				Val: stringPtr("bar"),
+		OfPieChart: []*cCharts{
+			{
+				OfPieType: &attrValString{
+					Val: stringPtr("bar"),
+				},
+				VaryColors: &attrValBool{
+					Val: opts.VaryColors,
+				},
+				SplitPos: splitPos,
+				Ser:      f.drawChartSeries(opts),
+				SerLines: &attrValString{},
 			},
-			VaryColors: &attrValBool{
-				Val: boolPtr(formatSet.VaryColors),
-			},
-			Ser:      f.drawChartSeries(formatSet),
-			SerLines: &attrValString{},
 		},
 	}
 }
 
 // drawRadarChart provides a function to draw the c:plotArea element for radar
 // chart by given format sets.
-func (f *File) drawRadarChart(formatSet *formatChart) *cPlotArea {
+func (f *File) drawRadarChart(pa *cPlotArea, opts *Chart) *cPlotArea {
 	return &cPlotArea{
-		RadarChart: &cCharts{
-			RadarStyle: &attrValString{
-				Val: stringPtr("marker"),
-			},
-			VaryColors: &attrValBool{
-				Val: boolPtr(false),
-			},
-			Ser:   f.drawChartSeries(formatSet),
-			DLbls: f.drawChartDLbls(formatSet),
-			AxID: []*attrValInt{
-				{Val: intPtr(754001152)},
-				{Val: intPtr(753999904)},
+		RadarChart: []*cCharts{
+			{
+				RadarStyle: &attrValString{
+					Val: stringPtr("marker"),
+				},
+				VaryColors: &attrValBool{
+					Val: boolPtr(false),
+				},
+				Ser:   f.drawChartSeries(opts),
+				DLbls: f.drawChartDLbls(opts),
+				AxID:  f.genAxID(opts),
 			},
 		},
-		CatAx: f.drawPlotAreaCatAx(formatSet),
-		ValAx: f.drawPlotAreaValAx(formatSet),
+		CatAx: f.drawPlotAreaCatAx(pa, opts),
+		ValAx: f.drawPlotAreaValAx(pa, opts),
 	}
 }
 
 // drawScatterChart provides a function to draw the c:plotArea element for
 // scatter chart by given format sets.
-func (f *File) drawScatterChart(formatSet *formatChart) *cPlotArea {
+func (f *File) drawScatterChart(pa *cPlotArea, opts *Chart) *cPlotArea {
 	return &cPlotArea{
-		ScatterChart: &cCharts{
-			ScatterStyle: &attrValString{
-				Val: stringPtr("smoothMarker"), // line,lineMarker,marker,none,smooth,smoothMarker
-			},
-			VaryColors: &attrValBool{
-				Val: boolPtr(false),
-			},
-			Ser:   f.drawChartSeries(formatSet),
-			DLbls: f.drawChartDLbls(formatSet),
-			AxID: []*attrValInt{
-				{Val: intPtr(754001152)},
-				{Val: intPtr(753999904)},
+		ScatterChart: []*cCharts{
+			{
+				ScatterStyle: &attrValString{
+					Val: stringPtr("smoothMarker"), // line,lineMarker,marker,none,smooth,smoothMarker
+				},
+				VaryColors: &attrValBool{
+					Val: boolPtr(false),
+				},
+				Ser:   f.drawChartSeries(opts),
+				DLbls: f.drawChartDLbls(opts),
+				AxID:  f.genAxID(opts),
 			},
 		},
-		CatAx: f.drawPlotAreaCatAx(formatSet),
-		ValAx: f.drawPlotAreaValAx(formatSet),
+		ValAx: append(f.drawPlotAreaCatAx(pa, opts), f.drawPlotAreaValAx(pa, opts)...),
 	}
 }
 
 // drawSurface3DChart provides a function to draw the c:surface3DChart element by
 // given format sets.
-func (f *File) drawSurface3DChart(formatSet *formatChart) *cPlotArea {
+func (f *File) drawSurface3DChart(pa *cPlotArea, opts *Chart) *cPlotArea {
 	plotArea := &cPlotArea{
-		Surface3DChart: &cCharts{
-			Ser: f.drawChartSeries(formatSet),
-			AxID: []*attrValInt{
-				{Val: intPtr(754001152)},
-				{Val: intPtr(753999904)},
-				{Val: intPtr(832256642)},
+		Surface3DChart: []*cCharts{
+			{
+				Ser: f.drawChartSeries(opts),
+				AxID: []*attrValInt{
+					{Val: intPtr(100000000)},
+					{Val: intPtr(100000001)},
+					{Val: intPtr(100000005)},
+				},
 			},
 		},
-		CatAx: f.drawPlotAreaCatAx(formatSet),
-		ValAx: f.drawPlotAreaValAx(formatSet),
-		SerAx: f.drawPlotAreaSerAx(formatSet),
+		CatAx: f.drawPlotAreaCatAx(pa, opts),
+		ValAx: f.drawPlotAreaValAx(pa, opts),
+		SerAx: f.drawPlotAreaSerAx(opts),
 	}
-	if formatSet.Type == WireframeSurface3D {
-		plotArea.Surface3DChart.Wireframe = &attrValBool{Val: boolPtr(true)}
+	if opts.Type == WireframeSurface3D {
+		plotArea.Surface3DChart[0].Wireframe = &attrValBool{Val: boolPtr(true)}
 	}
 	return plotArea
 }
 
 // drawSurfaceChart provides a function to draw the c:surfaceChart element by
 // given format sets.
-func (f *File) drawSurfaceChart(formatSet *formatChart) *cPlotArea {
+func (f *File) drawSurfaceChart(pa *cPlotArea, opts *Chart) *cPlotArea {
 	plotArea := &cPlotArea{
-		SurfaceChart: &cCharts{
-			Ser: f.drawChartSeries(formatSet),
-			AxID: []*attrValInt{
-				{Val: intPtr(754001152)},
-				{Val: intPtr(753999904)},
-				{Val: intPtr(832256642)},
+		SurfaceChart: []*cCharts{
+			{
+				Ser: f.drawChartSeries(opts),
+				AxID: []*attrValInt{
+					{Val: intPtr(100000000)},
+					{Val: intPtr(100000001)},
+					{Val: intPtr(100000005)},
+				},
 			},
 		},
-		CatAx: f.drawPlotAreaCatAx(formatSet),
-		ValAx: f.drawPlotAreaValAx(formatSet),
-		SerAx: f.drawPlotAreaSerAx(formatSet),
+		CatAx: f.drawPlotAreaCatAx(pa, opts),
+		ValAx: f.drawPlotAreaValAx(pa, opts),
+		SerAx: f.drawPlotAreaSerAx(opts),
 	}
-	if formatSet.Type == WireframeContour {
-		plotArea.SurfaceChart.Wireframe = &attrValBool{Val: boolPtr(true)}
+	if opts.Type == WireframeContour {
+		plotArea.SurfaceChart[0].Wireframe = &attrValBool{Val: boolPtr(true)}
 	}
 	return plotArea
 }
 
+// drawBubbleChart provides a function to draw the c:bubbleChart element by
+// given format sets.
+func (f *File) drawBubbleChart(pa *cPlotArea, opts *Chart) *cPlotArea {
+	plotArea := &cPlotArea{
+		BubbleChart: []*cCharts{
+			{
+				VaryColors: &attrValBool{
+					Val: opts.VaryColors,
+				},
+				Ser:   f.drawChartSeries(opts),
+				DLbls: f.drawChartDLbls(opts),
+				AxID:  f.genAxID(opts),
+			},
+		},
+		ValAx: append(f.drawPlotAreaCatAx(pa, opts), f.drawPlotAreaValAx(pa, opts)...),
+	}
+	if opts.BubbleSize > 0 && opts.BubbleSize <= 300 {
+		plotArea.BubbleChart[0].BubbleScale = &attrValFloat{Val: float64Ptr(float64(opts.BubbleSize))}
+	}
+	return plotArea
+}
+
+// drawChartGapWidth provides a function to draw the c:gapWidth element by given
+// format sets.
+func (f *File) drawChartGapWidth(opts *Chart) *attrValInt {
+	for _, t := range barColChartTypes {
+		if t == opts.Type && opts.GapWidth != nil && *opts.GapWidth != 150 && *opts.GapWidth <= 500 {
+			return &attrValInt{intPtr(int(*opts.GapWidth))}
+		}
+	}
+	return nil
+}
+
+// drawChartOverlap provides a function to draw the c:overlap element by given
+// format sets.
+func (f *File) drawChartOverlap(opts *Chart) *attrValInt {
+	var val *attrValInt
+	if _, ok := plotAreaChartOverlap[opts.Type]; ok {
+		val = &attrValInt{intPtr(100)}
+	}
+	if opts.Overlap != nil && -100 <= *opts.Overlap && *opts.Overlap <= 100 {
+		val = &attrValInt{intPtr(*opts.Overlap)}
+	}
+	for _, t := range barColChartTypes {
+		if t == opts.Type {
+			return val
+		}
+	}
+	return nil
+}
+
 // drawChartShape provides a function to draw the c:shape element by given
 // format sets.
-func (f *File) drawChartShape(formatSet *formatChart) *attrValString {
-	shapes := map[string]string{
+func (f *File) drawChartShape(opts *Chart) *attrValString {
+	shapes := map[ChartType]string{
 		Bar3DConeClustered:          "cone",
 		Bar3DConeStacked:            "cone",
 		Bar3DConePercentStacked:     "cone",
@@ -725,7 +750,7 @@ func (f *File) drawChartShape(formatSet *formatChart) *attrValString {
 		Col3DCylinderStacked:        "cylinder",
 		Col3DCylinderPercentStacked: "cylinder",
 	}
-	if shape, ok := shapes[formatSet.Type]; ok {
+	if shape, ok := shapes[opts.Type]; ok {
 		return &attrValString{Val: stringPtr(shape)}
 	}
 	return nil
@@ -733,58 +758,79 @@ func (f *File) drawChartShape(formatSet *formatChart) *attrValString {
 
 // drawChartSeries provides a function to draw the c:ser element by given
 // format sets.
-func (f *File) drawChartSeries(formatSet *formatChart) *[]cSer {
-	ser := []cSer{}
-	for k := range formatSet.Series {
+func (f *File) drawChartSeries(opts *Chart) *[]cSer {
+	var ser []cSer
+	for k := range opts.Series {
 		ser = append(ser, cSer{
-			IDx:   &attrValInt{Val: intPtr(k + formatSet.order)},
-			Order: &attrValInt{Val: intPtr(k + formatSet.order)},
+			IDx:   &attrValInt{Val: intPtr(k + opts.order)},
+			Order: &attrValInt{Val: intPtr(k + opts.order)},
 			Tx: &cTx{
 				StrRef: &cStrRef{
-					F: formatSet.Series[k].Name,
+					F: opts.Series[k].Name,
 				},
 			},
-			SpPr:             f.drawChartSeriesSpPr(k, formatSet),
-			Marker:           f.drawChartSeriesMarker(k, formatSet),
-			DPt:              f.drawChartSeriesDPt(k, formatSet),
-			DLbls:            f.drawChartSeriesDLbls(formatSet),
+			SpPr:             f.drawChartSeriesSpPr(k, opts),
+			Marker:           f.drawChartSeriesMarker(k, opts),
+			DPt:              f.drawChartSeriesDPt(k, opts),
+			DLbls:            f.drawChartSeriesDLbls(k, opts),
 			InvertIfNegative: &attrValBool{Val: boolPtr(false)},
-			Cat:              f.drawChartSeriesCat(formatSet.Series[k], formatSet),
-			Val:              f.drawChartSeriesVal(formatSet.Series[k], formatSet),
-			XVal:             f.drawChartSeriesXVal(formatSet.Series[k], formatSet),
-			YVal:             f.drawChartSeriesYVal(formatSet.Series[k], formatSet),
-			BubbleSize:       f.drawCharSeriesBubbleSize(formatSet.Series[k], formatSet),
-			Bubble3D:         f.drawCharSeriesBubble3D(formatSet),
+			Cat:              f.drawChartSeriesCat(opts.Series[k], opts),
+			Smooth:           &attrValBool{Val: boolPtr(opts.Series[k].Line.Smooth)},
+			Val:              f.drawChartSeriesVal(opts.Series[k], opts),
+			XVal:             f.drawChartSeriesXVal(opts.Series[k], opts),
+			YVal:             f.drawChartSeriesYVal(opts.Series[k], opts),
+			BubbleSize:       f.drawCharSeriesBubbleSize(opts.Series[k], opts),
+			Bubble3D:         f.drawCharSeriesBubble3D(opts),
 		})
 	}
 	return &ser
 }
 
+// drawShapeFill provides a function to draw the a:solidFill element by given
+// fill format sets.
+func (f *File) drawShapeFill(fill Fill, spPr *cSpPr) *cSpPr {
+	if fill.Type == "pattern" && fill.Pattern == 1 {
+		if spPr == nil {
+			spPr = &cSpPr{}
+		}
+		if len(fill.Color) == 1 {
+			spPr.SolidFill = &aSolidFill{SrgbClr: &attrValString{Val: stringPtr(strings.TrimPrefix(fill.Color[0], "#"))}}
+			return spPr
+		}
+		spPr.SolidFill = nil
+		spPr.NoFill = stringPtr("")
+	}
+	return spPr
+}
+
 // drawChartSeriesSpPr provides a function to draw the c:spPr element by given
 // format sets.
-func (f *File) drawChartSeriesSpPr(i int, formatSet *formatChart) *cSpPr {
-	spPrScatter := &cSpPr{
+func (f *File) drawChartSeriesSpPr(i int, opts *Chart) *cSpPr {
+	spPr := &cSpPr{SolidFill: &aSolidFill{SchemeClr: &aSchemeClr{Val: "accent" + strconv.Itoa((opts.order+i)%6+1)}}}
+	spPr = f.drawShapeFill(opts.Series[i].Fill, spPr)
+	solid := &cSpPr{
 		Ln: &aLn{
-			W:      25400,
-			NoFill: " ",
+			W:         f.ptToEMUs(opts.Series[i].Line.Width),
+			Cap:       "rnd", // rnd, sq, flat
+			SolidFill: spPr.SolidFill,
 		},
 	}
-	spPrLine := &cSpPr{
-		Ln: &aLn{
-			W:   f.ptToEMUs(formatSet.Series[i].Line.Width),
-			Cap: "rnd", // rnd, sq, flat
-			SolidFill: &aSolidFill{
-				SchemeClr: &aSchemeClr{Val: "accent" + strconv.Itoa((formatSet.order+i)%6+1)},
-			},
-		},
+	noLn := &cSpPr{Ln: &aLn{NoFill: &attrValString{}}}
+	if chartSeriesSpPr, ok := map[ChartType]map[ChartLineType]*cSpPr{
+		Line:    {ChartLineUnset: solid, ChartLineSolid: solid, ChartLineNone: noLn, ChartLineAutomatic: solid},
+		Scatter: {ChartLineUnset: noLn, ChartLineSolid: solid, ChartLineNone: noLn, ChartLineAutomatic: noLn},
+	}[opts.Type]; ok {
+		return chartSeriesSpPr[opts.Series[i].Line.Type]
 	}
-	chartSeriesSpPr := map[string]*cSpPr{Line: spPrLine, Scatter: spPrScatter}
-	return chartSeriesSpPr[formatSet.Type]
+	if spPr.SolidFill.SrgbClr != nil {
+		return spPr
+	}
+	return nil
 }
 
 // drawChartSeriesDPt provides a function to draw the c:dPt element by given
 // data index and format sets.
-func (f *File) drawChartSeriesDPt(i int, formatSet *formatChart) []*cDPt {
+func (f *File) drawChartSeriesDPt(i int, opts *Chart) []*cDPt {
 	dpt := []*cDPt{{
 		IDx:      &attrValInt{Val: intPtr(i)},
 		Bubble3D: &attrValBool{Val: boolPtr(false)},
@@ -807,20 +853,20 @@ func (f *File) drawChartSeriesDPt(i int, formatSet *formatChart) []*cDPt {
 			},
 		},
 	}}
-	chartSeriesDPt := map[string][]*cDPt{Pie: dpt, Pie3D: dpt}
-	return chartSeriesDPt[formatSet.Type]
+	chartSeriesDPt := map[ChartType][]*cDPt{Pie: dpt, Pie3D: dpt}
+	return chartSeriesDPt[opts.Type]
 }
 
 // drawChartSeriesCat provides a function to draw the c:cat element by given
 // chart series and format sets.
-func (f *File) drawChartSeriesCat(v formatChartSeries, formatSet *formatChart) *cCat {
+func (f *File) drawChartSeriesCat(v ChartSeries, opts *Chart) *cCat {
 	cat := &cCat{
 		StrRef: &cStrRef{
 			F: v.Categories,
 		},
 	}
-	chartSeriesCat := map[string]*cCat{Scatter: nil, Bubble: nil, Bubble3D: nil}
-	if _, ok := chartSeriesCat[formatSet.Type]; ok || v.Categories == "" {
+	chartSeriesCat := map[ChartType]*cCat{Scatter: nil, Bubble: nil, Bubble3D: nil}
+	if _, ok := chartSeriesCat[opts.Type]; ok || v.Categories == "" {
 		return nil
 	}
 	return cat
@@ -828,14 +874,14 @@ func (f *File) drawChartSeriesCat(v formatChartSeries, formatSet *formatChart) *
 
 // drawChartSeriesVal provides a function to draw the c:val element by given
 // chart series and format sets.
-func (f *File) drawChartSeriesVal(v formatChartSeries, formatSet *formatChart) *cVal {
+func (f *File) drawChartSeriesVal(v ChartSeries, opts *Chart) *cVal {
 	val := &cVal{
 		NumRef: &cNumRef{
 			F: v.Values,
 		},
 	}
-	chartSeriesVal := map[string]*cVal{Scatter: nil, Bubble: nil, Bubble3D: nil}
-	if _, ok := chartSeriesVal[formatSet.Type]; ok {
+	chartSeriesVal := map[ChartType]*cVal{Scatter: nil, Bubble: nil, Bubble3D: nil}
+	if _, ok := chartSeriesVal[opts.Type]; ok {
 		return nil
 	}
 	return val
@@ -843,16 +889,16 @@ func (f *File) drawChartSeriesVal(v formatChartSeries, formatSet *formatChart) *
 
 // drawChartSeriesMarker provides a function to draw the c:marker element by
 // given data index and format sets.
-func (f *File) drawChartSeriesMarker(i int, formatSet *formatChart) *cMarker {
-	defaultSymbol := map[string]*attrValString{Scatter: {Val: stringPtr("circle")}}
+func (f *File) drawChartSeriesMarker(i int, opts *Chart) *cMarker {
+	defaultSymbol := map[ChartType]*attrValString{Scatter: {Val: stringPtr("circle")}}
 	marker := &cMarker{
-		Symbol: defaultSymbol[formatSet.Type],
+		Symbol: defaultSymbol[opts.Type],
 		Size:   &attrValInt{Val: intPtr(5)},
 	}
-	if symbol := stringPtr(formatSet.Series[i].Marker.Symbol); *symbol != "" {
+	if symbol := stringPtr(opts.Series[i].Marker.Symbol); *symbol != "" {
 		marker.Symbol = &attrValString{Val: symbol}
 	}
-	if size := intPtr(formatSet.Series[i].Marker.Size); *size != 0 {
+	if size := intPtr(opts.Series[i].Marker.Size); *size != 0 {
 		marker.Size = &attrValInt{Val: size}
 	}
 	if i < 6 {
@@ -872,211 +918,315 @@ func (f *File) drawChartSeriesMarker(i int, formatSet *formatChart) *cMarker {
 			},
 		}
 	}
-	chartSeriesMarker := map[string]*cMarker{Scatter: marker, Line: marker}
-	return chartSeriesMarker[formatSet.Type]
+	marker.SpPr = f.drawShapeFill(opts.Series[i].Marker.Fill, marker.SpPr)
+	chartSeriesMarker := map[ChartType]*cMarker{Scatter: marker, Line: marker}
+	return chartSeriesMarker[opts.Type]
 }
 
 // drawChartSeriesXVal provides a function to draw the c:xVal element by given
 // chart series and format sets.
-func (f *File) drawChartSeriesXVal(v formatChartSeries, formatSet *formatChart) *cCat {
+func (f *File) drawChartSeriesXVal(v ChartSeries, opts *Chart) *cCat {
 	cat := &cCat{
 		StrRef: &cStrRef{
 			F: v.Categories,
 		},
 	}
-	chartSeriesXVal := map[string]*cCat{Scatter: cat}
-	return chartSeriesXVal[formatSet.Type]
+	chartSeriesXVal := map[ChartType]*cCat{Scatter: cat, Bubble: cat, Bubble3D: cat}
+	return chartSeriesXVal[opts.Type]
 }
 
 // drawChartSeriesYVal provides a function to draw the c:yVal element by given
 // chart series and format sets.
-func (f *File) drawChartSeriesYVal(v formatChartSeries, formatSet *formatChart) *cVal {
+func (f *File) drawChartSeriesYVal(v ChartSeries, opts *Chart) *cVal {
 	val := &cVal{
 		NumRef: &cNumRef{
 			F: v.Values,
 		},
 	}
-	chartSeriesYVal := map[string]*cVal{Scatter: val, Bubble: val, Bubble3D: val}
-	return chartSeriesYVal[formatSet.Type]
+	chartSeriesYVal := map[ChartType]*cVal{Scatter: val, Bubble: val, Bubble3D: val}
+	return chartSeriesYVal[opts.Type]
 }
 
 // drawCharSeriesBubbleSize provides a function to draw the c:bubbleSize
 // element by given chart series and format sets.
-func (f *File) drawCharSeriesBubbleSize(v formatChartSeries, formatSet *formatChart) *cVal {
-	if _, ok := map[string]bool{Bubble: true, Bubble3D: true}[formatSet.Type]; !ok {
+func (f *File) drawCharSeriesBubbleSize(v ChartSeries, opts *Chart) *cVal {
+	if _, ok := map[ChartType]bool{Bubble: true, Bubble3D: true}[opts.Type]; !ok {
 		return nil
+	}
+	fVal := v.Values
+	if v.Sizes != "" {
+		fVal = v.Sizes
 	}
 	return &cVal{
 		NumRef: &cNumRef{
-			F: v.Values,
+			F: fVal,
 		},
 	}
 }
 
 // drawCharSeriesBubble3D provides a function to draw the c:bubble3D element
 // by given format sets.
-func (f *File) drawCharSeriesBubble3D(formatSet *formatChart) *attrValBool {
-	if _, ok := map[string]bool{Bubble3D: true}[formatSet.Type]; !ok {
+func (f *File) drawCharSeriesBubble3D(opts *Chart) *attrValBool {
+	if _, ok := map[ChartType]bool{Bubble3D: true}[opts.Type]; !ok {
 		return nil
 	}
 	return &attrValBool{Val: boolPtr(true)}
 }
 
+// drawChartNumFmt provides a function to draw the c:numFmt element by given
+// data labels format sets.
+func (f *File) drawChartNumFmt(labels ChartNumFmt) *cNumFmt {
+	var numFmt *cNumFmt
+	if labels.CustomNumFmt != "" || labels.SourceLinked {
+		numFmt = &cNumFmt{
+			FormatCode:   labels.CustomNumFmt,
+			SourceLinked: labels.SourceLinked,
+		}
+	}
+	return numFmt
+}
+
 // drawChartDLbls provides a function to draw the c:dLbls element by given
 // format sets.
-func (f *File) drawChartDLbls(formatSet *formatChart) *cDLbls {
+func (f *File) drawChartDLbls(opts *Chart) *cDLbls {
 	return &cDLbls{
-		ShowLegendKey:   &attrValBool{Val: boolPtr(formatSet.Legend.ShowLegendKey)},
-		ShowVal:         &attrValBool{Val: boolPtr(formatSet.Plotarea.ShowVal)},
-		ShowCatName:     &attrValBool{Val: boolPtr(formatSet.Plotarea.ShowCatName)},
-		ShowSerName:     &attrValBool{Val: boolPtr(formatSet.Plotarea.ShowSerName)},
-		ShowBubbleSize:  &attrValBool{Val: boolPtr(formatSet.Plotarea.ShowBubbleSize)},
-		ShowPercent:     &attrValBool{Val: boolPtr(formatSet.Plotarea.ShowPercent)},
-		ShowLeaderLines: &attrValBool{Val: boolPtr(formatSet.Plotarea.ShowLeaderLines)},
+		NumFmt:          f.drawChartNumFmt(opts.PlotArea.NumFmt),
+		ShowLegendKey:   &attrValBool{Val: boolPtr(opts.Legend.ShowLegendKey)},
+		ShowVal:         &attrValBool{Val: boolPtr(opts.PlotArea.ShowVal)},
+		ShowCatName:     &attrValBool{Val: boolPtr(opts.PlotArea.ShowCatName)},
+		ShowSerName:     &attrValBool{Val: boolPtr(opts.PlotArea.ShowSerName)},
+		ShowBubbleSize:  &attrValBool{Val: boolPtr(opts.PlotArea.ShowBubbleSize)},
+		ShowPercent:     &attrValBool{Val: boolPtr(opts.PlotArea.ShowPercent)},
+		ShowLeaderLines: &attrValBool{Val: boolPtr(opts.PlotArea.ShowLeaderLines)},
 	}
+}
+
+// inSupportedChartDataLabelsPositionType provides a method to check if an
+// element is present in an array, and return the index of its location,
+// otherwise return -1.
+func inSupportedChartDataLabelsPositionType(a []ChartDataLabelPositionType, x ChartDataLabelPositionType) int {
+	for idx, n := range a {
+		if x == n {
+			return idx
+		}
+	}
+	return -1
 }
 
 // drawChartSeriesDLbls provides a function to draw the c:dLbls element by
 // given format sets.
-func (f *File) drawChartSeriesDLbls(formatSet *formatChart) *cDLbls {
-	dLbls := f.drawChartDLbls(formatSet)
-	chartSeriesDLbls := map[string]*cDLbls{
-		Scatter: nil, Surface3D: nil, WireframeSurface3D: nil, Contour: nil, WireframeContour: nil, Bubble: nil, Bubble3D: nil}
-	if _, ok := chartSeriesDLbls[formatSet.Type]; ok {
+func (f *File) drawChartSeriesDLbls(i int, opts *Chart) *cDLbls {
+	dLbls := f.drawChartDLbls(opts)
+	chartSeriesDLbls := map[ChartType]*cDLbls{
+		Scatter: nil, Surface3D: nil, WireframeSurface3D: nil, Contour: nil, WireframeContour: nil,
+	}
+	if _, ok := chartSeriesDLbls[opts.Type]; ok {
 		return nil
 	}
+	if types, ok := supportedChartDataLabelsPosition[opts.Type]; ok && opts.Series[i].DataLabelPosition != ChartDataLabelsPositionUnset {
+		if inSupportedChartDataLabelsPositionType(types, opts.Series[i].DataLabelPosition) != -1 {
+			dLbls.DLblPos = &attrValString{Val: stringPtr(chartDataLabelsPositionTypes[opts.Series[i].DataLabelPosition])}
+		}
+	}
+	dLbl := opts.Series[i].DataLabel
+	dLbls.SpPr = f.drawShapeFill(dLbl.Fill, dLbls.SpPr)
+	dLbls.TxPr = &cTxPr{BodyPr: aBodyPr{}, P: aP{PPr: &aPPr{DefRPr: aRPr{}}}}
+	drawChartFont(&dLbl.Font, &dLbls.TxPr.P.PPr.DefRPr)
 	return dLbls
 }
 
 // drawPlotAreaCatAx provides a function to draw the c:catAx element.
-func (f *File) drawPlotAreaCatAx(formatSet *formatChart) []*cAxs {
-	min := &attrValFloat{Val: float64Ptr(formatSet.XAxis.Minimum)}
-	max := &attrValFloat{Val: float64Ptr(formatSet.XAxis.Maximum)}
-	if formatSet.XAxis.Minimum == 0 {
-		min = nil
+func (f *File) drawPlotAreaCatAx(pa *cPlotArea, opts *Chart) []*cAxs {
+	maxVal := &attrValFloat{Val: opts.XAxis.Maximum}
+	minVal := &attrValFloat{Val: opts.XAxis.Minimum}
+	if opts.XAxis.Maximum == nil {
+		maxVal = nil
 	}
-	if formatSet.XAxis.Maximum == 0 {
-		max = nil
+	if opts.XAxis.Minimum == nil {
+		minVal = nil
 	}
-	axs := []*cAxs{
-		{
-			AxID: &attrValInt{Val: intPtr(754001152)},
-			Scaling: &cScaling{
-				Orientation: &attrValString{Val: stringPtr(orientation[formatSet.XAxis.ReverseOrder])},
-				Max:         max,
-				Min:         min,
-			},
-			Delete: &attrValBool{Val: boolPtr(formatSet.XAxis.None)},
-			AxPos:  &attrValString{Val: stringPtr(catAxPos[formatSet.XAxis.ReverseOrder])},
-			NumFmt: &cNumFmt{
-				FormatCode:   "General",
-				SourceLinked: true,
-			},
-			MajorTickMark: &attrValString{Val: stringPtr("none")},
-			MinorTickMark: &attrValString{Val: stringPtr("none")},
-			TickLblPos:    &attrValString{Val: stringPtr("nextTo")},
-			SpPr:          f.drawPlotAreaSpPr(),
-			TxPr:          f.drawPlotAreaTxPr(),
-			CrossAx:       &attrValInt{Val: intPtr(753999904)},
-			Crosses:       &attrValString{Val: stringPtr("autoZero")},
-			Auto:          &attrValBool{Val: boolPtr(true)},
-			LblAlgn:       &attrValString{Val: stringPtr("ctr")},
-			LblOffset:     &attrValInt{Val: intPtr(100)},
-			NoMultiLvlLbl: &attrValBool{Val: boolPtr(false)},
+	ax := &cAxs{
+		AxID: &attrValInt{Val: intPtr(100000000)},
+		Scaling: &cScaling{
+			Orientation: &attrValString{Val: stringPtr(orientation[opts.XAxis.ReverseOrder])},
+			Max:         maxVal,
+			Min:         minVal,
 		},
+		Delete:        &attrValBool{Val: boolPtr(opts.XAxis.None)},
+		AxPos:         &attrValString{Val: stringPtr(catAxPos[opts.XAxis.ReverseOrder])},
+		NumFmt:        &cNumFmt{FormatCode: "General"},
+		MajorTickMark: &attrValString{Val: stringPtr("none")},
+		MinorTickMark: &attrValString{Val: stringPtr("none")},
+		Title:         f.drawPlotAreaTitles(opts.XAxis.Title, ""),
+		TickLblPos:    &attrValString{Val: stringPtr(tickLblPosVal[opts.XAxis.TickLabelPosition])},
+		SpPr:          f.drawPlotAreaSpPr(),
+		TxPr:          f.drawPlotAreaTxPr(&opts.XAxis),
+		CrossAx:       &attrValInt{Val: intPtr(100000001)},
+		Crosses:       &attrValString{Val: stringPtr("autoZero")},
+		Auto:          &attrValBool{Val: boolPtr(true)},
+		LblAlgn:       &attrValString{Val: stringPtr("ctr")},
+		LblOffset:     &attrValInt{Val: intPtr(100)},
+		NoMultiLvlLbl: &attrValBool{Val: boolPtr(false)},
 	}
-	if formatSet.XAxis.MajorGridlines {
-		axs[0].MajorGridlines = &cChartLines{SpPr: f.drawPlotAreaSpPr()}
+	if numFmt := f.drawChartNumFmt(opts.XAxis.NumFmt); numFmt != nil {
+		ax.NumFmt = numFmt
 	}
-	if formatSet.XAxis.MinorGridlines {
-		axs[0].MinorGridlines = &cChartLines{SpPr: f.drawPlotAreaSpPr()}
+	if opts.XAxis.MajorGridLines {
+		ax.MajorGridlines = &cChartLines{SpPr: f.drawPlotAreaSpPr()}
 	}
-	if formatSet.XAxis.TickLabelSkip != 0 {
-		axs[0].TickLblSkip = &attrValInt{Val: intPtr(formatSet.XAxis.TickLabelSkip)}
+	if opts.XAxis.MinorGridLines {
+		ax.MinorGridlines = &cChartLines{SpPr: f.drawPlotAreaSpPr()}
 	}
-	return axs
+	if opts.XAxis.TickLabelSkip != 0 {
+		ax.TickLblSkip = &attrValInt{Val: intPtr(opts.XAxis.TickLabelSkip)}
+	}
+	if opts.order > 0 && opts.YAxis.Secondary && pa.CatAx != nil {
+		ax.AxID = &attrValInt{Val: intPtr(opts.XAxis.axID)}
+		ax.Delete = &attrValBool{Val: boolPtr(true)}
+		ax.Crosses = nil
+		ax.CrossAx = &attrValInt{Val: intPtr(opts.YAxis.axID)}
+		return []*cAxs{pa.CatAx[0], ax}
+	}
+	return []*cAxs{ax}
 }
 
 // drawPlotAreaValAx provides a function to draw the c:valAx element.
-func (f *File) drawPlotAreaValAx(formatSet *formatChart) []*cAxs {
-	min := &attrValFloat{Val: float64Ptr(formatSet.YAxis.Minimum)}
-	max := &attrValFloat{Val: float64Ptr(formatSet.YAxis.Maximum)}
-	if formatSet.YAxis.Minimum == 0 {
-		min = nil
+func (f *File) drawPlotAreaValAx(pa *cPlotArea, opts *Chart) []*cAxs {
+	maxVal := &attrValFloat{Val: opts.YAxis.Maximum}
+	minVal := &attrValFloat{Val: opts.YAxis.Minimum}
+	if opts.YAxis.Maximum == nil {
+		maxVal = nil
 	}
-	if formatSet.YAxis.Maximum == 0 {
-		max = nil
+	if opts.YAxis.Minimum == nil {
+		minVal = nil
 	}
 	var logBase *attrValFloat
-	if formatSet.YAxis.LogBase >= 2 && formatSet.YAxis.LogBase <= 1000 {
-		logBase = &attrValFloat{Val: float64Ptr(formatSet.YAxis.LogBase)}
+	if opts.YAxis.LogBase >= 2 && opts.YAxis.LogBase <= 1000 {
+		logBase = &attrValFloat{Val: float64Ptr(opts.YAxis.LogBase)}
 	}
-	axs := []*cAxs{
-		{
-			AxID: &attrValInt{Val: intPtr(753999904)},
-			Scaling: &cScaling{
-				LogBase:     logBase,
-				Orientation: &attrValString{Val: stringPtr(orientation[formatSet.YAxis.ReverseOrder])},
-				Max:         max,
-				Min:         min,
-			},
-			Delete: &attrValBool{Val: boolPtr(formatSet.YAxis.None)},
-			AxPos:  &attrValString{Val: stringPtr(valAxPos[formatSet.YAxis.ReverseOrder])},
-			NumFmt: &cNumFmt{
-				FormatCode:   chartValAxNumFmtFormatCode[formatSet.Type],
-				SourceLinked: true,
-			},
-			MajorTickMark: &attrValString{Val: stringPtr("none")},
-			MinorTickMark: &attrValString{Val: stringPtr("none")},
-			TickLblPos:    &attrValString{Val: stringPtr("nextTo")},
-			SpPr:          f.drawPlotAreaSpPr(),
-			TxPr:          f.drawPlotAreaTxPr(),
-			CrossAx:       &attrValInt{Val: intPtr(754001152)},
-			Crosses:       &attrValString{Val: stringPtr("autoZero")},
-			CrossBetween:  &attrValString{Val: stringPtr(chartValAxCrossBetween[formatSet.Type])},
+	ax := &cAxs{
+		AxID: &attrValInt{Val: intPtr(100000001)},
+		Scaling: &cScaling{
+			LogBase:     logBase,
+			Orientation: &attrValString{Val: stringPtr(orientation[opts.YAxis.ReverseOrder])},
+			Max:         maxVal,
+			Min:         minVal,
 		},
+		Delete: &attrValBool{Val: boolPtr(opts.YAxis.None)},
+		AxPos:  &attrValString{Val: stringPtr(valAxPos[opts.YAxis.ReverseOrder])},
+		Title:  f.drawPlotAreaTitles(opts.YAxis.Title, "horz"),
+		NumFmt: &cNumFmt{
+			FormatCode: chartValAxNumFmtFormatCode[opts.Type],
+		},
+		MajorTickMark: &attrValString{Val: stringPtr("none")},
+		MinorTickMark: &attrValString{Val: stringPtr("none")},
+		TickLblPos:    &attrValString{Val: stringPtr(tickLblPosVal[opts.YAxis.TickLabelPosition])},
+		SpPr:          f.drawPlotAreaSpPr(),
+		TxPr:          f.drawPlotAreaTxPr(&opts.YAxis),
+		CrossAx:       &attrValInt{Val: intPtr(100000000)},
+		Crosses:       &attrValString{Val: stringPtr("autoZero")},
+		CrossBetween:  &attrValString{Val: stringPtr(chartValAxCrossBetween[opts.Type])},
 	}
-	if formatSet.YAxis.MajorGridlines {
-		axs[0].MajorGridlines = &cChartLines{SpPr: f.drawPlotAreaSpPr()}
+	if numFmt := f.drawChartNumFmt(opts.YAxis.NumFmt); numFmt != nil {
+		ax.NumFmt = numFmt
 	}
-	if formatSet.YAxis.MinorGridlines {
-		axs[0].MinorGridlines = &cChartLines{SpPr: f.drawPlotAreaSpPr()}
+	if opts.YAxis.MajorGridLines {
+		ax.MajorGridlines = &cChartLines{SpPr: f.drawPlotAreaSpPr()}
 	}
-	if pos, ok := valTickLblPos[formatSet.Type]; ok {
-		axs[0].TickLblPos.Val = stringPtr(pos)
+	if opts.YAxis.MinorGridLines {
+		ax.MinorGridlines = &cChartLines{SpPr: f.drawPlotAreaSpPr()}
 	}
-	if formatSet.YAxis.MajorUnit != 0 {
-		axs[0].MajorUnit = &attrValFloat{Val: float64Ptr(formatSet.YAxis.MajorUnit)}
+	if pos, ok := tickLblPosNone[opts.Type]; ok {
+		ax.TickLblPos.Val = stringPtr(pos)
 	}
-	return axs
+	if opts.YAxis.MajorUnit != 0 {
+		ax.MajorUnit = &attrValFloat{Val: float64Ptr(opts.YAxis.MajorUnit)}
+	}
+	if opts.order > 0 && opts.YAxis.Secondary && pa.ValAx != nil {
+		ax.AxID = &attrValInt{Val: intPtr(opts.YAxis.axID)}
+		ax.AxPos = &attrValString{Val: stringPtr("r")}
+		ax.Crosses = &attrValString{Val: stringPtr("max")}
+		ax.CrossAx = &attrValInt{Val: intPtr(opts.XAxis.axID)}
+		return []*cAxs{pa.ValAx[0], ax}
+	}
+	return []*cAxs{ax}
 }
 
 // drawPlotAreaSerAx provides a function to draw the c:serAx element.
-func (f *File) drawPlotAreaSerAx(formatSet *formatChart) []*cAxs {
-	min := &attrValFloat{Val: float64Ptr(formatSet.YAxis.Minimum)}
-	max := &attrValFloat{Val: float64Ptr(formatSet.YAxis.Maximum)}
-	if formatSet.YAxis.Minimum == 0 {
-		min = nil
+func (f *File) drawPlotAreaSerAx(opts *Chart) []*cAxs {
+	maxVal := &attrValFloat{Val: opts.YAxis.Maximum}
+	minVal := &attrValFloat{Val: opts.YAxis.Minimum}
+	if opts.YAxis.Maximum == nil {
+		maxVal = nil
 	}
-	if formatSet.YAxis.Maximum == 0 {
-		max = nil
+	if opts.YAxis.Minimum == nil {
+		minVal = nil
 	}
 	return []*cAxs{
 		{
-			AxID: &attrValInt{Val: intPtr(832256642)},
+			AxID: &attrValInt{Val: intPtr(100000005)},
 			Scaling: &cScaling{
-				Orientation: &attrValString{Val: stringPtr(orientation[formatSet.YAxis.ReverseOrder])},
-				Max:         max,
-				Min:         min,
+				Orientation: &attrValString{Val: stringPtr(orientation[opts.YAxis.ReverseOrder])},
+				Max:         maxVal,
+				Min:         minVal,
 			},
-			Delete:     &attrValBool{Val: boolPtr(formatSet.YAxis.None)},
-			AxPos:      &attrValString{Val: stringPtr(catAxPos[formatSet.XAxis.ReverseOrder])},
-			TickLblPos: &attrValString{Val: stringPtr("nextTo")},
+			Delete:     &attrValBool{Val: boolPtr(opts.YAxis.None)},
+			AxPos:      &attrValString{Val: stringPtr(catAxPos[opts.XAxis.ReverseOrder])},
+			TickLblPos: &attrValString{Val: stringPtr(tickLblPosVal[opts.YAxis.TickLabelPosition])},
 			SpPr:       f.drawPlotAreaSpPr(),
-			TxPr:       f.drawPlotAreaTxPr(),
-			CrossAx:    &attrValInt{Val: intPtr(753999904)},
+			TxPr:       f.drawPlotAreaTxPr(nil),
+			CrossAx:    &attrValInt{Val: intPtr(100000001)},
 		},
 	}
+}
+
+// drawChartFont provides a function to draw the a:rPr element.
+func drawChartFont(fnt *Font, r *aRPr) {
+	if fnt == nil {
+		return
+	}
+	r.B = fnt.Bold
+	r.I = fnt.Italic
+	if idx := inStrSlice(supportedDrawingUnderlineTypes, fnt.Underline, true); idx != -1 {
+		r.U = supportedDrawingUnderlineTypes[idx]
+	}
+	if fnt.Color != "" {
+		if r.SolidFill == nil {
+			r.SolidFill = &aSolidFill{}
+		}
+		r.SolidFill.SchemeClr = nil
+		r.SolidFill.SrgbClr = &attrValString{Val: stringPtr(strings.ReplaceAll(strings.ToUpper(fnt.Color), "#", ""))}
+	}
+	if fnt.Family != "" {
+		r.Latin.Typeface = fnt.Family
+	}
+	if fnt.Size > 0 {
+		r.Sz = fnt.Size * 100
+	}
+	if fnt.Strike {
+		r.Strike = "sngStrike"
+	}
+}
+
+// drawPlotAreaTitles provides a function to draw the c:title element.
+func (f *File) drawPlotAreaTitles(runs []RichTextRun, vert string) *cTitle {
+	if len(runs) == 0 {
+		return nil
+	}
+	title := &cTitle{Tx: cTx{Rich: &cRich{}}, Overlay: &attrValBool{Val: boolPtr(false)}}
+	for _, run := range runs {
+		r := &aR{T: run.Text}
+		drawChartFont(run.Font, &r.RPr)
+		title.Tx.Rich.P = append(title.Tx.Rich.P, aP{
+			PPr:        &aPPr{DefRPr: aRPr{}},
+			R:          r,
+			EndParaRPr: &aEndParaRPr{Lang: "en-US", AltLang: "en-US"},
+		})
+	}
+	if vert == "horz" {
+		title.Tx.Rich.BodyPr = aBodyPr{Rot: -5400000, Vert: vert}
+	}
+	return title
 }
 
 // drawPlotAreaSpPr provides a function to draw the c:spPr element.
@@ -1099,8 +1249,8 @@ func (f *File) drawPlotAreaSpPr() *cSpPr {
 }
 
 // drawPlotAreaTxPr provides a function to draw the c:txPr element.
-func (f *File) drawPlotAreaTxPr() *cTxPr {
-	return &cTxPr{
+func (f *File) drawPlotAreaTxPr(opts *ChartAxis) *cTxPr {
+	cTxPr := &cTxPr{
 		BodyPr: aBodyPr{
 			Rot:              -60000000,
 			SpcFirstLastPara: true,
@@ -1127,7 +1277,7 @@ func (f *File) drawPlotAreaTxPr() *cTxPr {
 							LumOff: &attrValInt{Val: intPtr(85000)},
 						},
 					},
-					Latin: &aLatin{Typeface: "+mn-lt"},
+					Latin: &xlsxCTTextFont{Typeface: "+mn-lt"},
 					Ea:    &aEa{Typeface: "+mn-ea"},
 					Cs:    &aCs{Typeface: "+mn-cs"},
 				},
@@ -1135,29 +1285,77 @@ func (f *File) drawPlotAreaTxPr() *cTxPr {
 			EndParaRPr: &aEndParaRPr{Lang: "en-US"},
 		},
 	}
+	if opts != nil {
+		drawChartFont(&opts.Font, &cTxPr.P.PPr.DefRPr)
+		if -90 <= opts.Alignment.TextRotation && opts.Alignment.TextRotation <= 90 {
+			cTxPr.BodyPr.Rot = opts.Alignment.TextRotation * 60000
+		}
+		if idx := inStrSlice(supportedDrawingTextVerticalType, opts.Alignment.Vertical, true); idx != -1 {
+			cTxPr.BodyPr.Vert = supportedDrawingTextVerticalType[idx]
+		}
+	}
+	return cTxPr
+}
+
+// drawChartLn provides a function to draw the a:ln element.
+func (f *File) drawChartLn(opts *ChartLine) *aLn {
+	ln := &aLn{
+		W:    f.ptToEMUs(opts.Width),
+		Cap:  "flat",
+		Cmpd: "sng",
+		Algn: "ctr",
+	}
+	switch opts.Type {
+	case ChartLineSolid:
+		ln.SolidFill = &aSolidFill{
+			SchemeClr: &aSchemeClr{
+				Val: "tx1",
+				LumMod: &attrValInt{
+					Val: intPtr(15000),
+				},
+				LumOff: &attrValInt{
+					Val: intPtr(85000),
+				},
+			},
+		}
+		return ln
+	case ChartLineNone:
+		ln.NoFill = &attrValString{}
+		return ln
+	default:
+		return nil
+	}
 }
 
 // drawingParser provides a function to parse drawingXML. In order to solve
 // the problem that the label structure is changed after serialization and
 // deserialization, two different structures: decodeWsDr and encodeWsDr are
 // defined.
-func (f *File) drawingParser(path string) (*xlsxWsDr, int) {
+func (f *File) drawingParser(path string) (*xlsxWsDr, int, error) {
 	var (
 		err error
 		ok  bool
 	)
-
-	if f.Drawings[path] == nil {
-		content := xlsxWsDr{}
-		content.A = NameSpaceDrawingML.Value
-		content.Xdr = NameSpaceDrawingMLSpreadSheet.Value
-		if _, ok = f.XLSX[path]; ok { // Append Model
+	_, ok = f.Drawings.Load(path)
+	if !ok {
+		content := xlsxWsDr{
+			NS:  NameSpaceDrawingMLSpreadSheet.Value,
+			Xdr: NameSpaceDrawingMLSpreadSheet.Value,
+			A:   NameSpaceDrawingML.Value,
+		}
+		if _, ok = f.Pkg.Load(path); ok { // Append Model
 			decodeWsDr := decodeWsDr{}
 			if err = f.xmlNewDecoder(bytes.NewReader(namespaceStrictToTransitional(f.readXML(path)))).
 				Decode(&decodeWsDr); err != nil && err != io.EOF {
-				log.Printf("xml decode error: %s", err)
+				return nil, 0, err
 			}
 			content.R = decodeWsDr.R
+			for _, v := range decodeWsDr.AlternateContent {
+				content.AlternateContent = append(content.AlternateContent, &xlsxAlternateContent{
+					Content: v.Content,
+					XMLNSMC: SourceRelationshipCompatibility.Value,
+				})
+			}
 			for _, v := range decodeWsDr.OneCellAnchor {
 				content.OneCellAnchor = append(content.OneCellAnchor, &xdrCellAnchor{
 					EditAs:       v.EditAs,
@@ -1171,34 +1369,38 @@ func (f *File) drawingParser(path string) (*xlsxWsDr, int) {
 				})
 			}
 		}
-		f.Drawings[path] = &content
+		f.Drawings.Store(path, &content)
 	}
-	wsDr := f.Drawings[path]
-	return wsDr, len(wsDr.OneCellAnchor) + len(wsDr.TwoCellAnchor) + 2
+	var wsDr *xlsxWsDr
+	if drawing, ok := f.Drawings.Load(path); ok && drawing != nil {
+		wsDr = drawing.(*xlsxWsDr)
+	}
+	wsDr.mu.Lock()
+	defer wsDr.mu.Unlock()
+	return wsDr, len(wsDr.OneCellAnchor) + len(wsDr.TwoCellAnchor) + 2, nil
 }
 
 // addDrawingChart provides a function to add chart graphic frame by given
 // sheet, drawingXML, cell, width, height, relationship index and format sets.
-func (f *File) addDrawingChart(sheet, drawingXML, cell string, width, height, rID int, formatSet *formatPicture) error {
+func (f *File) addDrawingChart(sheet, drawingXML, cell string, width, height, rID int, opts *GraphicOptions) error {
 	col, row, err := CellNameToCoordinates(cell)
 	if err != nil {
 		return err
 	}
-	colIdx := col - 1
-	rowIdx := row - 1
-
-	width = int(float64(width) * formatSet.XScale)
-	height = int(float64(height) * formatSet.YScale)
-	colStart, rowStart, colEnd, rowEnd, x2, y2 :=
-		f.positionObjectPixels(sheet, colIdx, rowIdx, formatSet.OffsetX, formatSet.OffsetY, width, height)
-	content, cNvPrID := f.drawingParser(drawingXML)
+	width = int(float64(width) * opts.ScaleX)
+	height = int(float64(height) * opts.ScaleY)
+	colStart, rowStart, colEnd, rowEnd, x2, y2 := f.positionObjectPixels(sheet, col, row, opts.OffsetX, opts.OffsetY, width, height)
+	content, cNvPrID, err := f.drawingParser(drawingXML)
+	if err != nil {
+		return err
+	}
 	twoCellAnchor := xdrCellAnchor{}
-	twoCellAnchor.EditAs = formatSet.Positioning
+	twoCellAnchor.EditAs = opts.Positioning
 	from := xlsxFrom{}
 	from.Col = colStart
-	from.ColOff = formatSet.OffsetX * EMU
+	from.ColOff = opts.OffsetX * EMU
 	from.Row = rowStart
-	from.RowOff = formatSet.OffsetY * EMU
+	from.RowOff = opts.OffsetY * EMU
 	to := xlsxTo{}
 	to.Col = colEnd
 	to.ColOff = x2 * EMU
@@ -1228,23 +1430,26 @@ func (f *File) addDrawingChart(sheet, drawingXML, cell string, width, height, rI
 	graphic, _ := xml.Marshal(graphicFrame)
 	twoCellAnchor.GraphicFrame = string(graphic)
 	twoCellAnchor.ClientData = &xdrClientData{
-		FLocksWithSheet:  formatSet.FLocksWithSheet,
-		FPrintsWithSheet: formatSet.FPrintsWithSheet,
+		FLocksWithSheet:  *opts.Locked,
+		FPrintsWithSheet: *opts.PrintObject,
 	}
 	content.TwoCellAnchor = append(content.TwoCellAnchor, &twoCellAnchor)
-	f.Drawings[drawingXML] = content
+	f.Drawings.Store(drawingXML, content)
 	return err
 }
 
 // addSheetDrawingChart provides a function to add chart graphic frame for
 // chartsheet by given sheet, drawingXML, width, height, relationship index
 // and format sets.
-func (f *File) addSheetDrawingChart(drawingXML string, rID int, formatSet *formatPicture) {
-	content, cNvPrID := f.drawingParser(drawingXML)
+func (f *File) addSheetDrawingChart(drawingXML string, rID int, opts *GraphicOptions) error {
+	content, cNvPrID, err := f.drawingParser(drawingXML)
+	if err != nil {
+		return err
+	}
 	absoluteAnchor := xdrCellAnchor{
-		EditAs: formatSet.Positioning,
+		EditAs: opts.Positioning,
 		Pos:    &xlsxPoint2D{},
-		Ext:    &xlsxExt{},
+		Ext:    &aExt{},
 	}
 
 	graphicFrame := xlsxGraphicFrame{
@@ -1268,51 +1473,126 @@ func (f *File) addSheetDrawingChart(drawingXML string, rID int, formatSet *forma
 	graphic, _ := xml.Marshal(graphicFrame)
 	absoluteAnchor.GraphicFrame = string(graphic)
 	absoluteAnchor.ClientData = &xdrClientData{
-		FLocksWithSheet:  formatSet.FLocksWithSheet,
-		FPrintsWithSheet: formatSet.FPrintsWithSheet,
+		FLocksWithSheet:  *opts.Locked,
+		FPrintsWithSheet: *opts.PrintObject,
 	}
 	content.AbsoluteAnchor = append(content.AbsoluteAnchor, &absoluteAnchor)
-	f.Drawings[drawingXML] = content
+	f.Drawings.Store(drawingXML, content)
+	return err
 }
 
-// deleteDrawing provides a function to delete chart graphic frame by given by
+// deleteDrawing provides a function to delete the chart graphic frame and
+// returns deleted embed relationships ID (for unique picture cell anchor) by
 // given coordinates and graphic type.
-func (f *File) deleteDrawing(col, row int, drawingXML, drawingType string) (err error) {
+func (f *File) deleteDrawing(col, row int, drawingXML, drawingType string) ([]string, error) {
 	var (
-		wsDr            *xlsxWsDr
-		deTwoCellAnchor *decodeTwoCellAnchor
+		err            error
+		rID            string
+		delRID, refRID []string
+		rIDMaps        = map[string]int{}
+		wsDr           *xlsxWsDr
+		deCellAnchor   *decodeCellAnchor
 	)
 	xdrCellAnchorFuncs := map[string]func(anchor *xdrCellAnchor) bool{
 		"Chart": func(anchor *xdrCellAnchor) bool { return anchor.Pic == nil },
 		"Pic":   func(anchor *xdrCellAnchor) bool { return anchor.Pic != nil },
 	}
-	decodeTwoCellAnchorFuncs := map[string]func(anchor *decodeTwoCellAnchor) bool{
-		"Chart": func(anchor *decodeTwoCellAnchor) bool { return anchor.Pic == nil },
-		"Pic":   func(anchor *decodeTwoCellAnchor) bool { return anchor.Pic != nil },
+	decodeCellAnchorFuncs := map[string]func(anchor *decodeCellAnchor) bool{
+		"Chart": func(anchor *decodeCellAnchor) bool { return anchor.Pic == nil },
+		"Pic":   func(anchor *decodeCellAnchor) bool { return anchor.Pic != nil },
 	}
-	wsDr, _ = f.drawingParser(drawingXML)
-	for idx := 0; idx < len(wsDr.TwoCellAnchor); idx++ {
-		if err = nil; wsDr.TwoCellAnchor[idx].From != nil && xdrCellAnchorFuncs[drawingType](wsDr.TwoCellAnchor[idx]) {
-			if wsDr.TwoCellAnchor[idx].From.Col == col && wsDr.TwoCellAnchor[idx].From.Row == row {
-				wsDr.TwoCellAnchor = append(wsDr.TwoCellAnchor[:idx], wsDr.TwoCellAnchor[idx+1:]...)
-				idx--
+	onAnchorCell := func(c, r int) bool { return c == col && r == row }
+	if wsDr, _, err = f.drawingParser(drawingXML); err != nil {
+		return delRID, err
+	}
+	deleteCellAnchor := func(ca []*xdrCellAnchor) ([]*xdrCellAnchor, error) {
+		for idx := 0; idx < len(ca); idx++ {
+			if err = nil; ca[idx].From != nil && xdrCellAnchorFuncs[drawingType](ca[idx]) {
+				rID = extractEmbedRID(ca[idx].Pic, nil)
+				rIDMaps[rID]++
+				if onAnchorCell(ca[idx].From.Col, ca[idx].From.Row) {
+					refRID = append(refRID, rID)
+					ca = append(ca[:idx], ca[idx+1:]...)
+					idx--
+					rIDMaps[rID]--
+				}
+				continue
+			}
+			deCellAnchor = new(decodeCellAnchor)
+			if err = f.xmlNewDecoder(strings.NewReader("<decodeCellAnchor>" + ca[idx].GraphicFrame + "</decodeCellAnchor>")).
+				Decode(deCellAnchor); err != nil && err != io.EOF {
+				return ca, err
+			}
+			if err = nil; deCellAnchor.From != nil && decodeCellAnchorFuncs[drawingType](deCellAnchor) {
+				rID = extractEmbedRID(nil, deCellAnchor.Pic)
+				rIDMaps[rID]++
+				if onAnchorCell(deCellAnchor.From.Col, deCellAnchor.From.Row) {
+					refRID = append(refRID, rID)
+					ca = append(ca[:idx], ca[idx+1:]...)
+					idx--
+					rIDMaps[rID]--
+				}
 			}
 		}
+		return ca, err
 	}
-	for idx := 0; idx < len(wsDr.TwoCellAnchor); idx++ {
-		deTwoCellAnchor = new(decodeTwoCellAnchor)
-		if err = f.xmlNewDecoder(strings.NewReader("<decodeTwoCellAnchor>" + wsDr.TwoCellAnchor[idx].GraphicFrame + "</decodeTwoCellAnchor>")).
-			Decode(deTwoCellAnchor); err != nil && err != io.EOF {
-			err = fmt.Errorf("xml decode error: %s", err)
-			return
-		}
-		if err = nil; deTwoCellAnchor.From != nil && decodeTwoCellAnchorFuncs[drawingType](deTwoCellAnchor) {
-			if deTwoCellAnchor.From.Col == col && deTwoCellAnchor.From.Row == row {
-				wsDr.TwoCellAnchor = append(wsDr.TwoCellAnchor[:idx], wsDr.TwoCellAnchor[idx+1:]...)
-				idx--
-			}
+	if wsDr.OneCellAnchor, err = deleteCellAnchor(wsDr.OneCellAnchor); err != nil {
+		return delRID, err
+	}
+	if wsDr.TwoCellAnchor, err = deleteCellAnchor(wsDr.TwoCellAnchor); err != nil {
+		return delRID, err
+	}
+	f.Drawings.Store(drawingXML, wsDr)
+	return getUnusedCellAnchorRID(delRID, refRID, rIDMaps), err
+}
+
+// extractEmbedRID returns embed relationship ID by giving cell anchor.
+func extractEmbedRID(pic *xlsxPic, decodePic *decodePic) string {
+	var rID string
+	if pic != nil {
+		rID = pic.BlipFill.Blip.Embed
+	}
+	if decodePic != nil {
+		rID = decodePic.BlipFill.Blip.Embed
+	}
+	return rID
+}
+
+// getUnusedCellAnchorRID returns relationship ID lists in the cell anchor which
+// for remove.
+func getUnusedCellAnchorRID(delRID, refRID []string, rIDMaps map[string]int) []string {
+	for _, rID := range refRID {
+		if rIDMaps[rID] == 0 && inStrSlice(delRID, rID, false) == -1 {
+			delRID = append(delRID, rID)
 		}
 	}
-	f.Drawings[drawingXML] = wsDr
-	return err
+	return delRID
+}
+
+// deleteDrawingRels provides a function to delete relationships in
+// xl/drawings/_rels/drawings%d.xml.rels by giving drawings relationships path
+// and relationship ID.
+func (f *File) deleteDrawingRels(rels, rID string) {
+	drawingRels, _ := f.relsReader(rels)
+	if drawingRels == nil {
+		drawingRels = &xlsxRelationships{}
+	}
+	drawingRels.mu.Lock()
+	defer drawingRels.mu.Unlock()
+	for k, v := range drawingRels.Relationships {
+		if v.ID == rID {
+			drawingRels.Relationships = append(drawingRels.Relationships[:k], drawingRels.Relationships[k+1:]...)
+		}
+	}
+	f.Relationships.Store(rels, drawingRels)
+}
+
+// genAxID provides a function to generate ID for primary and secondary
+// horizontal or vertical axis.
+func (f *File) genAxID(opts *Chart) []*attrValInt {
+	opts.XAxis.axID, opts.YAxis.axID = 100000000, 100000001
+	if opts.order > 0 && opts.YAxis.Secondary {
+		opts.XAxis.axID, opts.YAxis.axID = 100000003, 100000004
+	}
+	return []*attrValInt{{Val: intPtr(opts.XAxis.axID)}, {Val: intPtr(opts.YAxis.axID)}}
 }
