@@ -3,10 +3,14 @@ package excelize
 import (
 	"bufio"
 	"bytes"
+	"os"
+	"path/filepath"
 	"strings"
+	"sync"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func BenchmarkWrite(b *testing.B) {
@@ -33,16 +37,61 @@ func BenchmarkWrite(b *testing.B) {
 }
 
 func TestWriteTo(t *testing.T) {
-	f := File{}
-	buf := bytes.Buffer{}
-	f.XLSX = make(map[string][]byte)
-	f.XLSX["/d/"] = []byte("s")
-	_, err := f.WriteTo(bufio.NewWriter(&buf))
-	assert.EqualError(t, err, "zip: write to directory")
-	delete(f.XLSX, "/d/")
+	// Test WriteToBuffer err
+	{
+		f, buf := File{Pkg: sync.Map{}}, bytes.Buffer{}
+		f.Pkg.Store("/d/", []byte("s"))
+		_, err := f.WriteTo(bufio.NewWriter(&buf))
+		assert.EqualError(t, err, "zip: write to directory")
+		f.Pkg.Delete("/d/")
+	}
 	// Test file path overflow
-	const maxUint16 = 1<<16 - 1
-	f.XLSX[strings.Repeat("s", maxUint16+1)] = nil
-	_, err = f.WriteTo(bufio.NewWriter(&buf))
-	assert.EqualError(t, err, "zip: FileHeader.Name too long")
+	{
+		f, buf := File{Pkg: sync.Map{}}, bytes.Buffer{}
+		const maxUint16 = 1<<16 - 1
+		f.Pkg.Store(strings.Repeat("s", maxUint16+1), nil)
+		_, err := f.WriteTo(bufio.NewWriter(&buf))
+		assert.EqualError(t, err, "zip: FileHeader.Name too long")
+	}
+	// Test StreamsWriter err
+	{
+		f, buf := File{Pkg: sync.Map{}}, bytes.Buffer{}
+		f.Pkg.Store("s", nil)
+		f.streams = make(map[string]*StreamWriter)
+		file, _ := os.Open("123")
+		f.streams["s"] = &StreamWriter{rawData: bufferedWriter{tmp: file}}
+		_, err := f.WriteTo(bufio.NewWriter(&buf))
+		assert.Nil(t, err)
+	}
+	// Test write with temporary file
+	{
+		f, buf := File{tempFiles: sync.Map{}}, bytes.Buffer{}
+		const maxUint16 = 1<<16 - 1
+		f.tempFiles.Store("s", "")
+		f.tempFiles.Store(strings.Repeat("s", maxUint16+1), "")
+		_, err := f.WriteTo(bufio.NewWriter(&buf))
+		assert.EqualError(t, err, "zip: FileHeader.Name too long")
+	}
+	// Test write with unsupported workbook file format
+	{
+		f, buf := File{Pkg: sync.Map{}}, bytes.Buffer{}
+		f.Pkg.Store("/d", []byte("s"))
+		f.Path = "Book1.xls"
+		_, err := f.WriteTo(bufio.NewWriter(&buf))
+		assert.EqualError(t, err, ErrWorkbookFileFormat.Error())
+	}
+	// Test write with unsupported charset content types.
+	{
+		f, buf := NewFile(), bytes.Buffer{}
+		f.ContentTypes, f.Path = nil, filepath.Join("test", "TestWriteTo.xlsx")
+		f.Pkg.Store(defaultXMLPathContentTypes, MacintoshCyrillicCharset)
+		_, err := f.WriteTo(bufio.NewWriter(&buf))
+		assert.EqualError(t, err, "XML syntax error on line 1: invalid UTF-8")
+	}
+}
+
+func TestClose(t *testing.T) {
+	f := NewFile()
+	f.tempFiles.Store("/d/", "/d/")
+	require.Error(t, f.Close())
 }
