@@ -1,76 +1,134 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
 import (
 	"bytes"
 	"encoding/xml"
-	"fmt"
 	"io"
-	"log"
 	"math"
+	"os"
 	"strconv"
+	"strings"
 
-	"github.com/mohae/deepcopy"
+	"github.com/tiendc/go-deepcopy"
 )
 
-// GetRows return all the rows in a sheet by given worksheet name (case
-// sensitive). For example:
+// duplicateHelperFunc defines functions to duplicate helper.
+var duplicateHelperFunc = [3]func(*File, *xlsxWorksheet, string, int, int) error{
+	func(f *File, ws *xlsxWorksheet, sheet string, row, row2 int) error {
+		return f.duplicateConditionalFormat(ws, sheet, row, row2)
+	},
+	func(f *File, ws *xlsxWorksheet, sheet string, row, row2 int) error {
+		return f.duplicateDataValidations(ws, sheet, row, row2)
+	},
+	func(f *File, ws *xlsxWorksheet, sheet string, row, row2 int) error {
+		return f.duplicateMergeCells(ws, sheet, row, row2)
+	},
+}
+
+// GetRows return all the rows in a sheet by given worksheet name, returned as
+// a two-dimensional array, where the value of the cell is converted to the
+// string type. If the cell format can be applied to the value of the cell,
+// the applied value will be used, otherwise the original value will be used.
+// GetRows fetched the rows with value or formula cells, the continually blank
+// cells in the tail of each row will be skipped, so the length of each row
+// may be inconsistent.
 //
-//    rows, err := f.GetRows("Sheet1")
-//    if err != nil {
-//        fmt.Println(err)
-//        return
-//    }
-//    for _, row := range rows {
-//        for _, colCell := range row {
-//            fmt.Print(colCell, "\t")
-//        }
-//        fmt.Println()
-//    }
+// For example, get and traverse the value of all cells by rows on a worksheet
+// named 'Sheet1':
 //
-func (f *File) GetRows(sheet string) ([][]string, error) {
+//	rows, err := f.GetRows("Sheet1")
+//	if err != nil {
+//	    fmt.Println(err)
+//	    return
+//	}
+//	for _, row := range rows {
+//	    for _, colCell := range row {
+//	        fmt.Print(colCell, "\t")
+//	    }
+//	    fmt.Println()
+//	}
+func (f *File) GetRows(sheet string, opts ...Options) ([][]string, error) {
 	rows, err := f.Rows(sheet)
 	if err != nil {
 		return nil, err
 	}
-	results, cur, max := make([][]string, 0, 64), 0, 0
+	results, cur, maxVal := make([][]string, 0, 64), 0, 0
 	for rows.Next() {
 		cur++
-		row, err := rows.Columns()
+		row, err := rows.Columns(opts...)
 		if err != nil {
 			break
 		}
-		results = append(results, row)
 		if len(row) > 0 {
-			max = cur
+			if emptyRows := cur - maxVal - 1; emptyRows > 0 {
+				results = append(results, make([][]string, emptyRows)...)
+			}
+			results = append(results, row)
+			maxVal = cur
 		}
 	}
-	return results[:max], nil
+	return results[:maxVal], rows.Close()
 }
 
 // Rows defines an iterator to a sheet.
 type Rows struct {
-	err                        error
-	curRow, totalRow, stashRow int
-	sheet                      string
-	f                          *File
-	decoder                    *xml.Decoder
+	err                     error
+	curRow, seekRow         int
+	needClose, rawCellValue bool
+	sheet                   string
+	f                       *File
+	tempFile                *os.File
+	sst                     *xlsxSST
+	decoder                 *xml.Decoder
+	token                   xml.Token
+	curRowOpts, seekRowOpts RowOpts
 }
 
-// Next will return true if find the next row element.
+// Next will return true if it finds the next row element.
 func (rows *Rows) Next() bool {
-	rows.curRow++
-	return rows.curRow <= rows.totalRow
+	rows.seekRow++
+	if rows.curRow >= rows.seekRow {
+		rows.curRowOpts = rows.seekRowOpts
+		return true
+	}
+	for {
+		token, _ := rows.decoder.Token()
+		if token == nil {
+			return false
+		}
+		switch xmlElement := token.(type) {
+		case xml.StartElement:
+			if xmlElement.Name.Local == "row" {
+				rows.curRow++
+				if rowNum, _ := attrValToInt("r", xmlElement.Attr); rowNum != 0 {
+					rows.curRow = rowNum
+				}
+				rows.token = token
+				rows.curRowOpts = extractRowOpts(xmlElement.Attr)
+				return true
+			}
+		case xml.EndElement:
+			if xmlElement.Name.Local == "sheetData" {
+				return false
+			}
+		}
+	}
+}
+
+// GetRowOpts will return the RowOpts of the current row.
+func (rows *Rows) GetRowOpts() RowOpts {
+	return rows.curRowOpts
 }
 
 // Error will return the error when the error occurs.
@@ -78,50 +136,78 @@ func (rows *Rows) Error() error {
 	return rows.err
 }
 
-// Columns return the current row's column values.
-func (rows *Rows) Columns() ([]string, error) {
-	var rowIterator rowXMLIterator
-	if rows.stashRow >= rows.curRow {
-		return rowIterator.columns, rowIterator.err
+// Close closes the open worksheet XML file in the system temporary
+// directory.
+func (rows *Rows) Close() error {
+	if rows.tempFile != nil {
+		return rows.tempFile.Close()
 	}
-	rowIterator.rows = rows
-	rowIterator.d = rows.f.sharedStringsReader()
+	return nil
+}
+
+// Columns return the current row's column values. This fetches the worksheet
+// data as a stream, returns each cell in a row as is, and will not skip empty
+// rows in the tail of the worksheet.
+func (rows *Rows) Columns(opts ...Options) ([]string, error) {
+	if rows.curRow > rows.seekRow {
+		return nil, nil
+	}
+	var rowIterator rowXMLIterator
+	var token xml.Token
+	rows.rawCellValue = rows.f.getOptions(opts...).RawCellValue
+	if rows.sst, rowIterator.err = rows.f.sharedStringsReader(); rowIterator.err != nil {
+		return rowIterator.cells, rowIterator.err
+	}
 	for {
-		token, _ := rows.decoder.Token()
-		if token == nil {
+		if rows.token != nil {
+			token = rows.token
+		} else if token, _ = rows.decoder.Token(); token == nil {
 			break
 		}
 		switch xmlElement := token.(type) {
 		case xml.StartElement:
 			rowIterator.inElement = xmlElement.Name.Local
 			if rowIterator.inElement == "row" {
-				rowIterator.row++
-				if rowIterator.attrR, rowIterator.err = attrValToInt("r", xmlElement.Attr); rowIterator.attrR != 0 {
-					rowIterator.row = rowIterator.attrR
+				rowNum := 0
+				if rowNum, rowIterator.err = attrValToInt("r", xmlElement.Attr); rowNum != 0 {
+					rows.curRow = rowNum
+				} else if rows.token == nil {
+					rows.curRow++
 				}
-				if rowIterator.row > rowIterator.rows.curRow {
-					rowIterator.rows.stashRow = rowIterator.row - 1
-					return rowIterator.columns, rowIterator.err
+				rows.token = token
+				rows.seekRowOpts = extractRowOpts(xmlElement.Attr)
+				if rows.curRow > rows.seekRow {
+					rows.token = nil
+					return rowIterator.cells, rowIterator.err
 				}
 			}
-			rowXMLHandler(&rowIterator, &xmlElement)
-			if rowIterator.err != nil {
-				return rowIterator.columns, rowIterator.err
+			if rows.rowXMLHandler(&rowIterator, &xmlElement, rows.rawCellValue); rowIterator.err != nil {
+				rows.token = nil
+				return rowIterator.cells, rowIterator.err
 			}
+			rows.token = nil
 		case xml.EndElement:
-			rowIterator.inElement = xmlElement.Name.Local
-			if rowIterator.row == 0 {
-				rowIterator.row = rowIterator.rows.curRow
-			}
-			if rowIterator.inElement == "row" && rowIterator.row+1 < rowIterator.rows.curRow {
-				return rowIterator.columns, rowIterator.err
-			}
-			if rowIterator.inElement == "sheetData" {
-				return rowIterator.columns, rowIterator.err
+			if xmlElement.Name.Local == "sheetData" {
+				return rowIterator.cells, rowIterator.err
 			}
 		}
 	}
-	return rowIterator.columns, rowIterator.err
+	return rowIterator.cells, rowIterator.err
+}
+
+// extractRowOpts extract row element attributes.
+func extractRowOpts(attrs []xml.Attr) RowOpts {
+	rowOpts := RowOpts{Height: defaultRowHeight}
+	if styleID, err := attrValToInt("s", attrs); err == nil && styleID > 0 && styleID < MaxCellStyles {
+		rowOpts.StyleID = styleID
+	}
+	if hidden, err := attrValToBool("hidden", attrs); err == nil {
+		rowOpts.Hidden = hidden
+	}
+	if height, err := attrValToFloat("ht", attrs); err == nil {
+		rowOpts.Height = height
+	}
+	return rowOpts
 }
 
 // appendSpace append blank characters to slice by given length and source slice.
@@ -132,81 +218,103 @@ func appendSpace(l int, s []string) []string {
 	return s
 }
 
-// ErrSheetNotExist defines an error of sheet is not exist
-type ErrSheetNotExist struct {
-	SheetName string
-}
-
-func (err ErrSheetNotExist) Error() string {
-	return fmt.Sprintf("sheet %s is not exist", string(err.SheetName))
-}
-
 // rowXMLIterator defined runtime use field for the worksheet row SAX parser.
 type rowXMLIterator struct {
-	err                 error
-	inElement           string
-	attrR, cellCol, row int
-	columns             []string
-	rows                *Rows
-	d                   *xlsxSST
+	err              error
+	inElement        string
+	cellCol, cellRow int
+	cells            []string
 }
 
 // rowXMLHandler parse the row XML element of the worksheet.
-func rowXMLHandler(rowIterator *rowXMLIterator, xmlElement *xml.StartElement) {
-	rowIterator.err = nil
+func (rows *Rows) rowXMLHandler(rowIterator *rowXMLIterator, xmlElement *xml.StartElement, raw bool) {
 	if rowIterator.inElement == "c" {
 		rowIterator.cellCol++
 		colCell := xlsxC{}
-		_ = rowIterator.rows.decoder.DecodeElement(&colCell, xmlElement)
+		_ = rows.decoder.DecodeElement(&colCell, xmlElement)
 		if colCell.R != "" {
 			if rowIterator.cellCol, _, rowIterator.err = CellNameToCoordinates(colCell.R); rowIterator.err != nil {
 				return
 			}
 		}
-		blank := rowIterator.cellCol - len(rowIterator.columns)
-		val, _ := colCell.getValueFrom(rowIterator.rows.f, rowIterator.d)
-		if val != "" || colCell.F != nil {
-			rowIterator.columns = append(appendSpace(blank, rowIterator.columns), val)
+		blank := rowIterator.cellCol - len(rowIterator.cells)
+		if val, _ := colCell.getValueFrom(rows.f, rows.sst, raw); val != "" || colCell.F != nil {
+			rowIterator.cells = append(appendSpace(blank, rowIterator.cells), val)
 		}
 	}
 }
 
 // Rows returns a rows iterator, used for streaming reading data for a
-// worksheet with a large data. For example:
+// worksheet with a large data. This function is concurrency safe. For
+// example:
 //
-//    rows, err := f.Rows("Sheet1")
-//    if err != nil {
-//        fmt.Println(err)
-//        return
-//    }
-//    for rows.Next() {
-//        row, err := rows.Columns()
-//        if err != nil {
-//            fmt.Println(err)
-//        }
-//        for _, colCell := range row {
-//            fmt.Print(colCell, "\t")
-//        }
-//        fmt.Println()
-//    }
-//
+//	rows, err := f.Rows("Sheet1")
+//	if err != nil {
+//	    fmt.Println(err)
+//	    return
+//	}
+//	for rows.Next() {
+//	    row, err := rows.Columns()
+//	    if err != nil {
+//	        fmt.Println(err)
+//	    }
+//	    for _, colCell := range row {
+//	        fmt.Print(colCell, "\t")
+//	    }
+//	    fmt.Println()
+//	}
+//	if err = rows.Close(); err != nil {
+//	    fmt.Println(err)
+//	}
 func (f *File) Rows(sheet string) (*Rows, error) {
-	name, ok := f.sheetMap[trimSheetName(sheet)]
+	if err := checkSheetName(sheet); err != nil {
+		return nil, err
+	}
+	name, ok := f.getSheetXMLPath(sheet)
 	if !ok {
 		return nil, ErrSheetNotExist{sheet}
 	}
-	if f.Sheet[name] != nil {
-		// flush data
-		output, _ := xml.Marshal(f.Sheet[name])
+	if worksheet, ok := f.Sheet.Load(name); ok && worksheet != nil {
+		ws := worksheet.(*xlsxWorksheet)
+		ws.mu.Lock()
+		defer ws.mu.Unlock()
+		// Flush data
+		output, _ := xml.Marshal(ws)
 		f.saveFileList(name, f.replaceNameSpaceBytes(name, output))
 	}
+	var err error
+	rows := Rows{f: f, sheet: name}
+	rows.needClose, rows.decoder, rows.tempFile, err = f.xmlDecoder(name)
+	return &rows, err
+}
+
+// getFromStringItem build shared string item offset list from system temporary
+// file at one time, and return value by given to string index.
+func (f *File) getFromStringItem(index int) string {
+	if f.sharedStringTemp != nil {
+		if len(f.sharedStringItem) <= index {
+			return strconv.Itoa(index)
+		}
+		offsetRange := f.sharedStringItem[index]
+		buf := make([]byte, offsetRange[1]-offsetRange[0])
+		if _, err := f.sharedStringTemp.ReadAt(buf, int64(offsetRange[0])); err != nil {
+			return strconv.Itoa(index)
+		}
+		return string(buf)
+	}
+	needClose, decoder, tempFile, err := f.xmlDecoder(defaultXMLPathSharedStrings)
+	if needClose && err == nil {
+		defer func() {
+			err = tempFile.Close()
+		}()
+	}
+	f.sharedStringItem = [][]uint{}
+	f.sharedStringTemp, _ = os.CreateTemp(os.TempDir(), "excelize-")
+	f.tempFiles.Store(defaultTempFileSST, f.sharedStringTemp.Name())
 	var (
-		err       error
 		inElement string
-		row       int
-		rows      Rows
+		i, offset uint
 	)
-	decoder := f.xmlNewDecoder(bytes.NewReader(f.readXML(name)))
 	for {
 		token, _ := decoder.Token()
 		if token == nil {
@@ -215,36 +323,42 @@ func (f *File) Rows(sheet string) (*Rows, error) {
 		switch xmlElement := token.(type) {
 		case xml.StartElement:
 			inElement = xmlElement.Name.Local
-			if inElement == "row" {
-				row++
-				for _, attr := range xmlElement.Attr {
-					if attr.Name.Local == "r" {
-						row, err = strconv.Atoi(attr.Value)
-						if err != nil {
-							return &rows, err
-						}
-					}
-				}
-				rows.totalRow = row
+			if inElement == "si" {
+				si := xlsxSI{}
+				_ = decoder.DecodeElement(&si, &xmlElement)
+
+				startIdx := offset
+				n, _ := f.sharedStringTemp.WriteString(si.String())
+				offset += uint(n)
+				f.sharedStringItem = append(f.sharedStringItem, []uint{startIdx, offset})
+				i++
 			}
-		case xml.EndElement:
-			if xmlElement.Name.Local == "sheetData" {
-				rows.f = f
-				rows.sheet = name
-				rows.decoder = f.xmlNewDecoder(bytes.NewReader(f.readXML(name)))
-				return &rows, nil
-			}
-		default:
 		}
 	}
-	return &rows, nil
+	return f.getFromStringItem(index)
 }
 
-// SetRowHeight provides a function to set the height of a single row. For
-// example, set the height of the first row in Sheet1:
+// xmlDecoder creates XML decoder by given path in the zip from memory data
+// or system temporary file.
+func (f *File) xmlDecoder(name string) (bool, *xml.Decoder, *os.File, error) {
+	var (
+		content  []byte
+		err      error
+		tempFile *os.File
+	)
+	if content = f.readXML(name); len(content) > 0 {
+		return false, f.xmlNewDecoder(bytes.NewReader(content)), tempFile, err
+	}
+	tempFile, err = f.readTemp(name)
+	return true, f.xmlNewDecoder(tempFile), tempFile, err
+}
+
+// SetRowHeight provides a function to set the height of a single row. If the
+// value of height is 0, will hide the specified row, if the value of height is
+// -1, will unset the custom row height. For example, set the height of the
+// first row in Sheet1:
 //
-//    err := f.SetRowHeight("Sheet1", 1, 50)
-//
+//	err := f.SetRowHeight("Sheet1", 1, 50)
 func (f *File) SetRowHeight(sheet string, row int, height float64) error {
 	if row < 1 {
 		return newInvalidRowNumberError(row)
@@ -252,43 +366,55 @@ func (f *File) SetRowHeight(sheet string, row int, height float64) error {
 	if height > MaxRowHeight {
 		return ErrMaxRowHeight
 	}
+	if height < -1 {
+		return ErrParameterInvalid
+	}
 	ws, err := f.workSheetReader(sheet)
 	if err != nil {
 		return err
 	}
 
-	prepareSheetXML(ws, 0, row)
+	ws.prepareSheetXML(0, row)
 
 	rowIdx := row - 1
-	ws.SheetData.Row[rowIdx].Ht = height
+	if height == -1 {
+		ws.SheetData.Row[rowIdx].Ht = nil
+		ws.SheetData.Row[rowIdx].CustomHeight = false
+		return err
+	}
+	ws.SheetData.Row[rowIdx].Ht = float64Ptr(height)
 	ws.SheetData.Row[rowIdx].CustomHeight = true
-	return nil
+	return err
 }
 
 // getRowHeight provides a function to get row height in pixels by given sheet
-// name and row index.
+// name and row number.
 func (f *File) getRowHeight(sheet string, row int) int {
 	ws, _ := f.workSheetReader(sheet)
+	ws.mu.Lock()
+	defer ws.mu.Unlock()
 	for i := range ws.SheetData.Row {
 		v := &ws.SheetData.Row[i]
-		if v.R == row+1 && v.Ht != 0 {
-			return int(convertRowHeightToPixels(v.Ht))
+		if v.R == row && v.Ht != nil {
+			return int(convertRowHeightToPixels(*v.Ht))
 		}
 	}
-	// Optimisation for when the row heights haven't changed.
+	if ws.SheetFormatPr != nil && ws.SheetFormatPr.DefaultRowHeight > 0 {
+		return int(convertRowHeightToPixels(ws.SheetFormatPr.DefaultRowHeight))
+	}
+	// Optimization for when the row heights haven't changed.
 	return int(defaultRowHeightPixels)
 }
 
 // GetRowHeight provides a function to get row height by given worksheet name
-// and row index. For example, get the height of the first row in Sheet1:
+// and row number. For example, get the height of the first row in Sheet1:
 //
-//    height, err := f.GetRowHeight("Sheet1", 1)
-//
+//	height, err := f.GetRowHeight("Sheet1", 1)
 func (f *File) GetRowHeight(sheet string, row int) (float64, error) {
 	if row < 1 {
-		return defaultRowHeightPixels, newInvalidRowNumberError(row)
+		return defaultRowHeight, newInvalidRowNumberError(row)
 	}
-	var ht = defaultRowHeight
+	ht := defaultRowHeight
 	ws, err := f.workSheetReader(sheet)
 	if err != nil {
 		return ht, err
@@ -300,27 +426,27 @@ func (f *File) GetRowHeight(sheet string, row int) (float64, error) {
 		return ht, nil // it will be better to use 0, but we take care with BC
 	}
 	for _, v := range ws.SheetData.Row {
-		if v.R == row && v.Ht != 0 {
-			return v.Ht, nil
+		if v.R == row && v.Ht != nil {
+			return *v.Ht, nil
 		}
 	}
-	// Optimisation for when the row heights haven't changed.
+	// Optimization for when the row heights haven't changed.
 	return ht, nil
 }
 
 // sharedStringsReader provides a function to get the pointer to the structure
 // after deserialization of xl/sharedStrings.xml.
-func (f *File) sharedStringsReader() *xlsxSST {
+func (f *File) sharedStringsReader() (*xlsxSST, error) {
 	var err error
-	f.Lock()
-	defer f.Unlock()
+	f.mu.Lock()
+	defer f.mu.Unlock()
 	relPath := f.getWorkbookRelsPath()
 	if f.SharedStrings == nil {
 		var sharedStrings xlsxSST
-		ss := f.readXML("xl/sharedStrings.xml")
+		ss := f.readXML(defaultXMLPathSharedStrings)
 		if err = f.xmlNewDecoder(bytes.NewReader(namespaceStrictToTransitional(ss))).
 			Decode(&sharedStrings); err != nil && err != io.EOF {
-			log.Printf("xml decode error: %s", err)
+			return f.SharedStrings, err
 		}
 		if sharedStrings.Count == 0 {
 			sharedStrings.Count = len(sharedStrings.SI)
@@ -334,70 +460,29 @@ func (f *File) sharedStringsReader() *xlsxSST {
 				f.sharedStringsMap[sharedStrings.SI[i].T.Val] = i
 			}
 		}
-		f.addContentTypePart(0, "sharedStrings")
-		rels := f.relsReader(relPath)
+		if err = f.addContentTypePart(0, "sharedStrings"); err != nil {
+			return f.SharedStrings, err
+		}
+		rels, err := f.relsReader(relPath)
+		if err != nil {
+			return f.SharedStrings, err
+		}
 		for _, rel := range rels.Relationships {
 			if rel.Target == "/xl/sharedStrings.xml" {
-				return f.SharedStrings
+				return f.SharedStrings, nil
 			}
 		}
 		// Update workbook.xml.rels
 		f.addRels(relPath, SourceRelationshipSharedStrings, "/xl/sharedStrings.xml", "")
 	}
 
-	return f.SharedStrings
-}
-
-// getValueFrom return a value from a column/row cell, this function is
-// inteded to be used with for range on rows an argument with the spreadsheet
-// opened file.
-func (c *xlsxC) getValueFrom(f *File, d *xlsxSST) (string, error) {
-	f.Lock()
-	defer f.Unlock()
-	switch c.T {
-	case "s":
-		if c.V != "" {
-			xlsxSI := 0
-			xlsxSI, _ = strconv.Atoi(c.V)
-			if len(d.SI) > xlsxSI {
-				return f.formattedValue(c.S, d.SI[xlsxSI].String()), nil
-			}
-		}
-		return f.formattedValue(c.S, c.V), nil
-	case "str":
-		return f.formattedValue(c.S, c.V), nil
-	case "inlineStr":
-		if c.IS != nil {
-			return f.formattedValue(c.S, c.IS.String()), nil
-		}
-		return f.formattedValue(c.S, c.V), nil
-	default:
-		isNum, precision := isNumeric(c.V)
-		if isNum && precision > 15 {
-			val, _ := roundPrecision(c.V)
-			if val != c.V {
-				return f.formattedValue(c.S, val), nil
-			}
-		}
-		return f.formattedValue(c.S, c.V), nil
-	}
-}
-
-// roundPrecision round precision for numeric.
-func roundPrecision(value string) (result string, err error) {
-	var num float64
-	if num, err = strconv.ParseFloat(value, 64); err != nil {
-		return
-	}
-	result = fmt.Sprintf("%g", math.Round(num*numericPrecision)/numericPrecision)
-	return
+	return f.SharedStrings, nil
 }
 
 // SetRowVisible provides a function to set visible of a single row by given
 // worksheet name and Excel row number. For example, hide row 2 in Sheet1:
 //
-//    err := f.SetRowVisible("Sheet1", 2, false)
-//
+//	err := f.SetRowVisible("Sheet1", 2, false)
 func (f *File) SetRowVisible(sheet string, row int, visible bool) error {
 	if row < 1 {
 		return newInvalidRowNumberError(row)
@@ -407,7 +492,7 @@ func (f *File) SetRowVisible(sheet string, row int, visible bool) error {
 	if err != nil {
 		return err
 	}
-	prepareSheetXML(ws, 0, row)
+	ws.prepareSheetXML(0, row)
 	ws.SheetData.Row[row-1].Hidden = !visible
 	return nil
 }
@@ -416,8 +501,7 @@ func (f *File) SetRowVisible(sheet string, row int, visible bool) error {
 // worksheet name and Excel row number. For example, get visible state of row
 // 2 in Sheet1:
 //
-//    visible, err := f.GetRowVisible("Sheet1", 2)
-//
+//	visible, err := f.GetRowVisible("Sheet1", 2)
 func (f *File) GetRowVisible(sheet string, row int) (bool, error) {
 	if row < 1 {
 		return false, newInvalidRowNumberError(row)
@@ -437,8 +521,7 @@ func (f *File) GetRowVisible(sheet string, row int) (bool, error) {
 // single row by given worksheet name and Excel row number. The value of
 // parameter 'level' is 1-7. For example, outline row 2 in Sheet1 to level 1:
 //
-//    err := f.SetRowOutlineLevel("Sheet1", 2, 1)
-//
+//	err := f.SetRowOutlineLevel("Sheet1", 2, 1)
 func (f *File) SetRowOutlineLevel(sheet string, row int, level uint8) error {
 	if row < 1 {
 		return newInvalidRowNumberError(row)
@@ -450,7 +533,7 @@ func (f *File) SetRowOutlineLevel(sheet string, row int, level uint8) error {
 	if err != nil {
 		return err
 	}
-	prepareSheetXML(ws, 0, row)
+	ws.prepareSheetXML(0, row)
 	ws.SheetData.Row[row-1].OutlineLevel = level
 	return nil
 }
@@ -459,8 +542,7 @@ func (f *File) SetRowOutlineLevel(sheet string, row int, level uint8) error {
 // single row by given worksheet name and Excel row number. For example, get
 // outline number of row 2 in Sheet1:
 //
-//    level, err := f.GetRowOutlineLevel("Sheet1", 2)
-//
+//	level, err := f.GetRowOutlineLevel("Sheet1", 2)
 func (f *File) GetRowOutlineLevel(sheet string, row int) (uint8, error) {
 	if row < 1 {
 		return 0, newInvalidRowNumberError(row)
@@ -478,7 +560,7 @@ func (f *File) GetRowOutlineLevel(sheet string, row int) (uint8, error) {
 // RemoveRow provides a function to remove single row by given worksheet name
 // and Excel row number. For example, remove row 3 in Sheet1:
 //
-//    err := f.RemoveRow("Sheet1", 3)
+//	err := f.RemoveRow("Sheet1", 3)
 //
 // Use this method with caution, which will affect changes in references such
 // as formulas, charts, and so on. If there is any referenced value of the
@@ -508,26 +590,32 @@ func (f *File) RemoveRow(sheet string, row int) error {
 	return f.adjustHelper(sheet, rows, row, -1)
 }
 
-// InsertRow provides a function to insert a new row after given Excel row
-// number starting from 1. For example, create a new row before row 3 in
-// Sheet1:
+// InsertRows provides a function to insert new rows after the given Excel row
+// number starting from 1 and number of rows. For example, create two rows
+// before row 3 in Sheet1:
 //
-//    err := f.InsertRow("Sheet1", 3)
+//	err := f.InsertRows("Sheet1", 3, 2)
 //
 // Use this method with caution, which will affect changes in references such
 // as formulas, charts, and so on. If there is any referenced value of the
 // worksheet, it will cause a file error when you open it. The excelize only
 // partially updates these references currently.
-func (f *File) InsertRow(sheet string, row int) error {
+func (f *File) InsertRows(sheet string, row, n int) error {
 	if row < 1 {
 		return newInvalidRowNumberError(row)
 	}
-	return f.adjustHelper(sheet, rows, row, 1)
+	if row >= TotalRows || n >= TotalRows {
+		return ErrMaxRows
+	}
+	if n < 1 {
+		return ErrParameterInvalid
+	}
+	return f.adjustHelper(sheet, rows, row, n)
 }
 
 // DuplicateRow inserts a copy of specified row (by its Excel row number) below
 //
-//    err := f.DuplicateRow("Sheet1", 2)
+//	err := f.DuplicateRow("Sheet1", 2)
 //
 // Use this method with caution, which will affect changes in references such
 // as formulas, charts, and so on. If there is any referenced value of the
@@ -540,7 +628,7 @@ func (f *File) DuplicateRow(sheet string, row int) error {
 // DuplicateRowTo inserts a copy of specified row by it Excel number
 // to specified row position moving down exists rows after target position
 //
-//    err := f.DuplicateRowTo("Sheet1", 2, 7)
+//	err := f.DuplicateRowTo("Sheet1", 2, 7)
 //
 // Use this method with caution, which will affect changes in references such
 // as formulas, charts, and so on. If there is any referenced value of the
@@ -555,8 +643,9 @@ func (f *File) DuplicateRowTo(sheet string, row, row2 int) error {
 	if err != nil {
 		return err
 	}
-	if row > len(ws.SheetData.Row) || row2 < 1 || row == row2 {
-		return nil
+
+	if row2 < 1 || row == row2 {
+		return err
 	}
 
 	var ok bool
@@ -564,16 +653,17 @@ func (f *File) DuplicateRowTo(sheet string, row, row2 int) error {
 
 	for i, r := range ws.SheetData.Row {
 		if r.R == row {
-			rowCopy = deepcopy.Copy(ws.SheetData.Row[i]).(xlsxRow)
+			deepcopy.Copy(&rowCopy, ws.SheetData.Row[i])
 			ok = true
 			break
 		}
 	}
-	if !ok {
-		return nil
-	}
 
 	if err := f.adjustHelper(sheet, rows, row2, 1); err != nil {
+		return err
+	}
+
+	if !ok {
 		return err
 	}
 
@@ -584,24 +674,106 @@ func (f *File) DuplicateRowTo(sheet string, row, row2 int) error {
 			break
 		}
 	}
-	if idx2 == -1 && len(ws.SheetData.Row) >= row2 {
-		return nil
-	}
-
 	rowCopy.C = append(make([]xlsxC, 0, len(rowCopy.C)), rowCopy.C...)
-	f.ajustSingleRowDimensions(&rowCopy, row2)
+	rowCopy.adjustSingleRowDimensions(row2 - row)
+	_ = f.adjustSingleRowFormulas(sheet, sheet, &rowCopy, row, row2-row, true)
 
 	if idx2 != -1 {
 		ws.SheetData.Row[idx2] = rowCopy
 	} else {
 		ws.SheetData.Row = append(ws.SheetData.Row, rowCopy)
 	}
-	return f.duplicateMergeCells(sheet, ws, row, row2)
+	for _, fn := range duplicateHelperFunc {
+		if err := fn(f, ws, sheet, row, row2); err != nil {
+			return err
+		}
+	}
+	return err
+}
+
+// duplicateSQRefHelper provides a function to adjust conditional formatting and
+// data validations cell reference when duplicate rows.
+func duplicateSQRefHelper(row, row2 int, ref string) (string, error) {
+	if !strings.Contains(ref, ":") {
+		ref += ":" + ref
+	}
+	abs := strings.Contains(ref, "$")
+	coordinates, err := rangeRefToCoordinates(ref)
+	if err != nil {
+		return "", err
+	}
+	x1, y1, x2, y2 := coordinates[0], coordinates[1], coordinates[2], coordinates[3]
+	if y1 == y2 && y1 == row {
+		if ref, err = coordinatesToRangeRef([]int{x1, row2, x2, row2}, abs); err != nil {
+			return "", err
+		}
+		return ref, err
+	}
+	return "", err
+}
+
+// duplicateConditionalFormat create conditional formatting for the destination
+// row if there are conditional formats in the copied row.
+func (f *File) duplicateConditionalFormat(ws *xlsxWorksheet, sheet string, row, row2 int) error {
+	var cfs []*xlsxConditionalFormatting
+	for _, cf := range ws.ConditionalFormatting {
+		if cf != nil {
+			var SQRef []string
+			for _, ref := range strings.Split(cf.SQRef, " ") {
+				coordinates, err := duplicateSQRefHelper(row, row2, ref)
+				if err != nil {
+					return err
+				}
+				if coordinates != "" {
+					SQRef = append(SQRef, coordinates)
+				}
+			}
+			if len(SQRef) > 0 {
+				var cfCopy xlsxConditionalFormatting
+				deepcopy.Copy(&cfCopy, *cf)
+				cfCopy.SQRef = strings.Join(SQRef, " ")
+				cfs = append(cfs, &cfCopy)
+			}
+		}
+	}
+	ws.ConditionalFormatting = append(ws.ConditionalFormatting, cfs...)
+	return nil
+}
+
+// duplicateDataValidations create data validations for the destination row if
+// there are data validation rules in the copied row.
+func (f *File) duplicateDataValidations(ws *xlsxWorksheet, sheet string, row, row2 int) error {
+	if ws.DataValidations == nil {
+		return nil
+	}
+	var dvs []*xlsxDataValidation
+	for _, dv := range ws.DataValidations.DataValidation {
+		if dv != nil {
+			var SQRef []string
+			for _, ref := range strings.Split(dv.Sqref, " ") {
+				coordinates, err := duplicateSQRefHelper(row, row2, ref)
+				if err != nil {
+					return err
+				}
+				if coordinates != "" {
+					SQRef = append(SQRef, coordinates)
+				}
+			}
+			if len(SQRef) > 0 {
+				var dvCopy xlsxDataValidation
+				deepcopy.Copy(&dvCopy, *dv)
+				dvCopy.Sqref = strings.Join(SQRef, " ")
+				dvs = append(dvs, &dvCopy)
+			}
+		}
+	}
+	ws.DataValidations.DataValidation = append(ws.DataValidations.DataValidation, dvs...)
+	return nil
 }
 
 // duplicateMergeCells merge cells in the destination row if there are single
 // row merged cells in the copied row.
-func (f *File) duplicateMergeCells(sheet string, ws *xlsxWorksheet, row, row2 int) error {
+func (f *File) duplicateMergeCells(ws *xlsxWorksheet, sheet string, row, row2 int) error {
 	if ws.MergeCells == nil {
 		return nil
 	}
@@ -609,7 +781,7 @@ func (f *File) duplicateMergeCells(sheet string, ws *xlsxWorksheet, row, row2 in
 		row++
 	}
 	for _, rng := range ws.MergeCells.Cells {
-		coordinates, err := f.areaRefToCoordinates(rng.Ref)
+		coordinates, err := rangeRefToCoordinates(rng.Ref)
 		if err != nil {
 			return err
 		}
@@ -618,8 +790,8 @@ func (f *File) duplicateMergeCells(sheet string, ws *xlsxWorksheet, row, row2 in
 		}
 	}
 	for i := 0; i < len(ws.MergeCells.Cells); i++ {
-		areaData := ws.MergeCells.Cells[i]
-		coordinates, _ := f.areaRefToCoordinates(areaData.Ref)
+		mergedCells := ws.MergeCells.Cells[i]
+		coordinates, _ := rangeRefToCoordinates(mergedCells.Ref)
 		x1, y1, x2, y2 := coordinates[0], coordinates[1], coordinates[2], coordinates[3]
 		if y1 == y2 && y1 == row {
 			from, _ := CoordinatesToCellName(x1, row2)
@@ -635,28 +807,28 @@ func (f *File) duplicateMergeCells(sheet string, ws *xlsxWorksheet, row, row2 in
 // checkRow provides a function to check and fill each column element for all
 // rows and make that is continuous in a worksheet of XML. For example:
 //
-//    <row r="15" spans="1:22" x14ac:dyDescent="0.2">
-//        <c r="A15" s="2" />
-//        <c r="B15" s="2" />
-//        <c r="F15" s="1" />
-//        <c r="G15" s="1" />
-//    </row>
+//	<row r="15">
+//	    <c r="A15" s="2" />
+//	    <c r="B15" s="2" />
+//	    <c r="F15" s="1" />
+//	    <c r="G15" s="1" />
+//	</row>
 //
 // in this case, we should to change it to
 //
-//    <row r="15" spans="1:22" x14ac:dyDescent="0.2">
-//        <c r="A15" s="2" />
-//        <c r="B15" s="2" />
-//        <c r="C15" s="2" />
-//        <c r="D15" s="2" />
-//        <c r="E15" s="2" />
-//        <c r="F15" s="1" />
-//        <c r="G15" s="1" />
-//    </row>
+//	<row r="15">
+//	    <c r="A15" s="2" />
+//	    <c r="B15" s="2" />
+//	    <c r="C15" s="2" />
+//	    <c r="D15" s="2" />
+//	    <c r="E15" s="2" />
+//	    <c r="F15" s="1" />
+//	    <c r="G15" s="1" />
+//	</row>
 //
-// Noteice: this method could be very slow for large spreadsheets (more than
+// Notice: this method could be very slow for large spreadsheets (more than
 // 3000 rows one sheet).
-func checkRow(ws *xlsxWorksheet) error {
+func (ws *xlsxWorksheet) checkRow() error {
 	for rowIdx := range ws.SheetData.Row {
 		rowData := &ws.SheetData.Row[rowIdx]
 
@@ -686,8 +858,8 @@ func checkRow(ws *xlsxWorksheet) error {
 		}
 
 		if colCount < lastCol {
-			oldList := rowData.C
-			newlist := make([]xlsxC, 0, lastCol)
+			sourceList := rowData.C
+			targetList := make([]xlsxC, 0, lastCol)
 
 			rowData.C = ws.SheetData.Row[rowIdx].C[:0]
 
@@ -696,13 +868,13 @@ func checkRow(ws *xlsxWorksheet) error {
 				if err != nil {
 					return err
 				}
-				newlist = append(newlist, xlsxC{R: cellName})
+				targetList = append(targetList, xlsxC{R: cellName})
 			}
 
-			rowData.C = newlist
+			rowData.C = targetList
 
-			for colIdx := range oldList {
-				colData := &oldList[colIdx]
+			for colIdx := range sourceList {
+				colData := &sourceList[colIdx]
 				colNum, _, err := CellNameToCoordinates(colData.R)
 				if err != nil {
 					return err
@@ -714,14 +886,66 @@ func checkRow(ws *xlsxWorksheet) error {
 	return nil
 }
 
+// hasAttr determine if row non-default attributes.
+func (r *xlsxRow) hasAttr() bool {
+	return r.Spans != "" || r.S != 0 || r.CustomFormat || r.Ht != nil ||
+		r.Hidden || r.CustomHeight || r.OutlineLevel != 0 || r.Collapsed ||
+		r.ThickTop || r.ThickBot || r.Ph
+}
+
+// SetRowStyle provides a function to set the style of rows by given worksheet
+// name, row range, and style ID. Note that this will overwrite the existing
+// styles for the rows, it won't append or merge style with existing styles.
+//
+// For example set style of row 1 on Sheet1:
+//
+//	err := f.SetRowStyle("Sheet1", 1, 1, styleID)
+//
+// Set style of rows 1 to 10 on Sheet1:
+//
+//	err := f.SetRowStyle("Sheet1", 1, 10, styleID)
+func (f *File) SetRowStyle(sheet string, start, end, styleID int) error {
+	if end < start {
+		start, end = end, start
+	}
+	if start < 1 {
+		return newInvalidRowNumberError(start)
+	}
+	if end > TotalRows {
+		return ErrMaxRows
+	}
+	s, err := f.stylesReader()
+	if err != nil {
+		return err
+	}
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	if styleID < 0 || s.CellXfs == nil || len(s.CellXfs.Xf) <= styleID {
+		return newInvalidStyleID(styleID)
+	}
+	ws, err := f.workSheetReader(sheet)
+	if err != nil {
+		return err
+	}
+	ws.prepareSheetXML(0, end)
+	for row := start - 1; row < end; row++ {
+		ws.SheetData.Row[row].S = styleID
+		ws.SheetData.Row[row].CustomFormat = true
+		for i := range ws.SheetData.Row[row].C {
+			if _, rowNum, err := CellNameToCoordinates(ws.SheetData.Row[row].C[i].R); err == nil && rowNum-1 == row {
+				ws.SheetData.Row[row].C[i].S = styleID
+			}
+		}
+	}
+	return nil
+}
+
 // convertRowHeightToPixels provides a function to convert the height of a
 // cell from user's units to pixels. If the height hasn't been set by the user
 // we use the default value. If the row is hidden it has a value of zero.
 func convertRowHeightToPixels(height float64) float64 {
-	var pixels float64
 	if height == 0 {
-		return pixels
+		return 0
 	}
-	pixels = math.Ceil(4.0 / 3.0 * height)
-	return pixels
+	return math.Ceil(4.0 / 3.4 * height)
 }
