@@ -1,182 +1,50 @@
 package excelize
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
 
-var _ = []SheetViewOption{
-	DefaultGridColor(true),
-	RightToLeft(false),
-	ShowFormulas(false),
-	ShowGridLines(true),
-	ShowRowColHeaders(true),
-	TopLeftCell("B2"),
-	// SheetViewOptionPtr are also SheetViewOption
-	new(DefaultGridColor),
-	new(RightToLeft),
-	new(ShowFormulas),
-	new(ShowGridLines),
-	new(ShowRowColHeaders),
-	new(TopLeftCell),
-}
-
-var _ = []SheetViewOptionPtr{
-	(*DefaultGridColor)(nil),
-	(*RightToLeft)(nil),
-	(*ShowFormulas)(nil),
-	(*ShowGridLines)(nil),
-	(*ShowRowColHeaders)(nil),
-	(*TopLeftCell)(nil),
-}
-
-func ExampleFile_SetSheetViewOptions() {
+func TestSetView(t *testing.T) {
 	f := NewFile()
-	const sheet = "Sheet1"
-
-	if err := f.SetSheetViewOptions(sheet, 0,
-		DefaultGridColor(false),
-		RightToLeft(false),
-		ShowFormulas(true),
-		ShowGridLines(true),
-		ShowRowColHeaders(true),
-		ZoomScale(80),
-		TopLeftCell("C3"),
-	); err != nil {
-		fmt.Println(err)
+	assert.NoError(t, f.SetSheetView("Sheet1", -1, nil))
+	ws, ok := f.Sheet.Load("xl/worksheets/sheet1.xml")
+	assert.True(t, ok)
+	ws.(*xlsxWorksheet).SheetViews = nil
+	expected := ViewOptions{
+		DefaultGridColor:  boolPtr(false),
+		RightToLeft:       boolPtr(false),
+		ShowFormulas:      boolPtr(false),
+		ShowGridLines:     boolPtr(false),
+		ShowRowColHeaders: boolPtr(false),
+		ShowRuler:         boolPtr(false),
+		ShowZeros:         boolPtr(false),
+		TopLeftCell:       stringPtr("A1"),
+		View:              stringPtr("normal"),
+		ZoomScale:         float64Ptr(120),
 	}
-
-	var zoomScale ZoomScale
-	fmt.Println("Default:")
-	fmt.Println("- zoomScale: 80")
-
-	if err := f.SetSheetViewOptions(sheet, 0, ZoomScale(500)); err != nil {
-		fmt.Println(err)
-	}
-
-	if err := f.GetSheetViewOptions(sheet, 0, &zoomScale); err != nil {
-		fmt.Println(err)
-	}
-
-	fmt.Println("Used out of range value:")
-	fmt.Println("- zoomScale:", zoomScale)
-
-	if err := f.SetSheetViewOptions(sheet, 0, ZoomScale(123)); err != nil {
-		fmt.Println(err)
-	}
-
-	if err := f.GetSheetViewOptions(sheet, 0, &zoomScale); err != nil {
-		fmt.Println(err)
-	}
-
-	fmt.Println("Used correct value:")
-	fmt.Println("- zoomScale:", zoomScale)
-
-	// Output:
-	// Default:
-	// - zoomScale: 80
-	// Used out of range value:
-	// - zoomScale: 80
-	// Used correct value:
-	// - zoomScale: 123
-
+	assert.NoError(t, f.SetSheetView("Sheet1", 0, &expected))
+	opts, err := f.GetSheetView("Sheet1", 0)
+	assert.NoError(t, err)
+	assert.Equal(t, expected, opts)
+	// Test set sheet view options with invalid view index
+	assert.EqualError(t, f.SetSheetView("Sheet1", 1, nil), "view index 1 out of range")
+	assert.EqualError(t, f.SetSheetView("Sheet1", -2, nil), "view index -2 out of range")
+	// Test set sheet view options on not exists worksheet
+	assert.EqualError(t, f.SetSheetView("SheetN", 0, nil), "sheet SheetN does not exist")
 }
 
-func ExampleFile_GetSheetViewOptions() {
+func TestGetView(t *testing.T) {
 	f := NewFile()
-	const sheet = "Sheet1"
-
-	var (
-		defaultGridColor  DefaultGridColor
-		rightToLeft       RightToLeft
-		showFormulas      ShowFormulas
-		showGridLines     ShowGridLines
-		showZeros         ShowZeros
-		showRowColHeaders ShowRowColHeaders
-		zoomScale         ZoomScale
-		topLeftCell       TopLeftCell
-	)
-
-	if err := f.GetSheetViewOptions(sheet, 0,
-		&defaultGridColor,
-		&rightToLeft,
-		&showFormulas,
-		&showGridLines,
-		&showZeros,
-		&showRowColHeaders,
-		&zoomScale,
-		&topLeftCell,
-	); err != nil {
-		fmt.Println(err)
-	}
-
-	fmt.Println("Default:")
-	fmt.Println("- defaultGridColor:", defaultGridColor)
-	fmt.Println("- rightToLeft:", rightToLeft)
-	fmt.Println("- showFormulas:", showFormulas)
-	fmt.Println("- showGridLines:", showGridLines)
-	fmt.Println("- showZeros:", showZeros)
-	fmt.Println("- showRowColHeaders:", showRowColHeaders)
-	fmt.Println("- zoomScale:", zoomScale)
-	fmt.Println("- topLeftCell:", `"`+topLeftCell+`"`)
-
-	if err := f.SetSheetViewOptions(sheet, 0, TopLeftCell("B2")); err != nil {
-		fmt.Println(err)
-	}
-
-	if err := f.GetSheetViewOptions(sheet, 0, &topLeftCell); err != nil {
-		fmt.Println(err)
-	}
-
-	if err := f.SetSheetViewOptions(sheet, 0, ShowGridLines(false)); err != nil {
-		fmt.Println(err)
-	}
-
-	if err := f.GetSheetViewOptions(sheet, 0, &showGridLines); err != nil {
-		fmt.Println(err)
-	}
-
-	if err := f.SetSheetViewOptions(sheet, 0, ShowZeros(false)); err != nil {
-		fmt.Println(err)
-	}
-
-	if err := f.GetSheetViewOptions(sheet, 0, &showZeros); err != nil {
-		fmt.Println(err)
-	}
-
-	fmt.Println("After change:")
-	fmt.Println("- showGridLines:", showGridLines)
-	fmt.Println("- showZeros:", showZeros)
-	fmt.Println("- topLeftCell:", topLeftCell)
-
-	// Output:
-	// Default:
-	// - defaultGridColor: true
-	// - rightToLeft: false
-	// - showFormulas: false
-	// - showGridLines: true
-	// - showZeros: true
-	// - showRowColHeaders: true
-	// - zoomScale: 0
-	// - topLeftCell: ""
-	// After change:
-	// - showGridLines: false
-	// - showZeros: false
-	// - topLeftCell: B2
-}
-
-func TestSheetViewOptionsErrors(t *testing.T) {
-	f := NewFile()
-	const sheet = "Sheet1"
-
-	assert.NoError(t, f.GetSheetViewOptions(sheet, 0))
-	assert.NoError(t, f.GetSheetViewOptions(sheet, -1))
-	assert.Error(t, f.GetSheetViewOptions(sheet, 1))
-	assert.Error(t, f.GetSheetViewOptions(sheet, -2))
-	assert.NoError(t, f.SetSheetViewOptions(sheet, 0))
-	assert.NoError(t, f.SetSheetViewOptions(sheet, -1))
-	assert.Error(t, f.SetSheetViewOptions(sheet, 1))
-	assert.Error(t, f.SetSheetViewOptions(sheet, -2))
+	_, err := f.getSheetView("SheetN", 0)
+	assert.EqualError(t, err, "sheet SheetN does not exist")
+	// Test get sheet view options with invalid view index
+	_, err = f.GetSheetView("Sheet1", 1)
+	assert.EqualError(t, err, "view index 1 out of range")
+	_, err = f.GetSheetView("Sheet1", -2)
+	assert.EqualError(t, err, "view index -2 out of range")
+	// Test get sheet view options on not exists worksheet
+	_, err = f.GetSheetView("SheetN", 0)
+	assert.EqualError(t, err, "sheet SheetN does not exist")
 }
