@@ -1,13 +1,13 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
@@ -74,7 +74,7 @@ type cTx struct {
 type cRich struct {
 	BodyPr   aBodyPr `xml:"a:bodyPr,omitempty"`
 	LstStyle string  `xml:"a:lstStyle,omitempty"`
-	P        aP      `xml:"a:p"`
+	P        []aP    `xml:"a:p"`
 }
 
 // aBodyPr (Body Properties) directly maps the a:bodyPr element. This element
@@ -171,12 +171,11 @@ type aEa struct {
 	Typeface string `xml:"typeface,attr"`
 }
 
-// aLatin (Latin Font) directly maps the a:latin element. This element
-// specifies that a Latin font be used for a specific run of text. This font is
-// specified with a typeface attribute much like the others but is specifically
-// classified as a Latin font.
-type aLatin struct {
-	Typeface string `xml:"typeface,attr"`
+type xlsxCTTextFont struct {
+	Typeface    string `xml:"typeface,attr"`
+	Panose      string `xml:"panose,attr,omitempty"`
+	PitchFamily string `xml:"pitchFamily,attr,omitempty"`
+	Charset     string `xml:"Charset,attr,omitempty"`
 }
 
 // aR directly maps the a:r element.
@@ -191,29 +190,29 @@ type aR struct {
 // properties are defined as direct formatting, since they are directly applied
 // to the run and supersede any formatting from styles.
 type aRPr struct {
-	AltLang    string      `xml:"altLang,attr,omitempty"`
-	B          bool        `xml:"b,attr"`
-	Baseline   int         `xml:"baseline,attr"`
-	Bmk        string      `xml:"bmk,attr,omitempty"`
-	Cap        string      `xml:"cap,attr,omitempty"`
-	Dirty      bool        `xml:"dirty,attr,omitempty"`
-	Err        bool        `xml:"err,attr,omitempty"`
-	I          bool        `xml:"i,attr"`
-	Kern       int         `xml:"kern,attr"`
-	Kumimoji   bool        `xml:"kumimoji,attr,omitempty"`
-	Lang       string      `xml:"lang,attr,omitempty"`
-	NoProof    bool        `xml:"noProof,attr,omitempty"`
-	NormalizeH bool        `xml:"normalizeH,attr,omitempty"`
-	SmtClean   bool        `xml:"smtClean,attr,omitempty"`
-	SmtID      uint64      `xml:"smtId,attr,omitempty"`
-	Spc        int         `xml:"spc,attr"`
-	Strike     string      `xml:"strike,attr,omitempty"`
-	Sz         float64     `xml:"sz,attr,omitempty"`
-	U          string      `xml:"u,attr,omitempty"`
-	SolidFill  *aSolidFill `xml:"a:solidFill"`
-	Latin      *aLatin     `xml:"a:latin"`
-	Ea         *aEa        `xml:"a:ea"`
-	Cs         *aCs        `xml:"a:cs"`
+	AltLang    string          `xml:"altLang,attr,omitempty"`
+	B          bool            `xml:"b,attr"`
+	Baseline   int             `xml:"baseline,attr"`
+	Bmk        string          `xml:"bmk,attr,omitempty"`
+	Cap        string          `xml:"cap,attr,omitempty"`
+	Dirty      bool            `xml:"dirty,attr,omitempty"`
+	Err        bool            `xml:"err,attr,omitempty"`
+	I          bool            `xml:"i,attr"`
+	Kern       int             `xml:"kern,attr"`
+	Kumimoji   bool            `xml:"kumimoji,attr,omitempty"`
+	Lang       string          `xml:"lang,attr,omitempty"`
+	NoProof    bool            `xml:"noProof,attr,omitempty"`
+	NormalizeH bool            `xml:"normalizeH,attr,omitempty"`
+	SmtClean   bool            `xml:"smtClean,attr,omitempty"`
+	SmtID      uint64          `xml:"smtId,attr,omitempty"`
+	Spc        int             `xml:"spc,attr"`
+	Strike     string          `xml:"strike,attr,omitempty"`
+	Sz         float64         `xml:"sz,attr,omitempty"`
+	U          string          `xml:"u,attr,omitempty"`
+	SolidFill  *aSolidFill     `xml:"a:solidFill"`
+	Latin      *xlsxCTTextFont `xml:"a:latin"`
+	Ea         *aEa            `xml:"a:ea"`
+	Cs         *aCs            `xml:"a:cs"`
 }
 
 // cSpPr (Shape Properties) directly maps the spPr element. This element
@@ -249,13 +248,13 @@ type aContourClr struct {
 // shapes and text. The line allows for the specifying of many different types
 // of outlines including even line dashes and bevels.
 type aLn struct {
-	Algn      string      `xml:"algn,attr,omitempty"`
-	Cap       string      `xml:"cap,attr,omitempty"`
-	Cmpd      string      `xml:"cmpd,attr,omitempty"`
-	W         int         `xml:"w,attr,omitempty"`
-	NoFill    string      `xml:"a:noFill,omitempty"`
-	Round     string      `xml:"a:round,omitempty"`
-	SolidFill *aSolidFill `xml:"a:solidFill"`
+	Algn      string         `xml:"algn,attr,omitempty"`
+	Cap       string         `xml:"cap,attr,omitempty"`
+	Cmpd      string         `xml:"cmpd,attr,omitempty"`
+	W         int            `xml:"w,attr,omitempty"`
+	NoFill    *attrValString `xml:"a:noFill"`
+	Round     string         `xml:"a:round,omitempty"`
+	SolidFill *aSolidFill    `xml:"a:solidFill"`
 }
 
 // cTxPr (Text Properties) directly maps the txPr element. This element
@@ -301,25 +300,26 @@ type cView3D struct {
 // cPlotArea directly maps the plotArea element. This element specifies the
 // plot area of the chart.
 type cPlotArea struct {
-	Layout         *string  `xml:"layout"`
-	AreaChart      *cCharts `xml:"areaChart"`
-	Area3DChart    *cCharts `xml:"area3DChart"`
-	BarChart       *cCharts `xml:"barChart"`
-	Bar3DChart     *cCharts `xml:"bar3DChart"`
-	BubbleChart    *cCharts `xml:"bubbleChart"`
-	DoughnutChart  *cCharts `xml:"doughnutChart"`
-	LineChart      *cCharts `xml:"lineChart"`
-	PieChart       *cCharts `xml:"pieChart"`
-	Pie3DChart     *cCharts `xml:"pie3DChart"`
-	OfPieChart     *cCharts `xml:"ofPieChart"`
-	RadarChart     *cCharts `xml:"radarChart"`
-	ScatterChart   *cCharts `xml:"scatterChart"`
-	Surface3DChart *cCharts `xml:"surface3DChart"`
-	SurfaceChart   *cCharts `xml:"surfaceChart"`
-	CatAx          []*cAxs  `xml:"catAx"`
-	ValAx          []*cAxs  `xml:"valAx"`
-	SerAx          []*cAxs  `xml:"serAx"`
-	SpPr           *cSpPr   `xml:"spPr"`
+	Layout         *string    `xml:"layout"`
+	AreaChart      []*cCharts `xml:"areaChart"`
+	Area3DChart    []*cCharts `xml:"area3DChart"`
+	BarChart       []*cCharts `xml:"barChart"`
+	Bar3DChart     []*cCharts `xml:"bar3DChart"`
+	BubbleChart    []*cCharts `xml:"bubbleChart"`
+	DoughnutChart  []*cCharts `xml:"doughnutChart"`
+	LineChart      []*cCharts `xml:"lineChart"`
+	Line3DChart    []*cCharts `xml:"line3DChart"`
+	PieChart       []*cCharts `xml:"pieChart"`
+	Pie3DChart     []*cCharts `xml:"pie3DChart"`
+	OfPieChart     []*cCharts `xml:"ofPieChart"`
+	RadarChart     []*cCharts `xml:"radarChart"`
+	ScatterChart   []*cCharts `xml:"scatterChart"`
+	Surface3DChart []*cCharts `xml:"surface3DChart"`
+	SurfaceChart   []*cCharts `xml:"surfaceChart"`
+	CatAx          []*cAxs    `xml:"catAx"`
+	ValAx          []*cAxs    `xml:"valAx"`
+	SerAx          []*cAxs    `xml:"serAx"`
+	SpPr           *cSpPr     `xml:"spPr"`
 }
 
 // cCharts specifies the common element of the chart.
@@ -333,8 +333,10 @@ type cCharts struct {
 	VaryColors   *attrValBool   `xml:"varyColors"`
 	Wireframe    *attrValBool   `xml:"wireframe"`
 	Ser          *[]cSer        `xml:"ser"`
+	SplitPos     *attrValInt    `xml:"splitPos"`
 	SerLines     *attrValString `xml:"serLines"`
 	DLbls        *cDLbls        `xml:"dLbls"`
+	GapWidth     *attrValInt    `xml:"gapWidth"`
 	Shape        *attrValString `xml:"shape"`
 	HoleSize     *attrValInt    `xml:"holeSize"`
 	Smooth       *attrValBool   `xml:"smooth"`
@@ -350,6 +352,7 @@ type cAxs struct {
 	AxPos          *attrValString `xml:"axPos"`
 	MajorGridlines *cChartLines   `xml:"majorGridlines"`
 	MinorGridlines *cChartLines   `xml:"minorGridlines"`
+	Title          *cTitle        `xml:"title"`
 	NumFmt         *cNumFmt       `xml:"numFmt"`
 	MajorTickMark  *attrValString `xml:"majorTickMark"`
 	MinorTickMark  *attrValString `xml:"minorTickMark"`
@@ -475,18 +478,22 @@ type cNumCache struct {
 	PtCount    *attrValInt `xml:"ptCount"`
 }
 
-// cDLbls (Data Lables) directly maps the dLbls element. This element serves
+// cDLbls (Data Labels) directly maps the dLbls element. This element serves
 // as a root element that specifies the settings for the data labels for an
 // entire series or the entire chart. It contains child elements that specify
 // the specific formatting and positioning settings.
 type cDLbls struct {
-	ShowLegendKey   *attrValBool `xml:"showLegendKey"`
-	ShowVal         *attrValBool `xml:"showVal"`
-	ShowCatName     *attrValBool `xml:"showCatName"`
-	ShowSerName     *attrValBool `xml:"showSerName"`
-	ShowPercent     *attrValBool `xml:"showPercent"`
-	ShowBubbleSize  *attrValBool `xml:"showBubbleSize"`
-	ShowLeaderLines *attrValBool `xml:"showLeaderLines"`
+	NumFmt          *cNumFmt       `xml:"numFmt"`
+	SpPr            *cSpPr         `xml:"spPr"`
+	TxPr            *cTxPr         `xml:"txPr"`
+	DLblPos         *attrValString `xml:"dLblPos"`
+	ShowLegendKey   *attrValBool   `xml:"showLegendKey"`
+	ShowVal         *attrValBool   `xml:"showVal"`
+	ShowCatName     *attrValBool   `xml:"showCatName"`
+	ShowSerName     *attrValBool   `xml:"showSerName"`
+	ShowPercent     *attrValBool   `xml:"showPercent"`
+	ShowBubbleSize  *attrValBool   `xml:"showBubbleSize"`
+	ShowLeaderLines *attrValBool   `xml:"showLeaderLines"`
 }
 
 // cLegend (Legend) directly maps the legend element. This element specifies
@@ -518,139 +525,109 @@ type cPageMargins struct {
 	T      float64 `xml:"t,attr"`
 }
 
-// formatChartAxis directly maps the format settings of the chart axis.
-type formatChartAxis struct {
-	None                bool    `json:"none"`
-	Crossing            string  `json:"crossing"`
-	MajorGridlines      bool    `json:"major_grid_lines"`
-	MinorGridlines      bool    `json:"minor_grid_lines"`
-	MajorTickMark       string  `json:"major_tick_mark"`
-	MinorTickMark       string  `json:"minor_tick_mark"`
-	MinorUnitType       string  `json:"minor_unit_type"`
-	MajorUnit           float64 `json:"major_unit"`
-	MajorUnitType       string  `json:"major_unit_type"`
-	TickLabelSkip       int     `json:"tick_label_skip"`
-	DisplayUnits        string  `json:"display_units"`
-	DisplayUnitsVisible bool    `json:"display_units_visible"`
-	DateAxis            bool    `json:"date_axis"`
-	ReverseOrder        bool    `json:"reverse_order"`
-	Maximum             float64 `json:"maximum"`
-	Minimum             float64 `json:"minimum"`
-	NumFormat           string  `json:"num_format"`
-	NumFont             struct {
-		Color     string `json:"color"`
-		Bold      bool   `json:"bold"`
-		Italic    bool   `json:"italic"`
-		Underline bool   `json:"underline"`
-	} `json:"num_font"`
-	LogBase    float64      `json:"logbase"`
-	NameLayout formatLayout `json:"name_layout"`
+// ChartNumFmt directly maps the number format settings of the chart.
+type ChartNumFmt struct {
+	CustomNumFmt string
+	SourceLinked bool
 }
 
-type formatChartDimension struct {
-	Width  int `json:"width"`
-	Height int `json:"height"`
+// ChartAxis directly maps the format settings of the chart axis.
+type ChartAxis struct {
+	None              bool
+	MajorGridLines    bool
+	MinorGridLines    bool
+	MajorUnit         float64
+	TickLabelPosition ChartTickLabelPositionType
+	TickLabelSkip     int
+	ReverseOrder      bool
+	Secondary         bool
+	Maximum           *float64
+	Minimum           *float64
+	Alignment         Alignment
+	Font              Font
+	LogBase           float64
+	NumFmt            ChartNumFmt
+	Title             []RichTextRun
+	axID              int
 }
 
-// formatChart directly maps the format settings of the chart.
-type formatChart struct {
-	Type       string               `json:"type"`
-	Series     []formatChartSeries  `json:"series"`
-	Format     formatPicture        `json:"format"`
-	Dimension  formatChartDimension `json:"dimension"`
-	Legend     formatChartLegend    `json:"legend"`
-	Title      formatChartTitle     `json:"title"`
-	VaryColors bool                 `json:"vary_colors"`
-	XAxis      formatChartAxis      `json:"x_axis"`
-	YAxis      formatChartAxis      `json:"y_axis"`
-	Chartarea  struct {
-		Border struct {
-			None bool `json:"none"`
-		} `json:"border"`
-		Fill struct {
-			Color string `json:"color"`
-		} `json:"fill"`
-		Pattern struct {
-			Pattern string `json:"pattern"`
-			FgColor string `json:"fg_color"`
-			BgColor string `json:"bg_color"`
-		} `json:"pattern"`
-	} `json:"chartarea"`
-	Plotarea struct {
-		ShowBubbleSize  bool `json:"show_bubble_size"`
-		ShowCatName     bool `json:"show_cat_name"`
-		ShowLeaderLines bool `json:"show_leader_lines"`
-		ShowPercent     bool `json:"show_percent"`
-		ShowSerName     bool `json:"show_series_name"`
-		ShowVal         bool `json:"show_val"`
-		Gradient        struct {
-			Colors []string `json:"colors"`
-		} `json:"gradient"`
-		Border struct {
-			Color    string `json:"color"`
-			Width    int    `json:"width"`
-			DashType string `json:"dash_type"`
-		} `json:"border"`
-		Fill struct {
-			Color string `json:"color"`
-		} `json:"fill"`
-		Layout formatLayout `json:"layout"`
-	} `json:"plotarea"`
-	ShowBlanksAs   string `json:"show_blanks_as"`
-	ShowHiddenData bool   `json:"show_hidden_data"`
-	SetRotation    int    `json:"set_rotation"`
-	SetHoleSize    int    `json:"set_hole_size"`
-	order          int
+// ChartDimension directly maps the dimension of the chart.
+type ChartDimension struct {
+	Width  uint
+	Height uint
 }
 
-// formatChartLegend directly maps the format settings of the chart legend.
-type formatChartLegend struct {
-	None            bool         `json:"none"`
-	DeleteSeries    []int        `json:"delete_series"`
-	Font            Font         `json:"font"`
-	Layout          formatLayout `json:"layout"`
-	Position        string       `json:"position"`
-	ShowLegendEntry bool         `json:"show_legend_entry"`
-	ShowLegendKey   bool         `json:"show_legend_key"`
+// ChartPlotArea directly maps the format settings of the plot area.
+type ChartPlotArea struct {
+	SecondPlotValues int
+	ShowBubbleSize   bool
+	ShowCatName      bool
+	ShowLeaderLines  bool
+	ShowPercent      bool
+	ShowSerName      bool
+	ShowVal          bool
+	Fill             Fill
+	NumFmt           ChartNumFmt
 }
 
-// formatChartSeries directly maps the format settings of the chart series.
-type formatChartSeries struct {
-	Name       string `json:"name"`
-	Categories string `json:"categories"`
-	Values     string `json:"values"`
-	Line       struct {
-		None  bool    `json:"none"`
-		Color string  `json:"color"`
-		Width float64 `json:"width"`
-	} `json:"line"`
-	Marker struct {
-		Symbol string  `json:"symbol"`
-		Size   int     `json:"size"`
-		Width  float64 `json:"width"`
-		Border struct {
-			Color string `json:"color"`
-			None  bool   `json:"none"`
-		} `json:"border"`
-		Fill struct {
-			Color string `json:"color"`
-			None  bool   `json:"none"`
-		} `json:"fill"`
-	} `json:"marker"`
+// Chart directly maps the format settings of the chart.
+type Chart struct {
+	Type         ChartType
+	Series       []ChartSeries
+	Format       GraphicOptions
+	Dimension    ChartDimension
+	Legend       ChartLegend
+	Title        []RichTextRun
+	VaryColors   *bool
+	XAxis        ChartAxis
+	YAxis        ChartAxis
+	PlotArea     ChartPlotArea
+	Fill         Fill
+	Border       ChartLine
+	ShowBlanksAs string
+	BubbleSize   int
+	HoleSize     int
+	GapWidth     *uint
+	Overlap      *int
+	order        int
 }
 
-// formatChartTitle directly maps the format settings of the chart title.
-type formatChartTitle struct {
-	None    bool         `json:"none"`
-	Name    string       `json:"name"`
-	Overlay bool         `json:"overlay"`
-	Layout  formatLayout `json:"layout"`
+// ChartLegend directly maps the format settings of the chart legend.
+type ChartLegend struct {
+	Position      string
+	ShowLegendKey bool
 }
 
-// formatLayout directly maps the format settings of the element layout.
-type formatLayout struct {
-	X      float64 `json:"x"`
-	Y      float64 `json:"y"`
-	Width  float64 `json:"width"`
-	Height float64 `json:"height"`
+// ChartMarker directly maps the format settings of the chart marker.
+type ChartMarker struct {
+	Fill   Fill
+	Symbol string
+	Size   int
+}
+
+// ChartLine directly maps the format settings of the chart line.
+type ChartLine struct {
+	Type   ChartLineType
+	Smooth bool
+	Width  float64
+}
+
+// ChartDataLabel directly maps the format settings of the chart labels.
+type ChartDataLabel struct {
+	Alignment Alignment
+	Font      Font
+	Fill      Fill
+}
+
+// ChartSeries directly maps the format settings of the chart series.
+type ChartSeries struct {
+	Name              string
+	Categories        string
+	Values            string
+	Sizes             string
+	Fill              Fill
+	Line              ChartLine
+	Marker            ChartMarker
+	DataLabel         ChartDataLabel
+	DataLabelPosition ChartDataLabelPositionType
 }
