@@ -1,13 +1,15 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
 // struct code generated by github.com/xuri/xgen
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX files. Support reads and writes XLSX file generated by
-// Microsoft Excel™ 2007 and later. Support save file without losing original
-// charts of XLSX. This library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
@@ -33,10 +35,10 @@ type xlsxChartsheet struct {
 
 // xlsxChartsheetPr specifies chart sheet properties.
 type xlsxChartsheetPr struct {
-	XMLName       xml.Name      `xml:"sheetPr"`
-	PublishedAttr bool          `xml:"published,attr,omitempty"`
-	CodeNameAttr  string        `xml:"codeName,attr,omitempty"`
-	TabColor      *xlsxTabColor `xml:"tabColor"`
+	XMLName       xml.Name   `xml:"sheetPr"`
+	PublishedAttr bool       `xml:"published,attr,omitempty"`
+	CodeNameAttr  string     `xml:"codeName,attr,omitempty"`
+	TabColor      *xlsxColor `xml:"tabColor"`
 }
 
 // xlsxChartsheetViews specifies chart sheet views.
