@@ -1,13 +1,13 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
@@ -72,16 +72,13 @@ type xlsxPhoneticRun struct {
 	T  string `xml:"t"`
 }
 
-// formatComment directly maps the format settings of the comment.
-type formatComment struct {
-	Author string `json:"author"`
-	Text   string `json:"text"`
-}
-
 // Comment directly maps the comment information.
 type Comment struct {
-	Author   string `json:"author"`
-	AuthorID int    `json:"author_id"`
-	Ref      string `json:"ref"`
-	Text     string `json:"text"`
+	Author    string
+	AuthorID  int
+	Cell      string
+	Text      string
+	Width     uint
+	Height    uint
+	Paragraph []RichTextRun
 }
