@@ -1,13 +1,13 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
@@ -18,27 +18,67 @@ import "encoding/xml"
 // specifies a two cell anchor placeholder for a group, a shape, or a drawing
 // element. It moves with cells and its extents are in EMU units.
 type decodeCellAnchor struct {
-	EditAs     string            `xml:"editAs,attr,omitempty"`
-	From       *decodeFrom       `xml:"from"`
-	To         *decodeTo         `xml:"to"`
-	Sp         *decodeSp         `xml:"sp"`
-	ClientData *decodeClientData `xml:"clientData"`
-	Content    string            `xml:",innerxml"`
+	EditAs           string                  `xml:"editAs,attr,omitempty"`
+	From             *decodeFrom             `xml:"from"`
+	To               *decodeTo               `xml:"to"`
+	Sp               *decodeSp               `xml:"sp"`
+	Pic              *decodePic              `xml:"pic"`
+	ClientData       *decodeClientData       `xml:"clientData"`
+	AlternateContent []*xlsxAlternateContent `xml:"AlternateContent"`
+	Content          string                  `xml:",innerxml"`
 }
 
-// xdrSp (Shape) directly maps the sp element. This element specifies the
-// existence of a single shape. A shape can either be a preset or a custom
-// geometry, defined using the SpreadsheetDrawingML framework. In addition to
-// a geometry each shape can have both visual and non-visual properties
-// attached. Text and corresponding styling information can also be attached
-// to a shape. This shape is specified along with all other shapes within
-// either the shape tree or group shape elements.
+// decodeCellAnchorPos defines the structure used to deserialize the cell anchor
+// for adjust drawing object on inserting/deleting column/rows.
+type decodeCellAnchorPos struct {
+	EditAs           string                  `xml:"editAs,attr,omitempty"`
+	From             *xlsxFrom               `xml:"from"`
+	To               *xlsxTo                 `xml:"to"`
+	Pos              *xlsxInnerXML           `xml:"pos"`
+	Ext              *xlsxInnerXML           `xml:"ext"`
+	Sp               *xlsxSp                 `xml:"sp"`
+	GrpSp            *xlsxInnerXML           `xml:"grpSp"`
+	GraphicFrame     *xlsxInnerXML           `xml:"graphicFrame"`
+	CxnSp            *xlsxInnerXML           `xml:"cxnSp"`
+	Pic              *xlsxInnerXML           `xml:"pic"`
+	ContentPart      *xlsxInnerXML           `xml:"contentPart"`
+	AlternateContent []*xlsxAlternateContent `xml:"AlternateContent"`
+	ClientData       *xlsxInnerXML           `xml:"clientData"`
+}
+
+// decodeChoice defines the structure used to deserialize the mc:Choice element.
+type decodeChoice struct {
+	XMLName      xml.Name           `xml:"Choice"`
+	XMLNSA14     string             `xml:"a14,attr"`
+	XMLNSSle15   string             `xml:"sle15,attr"`
+	Requires     string             `xml:"Requires,attr"`
+	GraphicFrame decodeGraphicFrame `xml:"graphicFrame"`
+}
+
+// decodeGraphicFrame defines the structure used to deserialize the
+// xdr:graphicFrame element.
+type decodeGraphicFrame struct {
+	Macro            string                 `xml:"macro,attr"`
+	NvGraphicFramePr decodeNvGraphicFramePr `xml:"nvGraphicFramePr"`
+}
+
+// decodeNvGraphicFramePr defines the structure used to deserialize the
+// xdr:nvGraphicFramePr element.
+type decodeNvGraphicFramePr struct {
+	CNvPr decodeCNvPr `xml:"cNvPr"`
+}
+
+// decodeSp defines the structure used to deserialize the sp element.
 type decodeSp struct {
-	NvSpPr *decodeNvSpPr `xml:"nvSpPr"`
-	SpPr   *decodeSpPr   `xml:"spPr"`
+	Macro      string        `xml:"macro,attr,omitempty"`
+	TextLink   string        `xml:"textlink,attr,omitempty"`
+	FLocksText bool          `xml:"fLocksText,attr,omitempty"`
+	FPublished *bool         `xml:"fPublished,attr"`
+	NvSpPr     *decodeNvSpPr `xml:"nvSpPr"`
+	SpPr       *decodeSpPr   `xml:"spPr"`
 }
 
-// decodeSp (Non-Visual Properties for a Shape) directly maps the nvSpPr
+// decodeNvSpPr (Non-Visual Properties for a Shape) directly maps the nvSpPr
 // element. This element specifies all non-visual properties for a shape. This
 // element is a container for the non-visual identification properties, shape
 // properties and application properties that are to be associated with a
@@ -46,7 +86,7 @@ type decodeSp struct {
 // appearance of the shape to be stored.
 type decodeNvSpPr struct {
 	CNvPr   *decodeCNvPr   `xml:"cNvPr"`
-	ExtLst  *decodeExt     `xml:"extLst"`
+	ExtLst  *decodeAExt    `xml:"extLst"`
 	CNvSpPr *decodeCNvSpPr `xml:"cNvSpPr"`
 }
 
@@ -63,23 +103,13 @@ type decodeCNvSpPr struct {
 // changed after serialization and deserialization, two different structures
 // are defined. decodeWsDr just for deserialization.
 type decodeWsDr struct {
-	A             string              `xml:"xmlns a,attr"`
-	Xdr           string              `xml:"xmlns xdr,attr"`
-	R             string              `xml:"xmlns r,attr"`
-	OneCellAnchor []*decodeCellAnchor `xml:"oneCellAnchor,omitempty"`
-	TwoCellAnchor []*decodeCellAnchor `xml:"twoCellAnchor,omitempty"`
-	XMLName       xml.Name            `xml:"http://schemas.openxmlformats.org/drawingml/2006/spreadsheetDrawing wsDr,omitempty"`
-}
-
-// decodeTwoCellAnchor directly maps the oneCellAnchor (One Cell Anchor Shape
-// Size) and twoCellAnchor (Two Cell Anchor Shape Size). This element
-// specifies a two cell anchor placeholder for a group, a shape, or a drawing
-// element. It moves with cells and its extents are in EMU units.
-type decodeTwoCellAnchor struct {
-	From       *decodeFrom       `xml:"from"`
-	To         *decodeTo         `xml:"to"`
-	Pic        *decodePic        `xml:"pic,omitempty"`
-	ClientData *decodeClientData `xml:"clientData"`
+	XMLName          xml.Name            `xml:"http://schemas.openxmlformats.org/drawingml/2006/spreadsheetDrawing wsDr"`
+	A                string              `xml:"xmlns a,attr"`
+	Xdr              string              `xml:"xmlns xdr,attr"`
+	R                string              `xml:"xmlns r,attr"`
+	AlternateContent []*xlsxInnerXML     `xml:"http://schemas.openxmlformats.org/markup-compatibility/2006 AlternateContent"`
+	OneCellAnchor    []*decodeCellAnchor `xml:"oneCellAnchor,omitempty"`
+	TwoCellAnchor    []*decodeCellAnchor `xml:"twoCellAnchor,omitempty"`
 }
 
 // decodeCNvPr directly maps the cNvPr (Non-Visual Drawing Properties). This
@@ -87,10 +117,11 @@ type decodeTwoCellAnchor struct {
 // information that does not affect the appearance of the picture to be
 // stored.
 type decodeCNvPr struct {
-	ID    int    `xml:"id,attr"`
-	Name  string `xml:"name,attr"`
-	Descr string `xml:"descr,attr"`
-	Title string `xml:"title,attr,omitempty"`
+	XMLName xml.Name `xml:"cNvPr"`
+	ID      int      `xml:"id,attr"`
+	Name    string   `xml:"name,attr"`
+	Descr   string   `xml:"descr,attr"`
+	Title   string   `xml:"title,attr,omitempty"`
 }
 
 // decodePicLocks directly maps the picLocks (Picture Locks). This element
@@ -111,10 +142,8 @@ type decodePicLocks struct {
 	NoSelect           bool `xml:"noSelect,attr,omitempty"`
 }
 
-// decodeBlip directly maps the blip element in the namespace
-// http://purl.oclc.org/ooxml/officeDoc ument/relationships - This element
-// specifies the existence of an image (binary large image or picture) and
-// contains a reference to the image data.
+// decodeBlip element specifies the existence of an image (binary large image
+// or picture) and contains a reference to the image data.
 type decodeBlip struct {
 	Embed  string `xml:"embed,attr"`
 	Cstate string `xml:"cstate,attr,omitempty"`
@@ -135,8 +164,8 @@ type decodeOff struct {
 	Y int `xml:"y,attr"`
 }
 
-// decodeExt directly maps the ext element.
-type decodeExt struct {
+// decodeAExt directly maps the a:ext element.
+type decodeAExt struct {
 	Cx int `xml:"cx,attr"`
 	Cy int `xml:"cy,attr"`
 }
@@ -154,8 +183,8 @@ type decodePrstGeom struct {
 // frame. This transformation is applied to the graphic frame just as it would
 // be for a shape or group shape.
 type decodeXfrm struct {
-	Off decodeOff `xml:"off"`
-	Ext decodeExt `xml:"ext"`
+	Off decodeOff  `xml:"off"`
+	Ext decodeAExt `xml:"ext"`
 }
 
 // decodeCNvPicPr directly maps the cNvPicPr (Non-Visual Picture Drawing
@@ -200,7 +229,7 @@ type decodeSpPr struct {
 // decodePic elements encompass the definition of pictures within the
 // DrawingML framework. While pictures are in many ways very similar to shapes
 // they have specific properties that are unique in order to optimize for
-// picture- specific scenarios.
+// picture-specific scenarios.
 type decodePic struct {
 	NvPicPr  decodeNvPicPr  `xml:"nvPicPr"`
 	BlipFill decodeBlipFill `xml:"blipFill"`
@@ -232,4 +261,16 @@ type decodeTo struct {
 type decodeClientData struct {
 	FLocksWithSheet  bool `xml:"fLocksWithSheet,attr"`
 	FPrintsWithSheet bool `xml:"fPrintsWithSheet,attr"`
+}
+
+// decodeCellImages directly maps the Kingsoft WPS Office embedded cell images.
+type decodeCellImages struct {
+	XMLName   xml.Name          `xml:"http://www.wps.cn/officeDocument/2017/etCustomData cellImages"`
+	CellImage []decodeCellImage `xml:"cellImage"`
+}
+
+// decodeCellImage defines the structure used to deserialize the Kingsoft WPS
+// Office embedded cell images.
+type decodeCellImage struct {
+	Pic decodePic `xml:"pic"`
 }
