@@ -1,114 +1,20 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
-import "encoding/xml"
-
-// Source relationship and namespace list, associated prefixes and schema in which it was
-// introduced.
-var (
-	SourceRelationship                = xml.Attr{Name: xml.Name{Local: "r", Space: "xmlns"}, Value: "http://schemas.openxmlformats.org/officeDocument/2006/relationships"}
-	SourceRelationshipCompatibility   = xml.Attr{Name: xml.Name{Local: "mc", Space: "xmlns"}, Value: "http://schemas.openxmlformats.org/markup-compatibility/2006"}
-	SourceRelationshipChart20070802   = xml.Attr{Name: xml.Name{Local: "c14", Space: "xmlns"}, Value: "http://schemas.microsoft.com/office/drawing/2007/8/2/chart"}
-	SourceRelationshipChart2014       = xml.Attr{Name: xml.Name{Local: "c16", Space: "xmlns"}, Value: "http://schemas.microsoft.com/office/drawing/2014/chart"}
-	SourceRelationshipChart201506     = xml.Attr{Name: xml.Name{Local: "c16r2", Space: "xmlns"}, Value: "http://schemas.microsoft.com/office/drawing/2015/06/chart"}
-	NameSpaceSpreadSheet              = xml.Attr{Name: xml.Name{Local: "xmlns"}, Value: "http://schemas.openxmlformats.org/spreadsheetml/2006/main"}
-	NameSpaceSpreadSheetX14           = xml.Attr{Name: xml.Name{Local: "x14", Space: "xmlns"}, Value: "http://schemas.microsoft.com/office/spreadsheetml/2009/9/main"}
-	NameSpaceDrawingML                = xml.Attr{Name: xml.Name{Local: "a", Space: "xmlns"}, Value: "http://schemas.openxmlformats.org/drawingml/2006/main"}
-	NameSpaceDrawingMLChart           = xml.Attr{Name: xml.Name{Local: "c", Space: "xmlns"}, Value: "http://schemas.openxmlformats.org/drawingml/2006/chart"}
-	NameSpaceDrawingMLSpreadSheet     = xml.Attr{Name: xml.Name{Local: "xdr", Space: "xmlns"}, Value: "http://schemas.openxmlformats.org/drawingml/2006/spreadsheetDrawing"}
-	NameSpaceSpreadSheetX15           = xml.Attr{Name: xml.Name{Local: "x15", Space: "xmlns"}, Value: "http://schemas.microsoft.com/office/spreadsheetml/2010/11/main"}
-	NameSpaceSpreadSheetExcel2006Main = xml.Attr{Name: xml.Name{Local: "xne", Space: "xmlns"}, Value: "http://schemas.microsoft.com/office/excel/2006/main"}
-	NameSpaceMacExcel2008Main         = xml.Attr{Name: xml.Name{Local: "mx", Space: "xmlns"}, Value: "http://schemas.microsoft.com/office/mac/excel/2008/main"}
+import (
+	"encoding/xml"
+	"sync"
 )
-
-// Source relationship and namespace.
-const (
-	SourceRelationshipOfficeDocument             = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/officeDocument"
-	SourceRelationshipChart                      = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/chart"
-	SourceRelationshipComments                   = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/comments"
-	SourceRelationshipImage                      = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/image"
-	SourceRelationshipTable                      = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/table"
-	SourceRelationshipDrawingML                  = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/drawing"
-	SourceRelationshipDrawingVML                 = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/vmlDrawing"
-	SourceRelationshipHyperLink                  = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/hyperlink"
-	SourceRelationshipWorkSheet                  = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/worksheet"
-	SourceRelationshipChartsheet                 = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/chartsheet"
-	SourceRelationshipDialogsheet                = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/dialogsheet"
-	SourceRelationshipPivotTable                 = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/pivotTable"
-	SourceRelationshipPivotCache                 = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/pivotCacheDefinition"
-	SourceRelationshipSharedStrings              = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/sharedStrings"
-	SourceRelationshipVBAProject                 = "http://schemas.microsoft.com/office/2006/relationships/vbaProject"
-	NameSpaceXML                                 = "http://www.w3.org/XML/1998/namespace"
-	NameSpaceXMLSchemaInstance                   = "http://www.w3.org/2001/XMLSchema-instance"
-	StrictSourceRelationship                     = "http://purl.oclc.org/ooxml/officeDocument/relationships"
-	StrictSourceRelationshipOfficeDocument       = "http://purl.oclc.org/ooxml/officeDocument/relationships/officeDocument"
-	StrictSourceRelationshipChart                = "http://purl.oclc.org/ooxml/officeDocument/relationships/chart"
-	StrictSourceRelationshipComments             = "http://purl.oclc.org/ooxml/officeDocument/relationships/comments"
-	StrictSourceRelationshipImage                = "http://purl.oclc.org/ooxml/officeDocument/relationships/image"
-	StrictNameSpaceSpreadSheet                   = "http://purl.oclc.org/ooxml/spreadsheetml/main"
-	NameSpaceDublinCore                          = "http://purl.org/dc/elements/1.1/"
-	NameSpaceDublinCoreTerms                     = "http://purl.org/dc/terms/"
-	NameSpaceDublinCoreMetadataIntiative         = "http://purl.org/dc/dcmitype/"
-	ContentTypeDrawing                           = "application/vnd.openxmlformats-officedocument.drawing+xml"
-	ContentTypeDrawingML                         = "application/vnd.openxmlformats-officedocument.drawingml.chart+xml"
-	ContentTypeMacro                             = "application/vnd.ms-excel.sheet.macroEnabled.main+xml"
-	ContentTypeSpreadSheetMLChartsheet           = "application/vnd.openxmlformats-officedocument.spreadsheetml.chartsheet+xml"
-	ContentTypeSpreadSheetMLComments             = "application/vnd.openxmlformats-officedocument.spreadsheetml.comments+xml"
-	ContentTypeSpreadSheetMLPivotCacheDefinition = "application/vnd.openxmlformats-officedocument.spreadsheetml.pivotCacheDefinition+xml"
-	ContentTypeSpreadSheetMLPivotTable           = "application/vnd.openxmlformats-officedocument.spreadsheetml.pivotTable+xml"
-	ContentTypeSpreadSheetMLSharedStrings        = "application/vnd.openxmlformats-officedocument.spreadsheetml.sharedStrings+xml"
-	ContentTypeSpreadSheetMLTable                = "application/vnd.openxmlformats-officedocument.spreadsheetml.table+xml"
-	ContentTypeSpreadSheetMLWorksheet            = "application/vnd.openxmlformats-officedocument.spreadsheetml.worksheet+xml"
-	ContentTypeVBA                               = "application/vnd.ms-office.vbaProject"
-	ContentTypeVML                               = "application/vnd.openxmlformats-officedocument.vmlDrawing"
-	// ExtURIConditionalFormattings is the extLst child element
-	// ([ISO/IEC29500-1:2016] section 18.2.10) of the worksheet element
-	// ([ISO/IEC29500-1:2016] section 18.3.1.99) is extended by the addition of
-	// new child ext elements ([ISO/IEC29500-1:2016] section 18.2.7)
-	ExtURIConditionalFormattings = "{78C0D931-6437-407D-A8EE-F0AAD7539E65}"
-	ExtURIDataValidations        = "{CCE6A557-97BC-4B89-ADB6-D9C93CAAB3DF}"
-	ExtURISparklineGroups        = "{05C60535-1F16-4fd2-B633-F4F36F0B64E0}"
-	ExtURISlicerListX14          = "{A8765BA9-456A-4DAB-B4F3-ACF838C121DE}"
-	ExtURISlicerCachesListX14    = "{BBE1A952-AA13-448e-AADC-164F8A28A991}"
-	ExtURISlicerListX15          = "{3A4CF648-6AED-40f4-86FF-DC5316D8AED3}"
-	ExtURIProtectedRanges        = "{FC87AEE6-9EDD-4A0A-B7FB-166176984837}"
-	ExtURIIgnoredErrors          = "{01252117-D84E-4E92-8308-4BE1C098FCBB}"
-	ExtURIWebExtensions          = "{F7C9EE02-42E1-4005-9D12-6889AFFD525C}"
-	ExtURITimelineRefs           = "{7E03D99C-DC04-49d9-9315-930204A7B6E9}"
-	ExtURIDrawingBlip            = "{28A0092B-C50C-407E-A947-70E740481C1C}"
-	ExtURIMacExcelMX             = "{64002731-A6B0-56B0-2670-7721B7C09600}"
-)
-
-// Excel specifications and limits
-const (
-	StreamChunkSize      = 1 << 24
-	MaxFontFamilyLength  = 31
-	MaxFontSize          = 409
-	MaxFileNameLength    = 207
-	MaxColumnWidth       = 255
-	MaxRowHeight         = 409
-	TotalRows            = 1048576
-	TotalColumns         = 16384
-	TotalSheetHyperlinks = 65529
-	TotalCellChars       = 32767
-	// pivotTableVersion should be greater than 3. One or more of the
-	// PivotTables chosen are created in a version of Excel earlier than
-	// Excel 2007 or in compatibility mode. Slicer can only be used with
-	// PivotTables created in Excel 2007 or a newer version of Excel.
-	pivotTableVersion = 3
-)
-
-var supportImageTypes = map[string]string{".gif": ".gif", ".jpg": ".jpeg", ".jpeg": ".jpeg", ".png": ".png", ".tif": ".tiff", ".tiff": ".tiff"}
 
 // xlsxCNvPr directly maps the cNvPr (Non-Visual Drawing Properties). This
 // element specifies non-visual canvas properties. This allows for additional
@@ -154,14 +60,13 @@ type xlsxPicLocks struct {
 	NoSelect           bool `xml:"noSelect,attr,omitempty"`
 }
 
-// xlsxBlip directly maps the blip element in the namespace
-// http://purl.oclc.org/ooxml/officeDoc ument/relationships - This element
-// specifies the existence of an image (binary large image or picture) and
-// contains a reference to the image data.
+// xlsxBlip element specifies the existence of an image (binary large image or
+// picture) and contains a reference to the image data.
 type xlsxBlip struct {
-	Embed  string `xml:"r:embed,attr"`
-	Cstate string `xml:"cstate,attr,omitempty"`
-	R      string `xml:"xmlns:r,attr"`
+	Embed   string                        `xml:"r:embed,attr"`
+	Cstate  string                        `xml:"cstate,attr,omitempty"`
+	R       string                        `xml:"xmlns:r,attr"`
+	ExtList *xlsxEGOfficeArtExtensionList `xml:"a:extLst"`
 }
 
 // xlsxStretch directly maps the stretch element. This element specifies that a
@@ -178,8 +83,8 @@ type xlsxOff struct {
 	Y int `xml:"y,attr"`
 }
 
-// xlsxExt directly maps the ext element.
-type xlsxExt struct {
+// aExt directly maps the a:ext element.
+type aExt struct {
 	Cx int `xml:"cx,attr"`
 	Cy int `xml:"cy,attr"`
 }
@@ -198,7 +103,7 @@ type xlsxPrstGeom struct {
 // be for a shape or group shape.
 type xlsxXfrm struct {
 	Off xlsxOff `xml:"a:off"`
-	Ext xlsxExt `xml:"a:ext"`
+	Ext aExt    `xml:"a:ext"`
 }
 
 // xlsxCNvPicPr directly maps the cNvPicPr (Non-Visual Picture Drawing
@@ -221,6 +126,28 @@ type xlsxNvPicPr struct {
 	CNvPicPr xlsxCNvPicPr `xml:"xdr:cNvPicPr"`
 }
 
+// xlsxCTSVGBlip specifies a graphic element in Scalable Vector Graphics (SVG)
+// format.
+type xlsxCTSVGBlip struct {
+	XMLNSaAVG string `xml:"xmlns:asvg,attr"`
+	Embed     string `xml:"r:embed,attr"`
+	Link      string `xml:"r:link,attr,omitempty"`
+}
+
+// xlsxCTOfficeArtExtension used for future extensibility and is seen elsewhere
+// throughout the drawing area.
+type xlsxCTOfficeArtExtension struct {
+	XMLName xml.Name      `xml:"a:ext"`
+	URI     string        `xml:"uri,attr"`
+	SVGBlip xlsxCTSVGBlip `xml:"asvg:svgBlip"`
+}
+
+// xlsxEGOfficeArtExtensionList used for future extensibility and is seen
+// elsewhere throughout the drawing area.
+type xlsxEGOfficeArtExtensionList struct {
+	Ext []xlsxCTOfficeArtExtension `xml:"ext"`
+}
+
 // xlsxBlipFill directly maps the blipFill (Picture Fill). This element
 // specifies the kind of picture fill that the picture object has. Because a
 // picture has a picture fill already by default, it is possible to have two
@@ -230,14 +157,24 @@ type xlsxBlipFill struct {
 	Stretch xlsxStretch `xml:"a:stretch"`
 }
 
+// xlsxLineProperties specifies the width of a line in EMUs. This simple type
+// has a minimum value of greater than or equal to 0. This simple type has a
+// maximum value of less than or equal to 20116800.
+type xlsxLineProperties struct {
+	W         int           `xml:"w,attr,omitempty"`
+	SolidFill *xlsxInnerXML `xml:"a:solidFill"`
+}
+
 // xlsxSpPr directly maps the spPr (Shape Properties). This element specifies
 // the visual shape properties that can be applied to a picture. These are the
 // same properties that are allowed to describe the visual properties of a shape
 // but are used here to describe the visual appearance of a picture within a
 // document.
 type xlsxSpPr struct {
-	Xfrm     xlsxXfrm     `xml:"a:xfrm"`
-	PrstGeom xlsxPrstGeom `xml:"a:prstGeom"`
+	Xfrm      xlsxXfrm           `xml:"a:xfrm"`
+	PrstGeom  xlsxPrstGeom       `xml:"a:prstGeom"`
+	SolidFill *xlsxInnerXML      `xml:"a:solidFill"`
+	Ln        xlsxLineProperties `xml:"a:ln"`
 }
 
 // xlsxPic elements encompass the definition of pictures within the DrawingML
@@ -277,20 +214,53 @@ type xdrClientData struct {
 	FPrintsWithSheet bool `xml:"fPrintsWithSheet,attr"`
 }
 
-// xdrCellAnchor directly maps the oneCellAnchor (One Cell Anchor Shape Size)
-// and twoCellAnchor (Two Cell Anchor Shape Size). This element specifies a two
-// cell anchor placeholder for a group, a shape, or a drawing element. It moves
-// with cells and its extents are in EMU units.
+// xdrCellAnchor specifies a oneCellAnchor (One Cell Anchor Shape Size) and
+// twoCellAnchor (Two Cell Anchor Shape Size) placeholder for a group, a shape,
+// or a drawing element. It moves with cells and its extents are in EMU units.
 type xdrCellAnchor struct {
-	EditAs       string         `xml:"editAs,attr,omitempty"`
-	Pos          *xlsxPoint2D   `xml:"xdr:pos"`
-	From         *xlsxFrom      `xml:"xdr:from"`
-	To           *xlsxTo        `xml:"xdr:to"`
-	Ext          *xlsxExt       `xml:"xdr:ext"`
-	Sp           *xdrSp         `xml:"xdr:sp"`
-	Pic          *xlsxPic       `xml:"xdr:pic,omitempty"`
-	GraphicFrame string         `xml:",innerxml"`
-	ClientData   *xdrClientData `xml:"xdr:clientData"`
+	EditAs           string                  `xml:"editAs,attr,omitempty"`
+	Pos              *xlsxPoint2D            `xml:"xdr:pos"`
+	From             *xlsxFrom               `xml:"xdr:from"`
+	To               *xlsxTo                 `xml:"xdr:to"`
+	Ext              *aExt                   `xml:"xdr:ext"`
+	Sp               *xdrSp                  `xml:"xdr:sp"`
+	Pic              *xlsxPic                `xml:"xdr:pic,omitempty"`
+	GraphicFrame     string                  `xml:",innerxml"`
+	AlternateContent []*xlsxAlternateContent `xml:"mc:AlternateContent"`
+	ClientData       *xdrClientData          `xml:"xdr:clientData"`
+}
+
+// xlsxCellAnchorPos defines the structure used to serialize the cell anchor for
+// adjust drawing object on inserting/deleting column/rows.
+type xlsxCellAnchorPos struct {
+	EditAs           string                  `xml:"editAs,attr,omitempty"`
+	From             *xlsxFrom               `xml:"xdr:from"`
+	To               *xlsxTo                 `xml:"xdr:to"`
+	Pos              *xlsxInnerXML           `xml:"xdr:pos"`
+	Ext              *xlsxInnerXML           `xml:"xdr:ext"`
+	Sp               *xlsxSp                 `xml:"xdr:sp"`
+	GrpSp            *xlsxInnerXML           `xml:"xdr:grpSp"`
+	GraphicFrame     *xlsxInnerXML           `xml:"xdr:graphicFrame"`
+	CxnSp            *xlsxInnerXML           `xml:"xdr:cxnSp"`
+	Pic              *xlsxInnerXML           `xml:"xdr:pic"`
+	ContentPart      *xlsxInnerXML           `xml:"xdr:contentPart"`
+	AlternateContent []*xlsxAlternateContent `xml:"mc:AlternateContent"`
+	ClientData       *xlsxInnerXML           `xml:"xdr:clientData"`
+}
+
+// xdrSp (Shape) directly maps the sp element. This element specifies the
+// existence of a single shape. A shape can either be a preset or a custom
+// geometry, defined using the SpreadsheetDrawingML framework. In addition to
+// a geometry each shape can have both visual and non-visual properties
+// attached. Text and corresponding styling information can also be attached
+// to a shape. This shape is specified along with all other shapes within
+// either the shape tree or group shape elements.
+type xlsxSp struct {
+	Macro      string `xml:"macro,attr,omitempty"`
+	TextLink   string `xml:"textlink,attr,omitempty"`
+	FLocksText bool   `xml:"fLocksText,attr,omitempty"`
+	FPublished *bool  `xml:"fPublished,attr"`
+	Content    string `xml:",innerxml"`
 }
 
 // xlsxPoint2D describes the position of a drawing element within a spreadsheet.
@@ -303,13 +273,16 @@ type xlsxPoint2D struct {
 // xlsxWsDr directly maps the root element for a part of this content type shall
 // wsDr.
 type xlsxWsDr struct {
-	XMLName        xml.Name         `xml:"xdr:wsDr"`
-	AbsoluteAnchor []*xdrCellAnchor `xml:"xdr:absoluteAnchor"`
-	OneCellAnchor  []*xdrCellAnchor `xml:"xdr:oneCellAnchor"`
-	TwoCellAnchor  []*xdrCellAnchor `xml:"xdr:twoCellAnchor"`
-	A              string           `xml:"xmlns:a,attr,omitempty"`
-	Xdr            string           `xml:"xmlns:xdr,attr,omitempty"`
-	R              string           `xml:"xmlns:r,attr,omitempty"`
+	mu               sync.Mutex
+	XMLName          xml.Name                `xml:"xdr:wsDr"`
+	NS               string                  `xml:"xmlns,attr,omitempty"`
+	A                string                  `xml:"xmlns:a,attr,omitempty"`
+	Xdr              string                  `xml:"xmlns:xdr,attr,omitempty"`
+	R                string                  `xml:"xmlns:r,attr,omitempty"`
+	AlternateContent []*xlsxAlternateContent `xml:"mc:AlternateContent"`
+	AbsoluteAnchor   []*xdrCellAnchor        `xml:"xdr:absoluteAnchor"`
+	OneCellAnchor    []*xdrCellAnchor        `xml:"xdr:oneCellAnchor"`
+	TwoCellAnchor    []*xdrCellAnchor        `xml:"xdr:twoCellAnchor"`
 }
 
 // xlsxGraphicFrame (Graphic Frame) directly maps the xdr:graphicFrame element.
@@ -352,6 +325,12 @@ type xlsxGraphic struct {
 type xlsxGraphicData struct {
 	URI   string     `xml:"uri,attr"`
 	Chart *xlsxChart `xml:"c:chart,omitempty"`
+	Sle   *xlsxSle   `xml:"sle:slicer"`
+}
+
+type xlsxSle struct {
+	XMLNS string `xml:"xmlns:sle,attr"`
+	Name  string `xml:"name,attr"`
 }
 
 // xlsxChart (Chart) directly maps the c:chart element.
@@ -369,6 +348,7 @@ type xlsxChart struct {
 // This shape is specified along with all other shapes within either the shape
 // tree or group shape elements.
 type xdrSp struct {
+	XMLName  xml.Name   `xml:"xdr:sp"`
 	Macro    string     `xml:"macro,attr"`
 	Textlink string     `xml:"textlink,attr"`
 	NvSpPr   *xdrNvSpPr `xml:"xdr:nvSpPr"`
@@ -442,41 +422,46 @@ type xdrTxBody struct {
 	P      []*aP    `xml:"a:p"`
 }
 
-// formatPicture directly maps the format settings of the picture.
-type formatPicture struct {
-	FPrintsWithSheet bool    `json:"print_obj"`
-	FLocksWithSheet  bool    `json:"locked"`
-	NoChangeAspect   bool    `json:"lock_aspect_ratio"`
-	Autofit          bool    `json:"autofit"`
-	OffsetX          int     `json:"x_offset"`
-	OffsetY          int     `json:"y_offset"`
-	XScale           float64 `json:"x_scale"`
-	YScale           float64 `json:"y_scale"`
-	Hyperlink        string  `json:"hyperlink"`
-	HyperlinkType    string  `json:"hyperlink_type"`
-	Positioning      string  `json:"positioning"`
+// Picture maps the format settings of the picture.
+type Picture struct {
+	Extension  string
+	File       []byte
+	Format     *GraphicOptions
+	InsertType PictureInsertType
 }
 
-// formatShape directly maps the format settings of the shape.
-type formatShape struct {
-	Type      string                 `json:"type"`
-	Width     int                    `json:"width"`
-	Height    int                    `json:"height"`
-	Format    formatPicture          `json:"format"`
-	Color     formatShapeColor       `json:"color"`
-	Paragraph []formatShapeParagraph `json:"paragraph"`
+// GraphicOptions directly maps the format settings of the picture.
+type GraphicOptions struct {
+	AltText             string
+	PrintObject         *bool
+	Locked              *bool
+	LockAspectRatio     bool
+	AutoFit             bool
+	AutoFitIgnoreAspect bool
+	OffsetX             int
+	OffsetY             int
+	ScaleX              float64
+	ScaleY              float64
+	Hyperlink           string
+	HyperlinkType       string
+	Positioning         string
 }
 
-// formatShapeParagraph directly maps the format settings of the paragraph in
-// the shape.
-type formatShapeParagraph struct {
-	Font Font   `json:"font"`
-	Text string `json:"text"`
+// Shape directly maps the format settings of the shape.
+type Shape struct {
+	Cell      string
+	Type      string
+	Macro     string
+	Width     uint
+	Height    uint
+	Format    GraphicOptions
+	Fill      Fill
+	Line      ShapeLine
+	Paragraph []RichTextRun
 }
 
-// formatShapeColor directly maps the color settings of the shape.
-type formatShapeColor struct {
-	Line   string `json:"line"`
-	Fill   string `json:"fill"`
-	Effect string `json:"effect"`
+// ShapeLine directly maps the line settings of the shape.
+type ShapeLine struct {
+	Color string
+	Width *float64
 }
