@@ -1,13 +1,13 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
@@ -79,8 +79,7 @@ type xlsxWorksheetSource struct {
 // PivotTable is a collection of ranges in the workbook. The ranges are
 // specified in the rangeSets collection. The logic for how the application
 // consolidates the data in the ranges is application- defined.
-type xlsxConsolidation struct {
-}
+type xlsxConsolidation struct{}
 
 // xlsxCacheFields represents the collection of field definitions in the
 // source data.
@@ -121,31 +120,30 @@ type xlsxCacheField struct {
 // those values that are referenced in multiple places across all the
 // PivotTable parts.
 type xlsxSharedItems struct {
-	ContainsSemiMixedTypes bool          `xml:"containsSemiMixedTypes,attr,omitempty"`
-	ContainsNonDate        bool          `xml:"containsNonDate,attr,omitempty"`
-	ContainsDate           bool          `xml:"containsDate,attr,omitempty"`
-	ContainsString         bool          `xml:"containsString,attr,omitempty"`
-	ContainsBlank          bool          `xml:"containsBlank,attr,omitempty"`
-	ContainsMixedTypes     bool          `xml:"containsMixedTypes,attr,omitempty"`
-	ContainsNumber         bool          `xml:"containsNumber,attr,omitempty"`
-	ContainsInteger        bool          `xml:"containsInteger,attr,omitempty"`
-	MinValue               float64       `xml:"minValue,attr,omitempty"`
-	MaxValue               float64       `xml:"maxValue,attr,omitempty"`
-	MinDate                string        `xml:"minDate,attr,omitempty"`
-	MaxDate                string        `xml:"maxDate,attr,omitempty"`
-	Count                  int           `xml:"count,attr"`
-	LongText               bool          `xml:"longText,attr,omitempty"`
-	M                      *xlsxMissing  `xml:"m"`
-	N                      *xlsxNumber   `xml:"n"`
-	B                      *xlsxBoolean  `xml:"b"`
-	E                      *xlsxError    `xml:"e"`
-	S                      *xlsxString   `xml:"s"`
-	D                      *xlsxDateTime `xml:"d"`
+	ContainsSemiMixedTypes bool           `xml:"containsSemiMixedTypes,attr,omitempty"`
+	ContainsNonDate        bool           `xml:"containsNonDate,attr,omitempty"`
+	ContainsDate           bool           `xml:"containsDate,attr,omitempty"`
+	ContainsString         bool           `xml:"containsString,attr,omitempty"`
+	ContainsBlank          bool           `xml:"containsBlank,attr,omitempty"`
+	ContainsMixedTypes     bool           `xml:"containsMixedTypes,attr,omitempty"`
+	ContainsNumber         bool           `xml:"containsNumber,attr,omitempty"`
+	ContainsInteger        bool           `xml:"containsInteger,attr,omitempty"`
+	MinValue               float64        `xml:"minValue,attr,omitempty"`
+	MaxValue               float64        `xml:"maxValue,attr,omitempty"`
+	MinDate                string         `xml:"minDate,attr,omitempty"`
+	MaxDate                string         `xml:"maxDate,attr,omitempty"`
+	Count                  int            `xml:"count,attr"`
+	LongText               bool           `xml:"longText,attr,omitempty"`
+	M                      []xlsxMissing  `xml:"m"`
+	N                      []xlsxNumber   `xml:"n"`
+	B                      []xlsxBoolean  `xml:"b"`
+	E                      []xlsxError    `xml:"e"`
+	S                      []xlsxString   `xml:"s"`
+	D                      []xlsxDateTime `xml:"d"`
 }
 
 // xlsxMissing represents a value that was not specified.
-type xlsxMissing struct {
-}
+type xlsxMissing struct{}
 
 // xlsxNumber represents a numeric value in the PivotTable.
 type xlsxNumber struct {
@@ -167,18 +165,15 @@ type xlsxNumber struct {
 
 // xlsxTuples represents members for the OLAP sheet data entry, also known as
 // a tuple.
-type xlsxTuples struct {
-}
+type xlsxTuples struct{}
 
 // xlsxBoolean represents a boolean value for an item in the PivotTable.
-type xlsxBoolean struct {
-}
+type xlsxBoolean struct{}
 
 // xlsxError represents an error value. The use of this item indicates that an
 // error value is present in the PivotTable source. The error is recorded in
 // the value attribute.
-type xlsxError struct {
-}
+type xlsxError struct{}
 
 // xlsxString represents a character value in a PivotTable.
 type xlsxString struct {
@@ -199,45 +194,49 @@ type xlsxString struct {
 }
 
 // xlsxDateTime represents a date-time value in the PivotTable.
-type xlsxDateTime struct {
-}
+type xlsxDateTime struct{}
 
 // xlsxFieldGroup represents the collection of properties for a field group.
-type xlsxFieldGroup struct {
-}
+type xlsxFieldGroup struct{}
 
 // xlsxCacheHierarchies represents the collection of OLAP hierarchies in the
 // PivotCache.
-type xlsxCacheHierarchies struct {
-}
+type xlsxCacheHierarchies struct{}
 
 // xlsxKpis represents the collection of Key Performance Indicators (KPIs)
 // defined on the OLAP server and stored in the PivotCache.
-type xlsxKpis struct {
-}
+type xlsxKpis struct{}
 
 // xlsxTupleCache represents the cache of OLAP sheet data members, or tuples.
-type xlsxTupleCache struct {
-}
+type xlsxTupleCache struct{}
 
 // xlsxCalculatedItems represents the collection of calculated items.
-type xlsxCalculatedItems struct {
-}
+type xlsxCalculatedItems struct{}
 
 // xlsxCalculatedMembers represents the collection of calculated members in an
 // OLAP PivotTable.
-type xlsxCalculatedMembers struct {
-}
+type xlsxCalculatedMembers struct{}
 
 // xlsxDimensions represents the collection of PivotTable OLAP dimensions.
-type xlsxDimensions struct {
-}
+type xlsxDimensions struct{}
 
 // xlsxMeasureGroups represents the collection of PivotTable OLAP measure
 // groups.
-type xlsxMeasureGroups struct {
-}
+type xlsxMeasureGroups struct{}
 
 // xlsxMaps represents the PivotTable OLAP measure group - Dimension maps.
-type xlsxMaps struct {
+type xlsxMaps struct{}
+
+// xlsxX14PivotCacheDefinition specifies the extended properties of a pivot
+// table cache definition.
+type xlsxX14PivotCacheDefinition struct {
+	XMLName      xml.Name `xml:"x14:pivotCacheDefinition"`
+	PivotCacheID int      `xml:"pivotCacheId,attr"`
+}
+
+// decodeX14PivotCacheDefinition defines the structure used to parse the
+// x14:pivotCacheDefinition element of a pivot table cache.
+type decodeX14PivotCacheDefinition struct {
+	XMLName      xml.Name `xml:"pivotCacheDefinition"`
+	PivotCacheID int      `xml:"pivotCacheId,attr"`
 }
