@@ -1,13 +1,13 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
@@ -56,25 +56,25 @@ type xlsxPivotTableDefinition struct {
 	EnableDrill             bool                     `xml:"enableDrill,attr,omitempty"`
 	EnableFieldProperties   bool                     `xml:"enableFieldProperties,attr,omitempty"`
 	PreserveFormatting      bool                     `xml:"preserveFormatting,attr,omitempty"`
-	UseAutoFormatting       *bool                    `xml:"useAutoFormatting,attr,omitempty"`
+	UseAutoFormatting       *bool                    `xml:"useAutoFormatting,attr"`
 	PageWrap                int                      `xml:"pageWrap,attr,omitempty"`
-	PageOverThenDown        *bool                    `xml:"pageOverThenDown,attr,omitempty"`
+	PageOverThenDown        *bool                    `xml:"pageOverThenDown,attr"`
 	SubtotalHiddenItems     bool                     `xml:"subtotalHiddenItems,attr,omitempty"`
-	RowGrandTotals          *bool                    `xml:"rowGrandTotals,attr,omitempty"`
-	ColGrandTotals          *bool                    `xml:"colGrandTotals,attr,omitempty"`
+	RowGrandTotals          *bool                    `xml:"rowGrandTotals,attr"`
+	ColGrandTotals          *bool                    `xml:"colGrandTotals,attr"`
 	FieldPrintTitles        bool                     `xml:"fieldPrintTitles,attr,omitempty"`
 	ItemPrintTitles         bool                     `xml:"itemPrintTitles,attr,omitempty"`
-	MergeItem               *bool                    `xml:"mergeItem,attr,omitempty"`
+	MergeItem               *bool                    `xml:"mergeItem,attr"`
 	ShowDropZones           bool                     `xml:"showDropZones,attr,omitempty"`
 	CreatedVersion          int                      `xml:"createdVersion,attr,omitempty"`
 	Indent                  int                      `xml:"indent,attr,omitempty"`
 	ShowEmptyRow            bool                     `xml:"showEmptyRow,attr,omitempty"`
 	ShowEmptyCol            bool                     `xml:"showEmptyCol,attr,omitempty"`
 	ShowHeaders             bool                     `xml:"showHeaders,attr,omitempty"`
-	Compact                 bool                     `xml:"compact,attr"`
-	Outline                 bool                     `xml:"outline,attr"`
+	Compact                 *bool                    `xml:"compact,attr"`
+	Outline                 *bool                    `xml:"outline,attr"`
 	OutlineData             bool                     `xml:"outlineData,attr,omitempty"`
-	CompactData             *bool                    `xml:"compactData,attr,omitempty"`
+	CompactData             *bool                    `xml:"compactData,attr"`
 	Published               bool                     `xml:"published,attr,omitempty"`
 	GridDropZones           bool                     `xml:"gridDropZones,attr,omitempty"`
 	Immersive               bool                     `xml:"immersive,attr,omitempty"`
@@ -125,10 +125,10 @@ type xlsxPivotField struct {
 	ShowDropDowns                bool               `xml:"showDropDowns,attr,omitempty"`
 	HiddenLevel                  bool               `xml:"hiddenLevel,attr,omitempty"`
 	UniqueMemberProperty         string             `xml:"uniqueMemberProperty,attr,omitempty"`
-	Compact                      bool               `xml:"compact,attr"`
+	Compact                      *bool              `xml:"compact,attr"`
 	AllDrilled                   bool               `xml:"allDrilled,attr,omitempty"`
 	NumFmtID                     string             `xml:"numFmtId,attr,omitempty"`
-	Outline                      bool               `xml:"outline,attr"`
+	Outline                      *bool              `xml:"outline,attr"`
 	SubtotalTop                  bool               `xml:"subtotalTop,attr,omitempty"`
 	DragToRow                    bool               `xml:"dragToRow,attr,omitempty"`
 	DragToCol                    bool               `xml:"dragToCol,attr,omitempty"`
@@ -150,7 +150,7 @@ type xlsxPivotField struct {
 	DataSourceSort               bool               `xml:"dataSourceSort,attr,omitempty"`
 	NonAutoSortDefault           bool               `xml:"nonAutoSortDefault,attr,omitempty"`
 	RankBy                       int                `xml:"rankBy,attr,omitempty"`
-	DefaultSubtotal              *bool              `xml:"defaultSubtotal,attr,omitempty"`
+	DefaultSubtotal              *bool              `xml:"defaultSubtotal,attr"`
 	SumSubtotal                  bool               `xml:"sumSubtotal,attr,omitempty"`
 	CountASubtotal               bool               `xml:"countASubtotal,attr,omitempty"`
 	AvgSubtotal                  bool               `xml:"avgSubtotal,attr,omitempty"`
@@ -195,8 +195,7 @@ type xlsxItem struct {
 }
 
 // xlsxAutoSortScope represents the sorting scope for the PivotTable.
-type xlsxAutoSortScope struct {
-}
+type xlsxAutoSortScope struct{}
 
 // xlsxRowFields represents the collection of row fields for the PivotTable.
 type xlsxRowFields struct {
@@ -225,8 +224,7 @@ type xlsxI struct {
 }
 
 // xlsxX represents an array of indexes to cached shared item values.
-type xlsxX struct {
-}
+type xlsxX struct{}
 
 // xlsxColFields represents the collection of fields that are on the column
 // axis of the PivotTable.
@@ -275,14 +273,13 @@ type xlsxDataField struct {
 	ShowDataAs string      `xml:"showDataAs,attr,omitempty"`
 	BaseField  int         `xml:"baseField,attr,omitempty"`
 	BaseItem   int64       `xml:"baseItem,attr,omitempty"`
-	NumFmtID   string      `xml:"numFmtId,attr,omitempty"`
+	NumFmtID   int         `xml:"numFmtId,attr,omitempty"`
 	ExtLst     *xlsxExtLst `xml:"extLst"`
 }
 
 // xlsxConditionalFormats represents the collection of conditional formats
 // applied to a PivotTable.
-type xlsxConditionalFormats struct {
-}
+type xlsxConditionalFormats struct{}
 
 // xlsxPivotTableStyleInfo represent information on style applied to the
 // PivotTable.
