@@ -1,19 +1,19 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
 import (
 	"encoding/xml"
-	"strings"
+	"sync"
 )
 
 // xlsxSST directly maps the sst element from the namespace
@@ -24,6 +24,7 @@ import (
 // is an indexed list of string values, shared across the workbook, which allows
 // implementations to store values only once.
 type xlsxSST struct {
+	mu          sync.Mutex
 	XMLName     xml.Name `xml:"http://schemas.openxmlformats.org/spreadsheetml/2006/main sst"`
 	Count       int      `xml:"count,attr"`
 	UniqueCount int      `xml:"uniqueCount,attr"`
@@ -44,30 +45,14 @@ type xlsxSI struct {
 	PhoneticPr *xlsxPhoneticPr    `xml:"phoneticPr"`
 }
 
-// String extracts characters from a string item.
-func (x xlsxSI) String() string {
-	if len(x.R) > 0 {
-		var rows strings.Builder
-		for _, s := range x.R {
-			if s.T != nil {
-				rows.WriteString(s.T.Val)
-			}
-		}
-		return bstrUnmarshal(rows.String())
-	}
-	if x.T != nil {
-		return bstrUnmarshal(x.T.Val)
-	}
-	return ""
-}
-
 // xlsxR represents a run of rich text. A rich text run is a region of text
 // that share a common set of properties, such as formatting properties. The
 // properties are defined in the rPr element, and the text displayed to the
 // user is defined in the Text (t) element.
 type xlsxR struct {
-	RPr *xlsxRPr `xml:"rPr"`
-	T   *xlsxT   `xml:"t"`
+	XMLName xml.Name `xml:"r"`
+	RPr     *xlsxRPr `xml:"rPr"`
+	T       *xlsxT   `xml:"t"`
 }
 
 // xlsxT directly maps the t element in the run properties.
