@@ -1,31 +1,35 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
-import "encoding/xml"
+import (
+	"encoding/xml"
+	"sync"
+)
 
 // xlsxStyleSheet is the root element of the Styles part.
 type xlsxStyleSheet struct {
+	mu           sync.Mutex
 	XMLName      xml.Name          `xml:"http://schemas.openxmlformats.org/spreadsheetml/2006/main styleSheet"`
-	NumFmts      *xlsxNumFmts      `xml:"numFmts,omitempty"`
-	Fonts        *xlsxFonts        `xml:"fonts,omitempty"`
-	Fills        *xlsxFills        `xml:"fills,omitempty"`
-	Borders      *xlsxBorders      `xml:"borders,omitempty"`
-	CellStyleXfs *xlsxCellStyleXfs `xml:"cellStyleXfs,omitempty"`
-	CellXfs      *xlsxCellXfs      `xml:"cellXfs,omitempty"`
-	CellStyles   *xlsxCellStyles   `xml:"cellStyles,omitempty"`
-	Dxfs         *xlsxDxfs         `xml:"dxfs,omitempty"`
-	TableStyles  *xlsxTableStyles  `xml:"tableStyles,omitempty"`
-	Colors       *xlsxStyleColors  `xml:"colors,omitempty"`
+	NumFmts      *xlsxNumFmts      `xml:"numFmts"`
+	Fonts        *xlsxFonts        `xml:"fonts"`
+	Fills        *xlsxFills        `xml:"fills"`
+	Borders      *xlsxBorders      `xml:"borders"`
+	CellStyleXfs *xlsxCellStyleXfs `xml:"cellStyleXfs"`
+	CellXfs      *xlsxCellXfs      `xml:"cellXfs"`
+	CellStyles   *xlsxCellStyles   `xml:"cellStyles"`
+	Dxfs         *xlsxDxfs         `xml:"dxfs"`
+	TableStyles  *xlsxTableStyles  `xml:"tableStyles"`
+	Colors       *xlsxStyleColors  `xml:"colors"`
 	ExtLst       *xlsxExtLst       `xml:"extLst"`
 }
 
@@ -56,15 +60,15 @@ type xlsxProtection struct {
 // xlsxLine expresses a single set of cell border.
 type xlsxLine struct {
 	Style string     `xml:"style,attr,omitempty"`
-	Color *xlsxColor `xml:"color,omitempty"`
+	Color *xlsxColor `xml:"color"`
 }
 
 // xlsxColor is a common mapping used for both the fgColor and bgColor elements.
 // Foreground color of the cell fill pattern. Cell fill patterns operate with
-// two colors: a background color and a foreground color. These combine together
+// two colors: a background color and a foreground color. These combine
 // to make a patterned cell fill. Background color of the cell fill pattern.
 // Cell fill patterns operate with two colors: a background color and a
-// foreground color. These combine together to make a patterned cell fill.
+// foreground color. These combine to make a patterned cell fill.
 type xlsxColor struct {
 	Auto    bool    `xml:"auto,attr,omitempty"`
 	RGB     string  `xml:"rgb,attr,omitempty"`
@@ -83,13 +87,13 @@ type xlsxFonts struct {
 // xlsxFont directly maps the font element. This element defines the
 // properties for one of the fonts used in this workbook.
 type xlsxFont struct {
-	B        *attrValBool   `xml:"b,omitempty"`
-	I        *attrValBool   `xml:"i,omitempty"`
-	Strike   *attrValBool   `xml:"strike,omitempty"`
-	Outline  *attrValBool   `xml:"outline,omitempty"`
-	Shadow   *attrValBool   `xml:"shadow,omitempty"`
-	Condense *attrValBool   `xml:"condense,omitempty"`
-	Extend   *attrValBool   `xml:"extend,omitempty"`
+	B        *attrValBool   `xml:"b"`
+	I        *attrValBool   `xml:"i"`
+	Strike   *attrValBool   `xml:"strike"`
+	Outline  *attrValBool   `xml:"outline"`
+	Shadow   *attrValBool   `xml:"shadow"`
+	Condense *attrValBool   `xml:"condense"`
+	Extend   *attrValBool   `xml:"extend"`
 	U        *attrValString `xml:"u"`
 	Sz       *attrValFloat  `xml:"sz"`
 	Color    *xlsxColor     `xml:"color"`
@@ -99,20 +103,20 @@ type xlsxFont struct {
 	Scheme   *attrValString `xml:"scheme"`
 }
 
-// xlsxFills directly maps the fills element. This element defines the cell
+// xlsxFills directly maps the fills' element. This element defines the cell
 // fills portion of the Styles part, consisting of a sequence of fill records. A
 // cell fill consists of a background color, foreground color, and pattern to be
 // applied across the cell.
 type xlsxFills struct {
 	Count int         `xml:"count,attr"`
-	Fill  []*xlsxFill `xml:"fill,omitempty"`
+	Fill  []*xlsxFill `xml:"fill"`
 }
 
 // xlsxFill directly maps the fill element. This element specifies fill
 // formatting.
 type xlsxFill struct {
-	PatternFill  *xlsxPatternFill  `xml:"patternFill,omitempty"`
-	GradientFill *xlsxGradientFill `xml:"gradientFill,omitempty"`
+	PatternFill  *xlsxPatternFill  `xml:"patternFill"`
+	GradientFill *xlsxGradientFill `xml:"gradientFill"`
 }
 
 // xlsxPatternFill is used to specify cell fill information for pattern and
@@ -134,7 +138,7 @@ type xlsxGradientFill struct {
 	Right  float64                 `xml:"right,attr,omitempty"`
 	Top    float64                 `xml:"top,attr,omitempty"`
 	Type   string                  `xml:"type,attr,omitempty"`
-	Stop   []*xlsxGradientFillStop `xml:"stop,omitempty"`
+	Stop   []*xlsxGradientFillStop `xml:"stop"`
 }
 
 // xlsxGradientFillStop directly maps the stop element.
@@ -143,26 +147,28 @@ type xlsxGradientFillStop struct {
 	Color    xlsxColor `xml:"color,omitempty"`
 }
 
-// xlsxBorders directly maps the borders element. This element contains borders
+// xlsxBorders directly maps the borders' element. This element contains borders
 // formatting information, specifying all border definitions for all cells in
 // the workbook.
 type xlsxBorders struct {
 	Count  int           `xml:"count,attr"`
-	Border []*xlsxBorder `xml:"border,omitempty"`
+	Border []*xlsxBorder `xml:"border"`
 }
 
 // xlsxBorder directly maps the border element. Expresses a single set of cell
 // border formats (left, right, top, bottom, diagonal). Color is optional. When
 // missing, 'automatic' is implied.
 type xlsxBorder struct {
-	DiagonalDown bool     `xml:"diagonalDown,attr,omitempty"`
-	DiagonalUp   bool     `xml:"diagonalUp,attr,omitempty"`
-	Outline      bool     `xml:"outline,attr,omitempty"`
-	Left         xlsxLine `xml:"left,omitempty"`
-	Right        xlsxLine `xml:"right,omitempty"`
-	Top          xlsxLine `xml:"top,omitempty"`
-	Bottom       xlsxLine `xml:"bottom,omitempty"`
-	Diagonal     xlsxLine `xml:"diagonal,omitempty"`
+	DiagonalDown bool      `xml:"diagonalDown,attr,omitempty"`
+	DiagonalUp   bool      `xml:"diagonalUp,attr,omitempty"`
+	Outline      bool      `xml:"outline,attr,omitempty"`
+	Left         *xlsxLine `xml:"left"`
+	Right        *xlsxLine `xml:"right"`
+	Top          *xlsxLine `xml:"top"`
+	Bottom       *xlsxLine `xml:"bottom"`
+	Diagonal     *xlsxLine `xml:"diagonal"`
+	Vertical     *xlsxLine `xml:"vertical"`
+	Horizontal   *xlsxLine `xml:"horizontal"`
 }
 
 // xlsxCellStyles directly maps the cellStyles element. This element contains
@@ -173,7 +179,7 @@ type xlsxBorder struct {
 type xlsxCellStyles struct {
 	XMLName   xml.Name         `xml:"cellStyles"`
 	Count     int              `xml:"count,attr"`
-	CellStyle []*xlsxCellStyle `xml:"cellStyle,omitempty"`
+	CellStyle []*xlsxCellStyle `xml:"cellStyle"`
 }
 
 // xlsxCellStyle directly maps the cellStyle element. This element represents
@@ -183,17 +189,17 @@ type xlsxCellStyle struct {
 	XMLName       xml.Name `xml:"cellStyle"`
 	Name          string   `xml:"name,attr"`
 	XfID          int      `xml:"xfId,attr"`
-	BuiltInID     *int     `xml:"builtinId,attr,omitempty"`
-	ILevel        *int     `xml:"iLevel,attr,omitempty"`
-	Hidden        *bool    `xml:"hidden,attr,omitempty"`
-	CustomBuiltIn *bool    `xml:"customBuiltin,attr,omitempty"`
+	BuiltInID     *int     `xml:"builtinId,attr"`
+	ILevel        *int     `xml:"iLevel,attr"`
+	Hidden        *bool    `xml:"hidden,attr"`
+	CustomBuiltIn *bool    `xml:"customBuiltin,attr"`
 }
 
 // xlsxCellStyleXfs directly maps the cellStyleXfs element. This element
 // contains the master formatting records (xf's) which define the formatting for
 // all named cell styles in this workbook. Master formatting records reference
 // individual elements of formatting (e.g., number format, font definitions,
-// cell fills, etc) by specifying a zero-based index into those collections.
+// cell fills, etc.) by specifying a zero-based index into those collections.
 // Master formatting records also specify whether to apply or ignore particular
 // aspects of formatting.
 type xlsxCellStyleXfs struct {
@@ -201,7 +207,7 @@ type xlsxCellStyleXfs struct {
 	Xf    []xlsxXf `xml:"xf,omitempty"`
 }
 
-// xlsxXf directly maps the xf element. A single xf element describes all of the
+// xlsxXf directly maps the xf element. A single xf element describes all the
 // formatting for a cell.
 type xlsxXf struct {
 	NumFmtID          *int            `xml:"numFmtId,attr"`
@@ -232,8 +238,8 @@ type xlsxCellXfs struct {
 }
 
 // xlsxDxfs directly maps the dxfs element. This element contains the master
-// differential formatting records (dxf's) which define formatting for all non-
-// cell formatting in this workbook. Whereas xf records fully specify a
+// differential formatting records (dxf's) which define formatting for all
+// non-cell formatting in this workbook. Whereas xf records fully specify a
 // particular aspect of formatting (e.g., cell borders) by referencing those
 // formatting definitions elsewhere in the Styles part, dxf records specify
 // incremental (or differential) aspects of formatting directly inline within
@@ -241,24 +247,19 @@ type xlsxCellXfs struct {
 // to any formatting already present on the object using the dxf record.
 type xlsxDxfs struct {
 	Count int        `xml:"count,attr"`
-	Dxfs  []*xlsxDxf `xml:"dxf,omitempty"`
+	Dxfs  []*xlsxDxf `xml:"dxf"`
 }
 
 // xlsxDxf directly maps the dxf element. A single dxf record, expressing
 // incremental formatting to be applied.
 type xlsxDxf struct {
-	Dxf string `xml:",innerxml"`
-}
-
-// dxf directly maps the dxf element.
-type dxf struct {
 	Font       *xlsxFont       `xml:"font"`
 	NumFmt     *xlsxNumFmt     `xml:"numFmt"`
 	Fill       *xlsxFill       `xml:"fill"`
 	Alignment  *xlsxAlignment  `xml:"alignment"`
 	Border     *xlsxBorder     `xml:"border"`
 	Protection *xlsxProtection `xml:"protection"`
-	ExtLst     *xlsxExt        `xml:"extLst"`
+	ExtLst     *aExt           `xml:"extLst"`
 }
 
 // xlsxTableStyles directly maps the tableStyles element. This element
@@ -269,7 +270,7 @@ type xlsxTableStyles struct {
 	Count             int               `xml:"count,attr"`
 	DefaultPivotStyle string            `xml:"defaultPivotStyle,attr"`
 	DefaultTableStyle string            `xml:"defaultTableStyle,attr"`
-	TableStyles       []*xlsxTableStyle `xml:"tableStyle,omitempty"`
+	TableStyles       []*xlsxTableStyle `xml:"tableStyle"`
 }
 
 // xlsxTableStyle directly maps the tableStyle element. This element represents
@@ -289,80 +290,91 @@ type xlsxTableStyle struct {
 // to format and render the numeric value of a cell.
 type xlsxNumFmts struct {
 	Count  int           `xml:"count,attr"`
-	NumFmt []*xlsxNumFmt `xml:"numFmt,omitempty"`
+	NumFmt []*xlsxNumFmt `xml:"numFmt"`
 }
 
 // xlsxNumFmt directly maps the numFmt element. This element specifies number
 // format properties which indicate how to format and render the numeric value
 // of a cell.
 type xlsxNumFmt struct {
-	NumFmtID   int    `xml:"numFmtId,attr"`
-	FormatCode string `xml:"formatCode,attr,omitempty"`
+	NumFmtID     int    `xml:"numFmtId,attr"`
+	FormatCode   string `xml:"formatCode,attr"`
+	FormatCode16 string `xml:"http://schemas.microsoft.com/office/spreadsheetml/2015/02/main formatCode16,attr,omitempty"`
 }
 
-// xlsxStyleColors directly maps the colors element. Color information
-// associated with this stylesheet. This collection is written whenever the
+// xlsxIndexedColors directly maps the single ARGB entry for the corresponding
+// color index.
+type xlsxIndexedColors struct {
+	RgbColor []xlsxColor `xml:"rgbColor"`
+}
+
+// xlsxStyleColors directly maps the colors' element. Color information
+// associated with this style sheet. This collection is written whenever the
 // legacy color palette has been modified (backwards compatibility settings) or
 // a custom color has been selected while using this workbook.
 type xlsxStyleColors struct {
-	Color string `xml:",innerxml"`
+	IndexedColors *xlsxIndexedColors `xml:"indexedColors"`
+	MruColors     *xlsxInnerXML      `xml:"mruColors"`
 }
 
 // Alignment directly maps the alignment settings of the cells.
 type Alignment struct {
-	Horizontal      string `json:"horizontal"`
-	Indent          int    `json:"indent"`
-	JustifyLastLine bool   `json:"justify_last_line"`
-	ReadingOrder    uint64 `json:"reading_order"`
-	RelativeIndent  int    `json:"relative_indent"`
-	ShrinkToFit     bool   `json:"shrink_to_fit"`
-	TextRotation    int    `json:"text_rotation"`
-	Vertical        string `json:"vertical"`
-	WrapText        bool   `json:"wrap_text"`
+	Horizontal      string
+	Indent          int
+	JustifyLastLine bool
+	ReadingOrder    uint64
+	RelativeIndent  int
+	ShrinkToFit     bool
+	TextRotation    int
+	Vertical        string
+	WrapText        bool
 }
 
 // Border directly maps the border settings of the cells.
 type Border struct {
-	Type  string `json:"type"`
-	Color string `json:"color"`
-	Style int    `json:"style"`
+	Type  string
+	Color string
+	Style int
 }
 
 // Font directly maps the font settings of the fonts.
 type Font struct {
-	Bold      bool    `json:"bold"`
-	Italic    bool    `json:"italic"`
-	Underline string  `json:"underline"`
-	Family    string  `json:"family"`
-	Size      float64 `json:"size"`
-	Strike    bool    `json:"strike"`
-	Color     string  `json:"color"`
+	Bold         bool
+	Italic       bool
+	Underline    string
+	Family       string
+	Size         float64
+	Strike       bool
+	Color        string
+	ColorIndexed int
+	ColorTheme   *int
+	ColorTint    float64
+	VertAlign    string
 }
 
 // Fill directly maps the fill settings of the cells.
 type Fill struct {
-	Type    string   `json:"type"`
-	Pattern int      `json:"pattern"`
-	Color   []string `json:"color"`
-	Shading int      `json:"shading"`
+	Type    string
+	Pattern int
+	Color   []string
+	Shading int
 }
 
 // Protection directly maps the protection settings of the cells.
 type Protection struct {
-	Hidden bool `json:"hidden"`
-	Locked bool `json:"locked"`
+	Hidden bool
+	Locked bool
 }
 
 // Style directly maps the style settings of the cells.
 type Style struct {
-	Border        []Border    `json:"border"`
-	Fill          Fill        `json:"fill"`
-	Font          *Font       `json:"font"`
-	Alignment     *Alignment  `json:"alignment"`
-	Protection    *Protection `json:"protection"`
-	NumFmt        int         `json:"number_format"`
-	DecimalPlaces int         `json:"decimal_places"`
-	CustomNumFmt  *string     `json:"custom_number_format"`
-	Lang          string      `json:"lang"`
-	NegRed        bool        `json:"negred"`
+	Border        []Border
+	Fill          Fill
+	Font          *Font
+	Alignment     *Alignment
+	Protection    *Protection
+	NumFmt        int
+	DecimalPlaces *int
+	CustomNumFmt  *string
+	NegRed        bool
 }
