@@ -1,42 +1,48 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
 import "encoding/xml"
 
 // xlsxTable directly maps the table element. A table helps organize and provide
-// structure to lists of information in a worksheet. Tables have clearly labeled
+// structure to list of information in a worksheet. Tables have clearly labeled
 // columns, rows, and data regions. Tables make it easier for users to sort,
 // analyze, format, manage, add, and delete information. This element is the
 // root element for a table that is not a single cell XML table.
 type xlsxTable struct {
 	XMLName              xml.Name            `xml:"table"`
 	XMLNS                string              `xml:"xmlns,attr"`
-	DataCellStyle        string              `xml:"dataCellStyle,attr,omitempty"`
-	DataDxfID            int                 `xml:"dataDxfId,attr,omitempty"`
-	DisplayName          string              `xml:"displayName,attr,omitempty"`
-	HeaderRowBorderDxfID int                 `xml:"headerRowBorderDxfId,attr,omitempty"`
-	HeaderRowCellStyle   string              `xml:"headerRowCellStyle,attr,omitempty"`
-	HeaderRowCount       int                 `xml:"headerRowCount,attr,omitempty"`
-	HeaderRowDxfID       int                 `xml:"headerRowDxfId,attr,omitempty"`
 	ID                   int                 `xml:"id,attr"`
+	Name                 string              `xml:"name,attr"`
+	DisplayName          string              `xml:"displayName,attr,omitempty"`
+	Comment              string              `xml:"comment,attr,omitempty"`
+	Ref                  string              `xml:"ref,attr"`
+	TableType            string              `xml:"tableType,attr,omitempty"`
+	HeaderRowCount       *int                `xml:"headerRowCount,attr"`
 	InsertRow            bool                `xml:"insertRow,attr,omitempty"`
 	InsertRowShift       bool                `xml:"insertRowShift,attr,omitempty"`
-	Name                 string              `xml:"name,attr"`
-	Published            bool                `xml:"published,attr,omitempty"`
-	Ref                  string              `xml:"ref,attr"`
 	TotalsRowCount       int                 `xml:"totalsRowCount,attr,omitempty"`
+	TotalsRowShown       *bool               `xml:"totalsRowShown,attr"`
+	Published            bool                `xml:"published,attr,omitempty"`
+	HeaderRowDxfID       int                 `xml:"headerRowDxfId,attr,omitempty"`
+	DataDxfID            int                 `xml:"dataDxfId,attr,omitempty"`
 	TotalsRowDxfID       int                 `xml:"totalsRowDxfId,attr,omitempty"`
-	TotalsRowShown       bool                `xml:"totalsRowShown,attr"`
+	HeaderRowBorderDxfID int                 `xml:"headerRowBorderDxfId,attr,omitempty"`
+	TableBorderDxfID     int                 `xml:"tableBorderDxfId,attr,omitempty"`
+	TotalsRowBorderDxfID int                 `xml:"totalsRowBorderDxfId,attr,omitempty"`
+	HeaderRowCellStyle   string              `xml:"headerRowCellStyle,attr,omitempty"`
+	DataCellStyle        string              `xml:"dataCellStyle,attr,omitempty"`
+	TotalsRowCellStyle   string              `xml:"totalsRowCellStyle,attr,omitempty"`
+	ConnectionID         int                 `xml:"connectionId,attr,omitempty"`
 	AutoFilter           *xlsxAutoFilter     `xml:"autoFilter"`
 	TableColumns         *xlsxTableColumns   `xml:"tableColumns"`
 	TableStyleInfo       *xlsxTableStyleInfo `xml:"tableStyleInfo"`
@@ -171,18 +177,18 @@ type xlsxTableColumns struct {
 // xlsxTableColumn directly maps the element representing a single column for
 // this table.
 type xlsxTableColumn struct {
-	DataCellStyle      string `xml:"dataCellStyle,attr,omitempty"`
-	DataDxfID          int    `xml:"dataDxfId,attr,omitempty"`
-	HeaderRowCellStyle string `xml:"headerRowCellStyle,attr,omitempty"`
-	HeaderRowDxfID     int    `xml:"headerRowDxfId,attr,omitempty"`
 	ID                 int    `xml:"id,attr"`
+	UniqueName         string `xml:"uniqueName,attr,omitempty"`
 	Name               string `xml:"name,attr"`
-	QueryTableFieldID  int    `xml:"queryTableFieldId,attr,omitempty"`
-	TotalsRowCellStyle string `xml:"totalsRowCellStyle,attr,omitempty"`
-	TotalsRowDxfID     int    `xml:"totalsRowDxfId,attr,omitempty"`
 	TotalsRowFunction  string `xml:"totalsRowFunction,attr,omitempty"`
 	TotalsRowLabel     string `xml:"totalsRowLabel,attr,omitempty"`
-	UniqueName         string `xml:"uniqueName,attr,omitempty"`
+	QueryTableFieldID  int    `xml:"queryTableFieldId,attr,omitempty"`
+	HeaderRowDxfID     int    `xml:"headerRowDxfId,attr,omitempty"`
+	DataDxfID          int    `xml:"dataDxfId,attr,omitempty"`
+	TotalsRowDxfID     int    `xml:"totalsRowDxfId,attr,omitempty"`
+	HeaderRowCellStyle string `xml:"headerRowCellStyle,attr,omitempty"`
+	DataCellStyle      string `xml:"dataCellStyle,attr,omitempty"`
+	TotalsRowCellStyle string `xml:"totalsRowCellStyle,attr,omitempty"`
 }
 
 // xlsxTableStyleInfo directly maps the tableStyleInfo element. This element
@@ -196,22 +202,52 @@ type xlsxTableStyleInfo struct {
 	ShowColumnStripes bool   `xml:"showColumnStripes,attr"`
 }
 
-// formatTable directly maps the format settings of the table.
-type formatTable struct {
-	TableName         string `json:"table_name"`
-	TableStyle        string `json:"table_style"`
-	ShowFirstColumn   bool   `json:"show_first_column"`
-	ShowLastColumn    bool   `json:"show_last_column"`
-	ShowRowStripes    bool   `json:"show_row_stripes"`
-	ShowColumnStripes bool   `json:"show_column_stripes"`
+// xlsxSingleXMLCells is a single cell table is generated from an XML mapping.
+// These really just look like regular cells to the spreadsheet user, but shall
+// be implemented as Tables "under the covers."
+type xlsxSingleXMLCells struct {
+	XMLName       xml.Name            `xml:"singleXmlCells"`
+	SingleXmlCell []xlsxSingleXMLCell `xml:"singleXmlCell"`
 }
 
-// formatAutoFilter directly maps the auto filter settings.
-type formatAutoFilter struct {
-	Column     string `json:"column"`
-	Expression string `json:"expression"`
-	FilterList []struct {
-		Column string `json:"column"`
-		Value  []int  `json:"value"`
-	} `json:"filter_list"`
+// xlsxSingleXMLCell is a element represents the table properties for a single
+// cell XML table.
+type xlsxSingleXMLCell struct {
+	XMLName      xml.Name      `xml:"singleXmlCell"`
+	ID           int           `xml:"id,attr"`
+	R            string        `xml:"r,attr"`
+	ConnectionID int           `xml:"connectionId,attr"`
+	XMLCellPr    xlsxXMLCellPr `xml:"xmlCellPr"`
+	ExtLst       *xlsxInnerXML `xml:"extLst"`
+}
+
+// xlsxXMLCellPr is a element stores the XML properties for the cell of a single
+// cell xml table.
+type xlsxXMLCellPr struct {
+	XMLName    xml.Name      `xml:"xmlCellPr"`
+	ID         int           `xml:"id,attr"`
+	UniqueName string        `xml:"uniqueName,attr,omitempty"`
+	XMLPr      *xlsxInnerXML `xml:"xmlPr"`
+	ExtLst     *xlsxInnerXML `xml:"extLst"`
+}
+
+// Table directly maps the format settings of the table.
+type Table struct {
+	tID               int
+	rID               string
+	tableXML          string
+	Range             string
+	Name              string
+	StyleName         string
+	ShowColumnStripes bool
+	ShowFirstColumn   bool
+	ShowHeaderRow     *bool
+	ShowLastColumn    bool
+	ShowRowStripes    *bool
+}
+
+// AutoFilterOptions directly maps the auto filter settings.
+type AutoFilterOptions struct {
+	Column     string
+	Expression string
 }
