@@ -1,13 +1,13 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
@@ -16,10 +16,57 @@ import "encoding/xml"
 // xlsxTheme directly maps the theme element in the namespace
 // http://schemas.openxmlformats.org/drawingml/2006/main
 type xlsxTheme struct {
-	ThemeElements     xlsxThemeElements     `xml:"themeElements"`
-	ObjectDefaults    xlsxObjectDefaults    `xml:"objectDefaults"`
-	ExtraClrSchemeLst xlsxExtraClrSchemeLst `xml:"extraClrSchemeLst"`
-	ExtLst            *xlsxExtLst           `xml:"extLst"`
+	XMLName           xml.Name              `xml:"a:theme"`
+	XMLNSa            string                `xml:"xmlns:a,attr"`
+	XMLNSr            string                `xml:"xmlns:r,attr"`
+	Name              string                `xml:"name,attr"`
+	ThemeElements     xlsxBaseStyles        `xml:"a:themeElements"`
+	ObjectDefaults    xlsxObjectDefaults    `xml:"a:objectDefaults"`
+	ExtraClrSchemeLst xlsxExtraClrSchemeLst `xml:"a:extraClrSchemeLst"`
+	CustClrLst        *xlsxInnerXML         `xml:"a:custClrLst"`
+	ExtLst            *xlsxExtLst           `xml:"a:extLst"`
+}
+
+// xlsxBaseStyles defines the theme elements for a theme, and is the workhorse
+// of the theme. The bulk of the shared theme information that is used by a
+// given document is defined here. Within this complex type is defined a color
+// scheme, a font scheme, and a style matrix (format scheme) that defines
+// different formatting options for different pieces of a document.
+type xlsxBaseStyles struct {
+	ClrScheme  xlsxColorScheme `xml:"a:clrScheme"`
+	FontScheme xlsxFontScheme  `xml:"a:fontScheme"`
+	FmtScheme  xlsxStyleMatrix `xml:"a:fmtScheme"`
+	ExtLst     *xlsxExtLst     `xml:"a:extLst"`
+}
+
+// xlsxCTColor holds the actual color values that are to be applied to a given
+// diagram and how those colors are to be applied.
+type xlsxCTColor struct {
+	ScrgbClr  *xlsxInnerXML  `xml:"a:scrgbClr"`
+	SrgbClr   *attrValString `xml:"a:srgbClr"`
+	HslClr    *xlsxInnerXML  `xml:"a:hslClr"`
+	SysClr    *xlsxSysClr    `xml:"a:sysClr"`
+	SchemeClr *xlsxInnerXML  `xml:"a:schemeClr"`
+	PrstClr   *xlsxInnerXML  `xml:"a:prstClr"`
+}
+
+// xlsxColorScheme defines a set of colors for the theme. The set of colors
+// consists of twelve color slots that can each hold a color of choice.
+type xlsxColorScheme struct {
+	Name     string      `xml:"name,attr"`
+	Dk1      xlsxCTColor `xml:"a:dk1"`
+	Lt1      xlsxCTColor `xml:"a:lt1"`
+	Dk2      xlsxCTColor `xml:"a:dk2"`
+	Lt2      xlsxCTColor `xml:"a:lt2"`
+	Accent1  xlsxCTColor `xml:"a:accent1"`
+	Accent2  xlsxCTColor `xml:"a:accent2"`
+	Accent3  xlsxCTColor `xml:"a:accent3"`
+	Accent4  xlsxCTColor `xml:"a:accent4"`
+	Accent5  xlsxCTColor `xml:"a:accent5"`
+	Accent6  xlsxCTColor `xml:"a:accent6"`
+	Hlink    xlsxCTColor `xml:"a:hlink"`
+	FolHlink xlsxCTColor `xml:"a:folHlink"`
+	ExtLst   *xlsxExtLst `xml:"a:extLst"`
 }
 
 // objectDefaults element allows for the definition of default shape, line,
@@ -35,24 +82,24 @@ type xlsxExtraClrSchemeLst struct {
 	ExtraClrSchemeLst string `xml:",innerxml"`
 }
 
-// xlsxThemeElements directly maps the element defines the theme formatting
-// options for the theme and is the workhorse of the theme. This is where the
-// bulk of the shared theme information is contained and used by a document.
-// This element contains the color scheme, font scheme, and format scheme
-// elements which define the different formatting aspects of what a theme
-// defines.
-type xlsxThemeElements struct {
-	ClrScheme  xlsxClrScheme  `xml:"clrScheme"`
-	FontScheme xlsxFontScheme `xml:"fontScheme"`
-	FmtScheme  xlsxFmtScheme  `xml:"fmtScheme"`
+// xlsxCTSupplementalFont defines an additional font that is used for language
+// specific fonts in themes. For example, one can specify a font that gets used
+// only within the Japanese language context.
+type xlsxCTSupplementalFont struct {
+	Script   string `xml:"script,attr"`
+	Typeface string `xml:"typeface,attr"`
 }
 
-// xlsxClrScheme element specifies the theme color, stored in the document's
-// Theme part to which the value of this theme color shall be mapped. This
-// mapping enables multiple theme colors to be chained together.
-type xlsxClrScheme struct {
-	Name     string            `xml:"name,attr"`
-	Children []xlsxClrSchemeEl `xml:",any"`
+// xlsxFontCollection defines a major and minor font which is used in the font
+// scheme. A font collection consists of a font definition for Latin, East
+// Asian, and complex script. On top of these three definitions, one can also
+// define a font for use in a specific language or languages.
+type xlsxFontCollection struct {
+	Latin  *xlsxCTTextFont          `xml:"a:latin"`
+	Ea     *xlsxCTTextFont          `xml:"a:ea"`
+	Cs     *xlsxCTTextFont          `xml:"a:cs"`
+	Font   []xlsxCTSupplementalFont `xml:"a:font"`
+	ExtLst *xlsxExtLst              `xml:"a:extLst"`
 }
 
 // xlsxFontScheme element defines the font scheme within the theme. The font
@@ -61,38 +108,23 @@ type xlsxClrScheme struct {
 // document, and the minor font corresponds well with the normal text or
 // paragraph areas.
 type xlsxFontScheme struct {
-	Name      string        `xml:"name,attr"`
-	MajorFont xlsxMajorFont `xml:"majorFont"`
-	MinorFont xlsxMinorFont `xml:"minorFont"`
-	ExtLst    *xlsxExtLst   `xml:"extLst"`
+	Name      string             `xml:"name,attr"`
+	MajorFont xlsxFontCollection `xml:"a:majorFont"`
+	MinorFont xlsxFontCollection `xml:"a:minorFont"`
+	ExtLst    *xlsxExtLst        `xml:"a:extLst"`
 }
 
-// xlsxMajorFont element defines the set of major fonts which are to be used
-// under different languages or locals.
-type xlsxMajorFont struct {
-	Children []xlsxFontSchemeEl `xml:",any"`
-}
-
-// xlsxMinorFont element defines the set of minor fonts that are to be used
-// under different languages or locals.
-type xlsxMinorFont struct {
-	Children []xlsxFontSchemeEl `xml:",any"`
-}
-
-// xlsxFmtScheme element contains the background fill styles, effect styles,
-// fill styles, and line styles which define the style matrix for a theme. The
-// style matrix consists of subtle, moderate, and intense fills, lines, and
-// effects. The background fills are not generally thought of to directly be
-// associated with the matrix, but do play a role in the style of the overall
-// document. Usually, a given object chooses a single line style, a single
-// fill style, and a single effect style in order to define the overall final
-// look of the object.
-type xlsxFmtScheme struct {
-	Name           string             `xml:"name,attr"`
-	FillStyleLst   xlsxFillStyleLst   `xml:"fillStyleLst"`
-	LnStyleLst     xlsxLnStyleLst     `xml:"lnStyleLst"`
-	EffectStyleLst xlsxEffectStyleLst `xml:"effectStyleLst"`
-	BgFillStyleLst xlsxBgFillStyleLst `xml:"bgFillStyleLst"`
+// xlsxStyleMatrix defines a set of formatting options, which can be referenced
+// by documents that apply a certain style to a given part of an object. For
+// example, in a given shape, say a rectangle, one can reference a themed line
+// style, themed effect, and themed fill that would be theme specific and
+// change when the theme is changed.
+type xlsxStyleMatrix struct {
+	Name           string             `xml:"name,attr,omitempty"`
+	FillStyleLst   xlsxFillStyleLst   `xml:"a:fillStyleLst"`
+	LnStyleLst     xlsxLnStyleLst     `xml:"a:lnStyleLst"`
+	EffectStyleLst xlsxEffectStyleLst `xml:"a:effectStyleLst"`
+	BgFillStyleLst xlsxBgFillStyleLst `xml:"a:bgFillStyleLst"`
 }
 
 // xlsxFillStyleLst element defines a set of three fill styles that are used
@@ -116,31 +148,11 @@ type xlsxEffectStyleLst struct {
 	EffectStyleLst string `xml:",innerxml"`
 }
 
-// xlsxBgFillStyleLst  element defines a list of background fills that are
+// xlsxBgFillStyleLst element defines a list of background fills that are
 // used within a theme. The background fills consist of three fills, arranged
 // in order from subtle to moderate to intense.
 type xlsxBgFillStyleLst struct {
 	BgFillStyleLst string `xml:",innerxml"`
-}
-
-// xlsxClrScheme specifies the theme color, stored in the document's Theme
-// part to which the value of this theme color shall be mapped. This mapping
-// enables multiple theme colors to be chained together.
-type xlsxClrSchemeEl struct {
-	XMLName xml.Name
-	SysClr  *xlsxSysClr    `xml:"sysClr"`
-	SrgbClr *attrValString `xml:"srgbClr"`
-}
-
-// xlsxFontSchemeEl directly maps the major and minor font of the style's font
-// scheme.
-type xlsxFontSchemeEl struct {
-	XMLName     xml.Name
-	Script      string `xml:"script,attr,omitempty"`
-	Typeface    string `xml:"typeface,attr"`
-	Panose      string `xml:"panose,attr,omitempty"`
-	PitchFamily string `xml:"pitchFamily,attr,omitempty"`
-	Charset     string `xml:"charset,attr,omitempty"`
 }
 
 // xlsxSysClr element specifies a color bound to predefined operating system
@@ -148,4 +160,86 @@ type xlsxFontSchemeEl struct {
 type xlsxSysClr struct {
 	Val     string `xml:"val,attr"`
 	LastClr string `xml:"lastClr,attr"`
+}
+
+// decodeTheme defines the structure used to parse the a:theme element for the
+// theme.
+type decodeTheme struct {
+	XMLName           xml.Name              `xml:"http://schemas.openxmlformats.org/drawingml/2006/main theme"`
+	Name              string                `xml:"name,attr"`
+	ThemeElements     decodeBaseStyles      `xml:"themeElements"`
+	ObjectDefaults    xlsxObjectDefaults    `xml:"objectDefaults"`
+	ExtraClrSchemeLst xlsxExtraClrSchemeLst `xml:"extraClrSchemeLst"`
+	CustClrLst        *xlsxInnerXML         `xml:"custClrLst"`
+	ExtLst            *xlsxExtLst           `xml:"extLst"`
+}
+
+// decodeBaseStyles defines the structure used to parse the theme elements for a
+// theme, and is the workhorse of the theme.
+type decodeBaseStyles struct {
+	ClrScheme  decodeColorScheme `xml:"clrScheme"`
+	FontScheme decodeFontScheme  `xml:"fontScheme"`
+	FmtScheme  decodeStyleMatrix `xml:"fmtScheme"`
+	ExtLst     *xlsxExtLst       `xml:"extLst"`
+}
+
+// decodeColorScheme defines the structure used to parse a set of colors for the
+// theme.
+type decodeColorScheme struct {
+	Name     string        `xml:"name,attr"`
+	Dk1      decodeCTColor `xml:"dk1"`
+	Lt1      decodeCTColor `xml:"lt1"`
+	Dk2      decodeCTColor `xml:"dk2"`
+	Lt2      decodeCTColor `xml:"lt2"`
+	Accent1  decodeCTColor `xml:"accent1"`
+	Accent2  decodeCTColor `xml:"accent2"`
+	Accent3  decodeCTColor `xml:"accent3"`
+	Accent4  decodeCTColor `xml:"accent4"`
+	Accent5  decodeCTColor `xml:"accent5"`
+	Accent6  decodeCTColor `xml:"accent6"`
+	Hlink    decodeCTColor `xml:"hlink"`
+	FolHlink decodeCTColor `xml:"folHlink"`
+	ExtLst   *xlsxExtLst   `xml:"extLst"`
+}
+
+// decodeFontScheme defines the structure used to parse font scheme within the
+// theme.
+type decodeFontScheme struct {
+	Name      string               `xml:"name,attr"`
+	MajorFont decodeFontCollection `xml:"majorFont"`
+	MinorFont decodeFontCollection `xml:"minorFont"`
+	ExtLst    *xlsxExtLst          `xml:"extLst"`
+}
+
+// decodeFontCollection defines the structure used to parse a major and minor
+// font which is used in the font scheme.
+type decodeFontCollection struct {
+	Latin  *xlsxCTTextFont          `xml:"latin"`
+	Ea     *xlsxCTTextFont          `xml:"ea"`
+	Cs     *xlsxCTTextFont          `xml:"cs"`
+	Font   []xlsxCTSupplementalFont `xml:"font"`
+	ExtLst *xlsxExtLst              `xml:"extLst"`
+}
+
+// decodeCTColor defines the structure used to parse the actual color values
+// that are to be applied to a given diagram and how those colors are to be
+// applied.
+type decodeCTColor struct {
+	ScrgbClr  *xlsxInnerXML  `xml:"scrgbClr"`
+	SrgbClr   *attrValString `xml:"srgbClr"`
+	HslClr    *xlsxInnerXML  `xml:"hslClr"`
+	SysClr    *xlsxSysClr    `xml:"sysClr"`
+	SchemeClr *xlsxInnerXML  `xml:"schemeClr"`
+	PrstClr   *xlsxInnerXML  `xml:"prstClr"`
+}
+
+// decodeStyleMatrix defines the structure used to parse a set of formatting
+// options, which can be referenced by documents that apply a certain style to
+// a given part of an object.
+type decodeStyleMatrix struct {
+	Name           string             `xml:"name,attr,omitempty"`
+	FillStyleLst   xlsxFillStyleLst   `xml:"fillStyleLst"`
+	LnStyleLst     xlsxLnStyleLst     `xml:"lnStyleLst"`
+	EffectStyleLst xlsxEffectStyleLst `xml:"effectStyleLst"`
+	BgFillStyleLst xlsxBgFillStyleLst `xml:"bgFillStyleLst"`
 }
