@@ -1,20 +1,25 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
-import "encoding/xml"
+import (
+	"encoding/xml"
+	"sync"
+)
 
-// xlsxRelationships describe references from parts to other internal resources in the package or to external resources.
+// xlsxRelationships describe references from parts to other internal resources
+// in the package or to external resources.
 type xlsxRelationships struct {
+	mu            sync.Mutex
 	XMLName       xml.Name           `xml:"http://schemas.openxmlformats.org/package/2006/relationships Relationships"`
 	Relationships []xlsxRelationship `xml:"Relationship"`
 }
@@ -31,27 +36,29 @@ type xlsxRelationship struct {
 // content of the workbook. The workbook's child elements each have their own
 // subclause references.
 type xlsxWorkbook struct {
-	XMLName             xml.Name                 `xml:"http://schemas.openxmlformats.org/spreadsheetml/2006/main workbook"`
-	Conformance         string                   `xml:"conformance,attr,omitempty"`
-	FileVersion         *xlsxFileVersion         `xml:"fileVersion"`
-	FileSharing         *xlsxExtLst              `xml:"fileSharing"`
-	WorkbookPr          *xlsxWorkbookPr          `xml:"workbookPr"`
-	WorkbookProtection  *xlsxWorkbookProtection  `xml:"workbookProtection"`
-	BookViews           *xlsxBookViews           `xml:"bookViews"`
-	Sheets              xlsxSheets               `xml:"sheets"`
-	FunctionGroups      *xlsxExtLst              `xml:"functionGroups"`
-	ExternalReferences  *xlsxExternalReferences  `xml:"externalReferences"`
-	DefinedNames        *xlsxDefinedNames        `xml:"definedNames"`
-	CalcPr              *xlsxCalcPr              `xml:"calcPr"`
-	OleSize             *xlsxExtLst              `xml:"oleSize"`
-	CustomWorkbookViews *xlsxCustomWorkbookViews `xml:"customWorkbookViews"`
-	PivotCaches         *xlsxPivotCaches         `xml:"pivotCaches"`
-	SmartTagPr          *xlsxExtLst              `xml:"smartTagPr"`
-	SmartTagTypes       *xlsxExtLst              `xml:"smartTagTypes"`
-	WebPublishing       *xlsxExtLst              `xml:"webPublishing"`
-	FileRecoveryPr      *xlsxFileRecoveryPr      `xml:"fileRecoveryPr"`
-	WebPublishObjects   *xlsxExtLst              `xml:"webPublishObjects"`
-	ExtLst              *xlsxExtLst              `xml:"extLst"`
+	XMLName                xml.Name                 `xml:"http://schemas.openxmlformats.org/spreadsheetml/2006/main workbook"`
+	Conformance            string                   `xml:"conformance,attr,omitempty"`
+	FileVersion            *xlsxFileVersion         `xml:"fileVersion"`
+	FileSharing            *xlsxExtLst              `xml:"fileSharing"`
+	WorkbookPr             *xlsxWorkbookPr          `xml:"workbookPr"`
+	AlternateContent       *xlsxAlternateContent    `xml:"mc:AlternateContent"`
+	DecodeAlternateContent *xlsxInnerXML            `xml:"http://schemas.openxmlformats.org/markup-compatibility/2006 AlternateContent"`
+	WorkbookProtection     *xlsxWorkbookProtection  `xml:"workbookProtection"`
+	BookViews              *xlsxBookViews           `xml:"bookViews"`
+	Sheets                 xlsxSheets               `xml:"sheets"`
+	FunctionGroups         *xlsxFunctionGroups      `xml:"functionGroups"`
+	ExternalReferences     *xlsxExternalReferences  `xml:"externalReferences"`
+	DefinedNames           *xlsxDefinedNames        `xml:"definedNames"`
+	CalcPr                 *xlsxCalcPr              `xml:"calcPr"`
+	OleSize                *xlsxExtLst              `xml:"oleSize"`
+	CustomWorkbookViews    *xlsxCustomWorkbookViews `xml:"customWorkbookViews"`
+	PivotCaches            *xlsxPivotCaches         `xml:"pivotCaches"`
+	SmartTagPr             *xlsxExtLst              `xml:"smartTagPr"`
+	SmartTagTypes          *xlsxExtLst              `xml:"smartTagTypes"`
+	WebPublishing          *xlsxExtLst              `xml:"webPublishing"`
+	FileRecoveryPr         *xlsxFileRecoveryPr      `xml:"fileRecoveryPr"`
+	WebPublishObjects      *xlsxExtLst              `xml:"webPublishObjects"`
+	ExtLst                 *xlsxExtLst              `xml:"extLst"`
 }
 
 // xlsxFileRecoveryPr maps sheet recovery information. This element defines
@@ -103,24 +110,24 @@ type xlsxFileVersion struct {
 // http://schemas.openxmlformats.org/spreadsheetml/2006/main This element
 // defines a collection of workbook properties.
 type xlsxWorkbookPr struct {
-	AllowRefreshQuery          bool   `xml:"allowRefreshQuery,attr,omitempty"`
-	AutoCompressPictures       bool   `xml:"autoCompressPictures,attr,omitempty"`
-	BackupFile                 bool   `xml:"backupFile,attr,omitempty"`
-	CheckCompatibility         bool   `xml:"checkCompatibility,attr,omitempty"`
-	CodeName                   string `xml:"codeName,attr,omitempty"`
 	Date1904                   bool   `xml:"date1904,attr,omitempty"`
-	DefaultThemeVersion        string `xml:"defaultThemeVersion,attr,omitempty"`
-	FilterPrivacy              bool   `xml:"filterPrivacy,attr,omitempty"`
-	HidePivotFieldList         bool   `xml:"hidePivotFieldList,attr,omitempty"`
-	PromptedSolutions          bool   `xml:"promptedSolutions,attr,omitempty"`
-	PublishItems               bool   `xml:"publishItems,attr,omitempty"`
-	RefreshAllConnections      bool   `xml:"refreshAllConnections,attr,omitempty"`
-	SaveExternalLinkValues     bool   `xml:"saveExternalLinkValues,attr,omitempty"`
-	ShowBorderUnselectedTables bool   `xml:"showBorderUnselectedTables,attr,omitempty"`
-	ShowInkAnnotation          bool   `xml:"showInkAnnotation,attr,omitempty"`
 	ShowObjects                string `xml:"showObjects,attr,omitempty"`
-	ShowPivotChartFilter       bool   `xml:"showPivotChartFilter,attr,omitempty"`
+	ShowBorderUnselectedTables *bool  `xml:"showBorderUnselectedTables,attr"`
+	FilterPrivacy              bool   `xml:"filterPrivacy,attr,omitempty"`
+	PromptedSolutions          bool   `xml:"promptedSolutions,attr,omitempty"`
+	ShowInkAnnotation          *bool  `xml:"showInkAnnotation,attr"`
+	BackupFile                 bool   `xml:"backupFile,attr,omitempty"`
+	SaveExternalLinkValues     *bool  `xml:"saveExternalLinkValues,attr"`
 	UpdateLinks                string `xml:"updateLinks,attr,omitempty"`
+	CodeName                   string `xml:"codeName,attr,omitempty"`
+	HidePivotFieldList         bool   `xml:"hidePivotFieldList,attr,omitempty"`
+	ShowPivotChartFilter       bool   `xml:"showPivotChartFilter,attr,omitempty"`
+	AllowRefreshQuery          bool   `xml:"allowRefreshQuery,attr,omitempty"`
+	PublishItems               bool   `xml:"publishItems,attr,omitempty"`
+	CheckCompatibility         bool   `xml:"checkCompatibility,attr,omitempty"`
+	AutoCompressPictures       *bool  `xml:"autoCompressPictures,attr"`
+	RefreshAllConnections      bool   `xml:"refreshAllConnections,attr,omitempty"`
+	DefaultThemeVersion        string `xml:"defaultThemeVersion,attr,omitempty"`
 }
 
 // xlsxBookViews directly maps the bookViews element. This element specifies the
@@ -135,19 +142,19 @@ type xlsxBookViews struct {
 // http://schemas.openxmlformats.org/spreadsheetml/2006/main This element
 // specifies a single Workbook view.
 type xlsxWorkBookView struct {
-	ActiveTab              int    `xml:"activeTab,attr,omitempty"`
-	AutoFilterDateGrouping bool   `xml:"autoFilterDateGrouping,attr,omitempty"`
-	FirstSheet             int    `xml:"firstSheet,attr,omitempty"`
-	Minimized              bool   `xml:"minimized,attr,omitempty"`
-	ShowHorizontalScroll   bool   `xml:"showHorizontalScroll,attr,omitempty"`
-	ShowSheetTabs          bool   `xml:"showSheetTabs,attr,omitempty"`
-	ShowVerticalScroll     bool   `xml:"showVerticalScroll,attr,omitempty"`
-	TabRatio               int    `xml:"tabRatio,attr,omitempty"`
-	Visibility             string `xml:"visibility,attr,omitempty"`
-	WindowHeight           int    `xml:"windowHeight,attr,omitempty"`
-	WindowWidth            int    `xml:"windowWidth,attr,omitempty"`
-	XWindow                string `xml:"xWindow,attr,omitempty"`
-	YWindow                string `xml:"yWindow,attr,omitempty"`
+	Visibility             string  `xml:"visibility,attr,omitempty"`
+	Minimized              bool    `xml:"minimized,attr,omitempty"`
+	ShowHorizontalScroll   *bool   `xml:"showHorizontalScroll,attr"`
+	ShowVerticalScroll     *bool   `xml:"showVerticalScroll,attr"`
+	ShowSheetTabs          *bool   `xml:"showSheetTabs,attr"`
+	XWindow                string  `xml:"xWindow,attr,omitempty"`
+	YWindow                string  `xml:"yWindow,attr,omitempty"`
+	WindowWidth            int     `xml:"windowWidth,attr,omitempty"`
+	WindowHeight           int     `xml:"windowHeight,attr,omitempty"`
+	TabRatio               float64 `xml:"tabRatio,attr,omitempty"`
+	FirstSheet             int     `xml:"firstSheet,attr,omitempty"`
+	ActiveTab              int     `xml:"activeTab,attr,omitempty"`
+	AutoFilterDateGrouping *bool   `xml:"autoFilterDateGrouping,attr"`
 }
 
 // xlsxSheets directly maps the sheets element from the namespace
@@ -163,6 +170,17 @@ type xlsxSheet struct {
 	SheetID int    `xml:"sheetId,attr,omitempty"`
 	ID      string `xml:"http://schemas.openxmlformats.org/officeDocument/2006/relationships id,attr"`
 	State   string `xml:"state,attr,omitempty"`
+}
+
+// xlsxFunctionGroup represents a single function group.
+type xlsxFunctionGroup struct {
+	Name string `xml:"name,attr"`
+}
+
+// xlsxFunctionGroups defines the collection of function groups for the workbook.
+type xlsxFunctionGroups struct {
+	BuiltInGroupCount *int                `xml:"builtInGroupCount,attr"`
+	FunctionGroup     []xlsxFunctionGroup `xml:"functionGroup"`
 }
 
 // xlsxExternalReferences directly maps the externalReferences element of the
@@ -195,10 +213,68 @@ type xlsxPivotCache struct {
 // document are specified in the markup specification and can be used to store
 // extensions to the markup specification, whether those are future version
 // extensions of the markup specification or are private extensions implemented
-// independently from the markup specification. Markup within an extension might
+// independently of the markup specification. Markup within an extension might
 // not be understood by a consumer.
 type xlsxExtLst struct {
 	Ext string `xml:",innerxml"`
+}
+
+// xlsxExt represents a the future feature data storage area. Each extension
+// within an extension list shall be contained within an ext element.
+// Extensions shall be versioned by namespace, using the uri attribute, and
+// shall be allowed to appear in any order within the extension list. Any
+// number of extensions shall be allowed within an extension list.
+type xlsxExt struct {
+	XMLName xml.Name `xml:"ext"`
+	URI     string   `xml:"uri,attr"`
+	Content string   `xml:",innerxml"`
+	xmlns   []xml.Attr
+}
+
+// xlsxAlternateContent is a container for a sequence of multiple
+// representations of a given piece of content. The program reading the file
+// should only process one of these, and the one chosen should be based on
+// which conditions match.
+type xlsxAlternateContent struct {
+	XMLNSMC string `xml:"xmlns:mc,attr,omitempty"`
+	Content string `xml:",innerxml"`
+}
+
+// xlsxChoice element shall be an element in the Markup Compatibility namespace
+// with local name "Choice". Parent elements of Choice elements shall be
+// AlternateContent elements.
+type xlsxChoice struct {
+	XMLName    xml.Name `xml:"mc:Choice"`
+	XMLNSA14   string   `xml:"xmlns:a14,attr,omitempty"`
+	XMLNSSle15 string   `xml:"xmlns:sle15,attr,omitempty"`
+	Requires   string   `xml:"Requires,attr,omitempty"`
+	Content    string   `xml:",innerxml"`
+}
+
+// xlsxFallback element shall be an element in the Markup Compatibility
+// namespace with local name "Fallback". Parent elements of Fallback elements
+// shall be AlternateContent elements.
+type xlsxFallback struct {
+	XMLName xml.Name `xml:"mc:Fallback"`
+	Content string   `xml:",innerxml"`
+}
+
+// xlsxInnerXML holds parts of XML content currently not unmarshal.
+type xlsxInnerXML struct {
+	Content string `xml:",innerxml"`
+}
+
+// decodeExtLst defines the structure used to parse the extLst element
+// of the future feature data storage area.
+type decodeExtLst struct {
+	XMLName xml.Name   `xml:"extLst"`
+	Ext     []*xlsxExt `xml:"ext"`
+}
+
+// decodeExt defines the structure used to parse the ext element.
+type decodeExt struct {
+	URI     string `xml:"uri,attr,omitempty"`
+	Content string `xml:",innerxml"`
 }
 
 // xlsxDefinedNames directly maps the definedNames element. This element defines
@@ -212,8 +288,8 @@ type xlsxDefinedNames struct {
 // xlsxDefinedName directly maps the definedName element from the namespace
 // http://schemas.openxmlformats.org/spreadsheetml/2006/main This element
 // defines a defined name within this workbook. A defined name is descriptive
-// text that is used to represents a cell, range of cells, formula, or constant
-// value. For a descriptions of the attributes see https://docs.microsoft.com/en-us/dotnet/api/documentformat.openxml.spreadsheet.definedname
+// text that is used to represent a cell, range of cells, formula, or constant
+// value. For a descriptions of the attributes see https://learn.microsoft.com/en-us/dotnet/api/documentformat.openxml.spreadsheet.definedname
 type xlsxDefinedName struct {
 	Comment           string `xml:"comment,attr,omitempty"`
 	CustomMenu        string `xml:"customMenu,attr,omitempty"`
@@ -273,30 +349,30 @@ type xlsxCustomWorkbookViews struct {
 // to implement configurable display modes, the customWorkbookView element
 // should be used to persist the settings for those display modes.
 type xlsxCustomWorkbookView struct {
-	ActiveSheetID        *int    `xml:"activeSheetId,attr"`
-	AutoUpdate           *bool   `xml:"autoUpdate,attr"`
-	ChangesSavedWin      *bool   `xml:"changesSavedWin,attr"`
-	GUID                 *string `xml:"guid,attr"`
-	IncludeHiddenRowCol  *bool   `xml:"includeHiddenRowCol,attr"`
-	IncludePrintSettings *bool   `xml:"includePrintSettings,attr"`
-	Maximized            *bool   `xml:"maximized,attr"`
-	MergeInterval        int     `xml:"mergeInterval,attr"`
-	Minimized            *bool   `xml:"minimized,attr"`
-	Name                 *string `xml:"name,attr"`
-	OnlySync             *bool   `xml:"onlySync,attr"`
-	PersonalView         *bool   `xml:"personalView,attr"`
-	ShowComments         *string `xml:"showComments,attr"`
-	ShowFormulaBar       *bool   `xml:"showFormulaBar,attr"`
-	ShowHorizontalScroll *bool   `xml:"showHorizontalScroll,attr"`
-	ShowObjects          *string `xml:"showObjects,attr"`
-	ShowSheetTabs        *bool   `xml:"showSheetTabs,attr"`
-	ShowStatusbar        *bool   `xml:"showStatusbar,attr"`
-	ShowVerticalScroll   *bool   `xml:"showVerticalScroll,attr"`
-	TabRatio             *int    `xml:"tabRatio,attr"`
-	WindowHeight         *int    `xml:"windowHeight,attr"`
-	WindowWidth          *int    `xml:"windowWidth,attr"`
-	XWindow              *int    `xml:"xWindow,attr"`
-	YWindow              *int    `xml:"yWindow,attr"`
+	ActiveSheetID        *int     `xml:"activeSheetId,attr"`
+	AutoUpdate           *bool    `xml:"autoUpdate,attr"`
+	ChangesSavedWin      *bool    `xml:"changesSavedWin,attr"`
+	GUID                 *string  `xml:"guid,attr"`
+	IncludeHiddenRowCol  *bool    `xml:"includeHiddenRowCol,attr"`
+	IncludePrintSettings *bool    `xml:"includePrintSettings,attr"`
+	Maximized            *bool    `xml:"maximized,attr"`
+	MergeInterval        int      `xml:"mergeInterval,attr"`
+	Minimized            *bool    `xml:"minimized,attr"`
+	Name                 *string  `xml:"name,attr"`
+	OnlySync             *bool    `xml:"onlySync,attr"`
+	PersonalView         *bool    `xml:"personalView,attr"`
+	ShowComments         *string  `xml:"showComments,attr"`
+	ShowFormulaBar       *bool    `xml:"showFormulaBar,attr"`
+	ShowHorizontalScroll *bool    `xml:"showHorizontalScroll,attr"`
+	ShowObjects          *string  `xml:"showObjects,attr"`
+	ShowSheetTabs        *bool    `xml:"showSheetTabs,attr"`
+	ShowStatusbar        *bool    `xml:"showStatusbar,attr"`
+	ShowVerticalScroll   *bool    `xml:"showVerticalScroll,attr"`
+	TabRatio             *float64 `xml:"tabRatio,attr"`
+	WindowHeight         *int     `xml:"windowHeight,attr"`
+	WindowWidth          *int     `xml:"windowWidth,attr"`
+	XWindow              *int     `xml:"xWindow,attr"`
+	YWindow              *int     `xml:"yWindow,attr"`
 }
 
 // DefinedName directly maps the name for a cell or cell range on a
@@ -306,4 +382,19 @@ type DefinedName struct {
 	Comment  string
 	RefersTo string
 	Scope    string
+}
+
+// WorkbookPropsOptions directly maps the settings of workbook proprieties.
+type WorkbookPropsOptions struct {
+	Date1904      *bool
+	FilterPrivacy *bool
+	CodeName      *string
+}
+
+// WorkbookProtectionOptions directly maps the settings of workbook protection.
+type WorkbookProtectionOptions struct {
+	AlgorithmName string
+	Password      string
+	LockStructure bool
+	LockWindows   bool
 }
