@@ -1,13 +1,13 @@
-// Copyright 2016 - 2021 The excelize Authors. All rights reserved. Use of
+// Copyright 2016 - 2025 The excelize Authors. All rights reserved. Use of
 // this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 //
-// Package excelize providing a set of functions that allow you to write to
-// and read from XLSX / XLSM / XLTM files. Supports reading and writing
-// spreadsheet documents generated by Microsoft Excel™ 2007 and later. Supports
-// complex components by high compatibility, and provided streaming API for
-// generating or reading data from a worksheet with huge amounts of data. This
-// library needs Go version 1.15 or later.
+// Package excelize providing a set of functions that allow you to write to and
+// read from XLAM / XLSM / XLSX / XLTM / XLTX files. Supports reading and
+// writing spreadsheet documents generated by Microsoft Excel™ 2007 and later.
+// Supports complex components by high compatibility, and provided streaming
+// API for generating or reading data from a worksheet with huge amounts of
+// data. This library needs Go version 1.20 or later.
 
 package excelize
 
@@ -19,47 +19,49 @@ import (
 // xlsxWorksheet directly maps the worksheet element in the namespace
 // http://schemas.openxmlformats.org/spreadsheetml/2006/main.
 type xlsxWorksheet struct {
-	sync.Mutex
-	XMLName               xml.Name                     `xml:"http://schemas.openxmlformats.org/spreadsheetml/2006/main worksheet"`
-	SheetPr               *xlsxSheetPr                 `xml:"sheetPr"`
-	Dimension             *xlsxDimension               `xml:"dimension"`
-	SheetViews            *xlsxSheetViews              `xml:"sheetViews"`
-	SheetFormatPr         *xlsxSheetFormatPr           `xml:"sheetFormatPr"`
-	Cols                  *xlsxCols                    `xml:"cols"`
-	SheetData             xlsxSheetData                `xml:"sheetData"`
-	SheetCalcPr           *xlsxInnerXML                `xml:"sheetCalcPr"`
-	SheetProtection       *xlsxSheetProtection         `xml:"sheetProtection"`
-	ProtectedRanges       *xlsxInnerXML                `xml:"protectedRanges"`
-	Scenarios             *xlsxInnerXML                `xml:"scenarios"`
-	AutoFilter            *xlsxAutoFilter              `xml:"autoFilter"`
-	SortState             *xlsxSortState               `xml:"sortState"`
-	DataConsolidate       *xlsxInnerXML                `xml:"dataConsolidate"`
-	CustomSheetViews      *xlsxCustomSheetViews        `xml:"customSheetViews"`
-	MergeCells            *xlsxMergeCells              `xml:"mergeCells"`
-	PhoneticPr            *xlsxPhoneticPr              `xml:"phoneticPr"`
-	ConditionalFormatting []*xlsxConditionalFormatting `xml:"conditionalFormatting"`
-	DataValidations       *xlsxDataValidations         `xml:"dataValidations"`
-	Hyperlinks            *xlsxHyperlinks              `xml:"hyperlinks"`
-	PrintOptions          *xlsxPrintOptions            `xml:"printOptions"`
-	PageMargins           *xlsxPageMargins             `xml:"pageMargins"`
-	PageSetUp             *xlsxPageSetUp               `xml:"pageSetup"`
-	HeaderFooter          *xlsxHeaderFooter            `xml:"headerFooter"`
-	RowBreaks             *xlsxBreaks                  `xml:"rowBreaks"`
-	ColBreaks             *xlsxBreaks                  `xml:"colBreaks"`
-	CustomProperties      *xlsxInnerXML                `xml:"customProperties"`
-	CellWatches           *xlsxInnerXML                `xml:"cellWatches"`
-	IgnoredErrors         *xlsxInnerXML                `xml:"ignoredErrors"`
-	SmartTags             *xlsxInnerXML                `xml:"smartTags"`
-	Drawing               *xlsxDrawing                 `xml:"drawing"`
-	LegacyDrawing         *xlsxLegacyDrawing           `xml:"legacyDrawing"`
-	LegacyDrawingHF       *xlsxLegacyDrawingHF         `xml:"legacyDrawingHF"`
-	DrawingHF             *xlsxDrawingHF               `xml:"drawingHF"`
-	Picture               *xlsxPicture                 `xml:"picture"`
-	OleObjects            *xlsxInnerXML                `xml:"oleObjects"`
-	Controls              *xlsxInnerXML                `xml:"controls"`
-	WebPublishItems       *xlsxInnerXML                `xml:"webPublishItems"`
-	TableParts            *xlsxTableParts              `xml:"tableParts"`
-	ExtLst                *xlsxExtLst                  `xml:"extLst"`
+	mu                     sync.Mutex
+	XMLName                xml.Name                     `xml:"http://schemas.openxmlformats.org/spreadsheetml/2006/main worksheet"`
+	SheetPr                *xlsxSheetPr                 `xml:"sheetPr"`
+	Dimension              *xlsxDimension               `xml:"dimension"`
+	SheetViews             *xlsxSheetViews              `xml:"sheetViews"`
+	SheetFormatPr          *xlsxSheetFormatPr           `xml:"sheetFormatPr"`
+	Cols                   *xlsxCols                    `xml:"cols"`
+	SheetData              xlsxSheetData                `xml:"sheetData"`
+	SheetCalcPr            *xlsxInnerXML                `xml:"sheetCalcPr"`
+	SheetProtection        *xlsxSheetProtection         `xml:"sheetProtection"`
+	ProtectedRanges        *xlsxInnerXML                `xml:"protectedRanges"`
+	Scenarios              *xlsxInnerXML                `xml:"scenarios"`
+	AutoFilter             *xlsxAutoFilter              `xml:"autoFilter"`
+	SortState              *xlsxSortState               `xml:"sortState"`
+	DataConsolidate        *xlsxInnerXML                `xml:"dataConsolidate"`
+	CustomSheetViews       *xlsxCustomSheetViews        `xml:"customSheetViews"`
+	MergeCells             *xlsxMergeCells              `xml:"mergeCells"`
+	PhoneticPr             *xlsxPhoneticPr              `xml:"phoneticPr"`
+	ConditionalFormatting  []*xlsxConditionalFormatting `xml:"conditionalFormatting"`
+	DataValidations        *xlsxDataValidations         `xml:"dataValidations"`
+	Hyperlinks             *xlsxHyperlinks              `xml:"hyperlinks"`
+	PrintOptions           *xlsxPrintOptions            `xml:"printOptions"`
+	PageMargins            *xlsxPageMargins             `xml:"pageMargins"`
+	PageSetUp              *xlsxPageSetUp               `xml:"pageSetup"`
+	HeaderFooter           *xlsxHeaderFooter            `xml:"headerFooter"`
+	RowBreaks              *xlsxRowBreaks               `xml:"rowBreaks"`
+	ColBreaks              *xlsxColBreaks               `xml:"colBreaks"`
+	CustomProperties       *xlsxInnerXML                `xml:"customProperties"`
+	CellWatches            *xlsxInnerXML                `xml:"cellWatches"`
+	IgnoredErrors          *xlsxIgnoredErrors           `xml:"ignoredErrors"`
+	SmartTags              *xlsxInnerXML                `xml:"smartTags"`
+	Drawing                *xlsxDrawing                 `xml:"drawing"`
+	LegacyDrawing          *xlsxLegacyDrawing           `xml:"legacyDrawing"`
+	LegacyDrawingHF        *xlsxLegacyDrawingHF         `xml:"legacyDrawingHF"`
+	DrawingHF              *xlsxDrawingHF               `xml:"drawingHF"`
+	Picture                *xlsxPicture                 `xml:"picture"`
+	OleObjects             *xlsxInnerXML                `xml:"oleObjects"`
+	Controls               *xlsxInnerXML                `xml:"controls"`
+	WebPublishItems        *xlsxInnerXML                `xml:"webPublishItems"`
+	AlternateContent       *xlsxAlternateContent        `xml:"mc:AlternateContent"`
+	TableParts             *xlsxTableParts              `xml:"tableParts"`
+	ExtLst                 *xlsxExtLst                  `xml:"extLst"`
+	DecodeAlternateContent *xlsxInnerXML                `xml:"http://schemas.openxmlformats.org/markup-compatibility/2006 AlternateContent"`
 }
 
 // xlsxDrawing change r:id to rid in the namespace.
@@ -76,18 +78,17 @@ type xlsxDrawing struct {
 // footers on the first page can differ from those on odd- and even-numbered
 // pages. In the latter case, the first page is not considered an odd page.
 type xlsxHeaderFooter struct {
-	XMLName          xml.Name       `xml:"headerFooter"`
-	AlignWithMargins bool           `xml:"alignWithMargins,attr,omitempty"`
-	DifferentFirst   bool           `xml:"differentFirst,attr,omitempty"`
-	DifferentOddEven bool           `xml:"differentOddEven,attr,omitempty"`
-	ScaleWithDoc     bool           `xml:"scaleWithDoc,attr,omitempty"`
-	OddHeader        string         `xml:"oddHeader,omitempty"`
-	OddFooter        string         `xml:"oddFooter,omitempty"`
-	EvenHeader       string         `xml:"evenHeader,omitempty"`
-	EvenFooter       string         `xml:"evenFooter,omitempty"`
-	FirstFooter      string         `xml:"firstFooter,omitempty"`
-	FirstHeader      string         `xml:"firstHeader,omitempty"`
-	DrawingHF        *xlsxDrawingHF `xml:"drawingHF"`
+	XMLName          xml.Name `xml:"headerFooter"`
+	DifferentOddEven bool     `xml:"differentOddEven,attr,omitempty"`
+	DifferentFirst   bool     `xml:"differentFirst,attr,omitempty"`
+	ScaleWithDoc     *bool    `xml:"scaleWithDoc,attr"`
+	AlignWithMargins *bool    `xml:"alignWithMargins,attr"`
+	OddHeader        string   `xml:"oddHeader,omitempty"`
+	OddFooter        string   `xml:"oddFooter,omitempty"`
+	EvenHeader       string   `xml:"evenHeader,omitempty"`
+	EvenFooter       string   `xml:"evenFooter,omitempty"`
+	FirstHeader      string   `xml:"firstHeader,omitempty"`
+	FirstFooter      string   `xml:"firstFooter,omitempty"`
 }
 
 // xlsxDrawingHF (Drawing Reference in Header Footer) specifies the usage of
@@ -111,20 +112,20 @@ type xlsxPageSetUp struct {
 	Copies             int      `xml:"copies,attr,omitempty"`
 	Draft              bool     `xml:"draft,attr,omitempty"`
 	Errors             string   `xml:"errors,attr,omitempty"`
-	FirstPageNumber    int      `xml:"firstPageNumber,attr,omitempty"`
-	FitToHeight        int      `xml:"fitToHeight,attr,omitempty"`
-	FitToWidth         int      `xml:"fitToWidth,attr,omitempty"`
-	HorizontalDPI      int      `xml:"horizontalDpi,attr,omitempty"`
+	FirstPageNumber    string   `xml:"firstPageNumber,attr,omitempty"`
+	FitToHeight        *int     `xml:"fitToHeight,attr"`
+	FitToWidth         *int     `xml:"fitToWidth,attr"`
+	HorizontalDPI      string   `xml:"horizontalDpi,attr,omitempty"`
 	RID                string   `xml:"http://schemas.openxmlformats.org/officeDocument/2006/relationships id,attr,omitempty"`
 	Orientation        string   `xml:"orientation,attr,omitempty"`
 	PageOrder          string   `xml:"pageOrder,attr,omitempty"`
 	PaperHeight        string   `xml:"paperHeight,attr,omitempty"`
-	PaperSize          int      `xml:"paperSize,attr,omitempty"`
+	PaperSize          *int     `xml:"paperSize,attr"`
 	PaperWidth         string   `xml:"paperWidth,attr,omitempty"`
 	Scale              int      `xml:"scale,attr,omitempty"`
 	UseFirstPageNumber bool     `xml:"useFirstPageNumber,attr,omitempty"`
 	UsePrinterDefaults bool     `xml:"usePrinterDefaults,attr,omitempty"`
-	VerticalDPI        int      `xml:"verticalDpi,attr,omitempty"`
+	VerticalDPI        string   `xml:"verticalDpi,attr,omitempty"`
 }
 
 // xlsxPrintOptions directly maps the printOptions element in the namespace
@@ -145,12 +146,12 @@ type xlsxPrintOptions struct {
 // a sheet or a custom sheet view.
 type xlsxPageMargins struct {
 	XMLName xml.Name `xml:"pageMargins"`
-	Bottom  float64  `xml:"bottom,attr"`
-	Footer  float64  `xml:"footer,attr"`
-	Header  float64  `xml:"header,attr"`
 	Left    float64  `xml:"left,attr"`
 	Right   float64  `xml:"right,attr"`
 	Top     float64  `xml:"top,attr"`
+	Bottom  float64  `xml:"bottom,attr"`
+	Header  float64  `xml:"header,attr"`
+	Footer  float64  `xml:"footer,attr"`
 }
 
 // xlsxSheetFormatPr directly maps the sheetFormatPr element in the namespace
@@ -191,6 +192,7 @@ type xlsxSheetView struct {
 	ShowZeros                *bool            `xml:"showZeros,attr,omitempty"`
 	RightToLeft              bool             `xml:"rightToLeft,attr,omitempty"`
 	TabSelected              bool             `xml:"tabSelected,attr,omitempty"`
+	ShowRuler                *bool            `xml:"showRuler,attr,omitempty"`
 	ShowWhiteSpace           *bool            `xml:"showWhiteSpace,attr"`
 	ShowOutlineSymbols       bool             `xml:"showOutlineSymbols,attr,omitempty"`
 	DefaultGridColor         *bool            `xml:"defaultGridColor,attr"`
@@ -239,33 +241,24 @@ type xlsxSheetPr struct {
 	CodeName                          string           `xml:"codeName,attr,omitempty"`
 	FilterMode                        bool             `xml:"filterMode,attr,omitempty"`
 	EnableFormatConditionsCalculation *bool            `xml:"enableFormatConditionsCalculation,attr"`
-	TabColor                          *xlsxTabColor    `xml:"tabColor,omitempty"`
-	OutlinePr                         *xlsxOutlinePr   `xml:"outlinePr,omitempty"`
-	PageSetUpPr                       *xlsxPageSetUpPr `xml:"pageSetUpPr,omitempty"`
+	TabColor                          *xlsxColor       `xml:"tabColor"`
+	OutlinePr                         *xlsxOutlinePr   `xml:"outlinePr"`
+	PageSetUpPr                       *xlsxPageSetUpPr `xml:"pageSetUpPr"`
 }
 
 // xlsxOutlinePr maps to the outlinePr element. SummaryBelow allows you to
 // adjust the direction of grouper controls.
 type xlsxOutlinePr struct {
 	ApplyStyles        *bool `xml:"applyStyles,attr"`
-	SummaryBelow       bool  `xml:"summaryBelow,attr"`
-	SummaryRight       bool  `xml:"summaryRight,attr"`
-	ShowOutlineSymbols bool  `xml:"showOutlineSymbols,attr"`
+	SummaryBelow       *bool `xml:"summaryBelow,attr"`
+	SummaryRight       *bool `xml:"summaryRight,attr"`
+	ShowOutlineSymbols *bool `xml:"showOutlineSymbols,attr"`
 }
 
 // xlsxPageSetUpPr expresses page setup properties of the worksheet.
 type xlsxPageSetUpPr struct {
 	AutoPageBreaks bool `xml:"autoPageBreaks,attr,omitempty"`
 	FitToPage      bool `xml:"fitToPage,attr,omitempty"`
-}
-
-// xlsxTabColor represents background color of the sheet tab.
-type xlsxTabColor struct {
-	Auto    bool    `xml:"auto,attr,omitempty"`
-	Indexed int     `xml:"indexed,attr,omitempty"`
-	RGB     string  `xml:"rgb,attr,omitempty"`
-	Theme   int     `xml:"theme,attr,omitempty"`
-	Tint    float64 `xml:"tint,attr,omitempty"`
 }
 
 // xlsxCols defines column width and column formatting for one or more columns
@@ -278,16 +271,16 @@ type xlsxCols struct {
 // xlsxCol directly maps the col (Column Width & Formatting). Defines column
 // width and column formatting for one or more columns of the worksheet.
 type xlsxCol struct {
-	BestFit      bool    `xml:"bestFit,attr,omitempty"`
-	Collapsed    bool    `xml:"collapsed,attr,omitempty"`
-	CustomWidth  bool    `xml:"customWidth,attr,omitempty"`
-	Hidden       bool    `xml:"hidden,attr,omitempty"`
-	Max          int     `xml:"max,attr"`
-	Min          int     `xml:"min,attr"`
-	OutlineLevel uint8   `xml:"outlineLevel,attr,omitempty"`
-	Phonetic     bool    `xml:"phonetic,attr,omitempty"`
-	Style        int     `xml:"style,attr,omitempty"`
-	Width        float64 `xml:"width,attr,omitempty"`
+	BestFit      bool     `xml:"bestFit,attr,omitempty"`
+	Collapsed    bool     `xml:"collapsed,attr,omitempty"`
+	CustomWidth  bool     `xml:"customWidth,attr,omitempty"`
+	Hidden       bool     `xml:"hidden,attr,omitempty"`
+	Max          int      `xml:"max,attr"`
+	Min          int      `xml:"min,attr"`
+	OutlineLevel uint8    `xml:"outlineLevel,attr,omitempty"`
+	Phonetic     bool     `xml:"phonetic,attr,omitempty"`
+	Style        int      `xml:"style,attr,omitempty"`
+	Width        *float64 `xml:"width,attr"`
 }
 
 // xlsxDimension directly maps the dimension element in the namespace
@@ -314,19 +307,19 @@ type xlsxSheetData struct {
 // about an entire row of a worksheet, and contains all cell definitions for a
 // particular row in the worksheet.
 type xlsxRow struct {
-	C            []xlsxC `xml:"c"`
-	R            int     `xml:"r,attr,omitempty"`
-	Spans        string  `xml:"spans,attr,omitempty"`
-	S            int     `xml:"s,attr,omitempty"`
-	CustomFormat bool    `xml:"customFormat,attr,omitempty"`
-	Ht           float64 `xml:"ht,attr,omitempty"`
-	Hidden       bool    `xml:"hidden,attr,omitempty"`
-	CustomHeight bool    `xml:"customHeight,attr,omitempty"`
-	OutlineLevel uint8   `xml:"outlineLevel,attr,omitempty"`
-	Collapsed    bool    `xml:"collapsed,attr,omitempty"`
-	ThickTop     bool    `xml:"thickTop,attr,omitempty"`
-	ThickBot     bool    `xml:"thickBot,attr,omitempty"`
-	Ph           bool    `xml:"ph,attr,omitempty"`
+	C            []xlsxC  `xml:"c"`
+	R            int      `xml:"r,attr,omitempty"`
+	Spans        string   `xml:"spans,attr,omitempty"`
+	S            int      `xml:"s,attr,omitempty"`
+	CustomFormat bool     `xml:"customFormat,attr,omitempty"`
+	Ht           *float64 `xml:"ht,attr"`
+	Hidden       bool     `xml:"hidden,attr,omitempty"`
+	CustomHeight bool     `xml:"customHeight,attr,omitempty"`
+	OutlineLevel uint8    `xml:"outlineLevel,attr,omitempty"`
+	Collapsed    bool     `xml:"collapsed,attr,omitempty"`
+	ThickTop     bool     `xml:"thickTop,attr,omitempty"`
+	ThickBot     bool     `xml:"thickBot,attr,omitempty"`
+	Ph           bool     `xml:"ph,attr,omitempty"`
 }
 
 // xlsxSortState directly maps the sortState element. This collection
@@ -354,6 +347,18 @@ type xlsxBrk struct {
 	Max int  `xml:"max,attr,omitempty"`
 	Man bool `xml:"man,attr,omitempty"`
 	Pt  bool `xml:"pt,attr,omitempty"`
+}
+
+// xlsxRowBreaks directly maps a collection of the row breaks.
+type xlsxRowBreaks struct {
+	XMLName xml.Name `xml:"rowBreaks"`
+	xlsxBreaks
+}
+
+// xlsxRowBreaks directly maps a collection of the column breaks.
+type xlsxColBreaks struct {
+	XMLName xml.Name `xml:"colBreaks"`
+	xlsxBreaks
 }
 
 // xlsxBreaks directly maps a collection of the row or column breaks.
@@ -399,7 +404,8 @@ type xlsxCustomSheetView struct {
 
 // xlsxMergeCell directly maps the mergeCell element. A single merged cell.
 type xlsxMergeCell struct {
-	Ref string `xml:"ref,attr,omitempty"`
+	Ref  string `xml:"ref,attr,omitempty"`
+	rect []int
 }
 
 // xlsxMergeCells directly maps the mergeCells element. This collection
@@ -413,31 +419,32 @@ type xlsxMergeCells struct {
 // xlsxDataValidations expresses all data validation information for cells in a
 // sheet which have data validation features applied.
 type xlsxDataValidations struct {
-	XMLName        xml.Name          `xml:"dataValidations"`
-	Count          int               `xml:"count,attr,omitempty"`
-	DisablePrompts bool              `xml:"disablePrompts,attr,omitempty"`
-	XWindow        int               `xml:"xWindow,attr,omitempty"`
-	YWindow        int               `xml:"yWindow,attr,omitempty"`
-	DataValidation []*DataValidation `xml:"dataValidation"`
+	XMLName        xml.Name              `xml:"dataValidations"`
+	Count          int                   `xml:"count,attr,omitempty"`
+	DisablePrompts bool                  `xml:"disablePrompts,attr,omitempty"`
+	XWindow        int                   `xml:"xWindow,attr,omitempty"`
+	YWindow        int                   `xml:"yWindow,attr,omitempty"`
+	DataValidation []*xlsxDataValidation `xml:"dataValidation"`
 }
 
-// DataValidation directly maps the a single item of data validation defined
+// DataValidation directly maps the single item of data validation defined
 // on a range of the worksheet.
-type DataValidation struct {
-	AllowBlank       bool    `xml:"allowBlank,attr"`
-	Error            *string `xml:"error,attr"`
-	ErrorStyle       *string `xml:"errorStyle,attr"`
-	ErrorTitle       *string `xml:"errorTitle,attr"`
-	Operator         string  `xml:"operator,attr,omitempty"`
-	Prompt           *string `xml:"prompt,attr"`
-	PromptTitle      *string `xml:"promptTitle,attr"`
-	ShowDropDown     bool    `xml:"showDropDown,attr,omitempty"`
-	ShowErrorMessage bool    `xml:"showErrorMessage,attr,omitempty"`
-	ShowInputMessage bool    `xml:"showInputMessage,attr,omitempty"`
-	Sqref            string  `xml:"sqref,attr"`
-	Type             string  `xml:"type,attr,omitempty"`
-	Formula1         string  `xml:",innerxml"`
-	Formula2         string  `xml:",innerxml"`
+type xlsxDataValidation struct {
+	AllowBlank       bool          `xml:"allowBlank,attr"`
+	Error            *string       `xml:"error,attr"`
+	ErrorStyle       *string       `xml:"errorStyle,attr"`
+	ErrorTitle       *string       `xml:"errorTitle,attr"`
+	Operator         string        `xml:"operator,attr,omitempty"`
+	Prompt           *string       `xml:"prompt,attr"`
+	PromptTitle      *string       `xml:"promptTitle,attr"`
+	ShowDropDown     bool          `xml:"showDropDown,attr,omitempty"`
+	ShowErrorMessage bool          `xml:"showErrorMessage,attr,omitempty"`
+	ShowInputMessage bool          `xml:"showInputMessage,attr,omitempty"`
+	Sqref            string        `xml:"sqref,attr"`
+	XMSqref          string        `xml:"sqref,omitempty"`
+	Type             string        `xml:"type,attr,omitempty"`
+	Formula1         *xlsxInnerXML `xml:"formula1"`
+	Formula2         *xlsxInnerXML `xml:"formula2"`
 }
 
 // xlsxC collection represents a cell in the worksheet. Information about the
@@ -446,39 +453,50 @@ type DataValidation struct {
 //
 // This simple type is restricted to the values listed in the following table:
 //
-//      Enumeration Value         | Description
-//     ---------------------------+---------------------------------
-//      b (Boolean)               | Cell containing a boolean.
-//      d (Date)                  | Cell contains a date in the ISO 8601 format.
-//      e (Error)                 | Cell containing an error.
-//      inlineStr (Inline String) | Cell containing an (inline) rich string, i.e., one not in the shared string table. If this cell type is used, then the cell value is in the is element rather than the v element in the cell (c element).
-//      n (Number)                | Cell containing a number.
-//      s (Shared String)         | Cell containing a shared string.
-//      str (String)              | Cell containing a formula string.
-//
+//	 Enumeration Value         | Description
+//	---------------------------+---------------------------------
+//	 b (Boolean)               | Cell containing a boolean.
+//	 d (Date)                  | Cell contains a date in the ISO 8601 format.
+//	 e (Error)                 | Cell containing an error.
+//	 inlineStr (Inline String) | Cell containing an (inline) rich string, i.e.,
+//	                           | one not in the shared string table. If this
+//	                           | cell type is used, then the cell value is in
+//	                           | the is element rather than the v element in
+//	                           | the cell (c element).
+//	 n (Number)                | Cell containing a number.
+//	 s (Shared String)         | Cell containing a shared string.
+//	 str (String)              | Cell containing a formula string.
 type xlsxC struct {
 	XMLName  xml.Name `xml:"c"`
 	XMLSpace xml.Attr `xml:"space,attr,omitempty"`
 	R        string   `xml:"r,attr,omitempty"` // Cell ID, e.g. A1
-	S        int      `xml:"s,attr,omitempty"` // Style reference.
-	// Str string `xml:"str,attr,omitempty"` // Style reference.
-	T  string  `xml:"t,attr,omitempty"` // Type.
-	F  *xlsxF  `xml:"f,omitempty"`      // Formula
-	V  string  `xml:"v,omitempty"`      // Value
-	IS *xlsxSI `xml:"is"`
-}
-
-func (c *xlsxC) hasValue() bool {
-	return c.S != 0 || c.V != "" || c.F != nil || c.T != ""
+	S        int      `xml:"s,attr,omitempty"` // Style reference
+	T        string   `xml:"t,attr,omitempty"` // Type
+	Cm       *uint    `xml:"cm,attr"`
+	Vm       *uint    `xml:"vm,attr"`
+	Ph       *bool    `xml:"ph,attr"`
+	F        *xlsxF   `xml:"f"`           // Formula
+	V        string   `xml:"v,omitempty"` // Value
+	IS       *xlsxSI  `xml:"is"`
+	f        string
 }
 
 // xlsxF represents a formula for the cell. The formula expression is
 // contained in the character node of this element.
 type xlsxF struct {
 	Content string `xml:",chardata"`
-	T       string `xml:"t,attr,omitempty"`   // Formula type
+	T       string `xml:"t,attr,omitempty"` // Formula type
+	Aca     bool   `xml:"aca,attr,omitempty"`
 	Ref     string `xml:"ref,attr,omitempty"` // Shared formula ref
-	Si      string `xml:"si,attr,omitempty"`  // Shared formula index
+	Dt2D    bool   `xml:"dt2D,attr,omitempty"`
+	Dtr     bool   `xml:"dtr,attr,omitempty"`
+	Del1    bool   `xml:"del1,attr,omitempty"`
+	Del2    bool   `xml:"del2,attr,omitempty"`
+	R1      string `xml:"r1,attr,omitempty"`
+	R2      string `xml:"r2,attr,omitempty"`
+	Ca      bool   `xml:"ca,attr,omitempty"`
+	Si      *int   `xml:"si,attr"` // Shared formula index
+	Bx      bool   `xml:"bx,attr,omitempty"`
 }
 
 // xlsxSheetProtection collection expresses the sheet protection options to
@@ -528,6 +546,7 @@ type xlsxPhoneticPr struct {
 // applied to a particular cell or range.
 type xlsxConditionalFormatting struct {
 	XMLName xml.Name      `xml:"conditionalFormatting"`
+	Pivot   bool          `xml:"pivot,attr,omitempty"`
 	SQRef   string        `xml:"sqref,attr,omitempty"`
 	CfRule  []*xlsxCfRule `xml:"cfRule"`
 }
@@ -535,19 +554,19 @@ type xlsxConditionalFormatting struct {
 // xlsxCfRule (Conditional Formatting Rule) represents a description of a
 // conditional formatting rule.
 type xlsxCfRule struct {
-	AboveAverage *bool           `xml:"aboveAverage,attr"`
-	Bottom       bool            `xml:"bottom,attr,omitempty"`
+	Type         string          `xml:"type,attr,omitempty"`
 	DxfID        *int            `xml:"dxfId,attr"`
-	EqualAverage bool            `xml:"equalAverage,attr,omitempty"`
-	Operator     string          `xml:"operator,attr,omitempty"`
-	Percent      bool            `xml:"percent,attr,omitempty"`
 	Priority     int             `xml:"priority,attr,omitempty"`
-	Rank         int             `xml:"rank,attr,omitempty"`
-	StdDev       int             `xml:"stdDev,attr,omitempty"`
 	StopIfTrue   bool            `xml:"stopIfTrue,attr,omitempty"`
+	AboveAverage *bool           `xml:"aboveAverage,attr"`
+	Percent      bool            `xml:"percent,attr,omitempty"`
+	Bottom       bool            `xml:"bottom,attr,omitempty"`
+	Operator     string          `xml:"operator,attr,omitempty"`
 	Text         string          `xml:"text,attr,omitempty"`
 	TimePeriod   string          `xml:"timePeriod,attr,omitempty"`
-	Type         string          `xml:"type,attr,omitempty"`
+	Rank         int             `xml:"rank,attr,omitempty"`
+	StdDev       int             `xml:"stdDev,attr,omitempty"`
+	EqualAverage bool            `xml:"equalAverage,attr,omitempty"`
 	Formula      []string        `xml:"formula,omitempty"`
 	ColorScale   *xlsxColorScale `xml:"colorScale"`
 	DataBar      *xlsxDataBar    `xml:"dataBar"`
@@ -566,7 +585,7 @@ type xlsxColorScale struct {
 type xlsxDataBar struct {
 	MaxLength int          `xml:"maxLength,attr,omitempty"`
 	MinLength int          `xml:"minLength,attr,omitempty"`
-	ShowValue bool         `xml:"showValue,attr,omitempty"`
+	ShowValue *bool        `xml:"showValue,attr"`
 	Cfvo      []*xlsxCfvo  `xml:"cfvo"`
 	Color     []*xlsxColor `xml:"color"`
 }
@@ -575,7 +594,7 @@ type xlsxDataBar struct {
 type xlsxIconSet struct {
 	Cfvo      []*xlsxCfvo `xml:"cfvo"`
 	IconSet   string      `xml:"iconSet,attr,omitempty"`
-	ShowValue bool        `xml:"showValue,attr,omitempty"`
+	ShowValue *bool       `xml:"showValue,attr"`
 	Percent   bool        `xml:"percent,attr,omitempty"`
 	Reverse   bool        `xml:"reverse,attr,omitempty"`
 }
@@ -632,13 +651,12 @@ type xlsxHyperlink struct {
 // size of the sample. To reference the table, just add the tableParts element,
 // of course after having created and stored the table part. For example:
 //
-//    <worksheet xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main">
-//        ...
-//        <tableParts count="1">
-// 		      <tablePart r:id="rId1" />
-//        </tableParts>
-//    </worksheet>
-//
+//	   <worksheet xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main">
+//	       ...
+//	       <tableParts count="1">
+//			      <tablePart r:id="rId1" />
+//	       </tableParts>
+//	   </worksheet>
 type xlsxTableParts struct {
 	XMLName    xml.Name         `xml:"tableParts"`
 	Count      int              `xml:"count,attr,omitempty"`
@@ -655,11 +673,32 @@ type xlsxTablePart struct {
 // http://schemas.openxmlformats.org/spreadsheetml/2006/main - Background sheet
 // image. For example:
 //
-//    <picture r:id="rId1"/>
-//
+//	<picture r:id="rId1"/>
 type xlsxPicture struct {
 	XMLName xml.Name `xml:"picture"`
 	RID     string   `xml:"http://schemas.openxmlformats.org/officeDocument/2006/relationships id,attr,omitempty"`
+}
+
+// xlsxIgnoredError specifies a single ignored error for a range of cells.
+type xlsxIgnoredError struct {
+	XMLName            xml.Name `xml:"ignoredError"`
+	Sqref              string   `xml:"sqref,attr"`
+	EvalError          bool     `xml:"evalError,attr,omitempty"`
+	TwoDigitTextYear   bool     `xml:"twoDigitTextYear,attr,omitempty"`
+	NumberStoredAsText bool     `xml:"numberStoredAsText,attr,omitempty"`
+	Formula            bool     `xml:"formula,attr,omitempty"`
+	FormulaRange       bool     `xml:"formulaRange,attr,omitempty"`
+	UnlockedFormula    bool     `xml:"unlockedFormula,attr,omitempty"`
+	EmptyCellReference bool     `xml:"emptyCellReference,attr,omitempty"`
+	ListDataValidation bool     `xml:"listDataValidation,attr,omitempty"`
+	CalculatedColumn   bool     `xml:"calculatedColumn,attr,omitempty"`
+}
+
+// xlsxIgnoredErrors specifies a collection of ignored errors, by cell range.
+type xlsxIgnoredErrors struct {
+	XMLName      xml.Name           `xml:"ignoredErrors"`
+	IgnoredError []xlsxIgnoredError `xml:"ignoredError"`
+	ExtLst       *xlsxExtLst        `xml:"extLst"`
 }
 
 // xlsxLegacyDrawing directly maps the legacyDrawing element in the namespace
@@ -684,28 +723,98 @@ type xlsxLegacyDrawingHF struct {
 	RID     string   `xml:"http://schemas.openxmlformats.org/officeDocument/2006/relationships id,attr,omitempty"`
 }
 
-type xlsxInnerXML struct {
-	Content string `xml:",innerxml"`
-}
-
-// xlsxWorksheetExt directly maps the ext element in the worksheet.
-type xlsxWorksheetExt struct {
-	XMLName xml.Name `xml:"ext"`
-	URI     string   `xml:"uri,attr"`
-	Content string   `xml:",innerxml"`
-}
-
-// decodeWorksheetExt directly maps the ext element.
-type decodeWorksheetExt struct {
-	XMLName xml.Name            `xml:"extLst"`
-	Ext     []*xlsxWorksheetExt `xml:"ext"`
-}
-
 // decodeX14SparklineGroups directly maps the sparklineGroups element.
 type decodeX14SparklineGroups struct {
 	XMLName xml.Name `xml:"sparklineGroups"`
 	XMLNSXM string   `xml:"xmlns:xm,attr"`
 	Content string   `xml:",innerxml"`
+}
+
+// decodeX14ConditionalFormattingExt directly maps the ext element.
+type decodeX14ConditionalFormattingExt struct {
+	XMLName xml.Name `xml:"ext"`
+	ID      string   `xml:"id"`
+}
+
+// decodeX14ConditionalFormattings directly maps the conditionalFormattings
+// element.
+type decodeX14ConditionalFormattings struct {
+	XMLName xml.Name `xml:"conditionalFormattings"`
+	XMLNSXM string   `xml:"xmlns:xm,attr"`
+	Content string   `xml:",innerxml"`
+}
+
+// decodeX14ConditionalFormattingRules directly maps the conditionalFormattings
+// element.
+type decodeX14ConditionalFormattingRules struct {
+	XMLName xml.Name                         `xml:"conditionalFormattings"`
+	XMLNSXM string                           `xml:"xmlns:xm,attr"`
+	CondFmt []decodeX14ConditionalFormatting `xml:"conditionalFormatting"`
+}
+
+// decodeX14ConditionalFormatting directly maps the conditionalFormatting
+// element.
+type decodeX14ConditionalFormatting struct {
+	XMLName xml.Name           `xml:"conditionalFormatting"`
+	CfRule  []*decodeX14CfRule `xml:"cfRule"`
+}
+
+// decodeX14CfRule directly maps the cfRule element.
+type decodeX14CfRule struct {
+	XMLName xml.Name          `xml:"cfRule"`
+	Type    string            `xml:"type,attr,omitempty"`
+	ID      string            `xml:"id,attr,omitempty"`
+	DataBar *decodeX14DataBar `xml:"dataBar"`
+}
+
+// decodeX14DataBar directly maps the dataBar element.
+type decodeX14DataBar struct {
+	XMLName           xml.Name    `xml:"dataBar"`
+	MaxLength         int         `xml:"maxLength,attr"`
+	MinLength         int         `xml:"minLength,attr"`
+	Border            bool        `xml:"border,attr,omitempty"`
+	Gradient          *bool       `xml:"gradient,attr"`
+	ShowValue         bool        `xml:"showValue,attr,omitempty"`
+	Direction         string      `xml:"direction,attr,omitempty"`
+	Cfvo              []*xlsxCfvo `xml:"cfvo"`
+	BorderColor       *xlsxColor  `xml:"borderColor"`
+	NegativeFillColor *xlsxColor  `xml:"negativeFillColor"`
+	AxisColor         *xlsxColor  `xml:"axisColor"`
+}
+
+// xlsxX14ConditionalFormattings directly maps the conditionalFormattings
+// element.
+type xlsxX14ConditionalFormattings struct {
+	XMLName xml.Name `xml:"x14:conditionalFormattings"`
+	Content string   `xml:",innerxml"`
+}
+
+// xlsxX14ConditionalFormatting directly maps the conditionalFormatting element.
+type xlsxX14ConditionalFormatting struct {
+	XMLName xml.Name         `xml:"x14:conditionalFormatting"`
+	XMLNSXM string           `xml:"xmlns:xm,attr"`
+	CfRule  []*xlsxX14CfRule `xml:"x14:cfRule"`
+}
+
+// xlsxX14CfRule directly maps the cfRule element.
+type xlsxX14CfRule struct {
+	Type    string         `xml:"type,attr,omitempty"`
+	ID      string         `xml:"id,attr,omitempty"`
+	DataBar *xlsx14DataBar `xml:"x14:dataBar"`
+}
+
+// xlsx14DataBar directly maps the dataBar element.
+type xlsx14DataBar struct {
+	MaxLength         int         `xml:"maxLength,attr"`
+	MinLength         int         `xml:"minLength,attr"`
+	Border            bool        `xml:"border,attr"`
+	Gradient          bool        `xml:"gradient,attr"`
+	ShowValue         bool        `xml:"showValue,attr,omitempty"`
+	Direction         string      `xml:"direction,attr,omitempty"`
+	Cfvo              []*xlsxCfvo `xml:"x14:cfvo"`
+	BorderColor       *xlsxColor  `xml:"x14:borderColor"`
+	NegativeFillColor *xlsxColor  `xml:"x14:negativeFillColor"`
+	AxisColor         *xlsxColor  `xml:"x14:axisColor"`
 }
 
 // xlsxX14SparklineGroups directly maps the sparklineGroups element.
@@ -736,14 +845,14 @@ type xlsxX14SparklineGroup struct {
 	MinAxisType         string            `xml:"minAxisType,attr,omitempty"`
 	MaxAxisType         string            `xml:"maxAxisType,attr,omitempty"`
 	RightToLeft         bool              `xml:"rightToLeft,attr,omitempty"`
-	ColorSeries         *xlsxTabColor     `xml:"x14:colorSeries"`
-	ColorNegative       *xlsxTabColor     `xml:"x14:colorNegative"`
+	ColorSeries         *xlsxColor        `xml:"x14:colorSeries"`
+	ColorNegative       *xlsxColor        `xml:"x14:colorNegative"`
 	ColorAxis           *xlsxColor        `xml:"x14:colorAxis"`
-	ColorMarkers        *xlsxTabColor     `xml:"x14:colorMarkers"`
-	ColorFirst          *xlsxTabColor     `xml:"x14:colorFirst"`
-	ColorLast           *xlsxTabColor     `xml:"x14:colorLast"`
-	ColorHigh           *xlsxTabColor     `xml:"x14:colorHigh"`
-	ColorLow            *xlsxTabColor     `xml:"x14:colorLow"`
+	ColorMarkers        *xlsxColor        `xml:"x14:colorMarkers"`
+	ColorFirst          *xlsxColor        `xml:"x14:colorFirst"`
+	ColorLast           *xlsxColor        `xml:"x14:colorLast"`
+	ColorHigh           *xlsxColor        `xml:"x14:colorHigh"`
+	ColorLow            *xlsxColor        `xml:"x14:colorLow"`
 	Sparklines          xlsxX14Sparklines `xml:"x14:sparklines"`
 }
 
@@ -758,8 +867,26 @@ type xlsxX14Sparkline struct {
 	Sqref string `xml:"xm:sqref"`
 }
 
-// SparklineOption directly maps the settings of the sparkline.
-type SparklineOption struct {
+// DataValidation directly maps the settings of the data validation rule.
+type DataValidation struct {
+	AllowBlank       bool
+	Error            *string
+	ErrorStyle       *string
+	ErrorTitle       *string
+	Operator         string
+	Prompt           *string
+	PromptTitle      *string
+	ShowDropDown     bool
+	ShowErrorMessage bool
+	ShowInputMessage bool
+	Sqref            string
+	Type             string
+	Formula1         string
+	Formula2         string
+}
+
+// SparklineOptions directly maps the settings of the sparkline.
+type SparklineOptions struct {
 	Location      []string
 	Range         []string
 	Max           int
@@ -789,48 +916,55 @@ type SparklineOption struct {
 	EmptyCells    string
 }
 
-// formatPanes directly maps the settings of the panes.
-type formatPanes struct {
-	Freeze      bool   `json:"freeze"`
-	Split       bool   `json:"split"`
-	XSplit      int    `json:"x_split"`
-	YSplit      int    `json:"y_split"`
-	TopLeftCell string `json:"top_left_cell"`
-	ActivePane  string `json:"active_pane"`
-	Panes       []struct {
-		SQRef      string `json:"sqref"`
-		ActiveCell string `json:"active_cell"`
-		Pane       string `json:"pane"`
-	} `json:"panes"`
+// Selection directly maps the settings of the worksheet selection.
+type Selection struct {
+	SQRef      string
+	ActiveCell string
+	Pane       string
 }
 
-// formatConditional directly maps the conditional format settings of the cells.
-type formatConditional struct {
-	Type         string `json:"type"`
-	AboveAverage bool   `json:"above_average"`
-	Percent      bool   `json:"percent"`
-	Format       int    `json:"format"`
-	Criteria     string `json:"criteria"`
-	Value        string `json:"value,omitempty"`
-	Minimum      string `json:"minimum,omitempty"`
-	Maximum      string `json:"maximum,omitempty"`
-	MinType      string `json:"min_type,omitempty"`
-	MidType      string `json:"mid_type,omitempty"`
-	MaxType      string `json:"max_type,omitempty"`
-	MinValue     string `json:"min_value,omitempty"`
-	MidValue     string `json:"mid_value,omitempty"`
-	MaxValue     string `json:"max_value,omitempty"`
-	MinColor     string `json:"min_color,omitempty"`
-	MidColor     string `json:"mid_color,omitempty"`
-	MaxColor     string `json:"max_color,omitempty"`
-	MinLength    string `json:"min_length,omitempty"`
-	MaxLength    string `json:"max_length,omitempty"`
-	MultiRange   string `json:"multi_range,omitempty"`
-	BarColor     string `json:"bar_color,omitempty"`
+// Panes directly maps the settings of the panes.
+type Panes struct {
+	Freeze      bool
+	Split       bool
+	XSplit      int
+	YSplit      int
+	TopLeftCell string
+	ActivePane  string
+	Selection   []Selection
 }
 
-// FormatSheetProtection directly maps the settings of worksheet protection.
-type FormatSheetProtection struct {
+// ConditionalFormatOptions directly maps the conditional format settings of the cells.
+type ConditionalFormatOptions struct {
+	Type           string
+	AboveAverage   bool
+	Percent        bool
+	Format         *int
+	Criteria       string
+	Value          string
+	MinType        string
+	MidType        string
+	MaxType        string
+	MinValue       string
+	MidValue       string
+	MaxValue       string
+	MinColor       string
+	MidColor       string
+	MaxColor       string
+	BarColor       string
+	BarBorderColor string
+	BarDirection   string
+	BarOnly        bool
+	BarSolid       bool
+	IconStyle      string
+	ReverseIcons   bool
+	IconsOnly      bool
+	StopIfTrue     bool
+}
+
+// SheetProtectionOptions directly maps the settings of worksheet protection.
+type SheetProtectionOptions struct {
+	AlgorithmName       string
 	AutoFilter          bool
 	DeleteColumns       bool
 	DeleteRows          bool
@@ -849,26 +983,148 @@ type FormatSheetProtection struct {
 	Sort                bool
 }
 
-// FormatHeaderFooter directly maps the settings of header and footer.
-type FormatHeaderFooter struct {
-	AlignWithMargins bool
+// HeaderFooterOptions directly maps the settings of header and footer.
+type HeaderFooterOptions struct {
+	AlignWithMargins *bool
 	DifferentFirst   bool
 	DifferentOddEven bool
-	ScaleWithDoc     bool
+	ScaleWithDoc     *bool
 	OddHeader        string
 	OddFooter        string
 	EvenHeader       string
 	EvenFooter       string
-	FirstFooter      string
 	FirstHeader      string
+	FirstFooter      string
 }
 
-// FormatPageMargins directly maps the settings of page margins
-type FormatPageMargins struct {
-	Bottom string
-	Footer string
-	Header string
-	Left   string
-	Right  string
-	Top    string
+// PageLayoutMarginsOptions directly maps the settings of page layout margins.
+type PageLayoutMarginsOptions struct {
+	Bottom       *float64
+	Footer       *float64
+	Header       *float64
+	Left         *float64
+	Right        *float64
+	Top          *float64
+	Horizontally *bool
+	Vertically   *bool
+}
+
+// PageLayoutOptions directly maps the settings of page layout.
+type PageLayoutOptions struct {
+	// Size defines the paper size of the worksheet.
+	Size *int
+	// Orientation defines the orientation of page layout for a worksheet.
+	Orientation *string
+	// FirstPageNumber specified the first printed page number. If no value is
+	// specified, then 'automatic' is assumed.
+	FirstPageNumber *uint
+	// AdjustTo defines the print scaling. This attribute is restricted to
+	// value ranging from 10 (10%) to 400 (400%). This setting is overridden
+	// when fitToWidth and/or fitToHeight are in use.
+	AdjustTo *uint
+	// FitToHeight specified the number of vertical pages to fit on.
+	FitToHeight *int
+	// FitToWidth specified the number of horizontal pages to fit on.
+	FitToWidth *int
+	// BlackAndWhite specified print black and white.
+	BlackAndWhite *bool
+	// PageOrder specifies the ordering of multiple pages. Values
+	// accepted: overThenDown and downThenOver
+	PageOrder *string
+}
+
+// ViewOptions directly maps the settings of sheet view.
+type ViewOptions struct {
+	// DefaultGridColor indicating that the consuming application should use
+	// the default grid lines color(system dependent). Overrides any color
+	// specified in colorId.
+	DefaultGridColor *bool
+	// RightToLeft indicating whether the sheet is in 'right to left' display
+	// mode. When in this mode, Column A is on the far right, Column B; is one
+	// column left of Column A, and so on. Also, information in cells is
+	// displayed in the Right to Left format.
+	RightToLeft *bool
+	// ShowFormulas indicating whether this sheet should display formulas.
+	ShowFormulas *bool
+	// ShowGridLines indicating whether this sheet should display grid lines.
+	ShowGridLines *bool
+	// ShowRowColHeaders indicating whether the sheet should display row and
+	// column headings.
+	ShowRowColHeaders *bool
+	// ShowRuler indicating this sheet should display ruler.
+	ShowRuler *bool
+	// ShowZeros indicating whether to "show a zero in cells that have zero
+	// value". When using a formula to reference another cell which is empty,
+	// the referenced value becomes 0 when the flag is true. (Default setting
+	// is true.)
+	ShowZeros *bool
+	// TopLeftCell specifies a location of the top left visible cell Location
+	// of the top left visible cell in the bottom right pane (when in
+	// Left-to-Right mode).
+	TopLeftCell *string
+	// View indicating how sheet is displayed, by default it uses empty string
+	// available options: normal, pageLayout, pageBreakPreview
+	View *string
+	// ZoomScale specifies a window zoom magnification for current view
+	// representing percent values. This attribute is restricted to values
+	// ranging from 10 to 400. Horizontal & Vertical scale together.
+	ZoomScale *float64
+}
+
+// SheetPropsOptions directly maps the settings of sheet view.
+type SheetPropsOptions struct {
+	// Specifies a stable name of the sheet, which should not change over time,
+	// and does not change from user input. This name should be used by code
+	// to reference a particular sheet.
+	CodeName *string
+	// EnableFormatConditionsCalculation indicating whether the conditional
+	// formatting calculations shall be evaluated. If set to false, then the
+	// min/max values of color scales or data bars or threshold values in Top N
+	// rules shall not be updated. Essentially the conditional
+	// formatting "calc" is off.
+	EnableFormatConditionsCalculation *bool
+	// Published indicating whether the worksheet is published.
+	Published *bool
+	// AutoPageBreaks indicating whether the sheet displays Automatic Page
+	// Breaks.
+	AutoPageBreaks *bool
+	// FitToPage indicating whether the Fit to Page print option is enabled.
+	FitToPage *bool
+	// TabColorIndexed represents the indexed color value.
+	TabColorIndexed *int
+	// TabColorRGB represents the standard Alpha Red Green Blue color value.
+	TabColorRGB *string
+	// TabColorTheme represents the zero-based index into the collection,
+	// referencing a particular value expressed in the Theme part.
+	TabColorTheme *int
+	// TabColorTint specifies the tint value applied to the color.
+	TabColorTint *float64
+	// OutlineSummaryBelow indicating whether summary rows appear below detail
+	// in an outline, when applying an outline.
+	OutlineSummaryBelow *bool
+	// OutlineSummaryRight indicating whether summary columns appear to the
+	// right of detail in an outline, when applying an outline.
+	OutlineSummaryRight *bool
+	// BaseColWidth specifies the number of characters of the maximum digit
+	// width of the normal style's font. This value does not include margin
+	// padding or extra padding for grid lines. It is only the number of
+	// characters.
+	BaseColWidth *uint8
+	// DefaultColWidth specifies the default column width measured as the
+	// number of characters of the maximum digit width of the normal style's
+	// font.
+	DefaultColWidth *float64
+	// DefaultRowHeight specifies the default row height measured in point
+	// size. Optimization so we don't have to write the height on all rows.
+	// This can be written out if most rows have custom height, to achieve the
+	// optimization.
+	DefaultRowHeight *float64
+	// CustomHeight specifies the custom height.
+	CustomHeight *bool
+	// ZeroHeight specifies if rows are hidden.
+	ZeroHeight *bool
+	// ThickTop specifies if rows have a thick top border by default.
+	ThickTop *bool
+	// ThickBottom specifies if rows have a thick bottom border by default.
+	ThickBottom *bool
 }
